Alternate Observations in a Data Frame over States

Split single-state process observations, apply multiple state projections, and combine these observations into a single data frame, arranged by process time or index.

dataStateSwitch(df, angles2 = list(yaw = 0, pitch = 90, roll = 30),
  scales2 = c(1, 0.5, 2), angles3 = list(yaw = 90, pitch = 0, roll = -30),
  scales3 = c(0.25, 0.1, 0.75))

Arguments

df
A data frame returned by processNOCdata() or faultSwitch().

angles2
Change the principal angles for State 2.

scales2
Change the principal scales for State 2.

angles3
Change the principal angles for State 3.

scales3
Change the principal scales for State 3.

Value

A data frame containing the time index, state, and feature values after state-specific rotation and scaling; this data frame also contains the other columns of df that aren't the feature values. This data frame has

dateTime - a POSIX column of the time stamps for each observation

state - column of state membership (1, 2, or 3)

x - the process values for the first feature, corresponding to t + random error

y - the process values for the second feature, corresponding to t ^ 2 - 3 * t + random error

z - the process values for the third feature, corresponding to -t ^ 3 + 3 * t ^ 2 + random error

t - the non-stationary and autocorrelated latent feature

err1 - a Gaussian white noise vector

err2 - a Gaussian white noise vector

err3 - a Gaussian white noise vector

Details

This function splits a process data frame by state, and rotates and scales the observations from states 2 and 3 by the scales and angles specified in the function arguments. After state-specific rotation and scaling, this function combines the observations back together and orders them by process time index. This function takes in data frame returned by processNOCdata() or faultSwitch(). This function calls rotateScale3D() and is called internally by mspProcessData().

See also

Calls: processNOCdata, faultSwitch, rotateScale3D. Called by: mspProcessData

Examples

nrml <- processNOCdata()
dataStateSwitch(nrml)

#>                  dateTime state            x             y             z
#> 1     2015-05-16 10:00:00     1  0.048215234 -0.0838415483  0.0895795221
#> 2     2015-05-16 10:01:00     1 -0.014453701 -0.0205706111  0.0349327346
#> 3     2015-05-16 10:02:00     1  0.125441410  0.1480926059 -0.0116241333
#> 4     2015-05-16 10:03:00     1  0.332361153 -0.2914041940  0.0815975420
#> 5     2015-05-16 10:04:00     1  0.146236063 -0.6281353488  0.1211055195
#> 6     2015-05-16 10:05:00     1 -0.084130327 -0.4730348572  0.0446131968
#> 7     2015-05-16 10:06:00     1  0.063210356 -0.3097323479  0.3186941977
#> 8     2015-05-16 10:07:00     1  0.016736430 -0.3830210909  0.0134367783
#> 9     2015-05-16 10:08:00     1  0.034379340 -0.6176175591  0.1700067963
#> 10    2015-05-16 10:09:00     1  0.098555571 -0.4850403170  0.1184397723
#> 11    2015-05-16 10:10:00     1  0.134841669 -0.5836909558 -0.0412002912
#> 12    2015-05-16 10:11:00     1  0.187239406 -0.8015469119  0.2306437194
#> 13    2015-05-16 10:12:00     1  0.277357791 -0.9343553959  0.5160699136
#> 14    2015-05-16 10:13:00     1  0.017209918 -0.7601475825  0.1895295655
#> 15    2015-05-16 10:14:00     1  0.418341810 -0.6628315651  0.4429876353
#> 16    2015-05-16 10:15:00     1  0.027203824 -0.4972037675 -0.0090372378
#> 17    2015-05-16 10:16:00     1  0.026041580 -0.5842305877  0.0707467769
#> 18    2015-05-16 10:17:00     1  0.136732439 -0.6087033809  0.0525407646
#> 19    2015-05-16 10:18:00     1  0.244394944 -0.6440936386  0.1985778176
#> 20    2015-05-16 10:19:00     1  0.188051491 -0.5804195421  0.0889359557
#> 21    2015-05-16 10:20:00     1  0.233188809 -0.8083672097  0.2467977267
#> 22    2015-05-16 10:21:00     1  0.263261718 -0.7949058967  0.2711562925
#> 23    2015-05-16 10:22:00     1  0.236929971 -0.7066259670  0.1464918112
#> 24    2015-05-16 10:23:00     1  0.171888612 -0.6905296948  0.2899271767
#> 25    2015-05-16 10:24:00     1  0.334354160 -0.9895187540  0.4109039299
#> 26    2015-05-16 10:25:00     1  0.304609882 -0.9396816436  0.2035590883
#> 27    2015-05-16 10:26:00     1  0.259546052 -0.8680445276  0.0854370262
#> 28    2015-05-16 10:27:00     1  0.262557288 -0.5986647440  0.0386150804
#> 29    2015-05-16 10:28:00     1  0.069159457 -0.7908548668  0.2013565628
#> 30    2015-05-16 10:29:00     1  0.287218408 -0.5868855777 -0.0175855652
#> 31    2015-05-16 10:30:00     1  0.356244732 -0.5273259430 -0.0133683006
#> 32    2015-05-16 10:31:00     1  0.407174442 -0.6291286763  0.2241561758
#> 33    2015-05-16 10:32:00     1  0.119903843 -0.8939248485  0.3508994112
#> 34    2015-05-16 10:33:00     1  0.635325877 -1.0820541247  0.3192711098
#> 35    2015-05-16 10:34:00     1  0.390578856 -0.9344416617  0.2745397259
#> 36    2015-05-16 10:35:00     1  0.222543581 -0.9428704095  0.2444252729
#> 37    2015-05-16 10:36:00     1  0.203018715 -0.6031427225  0.1356274976
#> 38    2015-05-16 10:37:00     1  0.153903203 -0.6500682449  0.1297842104
#> 39    2015-05-16 10:38:00     1  0.148549968 -0.6194929340  0.2280099348
#> 40    2015-05-16 10:39:00     1  0.184099015 -0.8278323185  0.2375388122
#> 41    2015-05-16 10:40:00     1  0.243445429 -0.6581022957  0.3213810470
#> 42    2015-05-16 10:41:00     1  0.340272824 -0.8043243948  0.1657903607
#> 43    2015-05-16 10:42:00     1  0.231785066 -0.7428043956  0.1403523355
#> 44    2015-05-16 10:43:00     1  0.329056527 -0.6639903865  0.1649302346
#> 45    2015-05-16 10:44:00     1  0.359651817 -0.7072817571  0.2803323477
#> 46    2015-05-16 10:45:00     1  0.406426306 -1.2488928033  0.5816649013
#> 47    2015-05-16 10:46:00     1  0.355775134 -1.2150099238  0.4338514412
#> 48    2015-05-16 10:47:00     1  0.277258961 -0.9353721502  0.3788924654
#> 49    2015-05-16 10:48:00     1  0.381075565 -1.0773544672  0.3318139274
#> 50    2015-05-16 10:49:00     1  0.113323327 -0.6208861103  0.0984795970
#> 51    2015-05-16 10:50:00     1  0.231707837 -0.9508240145  0.4382973498
#> 52    2015-05-16 10:51:00     1  0.299974149 -0.9567239998  0.1997061279
#> 53    2015-05-16 10:52:00     1  0.088509005 -0.7928239421  0.2093896822
#> 54    2015-05-16 10:53:00     1  0.325060006 -0.7700098815  0.2424203514
#> 55    2015-05-16 10:54:00     1  0.361065602 -0.7551410708  0.2594757090
#> 56    2015-05-16 10:55:00     1  0.139443400 -0.9460292527  0.1943159288
#> 57    2015-05-16 10:56:00     1  0.267826789 -0.4199070557  0.0962716025
#> 58    2015-05-16 10:57:00     1  0.574478521 -0.9371277752  0.4150819936
#> 59    2015-05-16 10:58:00     1  0.187316161 -0.8160785495  0.2102544666
#> 60    2015-05-16 10:59:00     1  0.386027561 -1.0418888441  0.2584215805
#> 61    2015-05-16 11:00:00     2  0.447106988 -0.2304243493 -1.3581134116
#> 62    2015-05-16 11:01:00     2  0.156776253 -0.2453362207 -1.5075675235
#> 63    2015-05-16 11:02:00     2  0.388055502 -0.3665900214 -1.8054747956
#> 64    2015-05-16 11:03:00     2  0.218441015 -0.3293490815 -1.4883602309
#> 65    2015-05-16 11:04:00     2  0.157641004 -0.2603927696 -1.2819433351
#> 66    2015-05-16 11:05:00     2  0.193459557 -0.1573409570 -0.8131579842
#> 67    2015-05-16 11:06:00     2  0.257139473 -0.4065154925 -1.5431566131
#> 68    2015-05-16 11:07:00     2  0.233547417 -0.2213891303 -1.5887421001
#> 69    2015-05-16 11:08:00     2  0.249002101 -0.4022525177 -1.5445709449
#> 70    2015-05-16 11:09:00     2  0.225492209 -0.2259218186 -0.6685760875
#> 71    2015-05-16 11:10:00     2  0.139308715 -0.1579195846 -0.8493743468
#> 72    2015-05-16 11:11:00     2  0.032755902 -0.1798738084 -0.8405124999
#> 73    2015-05-16 11:12:00     2  0.251784235 -0.2359165116 -1.5478473948
#> 74    2015-05-16 11:13:00     2  0.214980766 -0.2536269493 -1.4502207132
#> 75    2015-05-16 11:14:00     2  0.375019250 -0.2240165226 -1.3416883893
#> 76    2015-05-16 11:15:00     2  0.343478345 -0.3120801317 -1.5798383270
#> 77    2015-05-16 11:16:00     2  0.259600067 -0.3700029839 -1.9218507849
#> 78    2015-05-16 11:17:00     2  0.418030799 -0.3201539761 -1.7671972357
#> 79    2015-05-16 11:18:00     2  0.265508616 -0.3318148366 -1.2689714941
#> 80    2015-05-16 11:19:00     2  0.488325361 -0.2947406243 -1.5844223769
#> 81    2015-05-16 11:20:00     2  0.382562704 -0.3659693397 -1.6794178478
#> 82    2015-05-16 11:21:00     2  0.505164931 -0.4150913849 -1.8630110688
#> 83    2015-05-16 11:22:00     2  0.439777482 -0.3943524757 -2.1065159474
#> 84    2015-05-16 11:23:00     2  0.434934777 -0.3821311869 -1.8244056200
#> 85    2015-05-16 11:24:00     2  0.262709078 -0.2590179333 -1.4675116613
#> 86    2015-05-16 11:25:00     2  0.479973100 -0.3373378668 -1.7106764480
#> 87    2015-05-16 11:26:00     2  0.126753162 -0.2280640567 -1.0828622236
#> 88    2015-05-16 11:27:00     2  0.099589068 -0.1947597918 -1.2060715516
#> 89    2015-05-16 11:28:00     2  0.089417705 -0.2514974988 -1.3972723140
#> 90    2015-05-16 11:29:00     2  0.225087490 -0.2131103097 -0.7725546304
#> 91    2015-05-16 11:30:00     2  0.298136543 -0.2601903211 -1.5350721320
#> 92    2015-05-16 11:31:00     2  0.300947163 -0.3541542589 -1.3860512438
#> 93    2015-05-16 11:32:00     2  0.268034304 -0.3825422326 -1.3434031441
#> 94    2015-05-16 11:33:00     2  0.404874673 -0.3889845081 -1.5374689654
#> 95    2015-05-16 11:34:00     2  0.281718252 -0.3160364268 -1.7682713427
#> 96    2015-05-16 11:35:00     2  0.537704013 -0.3455182311 -1.5848479857
#> 97    2015-05-16 11:36:00     2  0.635748107 -0.3497907173 -1.5018849687
#> 98    2015-05-16 11:37:00     2  0.381501029 -0.3584279927 -1.4783349654
#> 99    2015-05-16 11:38:00     2  0.574031674 -0.2865411346 -1.8398219820
#> 100   2015-05-16 11:39:00     2  0.472552685 -0.3134193630 -1.9583882524
#> 101   2015-05-16 11:40:00     2  0.254691506 -0.2547492201 -1.5930939857
#> 102   2015-05-16 11:41:00     2  0.435021939 -0.2644580112 -1.3503751941
#> 103   2015-05-16 11:42:00     2  0.116986192 -0.2817146745 -0.7205630323
#> 104   2015-05-16 11:43:00     2  0.204958137 -0.1977293255 -1.0569865618
#> 105   2015-05-16 11:44:00     2  0.220298178 -0.1691497830 -0.4684829712
#> 106   2015-05-16 11:45:00     2  0.086705204 -0.1913353511 -1.1051970799
#> 107   2015-05-16 11:46:00     2  0.074527905 -0.1871138714 -0.7759726155
#> 108   2015-05-16 11:47:00     2  0.093387626 -0.1568013531 -0.7894090869
#> 109   2015-05-16 11:48:00     2  0.085356555 -0.2332964962 -0.9536666752
#> 110   2015-05-16 11:49:00     2  0.063870600 -0.1878360784 -1.2166895802
#> 111   2015-05-16 11:50:00     2  0.213494023 -0.2882480760 -1.0740905872
#>                t          err1          err2          err3
#> 1     0.01000000  3.821523e-02 -5.394155e-02  8.928052e-02
#> 2     0.09131685 -1.057706e-01  2.450412e-01  1.067790e-02
#> 3     0.02003700  1.054044e-01  2.078021e-01 -1.282053e-02
#> 4     0.08602981  2.463313e-01 -4.071589e-02  6.003088e-02
#> 5     0.16633122 -2.009516e-02 -1.568078e-01  4.270903e-02
#> 6     0.25243881 -3.365691e-01  2.205562e-01 -1.304761e-01
#> 7     0.17111334 -1.079030e-01  1.743279e-01  2.358650e-01
#> 8     0.18746392 -1.707275e-01  1.442279e-01 -8.540339e-02
#> 9     0.19988120 -1.655019e-01 -5.792645e-02  5.813506e-02
#> 10    0.20730353 -1.087480e-01  9.389551e-02 -1.575668e-03
#> 11    0.19955917 -6.471750e-02 -2.483731e-02 -1.527247e-01
#> 12    0.25184889 -6.460949e-02 -1.094281e-01  5.633436e-02
#> 13    0.36161015 -8.425236e-02  1.971315e-02  1.710690e-01
#> 14    0.26237261 -2.451627e-01 -4.186913e-02  1.072970e-03
#> 15    0.29323694  1.251049e-01  1.308913e-01  2.102388e-01
#> 16    0.19969762 -1.724938e-01  6.200994e-02 -1.207109e-01
#> 17    0.19541852 -1.693769e-01 -3.616343e-02 -3.635570e-02
#> 18    0.21526955 -7.853711e-02 -9.235712e-03 -7.650637e-02
#> 19    0.25941697 -1.502203e-02  6.686011e-02  1.414435e-02
#> 20    0.22095329 -3.290180e-02  3.361998e-02 -4.673810e-02
#> 21    0.26000657 -2.681776e-02 -9.595091e-02  6.156481e-02
#> 22    0.28413036 -2.086865e-02 -2.324487e-02  5.190396e-02
#> 23    0.22031252  1.661745e-02 -9.422602e-02  1.157244e-02
#> 24    0.27277607 -1.008875e-01  5.339174e-02  8.700321e-02
#> 25    0.36865303 -3.429887e-02 -1.946472e-02  5.329057e-02
#> 26    0.34300677 -3.839689e-02 -2.831498e-02 -1.090458e-01
#> 27    0.28258925 -2.304320e-02 -1.001335e-01 -1.315664e-01
#> 28    0.28246646 -1.990917e-02  1.689473e-01 -1.782096e-01
#> 29    0.24254570 -1.733862e-01 -1.220462e-01  3.913989e-02
#> 30    0.17892812  1.082903e-01 -8.211650e-02 -1.079029e-01
#> 31    0.25145898  1.047858e-01  1.638194e-01 -1.871630e-01
#> 32    0.26858130  1.385931e-01  1.044793e-01  2.712279e-02
#> 33    0.28472648 -1.648226e-01 -1.208146e-01  1.307744e-01
#> 34    0.36462110  2.707048e-01 -1.211394e-01 -3.109869e-02
#> 35    0.35029465  4.028421e-02 -6.264063e-03 -5.059592e-02
#> 36    0.32746653 -1.049229e-01 -6.770516e-02 -4.216205e-02
#> 37    0.22310118 -2.008247e-02  1.638669e-02 -2.590249e-03
#> 38    0.22494245 -7.103925e-02 -2.584000e-02 -1.063122e-02
#> 39    0.22464590 -7.609593e-02  3.978988e-03  8.794952e-02
#> 40    0.29225340 -1.081544e-01 -3.648418e-02  6.264630e-03
#> 41    0.29397297 -5.052754e-02  1.373965e-01  8.752590e-02
#> 42    0.26042976  7.984307e-02 -9.085878e-02 -2.001732e-02
#> 43    0.26421686 -3.243179e-02 -1.996437e-02 -5.063418e-02
#> 44    0.25716126  7.189527e-02  4.136148e-02 -1.645894e-02
#> 45    0.28607638  7.357543e-02  6.910769e-02  5.822566e-02
#> 46    0.42127098 -1.484467e-02 -1.625491e-01  1.240198e-01
#> 47    0.38960157 -3.382644e-02 -1.979946e-01  3.762067e-02
#> 48    0.39194618 -1.146872e-01  8.684458e-02 -2.176148e-02
#> 49    0.37113553  9.940033e-03 -1.016895e-01 -3.029002e-02
#> 50    0.25581555 -1.424922e-01  8.111894e-02 -8.110421e-02
#> 51    0.30558732 -7.387948e-02 -1.274457e-01  1.866834e-01
#> 52    0.28862312  1.135103e-02 -1.741579e-01 -2.616053e-02
#> 53    0.27772880 -1.892198e-01 -3.677083e-02 -5.880412e-04
#> 54    0.28062777  4.443224e-02 -6.878527e-03  2.826450e-02
#> 55    0.28331612  7.774948e-02  1.453926e-02  4.141287e-02
#> 56    0.31089364 -1.714502e-01 -1.100032e-01 -6.559925e-02
#> 57    0.19286021  7.496657e-02  1.214785e-01 -8.140137e-03
#> 58    0.34782648  2.266520e-01 -1.463159e-02  9.421339e-02
#> 59    0.31453574 -1.272196e-01  2.859593e-02 -5.542584e-02
#> 60    0.31779710  6.823046e-02 -1.894925e-01 -1.246750e-02
#> 61    0.36254212 -4.886114e-03  2.175545e-01  1.004480e-01
#> 62    0.31065908  9.680059e-02  3.364149e-02 -1.027695e-01
#> 63    0.35914341  5.606009e-02 -1.378750e-01  4.742729e-02
#> 64    0.32871139 -1.358159e-02 -6.445641e-02 -7.019487e-02
#> 65    0.26012580  3.457918e-02 -5.878737e-02 -2.775377e-02
#> 66    0.21860021 -2.383343e-02  1.322026e-01  6.054744e-02
#> 67    0.28771665 -2.602573e-02 -3.095256e-01  3.261430e-02
#> 68    0.29370813  1.728482e-01  1.421718e-02  9.059222e-05
#> 69    0.31574170 -4.917538e-02 -2.353323e-01 -1.859918e-02
#> 70    0.22478679 -1.612067e-01  6.537917e-02  8.526319e-02
#> 71    0.16781096  4.205934e-02 -1.059597e-02  5.955280e-02
#> 72    0.21128982 -2.721104e-02  6.754738e-02 -9.174157e-02
#> 73    0.29563929  1.386818e-01  3.934044e-03  1.541611e-02
#> 74    0.30484075  6.949629e-02  1.974442e-02 -3.547467e-02
#> 75    0.36487872 -7.927130e-03  2.380696e-01  2.418847e-02
#> 76    0.37089019  1.119744e-03  3.961280e-02 -1.818077e-02
#> 77    0.37947369  8.270913e-02 -1.269059e-01 -1.177565e-01
#> 78    0.35071130  9.435358e-02 -6.718679e-02  9.217249e-02
#> 79    0.34300375 -1.253378e-01  1.939665e-02 -4.709118e-02
#> 80    0.32433368  6.700071e-02 -3.880262e-02  2.068658e-01
#> 81    0.40930390 -4.806398e-02  6.650056e-03 -5.145578e-02
#> 82    0.44228891 -5.067284e-02 -5.346488e-02  4.826815e-03
#> 83    0.44003561  7.776008e-02 -8.319203e-02 -5.591184e-02
#> 84    0.42889227 -2.103265e-02 -1.524381e-02 -3.801684e-02
#> 85    0.36628542  1.014784e-02  1.491811e-01 -9.064325e-02
#> 86    0.40082154  2.585225e-03  2.985128e-02  6.239453e-02
#> 87    0.24385113 -3.022089e-03  6.355927e-03 -3.713675e-02
#> 88    0.24680751  8.067700e-02  4.065684e-02 -6.811875e-02
#> 89    0.27799344  7.554572e-02 -2.822456e-02 -1.209399e-01
#> 90    0.26131281 -1.398972e-01  1.533975e-01  3.807792e-02
#> 91    0.28840899  1.161064e-01 -5.238367e-02  7.258710e-02
#> 92    0.34012326 -9.409973e-02 -5.524003e-02 -6.757572e-03
#> 93    0.37820888 -1.790405e-01 -6.848689e-03 -1.069918e-01
#> 94    0.35339219 -7.663311e-02 -1.228176e-01  7.435031e-02
#> 95    0.36053945  8.910808e-02 -3.782934e-02 -6.138177e-02
#> 96    0.42420289 -8.346181e-02  9.799345e-02  7.419424e-02
#> 97    0.37237304 -7.182849e-02 -2.869095e-03  2.713969e-01
#> 98    0.39740262 -1.156928e-01  4.387978e-02 -2.952416e-02
#> 99    0.43438104  7.574412e-02  1.581969e-01  8.993302e-02
#> 100   0.38286365  1.517240e-01 -3.044894e-02  8.892086e-02
#> 101   0.33978369  9.529702e-02  6.438603e-02 -5.243833e-02
#> 102   0.28791937  3.235223e-02 -1.478797e-02  2.101971e-01
#> 103   0.23570130 -2.054030e-01 -1.653610e-02 -3.658470e-02
#> 104   0.20937601  5.058327e-02 -2.243416e-02  8.262189e-02
#> 105   0.14244543 -1.087361e-01 -3.051165e-03  1.623164e-01
#> 106   0.14460750  1.426215e-01 -1.947906e-01  2.699515e-02
#> 107   0.15272086 -3.828729e-03 -8.324498e-02  8.118934e-03
#> 108   0.15988844  2.513437e-02 -1.483917e-02  2.078212e-02
#> 109   0.20279178 -2.313849e-02 -7.524723e-02 -2.967725e-02
#> 110   0.27188478  6.712118e-02  1.122190e-01 -1.377953e-01
#> 111   0.26538688 -8.854009e-02 -4.205251e-02  2.089468e-02
#>  [ reached getOption("max.print") -- omitted 9969 rows ]