Process Fault Filtering

Flag and filter out observations beyond normal operating conditions, then return the observations within normal operating conditions.

faultFilter(trainData, testData, updateFreq, faultsToTriggerAlarm = 5, ...)

Arguments

trainData

An xts data matrix of initial training observations

testData

The data not included in the training data set

updateFreq

The number of observations from the test data matrix that must be returned to update the training data matrix and move it forward.

faultsToTriggerAlarm

Specifies how many sequential faults will cause an alarm to trigger. Defaults to 5.

...

Lazy dots for additional internal arguments

Value

A list of class "fault_ls" with the following:

faultObj --

An xts flagging matrix with the same number of rows as "testData". This flag matrix has the following five columns:

SPE --

The SPE statistic value for each observation in "testData". This statistic is defined as $$ SPE_i = (\textbf{X}_i - \textbf{Y}_i * \textbf{P}^T) * (\textbf{X}_i - \textbf{Y}_i * \textbf{P}^T)^T, $$ where \(\textbf{X}_i\) is the \(i^{th}\) observation vector, \(\textbf{Y}_i\) is the reduced-feature projection of the observation \(\textbf{X}_i\), and \(\textbf{P}\) is the projection matrix such that \(\textbf{X}_i\textbf{P} = \textbf{Y}_i\).

SPE_Flag --

A vector of SPE indicators recording 0 if the test statistic is less than or equal to the critical value passed through from the threshold object.

T2 --

The T2 statistic value for each observation in "testData". This statistic is defined as $$ T^2_i = \textbf{Y}_i * \textbf{D}^{-1} * \textbf{Y}_i^T, $$ where \(\textbf{Y}_i = \textbf{X}_i\textbf{P}\) is the reduced- feature projection of the observation \(\textbf{X}_i\), and \(\textbf{D}\) is the diagonal matrix of eigenvalues.

T2_Flag --

A vector of T2 fault indicators, defined like SPE_Flag.

Alarm --

A column indicating if there have been five flags in a row for either the SPE or T2 monitoring statistics or both. Alarm states are as follows: 0 = no alarm, 1 = Hotelling's T2 alarm, 2 = Squared Prediction Error alarm, and 3 = both alarms.

nonAlarmedTestObs --

An xts matrix of the first updateFreq number of rows of the training data which were not alarmed.

trainSpecs --

The threshold object returned by the internal threshold() function. See the threshold() function's help file for more details.

Details

This function is essentially a wrapper function to call and organize the output from these other internal functions: faultDetect(), threshold(), and pca(). It is applied over a rolling window, with observation width equal to updateFreq, of the larger full data matrix via the processMonitor() function, wherein the testing and training data sets move forward in time across the entire data matrix.

This internal function is called by processMonitor().

See also

Examples

nrml <- mspProcessData(faults = "NOC")
# Select the data under state 1
data <- nrml[nrml[,1] == 1]

faultFilter(trainData = data[1:672, -1],
            testData = data[673:3360, -1],
            updateFreq = 336)
#> Warning: In density.default(spe, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(t2, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> $faultObj
#>                             SPE SPE_Flag        T2 T2_Flag Alarm
#> 2015-05-17 19:12:00 0.114012628        0 5.2994154       0     0
#> 2015-05-17 19:13:00 0.002402309        0 3.0062106       0     0
#> 2015-05-17 19:14:00 0.007964447        0 3.1443622       0     0
#> 2015-05-17 19:15:00 0.038697964        0 2.1595448       0     0
#> 2015-05-17 19:16:00 0.036240875        0 2.8320322       0     0
#> 2015-05-17 19:17:00 0.011886945        0 6.5023520       0     0
#> 2015-05-17 19:18:00 0.728691885        0 5.4810056       0     0
#> 2015-05-17 19:19:00 0.028917390        0 5.5329110       0     0
#> 2015-05-17 19:20:00 0.093200071        0 8.9966250       0     0
#> 2015-05-17 19:21:00 0.233027339        0 6.8183716       0     0
#>                 ...                                             
#> 2015-05-23 07:50:00 0.128252934        0 0.5369437       0     0
#> 2015-05-23 07:51:00 0.026082268        0 0.7453808       0     0
#> 2015-05-23 07:52:00 0.001706097        0 1.8140347       0     0
#> 2015-05-23 07:53:00 0.037245955        0 0.2783575       0     0
#> 2015-05-23 07:54:00 0.015213025        0 4.3032937       0     0
#> 2015-05-23 07:55:00 0.108337104        0 0.1712543       0     0
#> 2015-05-23 07:56:00 0.139579831        0 0.6713787       0     0
#> 2015-05-23 07:57:00 0.131536167        0 1.4041508       0     0
#> 2015-05-23 07:58:00 0.202659348        0 2.1136380       0     0
#> 2015-05-23 07:59:00 0.020621403        0 3.2833027       0     0
#> 
#> $nonAlarmedTestObs
#>                             x         y        z
#> 2015-05-17 19:12:00 0.8831227 -1.720693 1.367046
#> 2015-05-17 19:13:00 0.7406452 -1.686064 1.182735
#> 2015-05-17 19:14:00 0.7331701 -1.691586 1.208045
#> 2015-05-17 19:15:00 0.6248452 -1.547574 1.104100
#> 2015-05-17 19:16:00 0.7730620 -1.565715 1.126214
#> 2015-05-17 19:17:00 0.6767583 -1.889323 1.317484
#> 2015-05-17 19:18:00 0.8578063 -1.680764 1.577371
#> 2015-05-17 19:19:00 0.8574237 -1.843360 1.413390
#> 2015-05-17 19:20:00 0.9700749 -1.572398 1.197798
#> 2015-05-17 19:21:00 0.9294983 -1.578787 1.290115
#>                 ...                             
#> 2015-05-21 10:51:00 1.1178958 -2.331058 2.573294
#> 2015-05-21 10:52:00 1.1817480 -2.252094 2.581455
#> 2015-05-21 13:00:00 1.0529113 -1.928097 2.132666
#> 2015-05-21 13:01:00 0.9667312 -1.781061 1.872213
#> 2015-05-21 13:02:00 1.1564076 -2.047757 2.189126
#> 2015-05-21 13:03:00 1.1420387 -2.130697 2.306182
#> 2015-05-21 13:04:00 1.0413285 -2.086854 2.314878
#> 2015-05-21 13:05:00 1.1878938 -2.076798 2.297967
#> 2015-05-21 13:06:00 1.0339233 -2.082500 2.532387
#> 2015-05-21 13:07:00 1.0091753 -2.044071 2.425581
#> 
#> $trainSpecs
#> $SPE_threshold
#>   99.9% 
#> 2.91448 
#> 
#> $T2_threshold
#>    99.9% 
#> 16.25952 
#> 
#> $projectionMatrix
#>            [,1]       [,2]
#> [1,]  0.5623506  0.8267589
#> [2,] -0.5842834  0.4103069
#> [3,]  0.5851280 -0.3848609
#> 
#> $LambdaInv
#>           [,1]     [,2]
#> [1,] 0.3768492 0.000000
#> [2,] 0.0000000 4.250532
#> 
#> $T2
#>   [1] 6.551522e+00 9.439233e+00 3.412084e+00 3.197090e+00 4.151433e+00
#>   [6] 3.859182e+00 2.689194e+00 1.758905e+00 1.248102e+00 2.881858e+00
#>  [11] 1.168388e+00 7.353899e-01 1.272114e+00 2.031356e+00 2.960847e+00
#>  [16] 1.327887e+00 3.691610e+00 4.504127e+00 2.863777e+00 9.338857e-01
#>  [21] 1.443889e+00 4.567256e-01 9.474019e-01 8.740649e-01 1.628690e+00
#>  [26] 3.567680e+00 2.565156e+00 3.759187e+00 3.622561e+00 2.117632e+00
#>  [31] 1.654545e+00 2.608662e+00 2.254980e+00 4.221710e+00 5.015415e-01
#>  [36] 3.215727e-01 1.017691e+00 1.807831e+00 6.281944e-01 3.994228e-01
#>  [41] 1.507773e+00 2.716527e+00 2.445038e+00 3.933016e+00 1.895610e+00
#>  [46] 1.112194e+00 4.170857e+00 7.155295e+00 1.404097e+00 7.993167e-01
#>  [51] 1.134670e+00 4.145171e+00 6.233571e-01 1.467529e+00 2.018774e+00
#>  [56] 1.306454e+00 1.026113e+00 1.595729e+00 1.498925e-01 1.728652e+00
#>  [61] 1.439703e+00 1.220500e+00 2.670491e-01 1.004152e+00 2.927902e+00
#>  [66] 5.868196e-02 3.372273e-01 1.218119e+00 6.969588e-01 1.226669e+00
#>  [71] 5.820296e-01 6.831005e-02 1.470058e+00 4.471172e+00 1.488455e+00
#>  [76] 1.684747e+00 5.883334e-02 8.088522e-01 1.298597e+00 3.400330e-01
#>  [81] 3.959631e+00 2.636906e-01 6.728037e-01 3.372647e-01 1.574515e+00
#>  [86] 1.881713e+00 2.535861e+00 6.394796e+00 1.059834e+01 1.432755e+00
#>  [91] 1.170329e+00 2.165453e+00 2.675062e+00 2.198405e+00 1.690016e+00
#>  [96] 4.074502e+00 2.588816e+00 1.750909e+00 1.985580e+00 1.256237e+00
#> [101] 4.344655e-01 7.571733e-01 5.594816e-01 7.598781e-01 9.775112e-01
#> [106] 3.722293e-01 2.018883e+00 2.768770e+00 1.440013e+00 2.677807e+00
#> [111] 4.836684e+00 6.847074e-01 3.460243e+00 3.066121e+00 1.628278e+00
#> [116] 2.274892e+00 1.525244e+00 1.816777e+00 1.962926e+00 7.450932e+00
#> [121] 3.120191e-01 5.737877e-01 4.872809e-01 7.377916e-01 8.013006e-01
#> [126] 8.065692e+00 6.245140e-01 4.152857e-01 8.104940e-01 2.894058e+00
#> [131] 7.641550e-01 1.414128e+00 1.101932e+00 2.006119e+00 1.503314e+00
#> [136] 4.017555e-01 6.896495e-01 9.566367e-01 1.529054e+00 3.323378e-01
#> [141] 1.088858e+00 5.415755e-01 6.117768e-01 3.472465e-01 4.016805e+00
#> [146] 1.984403e+00 1.260734e+00 1.438619e+00 6.275932e-01 3.830255e-01
#> [151] 5.306930e+00 1.130783e+00 7.210710e-01 1.076503e+00 3.877141e+00
#> [156] 2.720276e+00 9.786778e-01 1.011334e+00 3.453078e+00 2.273706e+00
#> [161] 1.161064e+01 1.038404e+01 2.286200e+00 2.797637e+00 1.307716e+00
#> [166] 2.414179e+00 4.612374e+00 3.494836e+00 2.279230e+00 1.203414e+00
#> [171] 3.076002e+00 9.266264e-01 1.626543e+00 1.853976e+00 9.927873e-01
#> [176] 3.183302e+00 3.548041e-01 2.286279e+00 7.957146e-01 1.158715e+00
#> [181] 2.447844e+00 2.224734e+00 3.600841e+00 8.395150e-01 5.417943e-01
#> [186] 2.815086e+00 2.028037e+00 1.766711e-01 1.160356e-01 1.108433e+00
#> [191] 6.687701e-01 7.195428e-01 1.117939e+00 5.750266e-01 1.002850e+00
#> [196] 1.760828e+00 5.219571e+00 2.965497e-01 5.884794e-01 1.098403e+00
#> [201] 3.982666e-01 8.278166e-01 3.930857e-01 7.642908e-01 2.199220e+00
#> [206] 3.185795e+00 8.986328e-02 3.957922e-01 5.921212e-01 1.009919e+00
#> [211] 7.364840e-01 1.426776e+00 3.066362e-01 1.813722e-01 6.018372e-01
#> [216] 1.867100e+00 8.153861e-01 4.277007e-01 1.024554e+00 1.420703e+00
#> [221] 4.980939e-01 2.802418e-01 2.346590e+00 2.464387e+00 8.663153e-01
#> [226] 2.769631e+00 3.976660e-01 6.409067e-01 5.057412e-02 5.733974e-01
#> [231] 1.167132e+00 1.794286e+00 6.250843e+00 1.769212e+00 1.825412e+00
#> [236] 1.975726e+00 2.851534e+00 3.233460e-01 6.148401e+00 5.676354e+00
#> [241] 1.052812e+00 3.768355e-01 1.223564e+00 1.556697e+00 6.292527e-01
#> [246] 1.972581e+00 4.817237e+00 9.649989e-01 9.193815e-01 2.629231e-01
#> [251] 8.797094e-01 2.422352e+00 4.986001e-01 1.139692e+00 2.065995e+00
#> [256] 2.260065e+00 7.683177e-01 1.517892e+00 3.661281e+00 1.891719e+00
#> [261] 1.004085e+00 1.704789e+00 3.570154e-01 1.246365e-01 5.769463e-01
#> [266] 1.464280e-01 1.439935e-01 1.506208e+00 9.297193e-02 5.279739e-01
#> [271] 1.786913e-01 6.378246e-01 1.233482e+00 8.632760e-01 7.175575e-01
#> [276] 2.592148e+00 8.004946e-02 3.379095e-02 3.506017e+00 1.747119e+00
#> [281] 1.342048e+00 6.875257e-01 9.906450e-01 6.593032e-01 5.336188e-01
#> [286] 9.089864e-01 1.308135e+00 4.456939e-01 5.929324e-01 4.239871e-01
#> [291] 1.934779e+00 2.087190e-01 9.108579e-01 5.147581e+00 6.856740e-01
#> [296] 2.077639e-01 3.690553e-01 1.382405e+00 1.774733e-01 2.846509e+00
#> [301] 6.850511e-01 5.694058e-01 2.581933e-01 2.330203e+00 3.780249e+00
#> [306] 5.464431e-01 1.738609e+00 3.155217e+00 2.490933e+00 6.655898e-01
#> [311] 1.423435e+00 4.439169e-01 2.609512e+00 2.009539e-01 4.208429e-01
#> [316] 1.252570e+00 1.636446e-01 1.679556e+00 8.775643e-01 9.563814e-01
#> [321] 5.838828e-01 2.308621e+00 8.920815e+00 5.421244e-01 1.785947e+00
#> [326] 2.576219e-01 6.692409e-01 1.464568e+00 4.849814e-01 3.412771e+00
#> [331] 3.867104e+00 6.474634e-01 2.956767e-01 1.525860e+00 1.858763e+00
#> [336] 1.044602e+01 7.787226e-01 2.497516e+00 3.949413e-01 3.227976e-01
#> [341] 4.232351e-02 1.705320e+00 4.663422e-01 3.987786e-01 1.501821e-01
#> [346] 1.019991e+00 9.982891e-01 4.120643e-01 5.085303e-01 8.828145e-01
#> [351] 1.602580e+00 3.297992e+00 7.465028e-01 5.704190e-01 3.934216e+00
#> [356] 3.486418e-01 8.698504e-02 1.138648e+00 6.601201e-01 2.257248e+00
#> [361] 1.457821e+00 6.227144e+00 9.234356e-01 7.946019e+00 6.580260e-01
#> [366] 1.384791e-01 1.305515e+00 2.173678e-01 3.043240e-01 1.533179e+00
#> [371] 3.935299e+00 1.170541e+00 2.165972e+00 3.408203e-01 8.405546e-01
#> [376] 2.434042e+00 2.768323e-01 3.734571e+00 2.544095e+00 1.571605e+00
#> [381] 6.778552e+00 5.266492e+00 1.987945e-01 2.493786e+00 3.518354e-01
#> [386] 4.514965e+00 5.327444e+00 1.362964e+00 7.833943e-01 2.928344e+00
#> [391] 1.399806e+00 8.159430e-01 5.467736e-01 8.081502e-02 5.436937e-01
#> [396] 2.722836e-01 1.131129e+00 2.324128e-01 2.241279e-01 2.646421e-01
#> [401] 1.880727e-01 1.483921e+00 2.164450e+00 9.196239e-01 6.483539e-01
#> [406] 7.586327e-01 1.271866e+00 6.252460e+00 1.379966e-01 6.012656e-01
#> [411] 3.525312e-01 6.856439e-02 1.795095e-02 4.509618e-02 5.918149e-02
#> [416] 6.058116e-02 4.306390e-01 8.605887e-01 4.001980e-01 1.694552e-02
#> [421] 4.390667e-01 5.834798e-02 5.265117e+00 4.226206e+00 1.278921e+00
#> [426] 4.494196e-01 4.030206e+00 1.441378e+00 2.029924e+00 4.881951e+00
#> [431] 9.316790e-01 1.633048e+00 1.338355e+00 1.728211e-01 1.657673e+00
#> [436] 1.468674e+00 2.370173e+00 3.688130e+00 1.407147e+00 2.001535e+00
#> [441] 9.149556e-01 1.063780e+00 4.552041e-01 4.892065e-01 5.126598e-01
#> [446] 1.178909e-01 1.713549e+00 1.581727e+00 5.377208e-01 1.616414e-02
#> [451] 1.901504e-01 1.997259e+00 1.312765e-01 7.147726e-01 1.278782e+00
#> [456] 1.295785e+00 1.291055e+00 3.079476e-01 1.280283e-01 6.750831e-01
#> [461] 5.252460e-01 7.482876e-01 3.487728e+00 1.160722e+00 1.556976e-01
#> [466] 1.039474e+00 1.857791e+00 1.270780e+00 1.410846e+00 6.512002e-01
#> [471] 5.443597e-01 7.102621e-04 8.014121e-01 1.159894e+00 1.051121e+00
#> [476] 2.378183e+00 3.292437e-01 2.369494e-01 3.854702e-01 6.631456e-01
#> [481] 4.438053e+00 6.151577e-01 5.408247e+00 3.448041e-02 1.453834e+00
#> [486] 1.456798e+00 2.441326e-01 4.264649e+00 4.928764e+00 1.331135e+00
#> [491] 7.102002e-01 1.000250e+00 3.683210e-01 2.721267e+00 7.126671e-01
#> [496] 4.595012e-01 4.737999e-01 6.033767e-02 2.388735e+00 1.605929e+00
#> [501] 4.278154e+00 2.688133e+00 1.457154e+00 1.902134e-01 3.357279e-01
#> [506] 2.922788e-01 4.545711e-01 8.207770e-02 1.984804e-01 3.317988e-02
#> [511] 6.019062e-02 1.622957e-01 2.707267e+00 1.561035e-02 2.057555e+00
#> [516] 1.447182e+00 1.181562e+00 1.008535e+00 3.685150e+00 1.976095e+00
#> [521] 1.186315e+00 8.242048e-01 4.775026e+00 7.072341e+00 1.959507e+00
#> [526] 2.221328e+00 2.577738e-01 1.121267e+00 2.166322e-01 3.651194e+00
#> [531] 3.437371e+00 2.545661e+00 9.108046e-01 1.530165e+00 1.401226e+00
#> [536] 4.280714e-01 2.237498e-01 3.276412e+00 1.683014e-01 6.372522e-01
#> [541] 8.602088e-01 1.182013e+00 1.124761e+00 2.078344e+00 9.450136e+00
#> [546] 3.514076e+00 2.242475e+00 3.521863e+00 4.837773e+00 3.865173e+00
#> [551] 3.493553e+00 3.128692e+00 1.377516e+00 1.021048e+01 1.110774e-01
#> [556] 4.913616e-02 1.018783e+00 1.010453e-01 3.888640e+00 1.627198e+00
#> [561] 3.954820e-01 8.245410e-01 1.178490e+00 8.409476e-01 2.221527e+00
#> [566] 6.400151e-01 1.787066e+00 4.275517e+00 2.502530e+00 1.995051e+00
#> [571] 9.457639e-01 7.326074e-01 3.181081e+00 3.145126e+00 3.664467e+00
#> [576] 5.169723e-01 1.114470e+00 5.192890e-01 5.075245e-01 9.301191e-01
#> [581] 2.130859e+00 1.580025e+00 8.801310e-01 4.411971e+00 2.378034e+00
#> [586] 5.186215e+00 5.472733e+00 2.645174e+00 7.134696e-01 2.485430e+00
#> [591] 1.104431e+00 1.676149e+00 5.371223e-01 1.325106e+00 1.019476e+00
#> [596] 1.996294e+00 4.749630e-01 1.604394e+00 1.839346e+00 4.126309e+00
#> [601] 4.376062e+00 5.879254e+00 3.342499e+00 5.476567e+00 5.684301e+00
#> [606] 9.051871e+00 2.569312e+00 2.866601e+00 2.198651e+00 4.308517e+00
#> [611] 2.451445e+00 8.275868e+00 3.961349e+00 5.789760e+00 5.680271e+00
#> [616] 5.906551e+00 5.907873e+00 1.980857e+00 1.126420e+00 1.983081e+00
#> [621] 2.297996e+00 1.869878e+00 3.232986e+00 2.851045e+00 2.449409e+00
#> [626] 1.844877e+00 2.982357e+00 1.878357e+00 4.436561e+00 4.421505e+00
#> [631] 3.082434e+00 3.421267e+00 3.421722e+00 2.396798e+00 1.332523e+00
#> [636] 3.515564e+00 2.147037e+00 1.576335e+00 1.444669e+00 1.342638e+00
#> [641] 6.921923e-01 3.150893e+00 8.642605e+00 2.173427e+00 6.959341e+00
#> [646] 2.136901e+00 5.775094e+00 2.434964e+00 3.920064e+00 7.583764e+00
#> [651] 5.733639e+00 6.686729e+00 6.100116e+00 5.787260e+00 5.770043e+00
#> [656] 3.250733e+00 3.486441e+00 4.125342e+00 8.131937e+00 4.295933e+00
#> [661] 1.093702e+01 1.633588e+01 7.941434e+00 6.193117e+00 7.232097e+00
#> [666] 5.585071e+00 3.788524e+00 1.989203e+00 2.519701e+00 2.497809e+00
#> [671] 9.166807e+00 8.246196e+00
#> 
#> $SPE
#>   [1] 2.917694e+00 1.542737e+00 6.667556e-02 4.292230e-01 2.800446e-02
#>   [6] 4.478200e-03 3.830670e-01 4.284324e-02 2.444803e-03 6.975674e-02
#>  [11] 6.254343e-03 1.153547e-01 7.313376e-03 6.271754e-02 9.278073e-02
#>  [16] 1.153251e-02 8.627169e-02 1.321549e-01 2.769889e-03 7.501178e-03
#>  [21] 5.461022e-03 3.460791e-01 3.143627e-03 1.233117e-01 3.161661e-02
#>  [26] 3.999730e-03 2.545403e-01 7.298060e-04 1.666228e-01 5.925645e-02
#>  [31] 7.433331e-03 9.498655e-04 1.905576e-02 4.293919e-01 1.176391e-02
#>  [36] 8.967562e-02 2.420068e-02 3.773701e-03 1.099646e-01 2.776154e-01
#>  [41] 7.083302e-03 3.763156e-04 1.168870e-02 2.234148e-02 2.359425e-01
#>  [46] 1.743915e-02 5.078988e-01 2.928711e-01 6.071188e-02 8.865929e-03
#>  [51] 5.242488e-05 1.157230e-01 2.834217e-01 2.667170e-02 3.737663e-02
#>  [56] 2.567560e-01 5.863307e-02 2.704229e-02 1.898144e-05 4.361913e-04
#>  [61] 9.338757e-02 3.107910e-04 1.239832e-02 4.479287e-02 4.606415e-01
#>  [66] 2.708165e-01 1.303387e-01 8.763995e-03 4.885737e-04 2.328872e-02
#>  [71] 1.334397e-01 9.684206e-03 2.085106e-01 9.389130e-02 4.817881e-01
#>  [76] 4.364740e-03 3.038964e-02 6.805130e-02 1.416865e-01 8.753768e-02
#>  [81] 4.755182e-04 1.703637e-01 1.478439e-02 1.826446e-04 4.459963e-02
#>  [86] 6.444892e-03 1.856806e-02 2.557697e-01 9.301638e-03 1.323996e-01
#>  [91] 6.081069e-03 9.306306e-02 1.412157e-02 4.060165e-02 2.938643e-02
#>  [96] 1.494928e-01 1.891524e-02 6.337726e-02 6.372240e-02 6.353526e-02
#> [101] 1.424403e-01 1.159741e-01 1.635105e-02 2.416519e-01 3.692654e-01
#> [106] 3.344734e-02 1.617459e-02 4.535980e-02 1.307677e-03 1.635219e-01
#> [111] 7.063455e-02 1.429885e-01 2.822888e-03 1.194539e-01 8.584604e-03
#> [116] 5.127832e-02 8.081391e-02 2.416997e-02 6.148951e-02 7.053945e-02
#> [121] 7.991214e-02 1.125798e-02 2.943356e-02 2.955178e-02 5.838242e-03
#> [126] 1.660477e-03 1.537970e-02 1.110006e-01 4.907263e-02 2.595071e-01
#> [131] 2.697099e-04 2.380370e-06 6.172927e-03 8.125854e-03 1.459506e-02
#> [136] 2.885376e-02 9.396685e-06 6.740226e-04 2.496466e-01 1.690416e-02
#> [141] 1.356141e-01 1.067541e-01 9.753507e-03 2.623505e-02 1.474783e-01
#> [146] 1.540132e-03 8.122340e-01 4.005660e-01 3.196950e-02 2.263790e-03
#> [151] 7.615108e-03 1.508617e-01 6.842036e-03 7.110354e-02 1.927949e-01
#> [156] 4.916352e-02 6.457362e-02 2.544612e-02 2.814452e-02 7.269296e-03
#> [161] 7.348588e-05 7.293471e-01 1.575940e-02 4.202831e-02 8.805943e-05
#> [166] 7.558769e-02 8.217202e-03 4.411994e-02 2.556386e-03 2.597461e-03
#> [171] 1.020657e-02 6.982435e-04 1.486804e-02 1.214976e-02 4.852257e-03
#> [176] 6.213006e-01 4.002359e-01 3.507918e-03 1.018420e-04 1.350873e-01
#> [181] 5.728792e-01 3.420974e-01 1.632444e-02 7.145210e-02 1.781108e-02
#> [186] 2.926086e-01 7.297702e-02 3.827843e-02 9.027936e-03 7.058162e-02
#> [191] 7.257659e-02 2.315756e-01 9.942867e-03 2.326036e-02 2.313014e-02
#> [196] 3.087531e-01 6.975784e-03 1.613103e-02 3.985503e-03 2.502181e-03
#> [201] 4.189106e-02 1.897679e-03 1.294359e-01 1.846964e-02 1.955928e-01
#> [206] 1.634565e-01 9.584356e-02 1.512611e-02 5.719601e-02 1.563610e-01
#> [211] 3.202079e-01 4.165327e-02 6.002261e-03 2.493104e-01 2.708529e-02
#> [216] 8.636568e-05 3.369923e-01 5.680479e-03 2.074944e-03 4.221464e-01
#> [221] 1.264007e-02 2.580920e-04 8.875152e-02 7.481038e-03 3.872921e-02
#> [226] 1.424337e-01 2.398114e-01 4.580308e-04 8.314970e-02 4.629173e-02
#> [231] 1.407839e-02 2.046024e-01 5.251002e-04 7.357833e-02 1.398301e-01
#> [236] 4.382134e-01 1.812419e-01 1.885475e-03 1.980806e-02 1.208085e-01
#> [241] 5.937632e-02 7.365529e-02 5.050014e-02 4.334037e-01 1.971687e-02
#> [246] 2.283705e-01 3.854177e-02 1.745543e-01 4.003144e-03 8.456887e-03
#> [251] 1.482197e-01 1.282228e-01 1.252037e-02 7.539086e-02 3.406500e-02
#> [256] 1.335128e-01 5.141781e-02 4.310411e-02 4.878722e-02 2.320361e-02
#> [261] 1.029821e-02 1.086061e-05 2.901743e-02 2.484915e-02 1.086299e-01
#> [266] 5.995841e-01 1.469989e-04 5.541212e-01 9.421039e-02 1.918394e-02
#> [271] 2.453647e-04 5.041711e-02 5.568236e-02 3.459620e-04 6.247357e-03
#> [276] 3.495184e-02 9.813580e-02 1.392833e-05 1.747388e-01 1.000184e-01
#> [281] 2.475988e-03 9.190331e-04 3.740372e-05 1.514727e-04 2.073543e-02
#> [286] 4.867943e-01 5.978015e-01 9.704992e-03 1.019366e-01 7.363245e-05
#> [291] 6.753558e-02 2.207427e-02 9.408563e-02 2.018462e-02 3.392878e-02
#> [296] 5.862459e-02 7.521651e-01 1.637628e-02 5.130121e-02 7.849326e-03
#> [301] 1.538381e-04 1.009467e-03 6.809741e-02 5.128690e-01 1.304420e-02
#> [306] 2.701786e-02 5.762861e-02 1.841500e-02 6.079727e-03 3.608860e-02
#> [311] 1.917356e-02 2.403072e-02 4.134486e-04 1.925537e-01 1.433362e-01
#> [316] 3.673217e-02 1.577277e-02 3.094569e-04 9.059035e-02 1.455836e-02
#> [321] 4.729174e-03 2.039746e-02 7.249118e-02 4.412648e-03 1.293820e+00
#> [326] 2.375436e-02 4.396263e-03 1.203192e-01 2.112019e-02 1.038840e-02
#> [331] 7.777048e-02 2.290360e-02 5.023014e-02 1.397263e-02 1.050001e-04
#> [336] 4.188142e-02 9.828664e-04 8.931296e-02 4.234223e-02 2.064019e-01
#> [341] 3.527952e-02 1.125211e-03 1.421964e-05 4.381176e-01 1.447218e-01
#> [346] 1.596678e-03 8.660316e-02 3.761382e-02 3.636643e-03 2.956953e-02
#> [351] 2.829591e-01 2.373518e-01 3.027580e-02 3.932934e-01 4.680185e-01
#> [356] 4.617129e-02 8.761008e-02 3.015358e-01 2.095291e-01 2.020248e-02
#> [361] 3.380568e-02 2.531604e-01 5.167552e-02 1.647995e-01 6.774497e-01
#> [366] 1.450828e-01 6.609532e-04 7.758910e-02 1.106743e-01 1.843073e-01
#> [371] 9.068630e-03 3.401529e-01 8.965139e-05 1.955822e-02 1.288295e-01
#> [376] 1.559322e-02 6.684598e-02 5.064621e-02 3.135139e-02 1.372105e-02
#> [381] 3.134132e-02 1.017858e-01 1.183251e-01 1.000344e-01 1.410132e-01
#> [386] 4.089367e-02 1.291090e-02 9.704655e-02 1.173336e-01 2.292508e-01
#> [391] 7.816099e-02 9.647221e-02 1.937136e-04 9.228372e-04 2.880736e-02
#> [396] 2.935475e-03 1.745490e-03 1.962184e-04 1.207803e+00 2.735934e-03
#> [401] 1.028716e-02 3.730359e-02 5.422113e-02 1.008990e-02 3.040232e-02
#> [406] 3.500258e-02 9.723482e-03 6.110615e-02 7.214554e-02 6.988587e-02
#> [411] 2.428549e-03 1.986913e-04 5.651283e-02 2.232477e-02 2.993349e-01
#> [416] 9.220739e-02 3.151131e-02 4.284275e-02 2.100561e-02 2.091836e-02
#> [421] 5.035850e-02 1.795323e-02 9.838998e-03 1.952595e-02 1.078815e-02
#> [426] 1.541101e-05 6.847035e-03 1.333863e-01 3.738800e-03 2.063249e-03
#> [431] 3.760917e-01 2.235286e-02 7.905486e-02 2.133412e-01 1.556827e-03
#> [436] 5.230021e-03 1.750975e-01 1.277394e-01 4.617108e-04 5.330260e-02
#> [441] 4.178657e-03 7.031983e-02 3.130715e-01 3.524151e-04 2.809182e-02
#> [446] 5.629887e-03 5.638584e-02 2.676918e-02 3.897482e-02 1.201156e-01
#> [451] 4.785149e-02 3.855420e-02 1.039109e-01 3.378282e-01 1.314615e-01
#> [456] 3.637000e-03 4.211893e-02 5.312462e-03 6.746316e-01 2.812040e-01
#> [461] 7.985293e-04 5.668204e-03 2.054860e-01 7.716215e-02 3.512918e-01
#> [466] 1.939325e-02 8.212298e-03 1.279722e-01 3.149467e-01 1.522865e-01
#> [471] 4.715577e-02 8.285818e-02 8.458204e-02 6.054623e-02 1.470955e-02
#> [476] 1.045359e-01 4.291528e-02 1.442152e-01 1.186629e-02 2.278272e-03
#> [481] 1.163293e-01 8.045556e-01 1.064099e-05 4.433089e-03 2.368152e-01
#> [486] 1.401681e-02 2.025344e-03 1.963364e-01 2.056409e-01 7.056276e-02
#> [491] 1.397273e-02 4.400548e-02 1.003673e-01 8.099170e-01 2.928813e-01
#> [496] 5.924737e-03 2.564193e-02 2.705543e-02 2.031015e-01 6.433706e-03
#> [501] 1.337861e-01 1.651816e-02 4.053646e-01 2.094480e-01 3.886415e-02
#> [506] 2.972876e-04 1.356418e-01 8.577358e-07 7.457513e-03 1.111397e-01
#> [511] 2.407625e-03 2.118841e-01 9.225266e-02 9.303113e-02 3.230671e-02
#> [516] 1.496395e-01 8.826710e-02 1.315819e-02 1.191603e-02 6.845142e-01
#> [521] 2.874012e-02 9.684287e-03 8.315145e-03 3.076873e-01 4.148008e-02
#> [526] 5.496001e-02 1.283373e-01 3.956688e-03 2.652163e-02 6.548874e-03
#> [531] 1.175673e-01 6.628871e-02 3.941116e-02 1.448797e-01 4.704319e-02
#> [536] 3.062137e-03 7.033533e-05 8.856010e-02 2.209004e-01 2.514308e-01
#> [541] 8.313714e-02 5.097552e-04 2.925875e-01 3.409587e-04 3.441703e-02
#> [546] 5.378989e-01 1.114617e-04 6.922915e-02 2.131356e-03 6.522646e-03
#> [551] 3.306621e-01 1.973110e-01 2.267840e-01 2.188613e-02 3.980497e-02
#> [556] 9.392256e-03 2.476072e-02 3.327912e-01 1.136783e-05 1.988466e-01
#> [561] 3.389733e-03 1.944779e-01 9.916577e-02 2.550061e-01 7.083697e-02
#> [566] 1.804081e-02 1.101262e-03 1.601144e-02 2.718453e-01 1.816128e-03
#> [571] 6.748546e-03 2.570747e-01 1.384563e-01 4.069339e-01 3.771216e-02
#> [576] 1.039819e-05 3.261888e-03 1.637724e-01 1.969674e-07 3.120793e-03
#> [581] 7.618055e-02 4.819658e-01 5.357757e-02 8.289066e-02 1.757695e-02
#> [586] 1.864402e-02 8.100095e-02 6.215484e-03 2.551308e-05 2.142563e-01
#> [591] 1.325618e-01 1.717159e-01 1.016466e-02 5.814643e-02 1.070216e-01
#> [596] 6.250508e-03 5.741948e-02 9.697984e-03 2.046423e-02 7.410872e-01
#> [601] 3.094976e-01 1.992845e-01 5.035046e-02 3.501052e-01 2.046500e-01
#> [606] 3.988314e-03 1.705515e-01 1.307674e-01 4.857705e-02 6.413319e-01
#> [611] 8.719199e-03 1.060286e-01 4.190951e-01 9.820191e-02 6.985709e-02
#> [616] 4.901433e-02 1.030304e-01 2.696289e-04 4.044369e-02 5.276622e-04
#> [621] 6.930278e-02 1.425392e-01 2.633019e-01 2.344436e-02 3.745397e-01
#> [626] 3.549106e-02 5.899364e-03 1.596725e-01 1.080903e-01 2.645562e-01
#> [631] 8.091863e-02 3.983128e-02 2.130739e-02 2.118112e-02 8.410764e-02
#> [636] 4.750461e-02 3.692911e-03 1.226899e-02 1.221036e-02 6.306115e-03
#> [641] 1.389107e-03 4.864757e-02 1.773619e-01 1.437213e-01 7.881893e-04
#> [646] 2.102891e-05 3.103670e-02 2.798554e-02 1.720443e-03 2.187656e-01
#> [651] 1.643462e-01 2.702200e-01 1.146448e-01 1.112599e-03 4.066179e-01
#> [656] 2.563982e-02 1.395682e-01 9.476803e-02 1.166768e-01 2.913521e-01
#> [661] 1.459940e+00 3.177181e-01 7.608325e-02 1.744309e-01 2.306552e-01
#> [666] 9.351454e-02 2.579032e-02 5.941472e-02 1.255435e-01 1.319636e-01
#> [671] 1.159556e-01 2.935587e-01
#> 
#> $muTrain
#>          x          y          z 
#>  0.4660535 -1.1510662  0.5814697 
#> 
#> $RootPrecisTrain
#>          [,1]    [,2]     [,3]
#> [1,] 5.452637 0.00000 0.000000
#> [2,] 0.000000 3.07409 0.000000
#> [3,] 0.000000 0.00000 2.850403
#> 
#> attr(,"class")
#> [1] "threshold" "pca"      
#> 
#> attr(,"class")
#> [1] "fault_ls"