Induce the Specified Fault on NOC Observations

Infect the input data frame with a specific fault, then return the infected data frame.

faultSwitch(df, fault, period = 7 * 24 * 60, faultStartIndex = round(0.8433
  * period), shift = 2, postStateSplit = FALSE)

Arguments

df
A data frame returned by the processNOCdata() function.

fault
A character string. Options are "NOC", "A1", "B1", "C1", "A2", "B2", "C2", "A3", "B3", or "C3". See "details" of mspProcessData() for more information.

period
The observation cycle length. Defaults to one week's worth of minute-level observations (10,080 observations).

faultStartIndex
An integer specifying the index at which the faults will start. Defaults to roughly 85 percent through the cycle.

shift
The fault parameter for faults "A1" and "B1" corresponding to the positive shock value added to features. Defaults to 2. See "details" of mspProcessData() for more information.

postStateSplit
Should we induce faults before or after state-splitting? Defaults to FALSE. Make this argument TRUE for faults 1C, 2C, 3C.

Value

A data frame with the same structure as df, but with faults induced across all observations. The mspProcessData() function then subsets the observations necessary to corrupt the normal data frame, and binds them together by row. This function is called by mspProcessData(). See ?mspProcessData for more details.

Details

The faults return data frames as follows:

A1 -- A data frame with 10080 rows and five columns, corresponding by default to one week worth of data recorded at a 1-minute interval (as defined by the "period" argument of this function and the "increment" argument of the processNOCdata() function). The fault is a system shift to each of the three features by 2 (the "shift" argument). The fault starts at row 8500 (specified by the argument "faultStartIndex"), and the five columns under the fault state are defined here:

dateTime : a POSIXct column

state : the state indicator for the multivariate system, with three levels when the argument "multiState" is TRUE and one level otherwise

x : x(t) = t + shift + error

y : y(t) = t ^ 2 - 3t + shift + error

z : z(t) = -t ^ 3 + 3t ^ 2 + shift + error

where t is a 10080-entry vector of autocorrelated and non-stationary hidden process realizations generated within the processNOCdata() function.

B1 -- A matrix as defined in A1, but with x, y, and z feature columns defined as follows:

x : x(t) = t + shift + error

y : y(t) = t ^ 2 - 3t + error

z : z(t) = -t ^ 3 + 3t ^ 2 + error

C1 -- A matrix as defined in A1, but with x, y, and z feature columns defined as follows:

x : x(t) = t + shift / 4 + error

y : y(t) = t ^ 2 - 3t + error

z : z(t) = -t ^ 3 + 3t ^ 2 + shift / 4 + error

This shift is applied only in State 3.

A2 -- The fault is a drift on each feature by (s - faultStartIndex / 10 ^ 3, where s is the observation index. The fault starts at "faultStartIndex", and the x, y, and z feature columns are defined as follows:

x : x(t) = t + drift + error

y : y(t) = t ^ 2 - 3t + drift + error

z : z(t) = -t ^ 3 + 3t ^ 2 + drift + error

B2 -- The fault is a drift a drift on the "y" and "z" feature by (s - faultStartIndex / 10 ^ 3, where s is the observation index. The fault starts at "faultStartIndex", and the x, y, and z feature columns are defined as follows:

x : x(t) = t + error

y : y(t) = t ^ 2 - 3t + drift + error

z : z(t) = -t ^ 3 + 3t ^ 2 + drift + error

C2 -- The fault is a negative drift on the "y" feature by 1.5 * (s - faultStartIndex) / (period - faultStartIndex). Thus,

x : x(t) = t + error

y : y(t) = t ^ 2 - 3t - drift + error

z : z(t) = -t ^ 3 + 3t ^ 2 + error

This drift is applied only in State 2.

A3 -- The fault is a signal amplificaton in the determining latent t vector. The fault starts at "faultStartIndex", and the x, y, and z features under the fault state are defined here:

x : x(t_*) = t_* + error

y : y(t_*) = (t_*) ^ 2 - 3t_* + error

z : z(t_*) = -(t_*) ^ 3 + 3(t_*) ^ 2 + error

where t_* = 5 x t x (period - s) / (period - faultStartIndex) and s is the observation index.

B3 -- The fault is a signal amplificaton in the determining latent t vector for the "z" feature only. The fault starts at "faultStartIndex", and the x, y, and z features under the fault state are defined here:

x : x(t) = t + error

y : y(t) = (t) ^ 2 - 3t + error

z : z(t_*) = -(t_*) ^ 3 + 3(t_*) ^ 2 + error

where t_* = 3 x t x (period - s) / (2 x period) and s is the observation index.

C3 -- This fault is a change in the error structure of feature "y". We let errorNew = 2 * error - 0.25, so that

x : x(t) = t + error

y : y(t) = t ^ 2 - 3t + errorNew

z : z(t) = -t ^ 3 + 3t ^ 2 + error

This new error structure is applied only in State 2.

See also

Called by: mspProcessData.

Examples

nrml <- processNOCdata()
faultSwitch(nrml, fault = "NOC")

#>                t          err1          err2          err3             x
#> 1     0.05216109  1.654803e-03 -6.410950e-02  6.007887e-02  5.381589e-02
#> 2     0.17481975  1.751476e-01 -6.743405e-02  1.532706e-01  3.499674e-01
#> 3     0.18648672  1.026655e-01 -1.573645e-02  1.449503e-01  2.891522e-01
#> 4     0.18474486  2.620330e-02  3.775899e-03  1.137589e-01  2.109482e-01
#> 5     0.18280758 -1.436657e-01  1.327616e-01 -2.212674e-01  3.914184e-02
#> 6     0.19835412  1.383920e-01  5.014908e-03 -6.326462e-02  3.367461e-01
#> 7     0.29160354 -2.129416e-01 -8.837956e-02 -1.748198e-01  7.866190e-02
#> 8     0.23540542 -8.799240e-02  4.187698e-02 -7.027085e-02  1.474130e-01
#> 9     0.25227175 -6.819602e-02 -6.740920e-02  1.922782e-01  1.840757e-01
#> 10    0.23357616 -7.812716e-03  5.808588e-02  5.079423e-03  2.257634e-01
#> 11    0.25980092 -2.583871e-01 -1.611995e-01 -2.085293e-02  1.413832e-03
#> 12    0.22816999  5.069082e-03  1.669900e-01  1.012231e-01  2.332391e-01
#> 13    0.18932971 -3.731243e-03 -9.716160e-02  2.038942e-01  1.855985e-01
#> 14    0.21053965 -4.813278e-02  1.025834e-01 -5.960064e-02  1.624069e-01
#> 15    0.12312917  4.915082e-02 -1.765777e-01  1.793727e-02  1.722800e-01
#> 16    0.11776185  3.036968e-02 -5.621336e-02  3.634789e-02  1.481315e-01
#> 17    0.12806638 -1.353608e-01  2.501334e-02 -4.662319e-03 -7.294382e-03
#> 18    0.15492769 -9.256065e-02 -4.420176e-02 -1.239921e-01  6.236704e-02
#> 19    0.12756807 -5.926917e-02 -2.779897e-03 -6.187339e-02  6.829890e-02
#> 20    0.14854137 -8.823591e-02  3.915211e-02 -9.879904e-02  6.030546e-02
#> 21    0.15366243 -2.106387e-01 -1.564349e-01 -3.563502e-02 -5.697630e-02
#> 22    0.18385240  4.551490e-03  1.641845e-02  1.047322e-01  1.884039e-01
#> 23    0.21684314 -1.270761e-01  1.703366e-02  1.027040e-01  8.976699e-02
#> 24    0.26312634  5.345050e-02  6.804011e-02  5.184745e-02  3.165768e-01
#> 25    0.22538031  5.665326e-02  2.663455e-01 -1.600472e-02  2.820336e-01
#> 26    0.16879762  6.124219e-02 -1.694388e-01 -9.249555e-02  2.300398e-01
#> 27    0.23181222 -5.343008e-02 -8.618127e-03  1.598446e-01  1.783821e-01
#> 28    0.21698616  1.684523e-01  4.662384e-02  3.735759e-02  3.854384e-01
#> 29    0.25259957  1.219834e-01  1.410162e-01  7.103503e-02  3.745830e-01
#> 30    0.22383156  1.088131e-01  1.450765e-01 -2.457085e-02  3.326447e-01
#> 31    0.16640338 -2.014684e-02  3.898110e-02  2.003424e-01  1.462565e-01
#> 32    0.18250327  2.852114e-01 -9.254174e-02 -2.865667e-02  4.677147e-01
#> 33    0.23197866 -6.987396e-02  1.564410e-01 -1.578631e-02  1.621047e-01
#> 34    0.25071814 -1.797035e-02 -1.183683e-01  1.293575e-01  2.327478e-01
#> 35    0.37769472  8.441193e-02  1.562688e-01  5.292760e-02  4.621067e-01
#> 36    0.33040606 -1.856650e-02  2.092980e-01 -7.721869e-03  3.118396e-01
#> 37    0.21865980  5.926980e-03  8.274327e-02  5.590479e-02  2.245868e-01
#> 38    0.25974647 -1.267305e-01  3.059519e-02  9.667269e-02  1.330160e-01
#> 39    0.20523403 -6.735981e-02 -8.523136e-02  2.990432e-02  1.378742e-01
#> 40    0.18072971  6.205100e-02  1.259100e-01  1.208721e-01  2.427807e-01
#> 41    0.22388810  3.530588e-02  3.307531e-02  2.454276e-01  2.591940e-01
#> 42    0.18202616 -1.751452e-01 -4.719239e-02  2.418899e-01  6.880972e-03
#> 43    0.21901051  1.033582e-01  4.641949e-03 -9.320367e-02  3.223687e-01
#> 44    0.22426583 -1.928186e-01 -3.060255e-02 -1.901378e-02  3.144721e-02
#> 45    0.18567018 -6.584405e-04 -1.382734e-01 -3.029529e-03  1.850117e-01
#> 46    0.22743480  2.018176e-01  9.790304e-02  1.242607e-01  4.292524e-01
#> 47    0.27674169 -1.184730e-02 -3.235066e-02  8.082901e-02  2.648944e-01
#> 48    0.37188651 -2.894964e-01  3.478806e-02  1.194688e-01  8.239012e-02
#> 49    0.33563094 -4.563078e-02 -2.922214e-02 -3.044229e-02  2.900002e-01
#> 50    0.36177393 -5.589941e-02 -1.166946e-01 -3.455022e-02  3.058745e-01
#> 51    0.41585981 -8.186427e-03 -5.157996e-02  1.176086e-01  4.076734e-01
#> 52    0.31209098 -9.825532e-02 -1.019380e-02  7.186917e-02  2.138357e-01
#> 53    0.27099187 -1.455124e-02 -6.937036e-02 -2.063635e-02  2.564406e-01
#> 54    0.33372833 -7.875378e-02 -1.137644e-01 -6.131106e-02  2.549746e-01
#> 55    0.15221091 -1.519438e-02 -6.525431e-02  7.580688e-02  1.370165e-01
#> 56    0.19071966 -1.784468e-01 -4.506130e-03  1.941729e-02  1.227286e-02
#> 57    0.28198129  9.026189e-02  2.078104e-01 -1.747611e-01  3.722432e-01
#> 58    0.25133104 -1.952194e-02  2.112333e-02 -2.025784e-01  2.318091e-01
#> 59    0.25940725 -1.737475e-01 -8.038930e-02 -5.563532e-02  8.565978e-02
#> 60    0.25721937  8.991934e-02 -2.366365e-01 -9.545741e-02  3.471387e-01
#> 61    0.23627476 -1.047916e-01 -9.586223e-03  4.030617e-02  1.314832e-01
#> 62    0.09959284  1.792294e-01 -4.108974e-02  1.416861e-01  2.788222e-01
#> 63    0.19329836  6.309029e-02  1.855059e-01 -1.630555e-02  2.563886e-01
#> 64    0.24147886 -7.093059e-03  2.152456e-01 -1.580449e-02  2.343858e-01
#> 65    0.32210861  2.422262e-02  1.214206e-01  7.151440e-02  3.463312e-01
#> 66    0.23297400 -4.776751e-02  6.095729e-02 -6.035194e-02  1.852065e-01
#> 67    0.36342313 -4.900767e-02  4.804685e-02  1.583387e-01  3.144155e-01
#> 68    0.36504695 -2.143558e-02 -1.499719e-01  1.249712e-02  3.436114e-01
#> 69    0.27198900  6.551856e-02  7.059923e-02 -4.053956e-03  3.375076e-01
#> 70    0.29432317  5.592512e-02 -8.420221e-02  1.950812e-01  3.502483e-01
#> 71    0.31274851 -3.065299e-02 -8.989487e-02 -1.007794e-01  2.820955e-01
#> 72    0.29818128 -1.016791e-01 -5.897948e-02  3.744293e-02  1.965022e-01
#> 73    0.34413112  1.180055e-01 -6.284064e-02 -2.549739e-03  4.621366e-01
#> 74    0.34532321  2.214300e-02 -3.360587e-02  2.578193e-02  3.674662e-01
#> 75    0.19977467  2.252163e-02  1.137745e-02  1.064833e-01  2.222963e-01
#> 76    0.14199452 -8.556447e-02 -6.914590e-02 -4.118403e-02  5.643005e-02
#> 77    0.21090447 -6.997116e-02 -1.478482e-01  9.424352e-02  1.409333e-01
#> 78    0.26856367  1.603929e-03 -2.149460e-02 -5.676591e-02  2.701676e-01
#> 79    0.21672714 -7.910288e-03 -2.235430e-02 -5.101468e-02  2.088169e-01
#> 80    0.15573590  1.800235e-01 -5.472951e-02 -6.497325e-02  3.357594e-01
#> 81    0.10405025  7.522646e-02 -5.164879e-03 -1.184415e-02  1.792767e-01
#> 82    0.18345997  5.148324e-02 -5.023004e-02  1.459946e-01  2.349432e-01
#> 83    0.20872455 -1.293244e-01 -4.229274e-02 -3.002617e-02  7.940015e-02
#> 84    0.24472797  1.225440e-01 -7.363627e-02 -4.377704e-02  3.672719e-01
#> 85    0.29046548  5.798680e-02  4.948554e-02  7.252343e-02  3.484523e-01
#> 86    0.20302491 -3.049426e-02  8.721045e-02 -6.513441e-02  1.725306e-01
#> 87    0.31567795  2.360596e-01  3.282920e-02  9.091618e-04  5.517375e-01
#> 88    0.35192132  5.033281e-02 -1.732565e-01  6.363448e-02  4.022541e-01
#> 89    0.32927604  7.145172e-02  1.346954e-01  1.211778e-01  4.007278e-01
#> 90    0.33879792 -8.610333e-03  6.373917e-02 -1.017512e-01  3.301876e-01
#> 91    0.28074074 -2.099633e-01 -1.059260e-01 -1.563774e-02  7.077739e-02
#> 92    0.32219099  3.149971e-02  7.433253e-03  1.170840e-01  3.536907e-01
#> 93    0.27152822  1.359551e-01  3.596544e-03 -4.903697e-03  4.074833e-01
#> 94    0.30904247 -7.785037e-02  1.489064e-01 -1.671977e-01  2.311921e-01
#> 95    0.26567023 -1.765174e-02 -2.112464e-02 -3.712928e-02  2.480185e-01
#> 96    0.28196906  1.453819e-01 -3.754966e-02 -1.412807e-01  4.273510e-01
#> 97    0.29240801 -1.524119e-01 -1.940036e-01  2.181554e-01  1.399961e-01
#> 98    0.24002831  1.127006e-01 -1.236371e-01  7.462745e-02  3.527289e-01
#> 99    0.26189333  2.709260e-01  1.220333e-01  3.227387e-02  5.328193e-01
#> 100   0.28347533  2.069836e-01  2.188621e-01  7.160896e-02  4.904589e-01
#> 101   0.29012649 -4.912399e-02 -6.992748e-02 -9.865345e-02  2.410025e-01
#> 102   0.28107540 -1.869908e-02 -6.646350e-02  1.007517e-01  2.623763e-01
#> 103   0.18785839 -1.754837e-01  1.698172e-01  9.786129e-02  1.237466e-02
#> 104   0.16025167  8.858998e-02  8.468867e-02  5.089884e-02  2.488417e-01
#> 105   0.20867673  2.275556e-02 -4.150453e-02 -3.364392e-03  2.314323e-01
#> 106   0.28453466  1.823680e-01  7.177413e-02  5.112056e-02  4.669027e-01
#> 107   0.17364964  6.536252e-02  3.207951e-02  9.983124e-02  2.390122e-01
#> 108   0.17902514 -1.412840e-01  2.061399e-02  3.823211e-02  3.774114e-02
#> 109   0.16963468 -2.099802e-02 -7.717366e-02  3.286250e-02  1.486367e-01
#> 110   0.18473381 -1.880955e-01 -6.612333e-02  2.244413e-02 -3.361650e-03
#> 111   0.14222675  1.859802e-01  5.699957e-02 -2.209816e-01  3.282069e-01
#>                 y             z            dateTime state
#> 1     -0.21787199  6.809929e-02 2015-05-16 10:00:00     1
#> 2     -0.56133136  2.396136e-01 2015-05-16 10:01:00     1
#> 3     -0.54041932  2.427967e-01 2015-05-16 10:02:00     1
#> 4     -0.51632801  2.098454e-01 2015-05-16 10:03:00     1
#> 5     -0.38224254 -1.271207e-01 2015-05-16 10:04:00     1
#> 6     -0.55070310  4.696433e-02 2015-05-16 10:05:00     1
#> 7     -0.87815756  5.548226e-02 2015-05-16 10:06:00     1
#> 8     -0.60892356  8.293112e-02 2015-05-16 10:07:00     1
#> 9     -0.76058343  3.671464e-01 2015-05-16 10:08:00     1
#> 10    -0.58808477  1.560095e-01 2015-05-16 10:09:00     1
#> 11    -0.87310576  1.641010e-01 2015-05-16 10:10:00     1
#> 12    -0.46545842  2.455289e-01 2015-05-16 10:11:00     1
#> 13    -0.62930499  3.046448e-01 2015-05-16 10:12:00     1
#> 14    -0.48470865  6.404762e-02 2015-05-16 10:13:00     1
#> 15    -0.53080446  6.155291e-02 2015-05-16 10:14:00     1
#> 16    -0.39563105  7.631835e-02 2015-05-16 10:15:00     1
#> 17    -0.34278480  4.244026e-02 2015-05-16 10:16:00     1
#> 18    -0.48498225 -5.570305e-02 2015-05-16 10:17:00     1
#> 19    -0.36921050 -1.512855e-02 2015-05-16 10:18:00     1
#> 20    -0.38440747 -3.588292e-02 2015-05-16 10:19:00     1
#> 21    -0.59381001  3.157310e-02 2015-05-16 10:20:00     1
#> 22    -0.50133705  1.999228e-01 2015-05-16 10:21:00     1
#> 23    -0.58647482  2.335707e-01 2015-05-16 10:22:00     1
#> 24    -0.65210343  2.413362e-01 2015-05-16 10:23:00     1
#> 25    -0.35899920  1.249357e-01 2015-05-16 10:24:00     1
#> 26    -0.64733905 -1.182713e-02 2015-05-16 10:25:00     1
#> 27    -0.65031788  3.085985e-01 2015-05-16 10:26:00     1
#> 28    -0.55725166  1.683902e-01 2015-05-16 10:27:00     1
#> 29    -0.55297599  2.463371e-01 2015-05-16 10:28:00     1
#> 30    -0.47631760  1.145168e-01 2015-05-16 10:29:00     1
#> 31    -0.43253896  2.788049e-01 2015-05-16 10:30:00     1
#> 32    -0.60674411  6.518694e-02 2015-05-16 10:31:00     1
#> 33    -0.48568085  1.331723e-01 2015-05-16 10:32:00     1
#> 34    -0.80766309  3.021762e-01 2015-05-16 10:33:00     1
#> 35    -0.83416211  4.270081e-01 2015-05-16 10:34:00     1
#> 36    -0.67275202  2.837128e-01 2015-05-16 10:35:00     1
#> 37    -0.52542402  1.888865e-01 2015-05-16 10:36:00     1
#> 38    -0.68117598  2.815527e-01 2015-05-16 10:37:00     1
#> 39    -0.65881243  1.476227e-01 2015-05-16 10:38:00     1
#> 40    -0.38361593  2.129585e-01 2015-05-16 10:39:00     1
#> 41    -0.58846311  3.845826e-01 2015-05-16 10:40:00     1
#> 42    -0.56013735  3.352593e-01 2015-05-16 10:41:00     1
#> 43    -0.60442397  4.018817e-02 2015-05-16 10:42:00     1
#> 44    -0.65310488  1.205922e-01 2015-05-16 10:43:00     1
#> 45    -0.66081050  9.399004e-02 2015-05-16 10:44:00     1
#> 46    -0.53267478  2.676760e-01 2015-05-16 10:45:00     1
#> 47    -0.78598977  2.893924e-01 2015-05-16 10:46:00     1
#> 48    -0.94257191  4.829358e-01 2015-05-16 10:47:00     1
#> 49    -0.92346684  2.696939e-01 2015-05-16 10:48:00     1
#> 50    -1.07113597  3.107418e-01 2015-05-16 10:49:00     1
#> 51    -1.12622001  5.645082e-01 2015-05-16 10:50:00     1
#> 52    -0.84906596  3.336736e-01 2015-05-16 10:51:00     1
#> 53    -0.80890938  1.797727e-01 2015-05-16 10:52:00     1
#> 54    -1.00357481  2.356439e-01 2015-05-16 10:53:00     1
#> 55    -0.49871888  1.417849e-01 2015-05-16 10:54:00     1
#> 56    -0.54029112  1.216020e-01 2015-05-16 10:55:00     1
#> 57    -0.55862004  4.135797e-02 2015-05-16 10:56:00     1
#> 58    -0.66970251 -2.895242e-02 2015-05-16 10:57:00     1
#> 59    -0.79131892  1.287850e-01 2015-05-16 10:58:00     1
#> 60    -0.94213277  8.600990e-02 2015-05-16 10:59:00     1
#> 61    -0.66258474  1.945932e-01 2015-05-16 11:00:00     2
#> 62    -0.32994954  1.704545e-01 2015-05-16 11:01:00     2
#> 63    -0.35702489  8.856477e-02 2015-05-16 11:02:00     2
#> 64    -0.45087890  1.450505e-01 2015-05-16 11:03:00     2
#> 65    -0.74115124  3.493562e-01 2015-05-16 11:04:00     2
#> 66    -0.58368782  8.983361e-02 2015-05-16 11:05:00     2
#> 67    -0.91014616  5.065682e-01 2015-05-16 11:06:00     2
#> 68    -1.11185350  3.636291e-01 2015-05-16 11:07:00     2
#> 69    -0.67138975  1.977589e-01 2015-05-16 11:08:00     2
#> 70    -0.88054560  4.294635e-01 2015-05-16 11:09:00     2
#> 71    -0.93032876  1.620650e-01 2015-05-16 11:10:00     2
#> 72    -0.86461125  2.776672e-01 2015-05-16 11:11:00     2
#> 73    -0.97680778  3.119748e-01 2015-05-16 11:12:00     2
#> 74    -0.95032738  3.423471e-01 2015-05-16 11:13:00     2
#> 75    -0.54803663  2.182400e-01 2015-05-16 11:14:00     2
#> 76    -0.47496702  1.644034e-02 2015-05-16 11:15:00     2
#> 77    -0.73608089  2.183044e-01 2015-05-16 11:16:00     2
#> 78    -0.75505917  1.402429e-01 2015-05-16 11:17:00     2
#> 79    -0.62556507  7.971747e-02 2015-05-16 11:18:00     2
#> 80    -0.49768354  4.010599e-03 2015-05-16 11:19:00     2
#> 81    -0.30648918  1.950872e-02 2015-05-16 11:20:00     2
#> 82    -0.56695240  2.407925e-01 2015-05-16 11:21:00     2
#> 83    -0.62490044  9.157836e-02 2015-05-16 11:22:00     2
#> 84    -0.74792840  1.212411e-01 2015-05-16 11:23:00     2
#> 85    -0.73754070  3.011274e-01 2015-05-16 11:24:00     2
#> 86    -0.48064517  5.015443e-02 2015-05-16 11:25:00     2
#> 87    -0.81455208  2.684087e-01 2015-05-16 11:26:00     2
#> 88    -1.10517182  3.915954e-01 2015-05-16 11:27:00     2
#> 89    -0.74471005  4.107449e-01 2015-05-16 11:28:00     2
#> 90    -0.83787057  2.037123e-01 2015-05-16 11:29:00     2
#> 91    -0.86933282  1.986817e-01 2015-05-16 11:30:00     2
#> 92    -0.85533269  3.950595e-01 2015-05-16 11:31:00     2
#> 93    -0.73726053  1.962599e-01 2015-05-16 11:32:00     2
#> 94    -0.68271377  8.980830e-02 2015-05-16 11:33:00     2
#> 95    -0.74755466  1.558615e-01 2015-05-16 11:34:00     2
#> 96    -0.80395030  7.482059e-02 2015-05-16 11:35:00     2
#> 97    -0.98572523  4.496611e-01 2015-05-16 11:36:00     2
#> 98    -0.78610847  2.336393e-01 2015-05-16 11:37:00     2
#> 99    -0.59505855  2.200755e-01 2015-05-16 11:38:00     2
#> 100   -0.55120563  2.899042e-01 2015-05-16 11:39:00     2
#> 101   -0.85613357  1.294458e-01 2015-05-16 11:40:00     2
#> 102   -0.83068631  3.155559e-01 2015-05-16 11:41:00     2
#> 103   -0.35846715  1.971039e-01 2015-05-16 11:42:00     2
#> 104   -0.37038575  1.238253e-01 2015-05-16 11:43:00     2
#> 105   -0.62398875  1.181865e-01 2015-05-16 11:44:00     2
#> 106   -0.70086989  2.709646e-01 2015-05-16 11:45:00     2
#> 107   -0.45871521  1.850576e-01 2015-05-16 11:46:00     2
#> 108   -0.48441144  1.286444e-01 2015-05-16 11:47:00     2
#> 109   -0.55730178  1.143089e-01 2015-05-16 11:48:00     2
#> 110   -0.58619819  1.185195e-01 2015-05-16 11:49:00     2
#> 111   -0.34945222 -1.631733e-01 2015-05-16 11:50:00     2
#>  [ reached getOption("max.print") -- omitted 9969 rows ]