A Package for Multivariate Statistical Process Monitoring

The mvMonitoring package has four main functions for external use, all of which begin with the string "msp" (for "multivariate statistical process") followed by the function use. Functions without this "msp" key are primarily internal functions. They are available to see and use, but will largely be uneccesary to call in common workflows.

mvMonitoring external functions

mspProcessData - A function for synthetic process data generation. Use this data to test new process monitoring methods.

mspTrain - A function to take in observations for training under normal conditions, and to return the training summary fromt these observations.

mspMonitor - A function to take in real-time process observations and detect system anomalies based on the training summary returned by mspTrain.

mspWarning - A function to take in observations returned by mspMonitor and check for alarms by measuring sequential anomalies. This function will also be equipped to send SMS notifications to process technicians in future versions.