PCA for Data Scatter Matrix

Calculate the principal component analysis for a data matrix, and also find the squared prediction error (SPE) and Hotelling's T2 test statistic values for each observation in this data matrix.

pca(data, var.amnt = 0.9, ...)

Arguments

data
A centred-and-scaled data matrix or xts matrix

var.amnt
The energy proportion to preserve in the projection, which dictates the number of principal components to keep. Defaults to 0.90.

...
Lazy dots for additional internal arguments

Value

A list of class "pca" with the following:

projectionMatrix -- the q eigenvectors corresponding to the q largest eigenvalues as a p x q projection matrix

LambdaInv -- the diagonal matrix of inverse eigenvalues

SPE -- the vector of SPE test statistic values for each of the n observations contained in "data"

T2 -- the vector of Hotelling's T2 test statistic for each of the same n observations

Details

This function takes in a training data matrix, without the label column, and the energy preservation proportion, which defaults to 95 percent per Kazor et al (2016). This proportion is the sum of the q largest eigenvalues divided by the sum of all p eigenvalues, where q is the number of columns of the p x q projection matrix P. This function then returns the projection matrix P, a diagonal matrix of the reciprocal eigenvalues (LambdaInv), a vector of the SPE test statistic values corresponding to the rows of the data matrix, and a T2 test statistic vector similar to the SPE vector.

This internal function is called by faultFilter().

See also

Called by: faultFilter.

Examples

nrml <- mspProcessData(faults = "NOC")
scaledData <- scale(nrml[,-1])
pca(scaledData)

#> $projectionMatrix
#>            [,1]       [,2]
#> [1,]  0.6632102  0.4338278
#> [2,]  0.1397011 -0.8723023
#> [3,] -0.7352794  0.2255706
#> 
#> $LambdaInv
#>           [,1]      [,2]
#> [1,] 0.6260268 0.0000000
#> [2,] 0.0000000 0.8191588
#> 
#> $SPE
#>    [1] 8.058740e-02 2.386571e-01 1.359440e-01 1.893450e-01 2.010592e-01
#>    [6] 3.169231e-01 1.586667e-01 2.618994e-01 5.002764e-02 2.890910e-01
#>   [11] 1.435361e-01 1.283893e-01 1.524650e-01 2.650472e-01 1.376000e-01
#>   [16] 1.097393e-01 3.474211e-01 1.227866e-01 1.587132e-01 2.361125e-01
#>   [21] 1.731504e-01 1.632963e-01 1.815461e-01 2.486854e-01 2.381990e-01
#>   [26] 3.533702e-01 1.629482e-01 9.716519e-02 6.990253e-02 1.710388e-01
#>   [31] 7.368942e-02 1.362220e-01 9.949512e-02 1.477370e-01 1.990349e-01
#>   [36] 9.211882e-02 2.695324e-01 3.962922e-01 9.677256e-02 1.703915e-01
#>   [41] 2.278936e-01 2.738227e-01 1.745606e-01 1.990543e-01 2.216943e-01
#>   [46] 2.259448e-01 2.003510e-01 1.687490e-01 2.564490e-01 1.671510e-01
#>   [51] 1.076563e-01 2.699322e-01 1.407393e-01 2.146885e-01 8.236246e-02
#>   [56] 1.052706e-01 1.724403e-01 2.542953e-01 1.328055e-01 1.735074e-01
#>   [61] 1.827728e-01 2.199460e-01 3.109767e-01 3.145781e-01 2.421651e-01
#>   [66] 2.072635e-01 4.417293e-01 2.919231e-01 4.053369e-01 2.813932e-01
#>   [71] 2.885671e-01 2.730644e-01 2.875943e-01 1.761656e-01 3.599124e-01
#>   [76] 3.455185e-01 3.268498e-01 3.977538e-01 1.723062e-01 2.920296e-01
#>   [81] 3.912288e-01 4.306288e-01 3.589002e-01 2.453706e-01 4.861408e-01
#>   [86] 2.309561e-01 3.709797e-01 1.757907e-01 3.960973e-01 3.691876e-01
#>   [91] 2.881461e-01 3.696437e-01 3.266585e-01 2.486677e-01 4.305374e-01
#>   [96] 1.870221e-01 1.467897e-01 2.762506e-01 3.324079e-01 3.218916e-01
#>  [101] 2.430166e-01 4.136631e-01 3.400763e-01 2.771385e-01 2.366767e-01
#>  [106] 2.431041e-01 3.343777e-01 2.130404e-01 1.655971e-01 2.600257e-01
#>  [111] 2.202443e-01 4.362393e-01 2.136997e-01 4.027620e-01 3.653432e-01
#>  [116] 2.914831e-01 3.180582e-01 3.235231e-01 4.772730e-01 2.932367e-01
#>  [121] 1.392148e-02 1.915394e-02 2.130473e-02 6.763571e-03 1.317926e-02
#>  [126] 4.015332e-03 2.274270e-03 1.655985e-02 3.656685e-02 2.877402e-02
#>  [131] 1.683726e-02 4.748579e-03 2.455226e-02 3.878168e-02 3.378357e-02
#>  [136] 3.308202e-02 1.710784e-02 2.956514e-02 3.068307e-02 9.810229e-03
#>  [141] 2.341352e-02 1.491027e-02 2.248345e-02 1.369729e-02 2.935155e-02
#>  [146] 2.756817e-02 3.344915e-02 2.171839e-02 5.226265e-02 4.432149e-02
#>  [151] 6.749229e-02 2.864274e-02 2.546169e-02 1.966734e-02 1.987066e-02
#>  [156] 1.643263e-02 2.969137e-02 2.398671e-02 1.173127e-02 4.626238e-02
#>  [161] 1.816484e-02 2.441140e-02 2.863838e-02 4.522316e-02 3.217744e-02
#>  [166] 3.233887e-02 8.206526e-03 2.878820e-02 1.977635e-03 1.678738e-02
#>  [171] 9.298960e-03 1.172184e-02 1.167835e-03 4.725741e-03 1.857525e-02
#>  [176] 2.005761e-02 1.134736e-02 1.897714e-02 2.400650e-02 1.473036e-02
#>  [181] 1.544607e-01 1.587492e-01 2.837638e-02 1.865389e-01 1.528715e-01
#>  [186] 4.568043e-02 3.648517e-02 1.390988e-01 7.139966e-02 9.000520e-02
#>  [191] 1.577658e-01 1.849272e-01 2.780299e-01 3.166126e-01 1.568824e-01
#>  [196] 1.663331e-01 1.044048e-01 1.773821e-01 1.607714e-01 2.464903e-01
#>  [201] 1.238852e-01 2.178085e-01 2.511207e-01 1.938563e-01 2.229297e-01
#>  [206] 1.775776e-01 2.080533e-01 1.720782e-01 1.576092e-01 2.256621e-01
#>  [211] 2.099108e-01 2.031767e-01 2.732415e-01 2.676160e-01 2.727782e-01
#>  [216] 2.052269e-01 2.586715e-01 1.333632e-01 1.555346e-01 2.503565e-01
#>  [221] 1.682162e-01 1.430889e-01 1.231896e-01 1.608225e-01 1.835248e-01
#>  [226] 9.569370e-02 1.861912e-01 9.520000e-02 1.434317e-01 1.993712e-01
#>  [231] 9.613421e-02 8.421226e-02 5.290227e-02 2.596203e-01 1.284800e-01
#>  [236] 1.084262e-01 1.570649e-01 1.963458e-01 2.730689e-01 3.253439e-01
#>  [241] 3.110070e-01 4.281416e-01 3.676705e-01 2.881547e-01 4.996899e-01
#>  [246] 2.171516e-01 4.731696e-01 3.709503e-01 3.310215e-01 3.661951e-01
#>  [251] 2.928757e-01 5.185623e-01 1.411266e-01 3.997183e-01 2.797489e-01
#>  [256] 2.143059e-01 2.203003e-01 2.569164e-01 3.446817e-01 3.252473e-01
#>  [261] 1.746975e-01 2.388809e-01 1.911894e-01 1.921334e-01 7.642009e-02
#>  [266] 1.623285e-02 2.022875e-01 2.893249e-01 6.706771e-02 2.474242e-01
#>  [271] 8.855400e-02 2.136202e-01 3.281126e-01 2.599926e-01 3.736801e-01
#>  [276] 2.578337e-01 2.467698e-01 2.860456e-01 2.569600e-01 3.147975e-01
#>  [281] 2.680434e-01 1.324585e-01 2.060637e-01 1.951548e-01 2.729647e-01
#>  [286] 2.548315e-01 4.815891e-01 2.991071e-01 5.433296e-01 4.034629e-01
#>  [291] 3.501460e-01 5.169517e-01 1.684221e-01 4.527371e-01 2.435742e-01
#>  [296] 3.516506e-01 1.896565e-01 4.173816e-01 1.911246e-01 2.283462e-01
#>  [301] 8.946825e-03 1.430383e-02 1.022011e-02 1.320094e-02 2.436856e-02
#>  [306] 5.314485e-02 1.692386e-02 3.574607e-03 3.739294e-03 1.911128e-02
#>  [311] 5.405407e-03 6.067863e-03 3.812893e-02 1.554549e-02 2.651006e-02
#>  [316] 3.237827e-02 3.377329e-02 3.907025e-02 1.424808e-02 2.205987e-02
#>  [321] 6.635510e-03 3.578459e-03 9.573373e-03 1.730436e-02 4.458200e-03
#>  [326] 4.406987e-03 9.953691e-03 8.068918e-03 1.520160e-02 2.367868e-02
#>  [331] 1.761446e-02 5.860215e-03 8.368777e-03 3.915648e-03 1.655276e-02
#>  [336] 5.643427e-03 2.073695e-03 2.889872e-03 1.170104e-02 3.096433e-03
#>  [341] 2.056184e-02 5.629314e-03 1.105205e-02 1.530754e-02 1.553898e-02
#>  [346] 1.979149e-02 3.509038e-03 8.053086e-05 1.737310e-02 7.616624e-03
#>  [351] 2.205126e-02 2.250370e-02 5.253787e-02 1.984666e-02 4.454732e-02
#>  [356] 3.327093e-02 2.319853e-02 7.310420e-03 2.134462e-02 4.940723e-02
#>  [361] 1.206959e-01 2.197814e-01 1.406313e-01 2.513633e-01 1.775013e-01
#>  [366] 2.559558e-01 2.503342e-01 2.751294e-01 1.877564e-01 1.807208e-01
#>  [371] 1.068321e-01 4.756421e-02 1.088244e-01 1.174375e-01 2.589036e-01
#>  [376] 1.221735e-01 1.586892e-01 2.146028e-01 2.663389e-01 1.697590e-01
#>  [381] 2.588632e-01 2.111887e-01 1.378853e-01 1.100625e-01 1.577461e-01
#>  [386] 2.683183e-01 8.564106e-02 1.295931e-01 2.331833e-01 1.096065e-01
#>  [391] 1.623955e-01 1.488949e-01 2.295171e-01 1.939574e-01 2.019239e-01
#>  [396] 2.789399e-01 2.042434e-01 3.178046e-01 3.239006e-01 1.261775e-01
#>  [401] 1.774176e-01 8.151867e-02 2.862502e-01 3.057155e-01 2.249454e-01
#>  [406] 3.824882e-01 2.515853e-01 2.097057e-01 1.640125e-01 1.457352e-01
#>  [411] 1.676226e-01 9.414207e-02 2.184216e-01 3.252284e-01 3.207009e-01
#>  [416] 2.433933e-01 2.608315e-01 1.634002e-01 2.880304e-01 1.169783e-01
#>  [421] 4.075338e-01 2.528716e-01 1.674983e-01 1.622819e-01 2.716973e-01
#>  [426] 2.899504e-01 1.847510e-01 2.148748e-01 2.838751e-01 2.267353e-01
#>  [431] 3.158245e-01 3.277259e-01 2.863130e-01 2.953629e-01 3.325884e-01
#>  [436] 2.435305e-01 6.451453e-01 3.489359e-01 4.930926e-01 2.687108e-01
#>  [441] 3.066216e-01 4.377630e-01 4.085161e-01 3.132610e-01 4.494374e-01
#>  [446] 2.715598e-01 3.937068e-01 4.016767e-01 2.067154e-01 3.832042e-01
#>  [451] 3.630694e-01 4.308279e-01 4.232495e-01 5.071036e-01 4.929628e-01
#>  [456] 1.582264e-01 2.232681e-01 2.495269e-01 2.337487e-01 1.718521e-01
#>  [461] 1.604074e-01 1.495452e-01 3.210636e-01 2.071696e-01 3.325936e-01
#>  [466] 2.510322e-01 2.736983e-01 3.741772e-01 2.451244e-01 4.609558e-01
#>  [471] 2.443559e-01 4.278253e-01 2.152746e-01 4.980712e-01 2.608834e-01
#>  [476] 2.243075e-01 7.614671e-01 3.839102e-01 2.108254e-01 2.869786e-01
#>  [481] 8.200525e-03 6.809055e-03 1.247646e-02 9.761923e-03 3.972420e-02
#>  [486] 2.506842e-02 2.037394e-02 1.857863e-02 1.037437e-02 7.218709e-03
#>  [491] 8.541222e-03 7.300019e-03 1.239008e-02 2.552670e-03 6.713304e-03
#>  [496] 3.026051e-03 1.387076e-02 2.378790e-02 2.089099e-02 1.785288e-02
#>  [501] 1.313799e-02 2.318354e-02 1.711518e-02 1.295360e-02 1.999845e-02
#>  [506] 5.478437e-02 1.792223e-02 2.656249e-02 2.900409e-02 1.673691e-02
#>  [511] 3.500260e-03 8.271804e-03 2.804200e-02 1.928735e-02 4.281813e-03
#>  [516] 8.084000e-03 2.814426e-03 1.112108e-04 8.741728e-04 6.072971e-03
#>  [521] 6.511160e-04 5.496287e-04 1.332819e-03 1.081138e-06 2.086136e-03
#>  [526] 2.071195e-03 1.827208e-02 1.328236e-02 3.448935e-03 1.286917e-02
#>  [531] 1.394085e-02 1.571200e-02 9.563451e-04 1.445006e-02 1.750387e-03
#>  [536] 1.916720e-04 3.700217e-03 1.353487e-02 1.083259e-02 1.206789e-02
#>  [541] 1.031873e-01 1.289807e-01 1.973328e-01 1.524107e-01 4.197690e-01
#>  [546] 2.028074e-01 2.312763e-01 1.542968e-01 1.804895e-01 1.846792e-01
#>  [551] 2.650060e-01 7.803390e-02 3.118412e-01 3.319709e-01 2.135918e-01
#>  [556] 1.862529e-01 1.096009e-01 2.714799e-01 2.384948e-01 1.080292e-01
#>  [561] 1.653596e-01 1.686147e-01 3.101788e-01 4.123672e-01 2.164835e-01
#>  [566] 2.737148e-01 1.902108e-01 2.448331e-01 2.674293e-01 1.591610e-01
#>  [571] 1.376923e-01 2.928810e-01 3.267491e-01 1.085330e-01 2.259336e-01
#>  [576] 3.269225e-01 1.864082e-01 1.449977e-01 3.023234e-01 9.771762e-02
#>  [581] 1.723585e-01 8.974424e-02 1.062638e-01 2.153176e-01 1.800461e-01
#>  [586] 2.505893e-01 2.278193e-01 1.027867e-01 2.085840e-01 1.188968e-01
#>  [591] 1.697028e-01 3.623663e-01 2.160098e-01 2.984121e-01 2.116112e-01
#>  [596] 2.376299e-01 3.310266e-01 9.107002e-02 3.492537e-01 3.415705e-01
#>  [601] 1.580051e-01 2.379113e-01 1.816476e-01 2.882130e-01 1.743916e-01
#>  [606] 2.551177e-01 1.460061e-01 6.963251e-02 3.816276e-01 3.353346e-01
#>  [611] 3.504931e-01 3.494645e-01 4.373783e-01 2.390877e-01 3.176484e-01
#>  [616] 2.508328e-01 4.455940e-01 2.701835e-01 3.619357e-01 3.293240e-01
#>  [621] 3.718607e-01 3.634302e-01 2.916659e-01 3.209086e-01 2.486971e-01
#>  [626] 3.490547e-01 2.862621e-01 2.611257e-01 2.403235e-01 3.058171e-01
#>  [631] 4.633354e-01 3.896625e-01 5.398944e-01 4.489262e-01 3.399087e-01
#>  [636] 3.012902e-01 5.979193e-01 2.487754e-01 3.408045e-01 3.280759e-01
#>  [641] 3.214676e-01 2.986592e-01 2.854221e-01 4.563105e-01 3.501666e-01
#>  [646] 3.148993e-01 2.070087e-01 2.772909e-01 3.590737e-01 3.179763e-01
#>  [651] 3.478596e-01 6.016996e-01 5.725432e-01 4.520231e-01 4.297302e-01
#>  [656] 2.930632e-01 1.525190e-01 3.909957e-01 4.042014e-01 4.609223e-01
#>  [661] 9.657176e-03 2.187761e-02 1.461643e-02 1.093505e-02 1.220460e-02
#>  [666] 2.919164e-03 1.747579e-04 9.655299e-04 6.843882e-03 4.131489e-03
#>  [671] 7.703612e-03 3.168969e-03 4.375641e-05 6.539110e-03 1.152269e-02
#>  [676] 9.913198e-03 2.770587e-03 1.575901e-02 1.106117e-02 1.791480e-02
#>  [681] 1.359078e-02 2.466687e-02 1.063843e-02 1.197440e-02 5.267044e-03
#>  [686] 1.662819e-02 1.329401e-02 1.156794e-02 1.245974e-02 2.838612e-02
#>  [691] 3.065103e-02 9.752653e-03 1.392580e-02 1.049402e-02 1.847322e-03
#>  [696] 5.920993e-03 1.151486e-02 8.083661e-03 1.395966e-03 4.436217e-03
#>  [701] 4.406014e-03 6.928953e-03 2.616835e-02 1.757357e-02 2.564583e-02
#>  [706] 4.459283e-02 1.805001e-02 2.822075e-02 3.246753e-03 1.711005e-03
#>  [711] 8.536509e-04 5.713548e-03 2.389709e-03 1.321376e-03 1.371602e-02
#>  [716] 2.263985e-02 3.139027e-02 2.120639e-02 4.639128e-03 1.674245e-02
#>  [721] 2.322756e-01 1.336668e-01 3.638505e-01 1.116159e-01 1.533997e-01
#>  [726] 1.407500e-01 2.208376e-01 1.422420e-01 3.002904e-01 2.373344e-01
#>  [731] 2.717164e-01 1.279207e-01 9.614855e-02 1.784291e-01 1.338716e-01
#>  [736] 2.184685e-01 2.451722e-01 1.375613e-01 2.762789e-01 2.667903e-01
#>  [741] 1.167459e-01 2.339502e-01 2.979414e-01 2.390518e-01 1.790762e-01
#>  [746] 2.734367e-01 1.153356e-01 1.687390e-01 1.982229e-01 9.615042e-02
#>  [751] 2.732186e-01 1.278358e-01 3.756920e-01 2.284123e-01 2.712590e-01
#>  [756] 2.212422e-01 1.690419e-01 1.967723e-01 2.008407e-01 1.739574e-01
#>  [761] 2.027894e-01 2.117607e-01 1.312764e-01 2.373253e-01 2.341326e-01
#>  [766] 2.472837e-01 2.956944e-01 2.331867e-01 1.972978e-01 1.855414e-01
#>  [771] 2.185085e-01 1.872869e-01 1.311984e-01 1.085890e-01 3.253597e-01
#>  [776] 1.312999e-01 2.444715e-01 6.812197e-02 3.540723e-01 1.120597e-01
#>  [781] 2.500591e-01 3.505884e-01 3.608817e-01 2.624236e-01 3.542910e-01
#>  [786] 2.129286e-01 2.576680e-01 1.812464e-01 4.265403e-01 3.095149e-01
#>  [791] 5.046286e-01 4.413030e-01 3.632236e-01 2.926048e-01 1.288487e-01
#>  [796] 1.063162e-01 5.774732e-01 3.354372e-01 3.168576e-01 4.999958e-01
#>  [801] 2.936351e-01 3.202430e-01 3.271353e-01 3.154458e-01 3.195420e-01
#>  [806] 2.900652e-01 3.260396e-01 2.198482e-01 3.964736e-01 3.397478e-01
#>  [811] 2.451727e-01 2.801843e-01 3.121109e-01 3.507698e-01 3.282742e-01
#>  [816] 3.421706e-01 1.848691e-01 4.338117e-01 3.100592e-01 1.819665e-01
#>  [821] 2.423892e-01 3.974317e-01 2.407576e-01 3.118587e-01 3.588347e-01
#>  [826] 3.585032e-01 2.560955e-01 2.686909e-01 2.341087e-01 4.661337e-01
#>  [831] 3.432508e-01 4.918855e-01 5.336876e-01 3.171818e-01 4.086522e-01
#>  [836] 5.925306e-01 2.939058e-01 3.728148e-01 3.589613e-01 4.846657e-01
#>  [841] 7.109317e-03 1.036225e-02 4.494584e-04 1.539256e-03 3.583463e-03
#>  [846] 9.188540e-05 1.093419e-03 9.576286e-04 1.964853e-03 2.153222e-03
#>  [851] 5.699190e-03 1.002391e-02 1.372560e-02 1.352359e-02 1.488416e-02
#>  [856] 2.522194e-02 1.060507e-02 7.798360e-03 5.413642e-04 1.574665e-03
#>  [861] 6.831788e-03 7.325884e-03 9.498259e-03 1.104101e-02 2.181686e-03
#>  [866] 9.645162e-03 1.556549e-02 8.899724e-04 7.121726e-03 7.441570e-04
#>  [871] 3.403105e-03 5.437058e-03 4.852193e-03 1.059700e-02 1.536466e-03
#>  [876] 3.410271e-03 2.452005e-02 1.182450e-02 1.012880e-04 1.936374e-03
#>  [881] 1.305081e-04 2.654382e-04 1.190601e-03 8.139646e-04 3.900280e-03
#>  [886] 1.738356e-02 1.848483e-02 1.326054e-02 4.597335e-03 1.991671e-02
#>  [891] 5.889433e-03 5.861710e-03 1.618190e-02 1.097489e-02 1.255512e-02
#>  [896] 2.051747e-03 1.409889e-02 2.476555e-02 1.189578e-02 9.568496e-03
#>  [901] 1.781875e-01 1.894899e-01 3.333928e-01 1.794645e-01 9.399536e-02
#>  [906] 2.668949e-01 3.585617e-01 1.914459e-01 2.181308e-01 2.045307e-01
#>  [911] 1.670534e-01 3.264982e-01 1.319482e-01 7.318083e-02 1.858604e-01
#>  [916] 1.245147e-01 1.772986e-01 2.381586e-01 3.632273e-01 1.990012e-01
#>  [921] 2.206752e-01 1.503358e-01 2.131715e-01 1.167689e-01 2.367469e-01
#>  [926] 2.243793e-01 2.122933e-01 2.859172e-01 3.216352e-01 1.605505e-01
#>  [931] 1.991960e-01 2.139264e-01 1.526011e-01 1.858389e-01 2.478449e-01
#>  [936] 2.104832e-01 1.248245e-01 3.412727e-01 2.032847e-01 2.791533e-01
#>  [941] 3.044255e-01 1.594183e-01 1.673521e-01 2.244262e-01 2.654485e-01
#>  [946] 2.599521e-01 2.955715e-01 2.374759e-01 1.315915e-01 3.289005e-01
#>  [951] 2.924980e-01 1.325473e-01 1.669685e-01 2.284665e-01 2.523986e-01
#>  [956] 2.090638e-01 1.904414e-01 2.117810e-01 1.687452e-01 2.262615e-01
#>  [961] 3.828768e-01 3.434565e-01 3.223854e-01 3.739081e-01 4.955369e-01
#>  [966] 3.382379e-01 4.320551e-01 2.596358e-01 3.457406e-01 4.562364e-01
#>  [971] 4.126577e-01 2.535710e-01 4.100633e-01 2.849888e-01 3.132407e-01
#>  [976] 4.036654e-01 1.401236e-01 3.235910e-01 3.132667e-01 4.431510e-01
#>  [981] 4.216137e-01 3.934417e-01 3.491767e-01 3.285078e-01 2.264699e-01
#>  [986] 3.848539e-01 2.930830e-01 2.395576e-01 3.292876e-01 4.488546e-01
#>  [991] 5.104107e-01 5.809710e-01 4.049679e-01 3.345620e-01 5.388607e-01
#>  [996] 2.721610e-01 4.304407e-01 2.164135e-01 2.969917e-01 3.584504e-01
#>  [ reached getOption("max.print") -- omitted 9080 entries ]
#> 
#> $T2
#>    [1] 0.2960406 0.5970471 0.2388344 0.3465329 0.2494634 0.3290543 0.2917948
#>    [8] 0.3330018 0.2722757 0.3078988 0.2463940 0.4012619 0.3337129 0.2880146
#>   [15] 0.3343376 0.4833398 0.3052189 0.4044452 0.3568397 0.5439571 0.3029531
#>   [22] 0.2789458 0.2036569 0.4114136 0.2893592 0.3386819 0.6494153 0.3463192
#>   [29] 0.3634043 0.6801174 1.1980064 0.8602388 1.1389372 0.7359873 0.4735117
#>   [36] 0.5424716 0.4183334 0.2957308 0.2766667 0.5342983 0.3916700 0.2954672
#>   [43] 0.2655374 0.2263055 0.4790266 0.3307536 0.2287844 0.3839718 0.5454168
#>   [50] 0.4683738 0.5141985 0.3494325 0.3044785 0.3558185 1.0844466 0.6542687
#>   [57] 0.3416593 0.7918932 0.8213614 0.6370602 0.7739958 0.7568661 1.0033241
#>   [64] 0.9167508 0.6484970 0.6661650 0.7875292 0.4902764 0.7881352 0.7669727
#>   [71] 0.5895974 0.6158209 0.4559403 0.7444574 0.6501934 0.6487543 0.7373629
#>   [78] 0.6235701 0.5543150 0.7125612 0.4444190 0.4650758 0.7305788 0.6855266
#>   [85] 0.6129492 0.4035046 0.6437031 0.4355794 0.9534996 0.9671315 0.5694799
#>   [92] 0.6180567 0.6583429 0.5140517 0.6026134 0.7532424 0.6117682 1.0433003
#>   [99] 0.7017126 0.7795578 0.6418206 0.8096442 0.9526149 0.6351465 0.7196410
#>  [106] 0.7776969 0.7653871 0.7415975 0.6860839 0.6428684 0.6673736 0.8992075
#>  [113] 0.7633470 0.3989805 0.6299310 0.6602272 0.6804816 0.5922163 0.6368522
#>  [120] 0.6729310 1.1141961 1.1485063 1.1355918 1.0632859 1.1797550 1.0737210
#>  [127] 1.1151017 1.1939009 1.2036362 1.1821192 1.1557341 1.0899507 1.1741985
#>  [134] 1.2508475 1.2317018 1.1687177 1.1591311 1.1834365 1.1598589 1.1168741
#>  [141] 1.1684522 1.1530851 1.1910387 1.1744215 1.1099691 1.1777699 1.2011069
#>  [148] 1.1110019 1.2162950 1.2552541 1.3371113 1.1707983 1.1909360 1.1593605
#>  [155] 1.1805712 1.1396706 1.1539259 1.1509660 1.1461730 1.1740438 1.1599294
#>  [162] 1.1681892 1.1670127 1.1557772 1.1110509 1.1893626 1.1350539 1.1545162
#>  [169] 1.0535189 1.0529315 1.1235889 1.0947745 1.0906814 1.1043805 1.1368415
#>  [176] 1.2054533 1.1180446 1.1617204 1.1432727 1.1040422 0.3647245 0.5286895
#>  [183] 0.8318660 0.3432366 0.6021821 0.9188669 0.6770091 0.6137567 0.7468560
#>  [190] 0.6186274 0.5890791 0.4777597 0.5121833 0.5088280 0.1885854 0.3398893
#>  [197] 0.5140060 0.2049756 0.3635096 0.4377594 0.3723669 0.4169850 0.3535843
#>  [204] 0.5783048 0.3208665 0.6190192 0.3273503 0.5486185 0.4433021 0.3556921
#>  [211] 0.3339420 0.3561771 0.3360460 0.7083063 0.3477392 0.2071189 0.3198630
#>  [218] 0.2581355 0.2410180 0.3180832 0.2433081 0.5495889 0.6363317 0.9072289
#>  [225] 0.4880857 0.6520819 0.8450130 0.4317474 0.7044692 0.6125264 0.2296716
#>  [232] 0.3886446 0.4833476 0.2501097 0.4187948 0.2295143 0.3968379 0.2482747
#>  [239] 0.4377605 0.3657592 0.7570694 0.6271416 0.4407126 0.6024508 0.6756501
#>  [246] 0.4959784 0.3817042 0.5193236 0.4513244 0.6579174 0.6623771 0.6128523
#>  [253] 0.4445466 0.6205170 0.5760190 0.4498245 0.7625415 1.1335907 1.1517217
#>  [260] 0.7733234 1.1243255 0.9490061 1.2647028 0.9217917 1.1815851 1.0294023
#>  [267] 1.2034914 1.0480818 1.1922568 1.3860933 0.7976687 0.7532007 0.8750414
#>  [274] 0.6918551 0.9157694 0.6862103 0.7083374 0.7401461 0.7646681 0.8096812
#>  [281] 1.0671954 1.0128190 0.8980222 0.6594243 0.6343073 0.5482137 0.6982742
#>  [288] 0.6712089 0.7921030 0.7084729 0.4659960 0.7359383 0.6454592 0.7656625
#>  [295] 0.6022457 0.6778406 0.5674947 0.6330220 0.6500963 0.5492983 1.1001552
#>  [302] 1.1249562 1.1898638 1.1463516 1.1676593 1.2005030 1.1388773 1.1468879
#>  [309] 1.0559415 1.1735210 1.0761908 1.0983691 1.1965560 1.1368358 1.1729798
#>  [316] 1.1891422 1.1676737 1.2301174 1.1029316 1.1529785 1.1128190 1.0919835
#>  [323] 1.1144537 1.1786108 1.1253873 1.0852550 1.1320483 1.0848441 1.1368236
#>  [330] 1.1363048 1.1811741 1.0903131 1.1320420 1.0524950 1.1309007 1.1400454
#>  [337] 1.0957525 1.0842696 1.1722233 1.0516317 1.1821436 1.0603091 1.1659383
#>  [344] 1.1321237 1.1402002 1.1757829 1.0863069 1.0474975 1.1360756 1.1097575
#>  [351] 1.1306911 1.1411015 1.2350328 1.1679910 1.2293725 1.2055291 1.1497190
#>  [358] 1.1424810 1.1515862 1.2780881 0.7263857 0.5049715 0.3598526 0.5056048
#>  [365] 0.4518285 0.5095205 0.4417027 0.5312355 0.2891076 0.2083237 0.2849108
#>  [372] 0.5198763 1.0522217 0.4463319 0.3098138 0.2124358 0.2070057 0.3621795
#>  [379] 0.3327584 0.2754951 0.3238303 0.3201916 0.3363380 0.4111586 0.3709085
#>  [386] 0.3022557 0.4663367 0.2339688 0.2871898 0.2250750 0.3202793 0.4874689
#>  [393] 0.2981995 0.2864257 0.3677315 0.2570844 0.3728765 0.3505167 0.3252731
#>  [400] 0.3412028 0.2760323 0.2079806 0.3788374 0.3840608 0.4923822 0.3861276
#>  [407] 0.3907334 0.2973375 0.3369552 0.4155455 0.4421750 0.2239159 0.2793168
#>  [414] 0.3476268 0.2579685 0.2565862 0.3182677 0.2532376 0.3991485 0.2329631
#>  [421] 0.9258895 0.7226914 0.8848808 0.5634589 0.7832626 0.8289730 0.9834852
#>  [428] 0.4719820 0.6485526 0.5949397 0.6803085 0.6896396 0.5699390 0.5662443
#>  [435] 0.5422050 0.5246938 0.6247353 0.5575160 0.4853677 0.6721802 0.6840016
#>  [442] 0.6714900 0.4690832 0.6140002 0.6199512 0.4807823 0.3772430 0.3759158
#>  [449] 0.4195949 0.5075206 0.5104203 0.7189318 0.5913782 0.6852460 0.6718704
#>  [456] 0.7726884 0.6776323 1.0119016 0.6776520 0.9617192 0.8802197 1.0967307
#>  [463] 0.7681118 0.4003666 0.4356907 0.5106761 0.6627450 0.5046660 0.3800322
#>  [470] 0.7741574 1.1475798 0.6857359 0.6886925 0.9387884 0.5453871 0.5136630
#>  [477] 0.8192197 0.5204676 0.7082316 0.8002276 1.1385391 1.0800361 1.1118264
#>  [484] 1.1274255 1.1804982 1.1606311 1.1368197 1.1241283 1.1135141 1.1499933
#>  [491] 1.1018444 1.1390717 1.1676625 1.0501828 1.1372930 1.1332626 1.1497078
#>  [498] 1.1331214 1.1860194 1.1307000 1.1171172 1.1185041 1.1835998 1.1501572
#>  [505] 1.1216100 1.1581073 1.1950350 1.2037299 1.1390493 1.1637781 1.0901449
#>  [512] 1.1238436 1.1145865 1.1670226 1.0872825 1.1012872 1.1140990 1.0846989
#>  [519] 1.0998377 1.1024281 1.1071290 1.0254176 1.0912154 1.0816727 1.0968167
#>  [526] 1.1245011 1.1019314 1.1322171 1.0989234 1.1441410 1.1334476 1.0790253
#>  [533] 1.0373894 1.1085285 1.0872388 1.0554977 1.1439506 1.1330269 1.0984659
#>  [540] 1.1369218 0.3220552 0.2560538 0.2663978 0.2337944 0.4035993 0.3766913
#>  [547] 0.2793038 0.2878852 0.5307770 0.2657092 0.2804405 0.2984328 0.3107466
#>  [554] 0.3560168 0.2580608 0.3055599 0.2981261 0.3384586 0.3242120 0.1768264
#>  [561] 0.3046941 0.2416141 0.3396639 0.3126531 0.2747987 0.3282853 0.3791498
#>  [568] 0.4503394 0.4364266 0.3019463 0.2432527 0.3149953 0.3158668 0.2688093
#>  [575] 0.3687940 0.2972745 0.3706121 0.2424540 0.2467122 0.2508709 0.3708490
#>  [582] 0.5098377 0.3463234 0.5121643 0.3725784 0.2970639 0.2734616 0.3302320
#>  [589] 0.5654888 0.6021437 0.3381426 0.4054841 0.4189343 0.3432761 0.3087719
#>  [596] 0.3454629 0.3476805 0.1625264 0.4641685 0.3327586 0.8322228 0.7187105
#>  [603] 0.7582209 0.8397148 0.6196428 0.8646155 0.7353718 1.0622228 0.7746223
#>  [610] 0.8291508 0.4755883 0.4792793 0.4833488 0.5536069 0.5211131 0.3858791
#>  [617] 0.5388922 0.4560265 0.4296814 0.4266707 0.3456953 0.4118434 0.6096897
#>  [624] 0.4548036 0.3640247 0.5461642 0.5956947 0.8272413 0.5492858 0.5129045
#>  [631] 0.4046836 0.3495624 0.2797099 0.4465286 0.6214684 0.7779797 0.6114111
#>  [638] 0.6728566 0.5625240 0.7293311 0.8678737 0.6083620 0.6370120 0.6276646
#>  [645] 0.7373275 0.6217791 0.8028639 0.8562869 0.4458242 0.5314464 0.3628713
#>  [652] 0.5821160 0.4031808 0.5908348 0.8451709 0.7259313 0.9403648 0.7266441
#>  [659] 0.5275538 0.6315653 1.0394857 1.1450918 1.0747506 1.1352104 1.1514935
#>  [666] 1.0946486 1.0213341 1.0964509 1.1519404 1.1439186 1.1258915 1.0844289
#>  [673] 1.1062741 1.1308895 1.1255190 1.1264979 1.1024894 1.0593814 1.1079141
#>  [680] 1.1185335 1.1024420 1.1639763 1.1070523 1.1416188 1.0770374 1.1788672
#>  [687] 1.1550353 1.1500204 1.1221100 1.2027467 1.1927768 1.1388169 1.1669372
#>  [694] 1.1173579 1.0836045 1.1030638 1.1436163 1.0985623 1.1215244 1.0967236
#>  [701] 1.1315302 1.1629671 1.1551770 1.1458261 1.1569455 1.2001916 1.1401589
#>  [708] 1.0739474 1.1160992 1.1173015 1.0621591 1.1230521 1.0950420 1.1045922
#>  [715] 1.1145644 1.2028228 1.2563381 1.1759424 1.0878043 1.1366903 0.3234085
#>  [722] 0.3358001 0.3269532 0.2852458 0.2635017 0.2286512 0.3760629 0.2305812
#>  [729] 0.2790112 0.3749557 0.2905110 0.1696306 0.3089339 0.6644081 0.4545884
#>  [736] 0.2834793 0.3031889 0.2595513 0.3487419 0.3932908 0.2409530 0.2253453
#>  [743] 0.3916974 0.3467616 0.3313710 0.4541476 0.2019083 0.2870339 0.2112386
#>  [750] 0.1946070 0.3147016 0.1689044 0.3048700 0.3210590 0.3406155 0.2890310
#>  [757] 0.2941168 0.3167095 0.2799984 0.2279355 0.3217180 0.2069568 0.3519338
#>  [764] 0.4126158 0.2851340 0.2310897 0.2696803 0.4289389 0.2727454 0.2955679
#>  [771] 0.3841047 0.2803750 0.3219852 0.2059443 0.3178887 0.2744618 0.2951609
#>  [778] 0.2392392 0.2921580 0.2947529 0.5386986 0.4732256 0.4053287 0.4325545
#>  [785] 0.4680762 0.5212782 0.5135027 0.6457067 0.7865099 0.7119284 0.6437381
#>  [792] 0.6169517 0.5672170 0.5750522 0.7017052 0.3731666 0.5419926 0.7994908
#>  [799] 0.6274490 0.8989296 0.7544518 0.9469125 0.6351656 0.3714993 0.6636923
#>  [806] 0.4902728 0.5890303 0.3919018 0.7200868 0.5540373 0.3928587 0.4094591
#>  [813] 0.3642848 0.4737403 0.5539069 0.6701255 0.4504047 0.5038681 0.4031589
#>  [820] 0.6950768 0.6434006 0.5419871 0.4532079 0.7545979 0.5717440 0.9497774
#>  [827] 0.7356668 0.9015197 0.5994780 0.5690008 0.5654709 0.4474675 0.5756155
#>  [834] 0.4096750 0.4177300 0.6129599 0.4640699 0.3751805 0.6336503 0.7328814
#>  [841] 1.0942324 1.1253875 1.0153452 1.0619555 1.0574834 1.0516871 1.0453846
#>  [848] 1.1422059 1.1136362 1.0765801 1.1055626 1.1436004 1.1420239 1.1282930
#>  [855] 1.1232789 1.1819920 1.1572182 1.1304537 1.0463667 1.0750167 1.1085543
#>  [862] 1.0795349 1.0883950 1.1231261 1.0132320 1.1357805 1.1410218 1.0898133
#>  [869] 1.1289582 1.0545519 1.1147953 1.0728328 1.1000303 1.0860808 1.1212003
#>  [876] 1.1192254 1.1698737 1.1153943 1.0915923 1.0469290 1.0254758 1.0101532
#>  [883] 1.0986979 1.0817891 1.0919672 1.1318803 1.1859721 1.1218352 1.1528469
#>  [890] 1.1521025 1.0487306 1.0774746 1.1659346 1.1008096 1.1960132 1.0925292
#>  [897] 1.1333657 1.1202534 1.1332484 1.1370084 0.2098248 0.2272124 0.3260441
#>  [904] 0.3253270 0.4555961 0.4784141 0.3099212 0.2616131 0.2251915 0.3434222
#>  [911] 0.2367337 0.3055668 0.3395170 0.2787719 0.2654575 0.2854387 0.2461109
#>  [918] 0.3214096 0.4773671 0.2534124 0.3221271 0.2707483 0.3264541 0.2747533
#>  [925] 0.3239419 0.3548048 0.2646010 0.2757355 0.3321274 0.2218735 0.2674210
#>  [932] 0.2555880 0.2314516 0.2791801 0.2801112 0.2171901 0.2415738 0.3520982
#>  [939] 0.2760048 0.3056633 0.4192470 0.2599583 0.3798357 0.3144619 0.3231364
#>  [946] 0.3487808 0.3119405 0.3338016 0.1691423 0.4138518 0.4165268 0.2393331
#>  [953] 0.3840254 0.3444130 0.3407853 0.3572938 0.3290988 0.2776367 0.3669346
#>  [960] 0.3333108 0.7426303 0.7783474 0.4523393 0.5280664 0.6613216 0.4169237
#>  [967] 0.3960546 0.6010614 0.4744402 0.4490505 0.4197837 0.4099990 0.7314379
#>  [974] 0.3779508 0.4880833 0.7026878 0.4416106 0.3366334 0.5554716 0.4077736
#>  [981] 0.5350902 0.5657788 0.6956007 0.5849944 0.4758488 0.5508807 0.3617635
#>  [988] 0.4649606 0.4097014 0.3575483 0.4713020 0.4468867 0.4292091 0.3833772
#>  [995] 0.3618892 0.3569443 0.3872962 0.5131522 0.8531446 0.6668093
#>  [ reached getOption("max.print") -- omitted 9080 entries ]
#> 
#> attr(,"class")
#> [1] "pca"