Adaptive Process Training

Apply Adaptive-Dynamic PCA to state-specific data matrices.

processMonitor(data, trainObs, updateFreq = ceiling(0.5 * trainObs),
  faultsToTriggerAlarm = 5, ...)

Arguments

data
An xts data matrix

trainObs
The number of training observations to be used

updateFreq
The number of non-flagged observations to collect before the function updates. Defaults to half as many observations as the number of training observations.

faultsToTriggerAlarm
The number of sequential faults needed to trigger an alarm. Defaults to 5.

...
Lazy dots for additional internal arguments

Value

A list with the following components:

FaultChecks -- a class-specific xts flagging matrix with the same number of rows as "data". This flag matrix has the following five columns:

SPE -- the SPE statistic value for each observation in "data"

SPE_Flag -- a vector of SPE indicators recording 0 if the test statistic is less than or equal to the critical value passed through from the threshold object

T2 -- the T2 statistic value for each observation in "data"

T2_Flag -- a vector of T2 fault indicators, defined like SPE_Flag

Alarm -- a column indicating if there have been three flags in a row for either the SPE or T2 monitoring statistics or both. Alarm states are as follows: 0 = no alarm, 1 = Hotelling's T2 alarm, 2 = Squared Prediction Error alarm, and 3 = both alarms.

Non_Alarmed_Obs -- a class-specific xts data matrix of all the non-alarmed observations (observations with alarm state equal to 0)

Alarms -- a class-specific xts data matrix of the features and specific alarms of Alarmed observations, where the alarm codes are listed above

trainSpecs -- a threshold object returned by the internal threshold() function. See the threshold() function's help file for more details.

Details

This function is the class-specific implementation of the Adaptive- Dynamic PCA described in the details of the mspTrain() function. See the mspTrain() function's help file for further details.

This internal function is called by mspTrain(). This function calls the faultFilter() function.

See also

Calls: faultFilter. Called by: mspTrain.

Examples

nrml <- mspProcessData(faults = "NOC")
data <- nrml[nrml[,1] == 1]

processMonitor(data = data[,-1], trainObs = 672)

#> $FaultChecks
#>                              SPE SPE_Flag           T2 T2_Flag Alarm
#> 2015-05-17 19:12:00 1.373589e-01        0 2.366453e+00       0     0
#> 2015-05-17 19:13:00 6.598896e-01        0 2.252538e+00       0     0
#> 2015-05-17 19:14:00 4.380608e-02        0 1.383148e+00       0     0
#> 2015-05-17 19:15:00 4.904451e-01        0 2.373219e+00       0     0
#> 2015-05-17 19:16:00 3.701806e-03        0 1.342934e+00       0     0
#> 2015-05-17 19:17:00 5.948814e-02        0 2.273244e+00       0     0
#> 2015-05-17 19:18:00 5.601749e-01        0 4.450382e+00       0     0
#> 2015-05-17 19:19:00 6.293680e-01        0 2.784704e+00       0     0
#> 2015-05-17 19:20:00 1.992169e-01        0 2.471552e+00       0     0
#> 2015-05-17 19:21:00 4.960258e-01        0 3.885684e+00       0     0
#> 2015-05-17 19:22:00 6.925554e-01        0 3.449012e+00       0     0
#> 2015-05-17 19:23:00 8.934860e-01        0 4.253107e+00       0     0
#> 2015-05-17 19:24:00 2.112123e-02        0 4.262298e+00       0     0
#> 2015-05-17 19:25:00 3.957155e-01        0 3.282353e+00       0     0
#> 2015-05-17 19:26:00 5.282496e-01        0 1.784025e+00       0     0
#> 2015-05-17 19:27:00 9.288253e-01        0 5.395212e+00       0     0
#> 2015-05-17 19:28:00 2.419697e-01        0 3.553766e+00       0     0
#> 2015-05-17 19:29:00 6.087128e-02        0 2.484521e+00       0     0
#> 2015-05-17 19:30:00 4.187146e-01        0 4.575818e+00       0     0
#> 2015-05-17 19:31:00 4.482594e-01        0 2.603749e+00       0     0
#> 2015-05-17 19:32:00 9.702553e-01        0 3.152080e+00       0     0
#> 2015-05-17 19:33:00 5.375954e-01        0 3.276036e+00       0     0
#> 2015-05-17 19:34:00 2.960856e-01        0 3.616396e+00       0     0
#> 2015-05-17 19:35:00 9.963324e-01        0 2.856150e+00       0     0
#> 2015-05-17 19:36:00 1.765527e-01        0 1.855128e+00       0     0
#> 2015-05-17 19:37:00 1.347460e-01        0 2.161293e+00       0     0
#> 2015-05-17 19:38:00 8.090856e-01        0 4.598090e+00       0     0
#> 2015-05-17 19:39:00 2.741150e-01        0 4.600827e+00       0     0
#> 2015-05-17 19:40:00 1.078168e-01        0 4.708701e+00       0     0
#> 2015-05-17 19:41:00 1.437937e-01        0 2.725853e+00       0     0
#> 2015-05-17 19:42:00 7.122179e-01        0 2.052205e+00       0     0
#> 2015-05-17 19:43:00 6.447573e-02        0 1.465197e+00       0     0
#> 2015-05-17 19:44:00 8.901610e-03        0 1.956096e+00       0     0
#> 2015-05-17 19:45:00 1.477720e-01        0 2.850563e+00       0     0
#> 2015-05-17 19:46:00 7.056355e-01        0 1.788104e+00       0     0
#> 2015-05-17 19:47:00 9.919606e-01        0 8.149419e+00       1     0
#> 2015-05-17 19:48:00 1.467127e-01        0 4.921393e+00       0     0
#> 2015-05-17 19:49:00 3.946079e-01        0 3.252075e+00       0     0
#> 2015-05-17 19:50:00 1.858370e-01        0 5.720159e+00       0     0
#> 2015-05-17 19:51:00 6.065547e-01        0 5.841785e+00       0     0
#> 2015-05-17 19:52:00 1.555818e-01        0 6.470362e+00       0     0
#> 2015-05-17 19:53:00 3.156260e-01        0 5.650137e+00       0     0
#> 2015-05-17 19:54:00 1.045164e-01        0 6.062905e+00       0     0
#> 2015-05-17 19:55:00 7.395074e-01        0 5.229003e+00       0     0
#> 2015-05-17 19:56:00 1.681720e-02        0 3.705228e+00       0     0
#> 2015-05-17 19:57:00 1.261317e-02        0 2.527208e+00       0     0
#> 2015-05-17 19:58:00 8.514218e-01        0 3.778509e-01       0     0
#> 2015-05-17 19:59:00 1.160150e-01        0 1.365165e+00       0     0
#> 2015-05-17 22:00:00 3.754256e-01        0 1.251705e+01       1     0
#> 2015-05-17 22:01:00 5.865851e-01        0 6.009061e+00       0     0
#> 2015-05-17 22:02:00 6.041092e-01        0 7.305685e+00       0     0
#> 2015-05-17 22:03:00 4.307226e-01        0 6.042943e+00       0     0
#> 2015-05-17 22:04:00 9.479559e-02        0 5.920820e+00       0     0
#> 2015-05-17 22:05:00 5.791847e-02        0 4.688712e+00       0     0
#> 2015-05-17 22:06:00 3.970058e-01        0 5.537973e+00       0     0
#> 2015-05-17 22:07:00 2.763457e-01        0 5.964094e+00       0     0
#> 2015-05-17 22:08:00 1.978958e-01        0 6.016621e+00       0     0
#> 2015-05-17 22:09:00 5.942632e-01        0 6.545575e+00       0     0
#> 2015-05-17 22:10:00 2.545825e-01        0 8.390888e+00       1     0
#> 2015-05-17 22:11:00 2.857872e-01        0 8.036993e+00       1     0
#> 2015-05-17 22:12:00 7.453579e-01        0 7.025367e+00       0     0
#> 2015-05-17 22:13:00 1.292382e+00        0 9.165686e+00       1     0
#> 2015-05-17 22:14:00 1.048177e+00        0 7.927745e+00       1     0
#> 2015-05-17 22:15:00 8.626342e-01        0 8.562979e+00       1     0
#> 2015-05-17 22:16:00 3.488343e-01        0 1.177592e+01       1     0
#> 2015-05-17 22:17:00 6.164003e-01        0 9.618046e+00       1     1
#> 2015-05-17 22:18:00 1.142361e-01        0 8.152410e+00       1     1
#> 2015-05-17 22:19:00 1.500728e+00        0 4.567305e+00       0     0
#> 2015-05-17 22:20:00 4.960883e-01        0 3.807034e+00       0     0
#> 2015-05-17 22:21:00 6.752976e-01        0 2.014464e+00       0     0
#> 2015-05-17 22:22:00 6.895545e-01        0 2.755549e+00       0     0
#> 2015-05-17 22:23:00 5.531777e-01        0 5.049380e+00       0     0
#> 2015-05-17 22:24:00 7.563366e-01        0 4.508433e+00       0     0
#> 2015-05-17 22:25:00 3.862508e-01        0 4.000206e+00       0     0
#> 2015-05-17 22:26:00 1.064565e-02        0 5.919991e+00       0     0
#> 2015-05-17 22:27:00 6.073149e-01        0 6.104407e+00       0     0
#> 2015-05-17 22:28:00 3.903237e-01        0 2.533987e+00       0     0
#> 2015-05-17 22:29:00 4.146299e-02        0 4.009453e+00       0     0
#> 2015-05-17 22:30:00 4.308342e-01        0 4.565230e+00       0     0
#> 2015-05-17 22:31:00 2.345712e-02        0 4.822153e+00       0     0
#> 2015-05-17 22:32:00 1.114100e+00        0 4.842374e+00       0     0
#> 2015-05-17 22:33:00 5.720700e-03        0 3.489478e+00       0     0
#> 2015-05-17 22:34:00 4.965796e-01        0 5.854044e+00       0     0
#> 2015-05-17 22:35:00 2.081255e-02        0 4.336035e+00       0     0
#> 2015-05-17 22:36:00 1.404491e-01        0 3.508048e+00       0     0
#> 2015-05-17 22:37:00 2.463849e-01        0 2.705469e+00       0     0
#> 2015-05-17 22:38:00 1.433998e-01        0 3.620371e+00       0     0
#> 2015-05-17 22:39:00 1.116796e-01        0 5.191942e+00       0     0
#> 2015-05-17 22:40:00 5.919498e-01        0 5.153451e+00       0     0
#> 2015-05-17 22:41:00 1.175371e-01        0 4.419764e+00       0     0
#> 2015-05-17 22:42:00 1.231838e-01        0 3.795792e+00       0     0
#> 2015-05-17 22:43:00 1.742186e-01        0 1.980143e+00       0     0
#> 2015-05-17 22:44:00 3.861567e-01        0 3.428309e+00       0     0
#> 2015-05-17 22:45:00 1.256809e-01        0 4.533996e+00       0     0
#> 2015-05-17 22:46:00 6.229459e-01        0 3.490059e+00       0     0
#> 2015-05-17 22:47:00 3.995255e-01        0 7.823709e+00       1     0
#> 2015-05-17 22:48:00 4.493759e-01        0 8.234401e+00       1     0
#> 2015-05-17 22:49:00 3.081371e-01        0 7.842999e+00       1     0
#> 2015-05-17 22:50:00 4.503264e-01        0 5.085895e+00       0     0
#> 2015-05-17 22:51:00 4.022634e-02        0 5.383396e+00       0     0
#> 2015-05-17 22:52:00 3.782628e-01        0 3.632956e+00       0     0
#> 2015-05-17 22:53:00 6.853981e-01        0 3.088736e+00       0     0
#> 2015-05-17 22:54:00 7.641215e-02        0 3.620033e+00       0     0
#> 2015-05-17 22:55:00 3.845735e-01        0 3.075495e+00       0     0
#> 2015-05-17 22:56:00 8.841779e-02        0 3.126937e+00       0     0
#> 2015-05-17 22:57:00 3.691514e-01        0 2.834443e+00       0     0
#> 2015-05-17 22:58:00 1.895070e-02        0 3.582837e+00       0     0
#> 2015-05-17 22:59:00 2.536900e-01        0 4.827402e+00       0     0
#> 2015-05-18 01:00:00 1.325997e+00        0 1.067163e+01       1     0
#> 2015-05-18 01:01:00 4.838348e-01        0 1.293017e+01       1     0
#> 2015-05-18 01:02:00 6.020627e-01        0 1.034050e+01       1     0
#> 2015-05-18 01:03:00 2.458727e+00        1 1.064876e+01       1     0
#> 2015-05-18 01:04:00 8.989961e-01        0 8.919222e+00       1     1
#> 2015-05-18 01:05:00 3.353734e-01        0 8.035936e+00       1     1
#> 2015-05-18 01:06:00 5.364450e-01        0 9.114321e+00       1     1
#> 2015-05-18 01:07:00 1.241583e+00        0 1.204553e+01       1     1
#> 2015-05-18 01:08:00 1.085474e+00        0 1.204183e+01       1     1
#> 2015-05-18 01:09:00 1.810269e+00        0 1.038520e+01       1     1
#> 2015-05-18 01:10:00 2.001899e+00        1 1.171841e+01       1     1
#> 2015-05-18 01:11:00 7.699784e-01        0 8.917823e+00       1     1
#> 2015-05-18 01:12:00 5.327415e-01        0 8.169432e+00       1     1
#> 2015-05-18 01:13:00 5.838712e-01        0 5.788298e+00       0     0
#> 2015-05-18 01:14:00 6.679072e-01        0 8.445962e+00       1     0
#> 2015-05-18 01:15:00 6.976756e-01        0 6.041168e+00       0     0
#> 2015-05-18 01:16:00 1.098201e-01        0 4.962349e+00       0     0
#> 2015-05-18 01:17:00 3.088988e-01        0 3.606474e+00       0     0
#> 2015-05-18 01:18:00 3.678167e-01        0 6.584924e+00       0     0
#> 2015-05-18 01:19:00 1.698391e-01        0 4.800437e+00       0     0
#> 2015-05-18 01:20:00 7.648519e-01        0 4.950914e+00       0     0
#> 2015-05-18 01:21:00 9.334053e-01        0 6.193329e+00       0     0
#> 2015-05-18 01:22:00 1.338723e-01        0 8.895487e+00       1     0
#> 2015-05-18 01:23:00 1.544913e+00        0 5.915276e+00       0     0
#> 2015-05-18 01:24:00 1.245397e+00        0 7.128119e+00       0     0
#> 2015-05-18 01:25:00 1.644791e+00        0 6.479314e+00       0     0
#> 2015-05-18 01:26:00 4.788914e-01        0 9.427100e+00       1     0
#> 2015-05-18 01:27:00 4.866173e-01        0 7.413192e+00       0     0
#> 2015-05-18 01:28:00 1.057515e+00        0 5.129325e+00       0     0
#> 2015-05-18 01:29:00 4.525177e-01        0 7.946663e+00       1     0
#> 2015-05-18 01:30:00 7.948858e-01        0 7.896067e+00       1     0
#> 2015-05-18 01:31:00 7.080458e-01        0 8.134120e+00       1     0
#> 2015-05-18 01:32:00 5.094920e-01        0 5.985209e+00       0     0
#> 2015-05-18 01:33:00 6.599689e-01        0 3.452813e+00       0     0
#> 2015-05-18 01:34:00 7.737165e-01        0 7.822099e+00       1     0
#> 2015-05-18 01:35:00 1.669604e+00        0 7.401744e+00       0     0
#> 2015-05-18 01:36:00 5.678410e-01        0 7.478511e+00       0     0
#> 2015-05-18 01:37:00 3.944319e-01        0 6.797540e+00       0     0
#> 2015-05-18 01:38:00 7.399370e-01        0 1.063717e+01       1     0
#> 2015-05-18 01:39:00 2.872269e-01        0 1.042655e+01       1     0
#> 2015-05-18 01:40:00 5.599879e-01        0 8.069968e+00       1     0
#> 2015-05-18 01:41:00 2.892932e-01        0 9.799184e+00       1     0
#> 2015-05-18 01:42:00 6.785141e-01        0 9.246130e+00       1     1
#> 2015-05-18 01:43:00 5.093461e-01        0 6.730358e+00       0     0
#> 2015-05-18 01:44:00 3.033322e-01        0 6.291642e+00       0     0
#> 2015-05-18 01:45:00 8.489072e-01        0 1.545384e+01       1     0
#> 2015-05-18 01:46:00 1.128882e+00        0 1.386677e+01       1     0
#> 2015-05-18 01:47:00 1.187501e+00        0 1.192652e+01       1     0
#> 2015-05-18 01:48:00 2.420407e+00        1 6.803471e+00       0     0
#> 2015-05-18 01:49:00 1.522337e+00        0 8.610134e+00       1     0
#> 2015-05-18 01:50:00 5.914381e-01        0 4.785292e+00       0     0
#> 2015-05-18 01:51:00 5.222246e-01        0 6.242934e+00       0     0
#> 2015-05-18 01:52:00 3.529407e-01        0 5.811664e+00       0     0
#> 2015-05-18 01:53:00 2.264368e+00        1 6.523826e+00       0     0
#> 2015-05-18 01:54:00 1.148550e+00        0 6.846113e+00       0     0
#> 2015-05-18 01:55:00 2.729466e-01        0 5.782372e+00       0     0
#> 2015-05-18 01:56:00 4.770255e-01        0 7.705274e+00       0     0
#> 2015-05-18 01:57:00 1.220790e+00        0 7.307010e+00       0     0
#> 2015-05-18 01:58:00 1.349665e+00        0 8.129593e+00       1     0
#> 2015-05-18 01:59:00 3.310237e-01        0 4.044988e+00       0     0
#> 2015-05-18 04:00:00 5.305662e-01        0 8.021678e+00       1     0
#> 2015-05-18 04:01:00 2.442602e+00        1 1.144365e+01       1     0
#> 2015-05-18 04:02:00 1.179993e+00        0 1.186365e+01       1     0
#> 2015-05-18 04:03:00 1.536292e+00        0 1.375330e+01       1     0
#> 2015-05-18 04:04:00 2.369945e+00        1 1.035741e+01       1     1
#> 2015-05-18 04:05:00 2.856446e+00        1 7.559885e+00       0     0
#> 2015-05-18 04:06:00 5.770022e-01        0 6.974735e+00       0     0
#> 2015-05-18 04:07:00 8.377565e-01        0 8.772345e+00       1     0
#> 2015-05-18 04:08:00 3.188200e-01        0 6.073827e+00       0     0
#> 2015-05-18 04:09:00 3.549410e-01        0 5.635486e+00       0     0
#> 2015-05-18 04:10:00 1.828205e-01        0 3.123149e+00       0     0
#> 2015-05-18 04:11:00 8.252857e-01        0 4.857243e+00       0     0
#> 2015-05-18 04:12:00 9.871082e-01        0 4.523701e+00       0     0
#> 2015-05-18 04:13:00 3.777005e-01        0 3.256787e+00       0     0
#> 2015-05-18 04:14:00 2.583129e-01        0 7.922342e+00       1     0
#> 2015-05-18 04:15:00 3.876956e-01        0 8.668598e+00       1     0
#> 2015-05-18 04:16:00 6.513451e-01        0 1.267887e+01       1     0
#> 2015-05-18 04:17:00 1.014412e+00        0 1.066915e+01       1     0
#> 2015-05-18 04:18:00 2.458031e+00        1 1.096397e+01       1     1
#> 2015-05-18 04:19:00 1.335709e+00        0 1.508687e+01       1     1
#> 2015-05-18 04:20:00 2.058177e+00        1 1.061961e+01       1     1
#> 2015-05-18 04:21:00 6.972530e-01        0 9.700285e+00       1     1
#> 2015-05-18 04:22:00 1.225052e+00        0 8.154723e+00       1     1
#> 2015-05-18 04:23:00 6.064139e-01        0 9.866564e+00       1     1
#> 2015-05-18 04:24:00 1.392600e+00        0 9.564861e+00       1     1
#> 2015-05-18 04:25:00 2.854080e-01        0 8.078900e+00       1     1
#> 2015-05-18 04:26:00 1.013030e+00        0 8.860166e+00       1     1
#> 2015-05-18 04:27:00 8.504004e-01        0 6.164668e+00       0     0
#> 2015-05-18 04:28:00 2.594341e-01        0 6.290993e+00       0     0
#> 2015-05-18 04:29:00 7.964691e-01        0 5.668254e+00       0     0
#> 2015-05-18 04:30:00 7.985101e-01        0 1.169454e+01       1     0
#> 2015-05-18 04:31:00 1.349224e+00        0 1.708027e+01       1     0
#>  [ reached getOption("max.print") -- omitted 2488 rows ]
#> 
#> $Non_Alarmed_Obs
#>                              x          y            z
#> 2015-05-17 19:12:00 0.69969016 -1.6372697  1.268822054
#> 2015-05-17 19:13:00 0.60444181 -1.7109133  1.330683646
#> 2015-05-17 19:14:00 0.63695286 -1.5691873  1.049103203
#> 2015-05-17 19:15:00 0.85244737 -1.6359295  0.979478553
#> 2015-05-17 19:16:00 0.66778364 -1.5516544  0.988942864
#> 2015-05-17 19:17:00 0.76732071 -1.5918329  1.152449884
#> 2015-05-17 19:18:00 0.84423619 -1.6872757  1.588739791
#> 2015-05-17 19:19:00 0.76148201 -1.5134016  1.430440956
#> 2015-05-17 19:20:00 0.68031459 -1.6950354  1.283073241
#> 2015-05-17 19:21:00 0.92962646 -1.6495333  1.307341619
#> 2015-05-17 19:22:00 0.80653878 -1.5633713  1.505374630
#> 2015-05-17 19:23:00 0.73470061 -1.7746138  1.652301715
#> 2015-05-17 19:24:00 0.84101472 -1.9036169  1.313547381
#> 2015-05-17 19:25:00 0.71085968 -1.7436662  1.445602954
#> 2015-05-17 19:26:00 0.71898843 -1.7758477  0.853608003
#> 2015-05-17 19:27:00 0.99144441 -1.6654594  1.572092767
#> 2015-05-17 19:28:00 0.73445822 -1.8055549  1.418096935
#> 2015-05-17 19:29:00 0.76880182 -1.6107742  1.207486244
#> 2015-05-17 19:30:00 0.89690314 -1.6944874  1.513581660
#> 2015-05-17 19:31:00 0.85180409 -1.5205849  1.186610821
#> 2015-05-17 19:32:00 0.62469659 -1.8559771  1.450230111
#> 2015-05-17 19:33:00 0.90815313 -1.5970872  1.220715851
#> 2015-05-17 19:34:00 0.82065854 -1.6519014  1.434837909
#> 2015-05-17 19:35:00 0.64777683 -1.6581186  1.519069948
#> 2015-05-17 19:36:00 0.64115205 -1.6316554  1.186395429
#> 2015-05-17 19:37:00 0.78387978 -1.5554122  1.116573264
#> 2015-05-17 19:38:00 0.81370443 -1.7034204  1.669732992
#> 2015-05-17 19:39:00 0.78447902 -1.8824898  1.536143764
#> 2015-05-17 19:40:00 0.87964816 -1.9696924  1.289172936
#> 2015-05-17 19:41:00 0.73603909 -1.6465684  1.317486708
#> 2015-05-17 19:42:00 0.62349506 -1.5599417  1.376870576
#> 2015-05-17 19:43:00 0.63928794 -1.5629377  1.090420762
#> 2015-05-17 19:44:00 0.72262692 -1.6362837  1.066707564
#> 2015-05-17 19:45:00 0.70774298 -1.7582938  1.295533015
#> 2015-05-17 19:46:00 0.56429413 -1.6849217  1.249028127
#> 2015-05-17 19:47:00 1.08829552 -1.8467915  1.799159521
#> 2015-05-17 19:48:00 0.90497922 -1.8023858  1.473655273
#> 2015-05-17 19:49:00 0.76889037 -1.6269468  1.448774754
#> 2015-05-17 19:50:00 0.87369332 -1.9076171  1.619142180
#> 2015-05-17 19:51:00 1.00394187 -1.7629263  1.551755854
#> 2015-05-17 19:52:00 0.98189039 -1.9220280  1.569210712
#> 2015-05-17 19:53:00 0.91427885 -1.8180890  1.619706243
#> 2015-05-17 19:54:00 0.94798715 -1.9094284  1.555000709
#> 2015-05-17 19:55:00 0.99232090 -1.6946000  1.498355473
#> 2015-05-17 19:56:00 0.83657342 -1.7852999  1.288864598
#> 2015-05-17 19:57:00 0.73831698 -1.6813173  1.206352338
#> 2015-05-17 19:58:00 0.41068721 -1.4475409  0.973518992
#> 2015-05-17 19:59:00 0.69747705 -1.4338394  1.068100780
#> 2015-05-17 22:00:00 1.05591341 -2.3404776  2.114448317
#> 2015-05-17 22:01:00 0.81631705 -1.9673596  1.733703088
#> 2015-05-17 22:02:00 0.98747826 -1.8714083  1.793636352
#> 2015-05-17 22:03:00 0.86332965 -1.9043213  1.718341106
#> 2015-05-17 22:04:00 0.88763736 -1.9493551  1.595350782
#> 2015-05-17 22:05:00 0.89094988 -1.8298814  1.410901050
#> 2015-05-17 22:06:00 0.81877349 -1.9285642  1.658576932
#> 2015-05-17 22:07:00 0.96896655 -1.8416252  1.563880608
#> 2015-05-17 22:08:00 0.87625492 -1.9421265  1.647220004
#> 2015-05-17 22:09:00 1.01011221 -1.8078619  1.653223388
#> 2015-05-17 22:10:00 1.00820435 -2.0264874  1.809647034
#> 2015-05-17 22:11:00 0.96367123 -2.0315497  1.819441932
#> 2015-05-17 22:12:00 0.94671254 -1.8544265  1.830324930
#> 2015-05-17 22:13:00 1.12165326 -1.8653441  1.912152533
#> 2015-05-17 22:14:00 0.87259828 -2.0354328  1.968140178
#> 2015-05-17 22:15:00 1.01995666 -1.9191864  1.934847896
#> 2015-05-17 22:16:00 1.11141313 -2.1973606  2.038982927
#> 2015-05-17 22:19:00 0.67362054 -1.9061057  1.714946477
#> 2015-05-17 22:20:00 0.74591860 -1.7448783  1.535806376
#> 2015-05-17 22:21:00 0.59230362 -1.8310906  1.133339086
#> 2015-05-17 22:22:00 0.89399851 -1.5176305  1.161556738
#> 2015-05-17 22:23:00 0.90267183 -1.7133084  1.605452715
#> 2015-05-17 22:24:00 0.75400072 -1.8061339  1.651202202
#> 2015-05-17 22:25:00 0.80014492 -1.7087955  1.525749857
#> 2015-05-17 22:26:00 0.92581319 -2.0045075  1.463215267
#> 2015-05-17 22:27:00 0.93318640 -1.7931795  1.716617869
#> 2015-05-17 22:28:00 0.65227518 -1.7295888  1.322633895
#> 2015-05-17 22:29:00 0.85364865 -1.7804831  1.349428065
#> 2015-05-17 22:30:00 0.78053978 -1.8314133  1.588553343
#> 2015-05-17 22:31:00 0.84184315 -1.9330236  1.430341870
#> 2015-05-17 22:32:00 0.74369646 -1.8266441  1.737132670
#> 2015-05-17 22:33:00 0.79195377 -1.8043794  1.289456155
#> 2015-05-17 22:34:00 0.83466597 -1.9172504  1.715517450
#> 2015-05-17 22:35:00 0.82183755 -1.8857446  1.389621442
#> 2015-05-17 22:36:00 0.81826913 -1.6850012  1.371535955
#> 2015-05-17 22:37:00 0.68990007 -1.7099769  1.331593804
#> 2015-05-17 22:38:00 0.75477031 -1.8102157  1.394170284
#> 2015-05-17 22:39:00 0.91887950 -1.8406377  1.474808461
#> 2015-05-17 22:40:00 0.83398388 -1.7914209  1.680496004
#> 2015-05-17 22:41:00 0.82496427 -1.8172892  1.479202184
#> 2015-05-17 22:42:00 0.87233574 -1.8170034  1.213122124
#> 2015-05-17 22:43:00 0.67874977 -1.5651208  1.236381617
#> 2015-05-17 22:44:00 0.75400163 -1.6821240  1.473597530
#> 2015-05-17 22:45:00 0.91190209 -1.8097931  1.351009465
#> 2015-05-17 22:46:00 0.69419945 -1.7767103  1.506730802
#> 2015-05-17 22:47:00 1.03514558 -1.9363011  1.740381022
#> 2015-05-17 22:48:00 1.01831725 -1.9615947  1.828332099
#> 2015-05-17 22:49:00 0.97802714 -1.9885802  1.798370478
#> 2015-05-17 22:50:00 0.92621172 -1.7281646  1.553697997
#> 2015-05-17 22:51:00 0.91317477 -1.8963516  1.473927012
#> 2015-05-17 22:52:00 0.89747003 -1.6373623  1.307785148
#> 2015-05-17 22:53:00 0.69104045 -1.6713747  1.498748313
#> 2015-05-17 22:54:00 0.77151148 -1.8031380  1.369454980
#> 2015-05-17 22:55:00 0.69918248 -1.7281559  1.418440119
#> 2015-05-17 22:56:00 0.78629579 -1.6763122  1.322960914
#> 2015-05-17 22:57:00 0.69830412 -1.6735634  1.398312677
#> 2015-05-17 22:58:00 0.81398184 -1.7613472  1.321197862
#> 2015-05-17 22:59:00 0.93683298 -1.7723699  1.420177365
#> 2015-05-18 01:00:00 1.13612021 -1.9537269  2.062876844
#> 2015-05-18 01:01:00 1.10925916 -2.2640407  2.159269433
#> 2015-05-18 01:02:00 1.02818588 -2.1145505  2.040897151
#> 2015-05-18 01:03:00 1.07656667 -1.8779498  2.253803452
#> 2015-05-18 01:13:00 0.78155467 -2.0669327  1.642791975
#> 2015-05-18 01:14:00 1.02584091 -1.9348588  1.884035078
#> 2015-05-18 01:15:00 0.91301154 -1.7887442  1.745454303
#> 2015-05-18 01:16:00 0.83198218 -1.9022166  1.518128460
#> 2015-05-18 01:17:00 0.88078252 -1.6428350  1.326077092
#> 2015-05-18 01:18:00 0.99220363 -1.8603550  1.640528384
#> 2015-05-18 01:19:00 0.84996411 -1.8182588  1.531155653
#> 2015-05-18 01:20:00 0.87467162 -1.6878788  1.661175870
#> 2015-05-18 01:21:00 0.80000106 -1.9635519  1.811558367
#> 2015-05-18 01:22:00 1.07395796 -2.1060440  1.696555236
#> 2015-05-18 01:23:00 0.71405879 -2.0775445  1.790980346
#> 2015-05-18 01:24:00 1.14205008 -1.8609796  1.470331923
#> 2015-05-18 01:25:00 0.75629094 -2.0087794  1.912079352
#> 2015-05-18 01:26:00 1.00657329 -2.0765766  1.957285336
#> 2015-05-18 01:27:00 1.04504140 -1.8932759  1.682435862
#> 2015-05-18 01:28:00 0.73205206 -1.9323823  1.720231274
#> 2015-05-18 01:29:00 0.90703497 -2.0838821  1.854329030
#> 2015-05-18 01:30:00 1.12539522 -1.9253513  1.593456147
#> 2015-05-18 01:31:00 1.03694632 -1.8974211  1.840849191
#> 2015-05-18 01:32:00 0.92842911 -1.8019897  1.688774611
#> 2015-05-18 01:33:00 0.67373155 -1.8362065  1.471373783
#> 2015-05-18 01:34:00 1.05442076 -1.8603556  1.783809187
#> 2015-05-18 01:35:00 0.97452660 -1.7525214  1.965098783
#> 2015-05-18 01:36:00 0.98775938 -1.8931565  1.806167058
#> 2015-05-18 01:37:00 0.93514674 -1.9048784  1.749528812
#> 2015-05-18 01:38:00 1.15254405 -2.0359059  1.937944011
#> 2015-05-18 01:39:00 1.15362945 -2.1789494  1.746787004
#> 2015-05-18 01:40:00 0.95271125 -1.9909674  1.890255789
#> 2015-05-18 01:41:00 1.04036671 -2.1139288  1.921026531
#> 2015-05-18 01:43:00 0.95752419 -1.8551826  1.744704805
#> 2015-05-18 01:44:00 0.89203204 -1.9216027  1.701888693
#> 2015-05-18 01:45:00 1.17560622 -2.3221181  2.351978673
#> 2015-05-18 01:46:00 1.12731347 -2.2187504  2.318497375
#> 2015-05-18 01:47:00 1.09165205 -2.0949620  2.210724293
#> 2015-05-18 01:48:00 0.77949980 -1.8945058  2.060673511
#> 2015-05-18 01:49:00 0.86120584 -2.0838981  2.073633075
#> 2015-05-18 01:50:00 0.76297721 -1.8782697  1.630401881
#> 2015-05-18 01:51:00 0.94433447 -1.8144420  1.704352501
#> 2015-05-18 01:52:00 0.83936609 -1.9387458  1.673643002
#> 2015-05-18 01:53:00 0.79096795 -1.8455281  2.029048681
#> 2015-05-18 01:54:00 0.88536508 -1.8538268  1.909981059
#> 2015-05-18 01:55:00 0.94063075 -1.8286581  1.589370533
#> 2015-05-18 01:56:00 0.99264775 -1.9270915  1.807560940
#> 2015-05-18 01:57:00 0.86687499 -1.9346632  1.958189998
#> 2015-05-18 01:58:00 0.94073279 -1.8991980  2.022771455
#> 2015-05-18 01:59:00 0.76872620 -1.7862695  1.516396854
#> 2015-05-18 04:00:00 0.92176227 -2.0426970  1.884968996
#> 2015-05-18 04:01:00 1.25198760 -1.8675912  2.064249684
#> 2015-05-18 04:02:00 1.16262813 -2.0394042  2.123160611
#> 2015-05-18 04:03:00 1.10665273 -2.1872309  2.374292929
#> 2015-05-18 04:05:00 0.78816791 -1.9403471  2.156068641
#> 2015-05-18 04:06:00 0.93526643 -1.8888498  1.804796174
#> 2015-05-18 04:07:00 0.99494736 -1.9656623  1.973934948
#> 2015-05-18 04:08:00 0.83392936 -2.0201260  1.658782269
#> 2015-05-18 04:09:00 0.98419501 -1.8122407  1.488184599
#> 2015-05-18 04:10:00 0.71362696 -1.8192230  1.309281073
#> 2015-05-18 04:11:00 0.74916260 -1.8737013  1.680441772
#> 2015-05-18 04:12:00 0.75849359 -1.7567143  1.699221262
#> 2015-05-18 04:13:00 0.88955910 -1.6266346  1.218927848
#> 2015-05-18 04:14:00 0.98878508 -1.9987240  1.783022312
#> 2015-05-18 04:15:00 1.01744736 -2.0121485  1.861298310
#> 2015-05-18 04:16:00 1.14244915 -2.1816220  2.143202453
#> 2015-05-18 04:17:00 1.00232755 -2.1204773  2.141982607
#> 2015-05-18 04:27:00 1.06640004 -1.7957388  1.472000277
#> 2015-05-18 04:28:00 0.90935565 -1.9087562  1.682143731
#> 2015-05-18 04:29:00 0.76233460 -2.0246981  1.696042867
#> 2015-05-18 04:30:00 1.14139060 -2.0911145  2.080886217
#> 2015-05-18 04:31:00 1.27987825 -2.2759127  2.430629878
#> 2015-05-18 04:32:00 0.98882910 -2.1120239  2.590256988
#> 2015-05-18 04:33:00 1.17181648 -2.0905618  2.380477171
#> 2015-05-18 04:45:00 0.88547807 -1.8781155  1.881885823
#> 2015-05-18 04:46:00 0.98556843 -1.9663019  1.912005548
#> 2015-05-18 04:47:00 1.09278227 -2.0500571  2.106547270
#> 2015-05-18 04:48:00 0.99865126 -1.9614584  2.077747517
#> 2015-05-18 04:49:00 0.91702270 -2.0731965  1.743018849
#> 2015-05-18 04:50:00 1.13011465 -2.0076572  1.905021469
#> 2015-05-18 04:51:00 0.61559002 -2.0315828  1.893107028
#> 2015-05-18 04:52:00 0.87600890 -1.8714397  1.954477255
#> 2015-05-18 04:53:00 0.94063606 -1.9079787  2.080809768
#> 2015-05-18 04:54:00 1.18145677 -2.0937789  2.232173733
#> 2015-05-18 04:55:00 1.01925478 -2.2678320  2.207614077
#> 2015-05-18 04:56:00 1.06020730 -2.0589805  2.232408720
#> 2015-05-18 04:58:00 0.85987633 -1.8063230  2.063911674
#> 2015-05-18 04:59:00 1.04038245 -1.8944953  1.855361385
#> 2015-05-18 07:00:00 1.04979342 -1.7977702  2.198187648
#> 2015-05-18 07:01:00 1.07919508 -2.0352094  2.408688860
#> 2015-05-18 07:02:00 0.97906667 -2.0937792  1.953385451
#> 2015-05-18 07:03:00 0.87988567 -1.9855969  1.869523495
#> 2015-05-18 07:04:00 0.88472561 -1.8890324  2.039916021
#> 2015-05-18 07:05:00 0.97851344 -1.9459661  2.097293162
#> 2015-05-18 07:06:00 1.01395281 -1.9172998  2.250028995
#> 2015-05-18 07:07:00 1.19343864 -2.0565592  2.455882604
#> 2015-05-18 07:08:00 1.11781554 -2.0122336  2.232583437
#> 2015-05-21 13:03:00 0.81887554 -1.8527663  2.066412023
#> 2015-05-21 13:04:00 1.21711505 -2.0263208  1.966841458
#> 2015-05-21 13:05:00 1.10450108 -1.9181063  2.099304720
#> 2015-05-21 13:06:00 0.95246102 -2.0127656  1.949152418
#> 2015-05-21 13:07:00 0.91784280 -1.9975249  2.021430472
#> 2015-05-21 13:51:00 1.00918495 -1.8184105  1.758225451
#> 2015-05-21 13:52:00 0.93123730 -1.9247048  1.865641376
#> 2015-05-21 13:53:00 1.05096373 -2.1464890  2.100979917
#> 2015-05-21 13:54:00 0.86263365 -2.0106665  1.856566812
#> 2015-05-21 13:55:00 1.10868796 -1.9393935  2.159969439
#> 2015-05-21 13:56:00 1.17824522 -2.2431346  2.407681896
#> 2015-05-21 13:57:00 1.03490147 -2.1341495  2.316468922
#> 2015-05-21 13:58:00 1.07052589 -2.0734266  2.409037794
#> 2015-05-21 16:00:00 0.86920845 -1.8615281  1.804438753
#> 2015-05-21 16:01:00 0.88991657 -2.0042157  1.904105040
#> 2015-05-21 16:02:00 1.04629423 -1.9796163  2.295805131
#> 2015-05-21 16:03:00 0.97661215 -1.9436865  2.147413526
#> 2015-05-21 16:04:00 1.11233786 -1.9980928  1.688731302
#> 2015-05-21 16:05:00 0.93247013 -1.9995212  2.003783417
#> 2015-05-21 16:07:00 0.96431746 -1.7951915  1.811275403
#> 2015-05-21 16:08:00 1.06129677 -1.9076818  2.034315093
#> 2015-05-21 16:09:00 0.95954370 -1.8761920  1.958214039
#> 2015-05-21 16:10:00 0.87105064 -1.9893877  1.949979583
#> 2015-05-21 16:11:00 1.15233227 -2.0283145  1.854277507
#> 2015-05-21 16:12:00 0.87803305 -1.9199534  2.039081415
#> 2015-05-21 16:13:00 1.23413769 -2.0897405  2.153648817
#> 2015-05-21 16:14:00 1.00765561 -2.0035843  2.076619321
#> 2015-05-21 16:15:00 1.14477225 -2.0529732  2.232970738
#> 2015-05-21 16:16:00 0.72395119 -2.1409677  2.235732191
#> 2015-05-21 16:33:00 0.89726831 -2.0247400  1.756855947
#> 2015-05-21 16:34:00 1.00593168 -1.9246233  1.575700863
#> 2015-05-21 16:35:00 1.03125025 -1.9127713  1.922064372
#> 2015-05-21 16:36:00 1.15978663 -2.1963811  1.945483009
#> 2015-05-21 16:37:00 1.24325209 -1.9160788  2.062385717
#> 2015-05-21 16:38:00 1.09186220 -2.0577496  1.967698446
#> 2015-05-21 16:48:00 0.89914945 -1.9362020  1.899533390
#> 2015-05-21 16:49:00 0.93599900 -2.0037610  1.785088632
#> 2015-05-21 16:50:00 0.80130778 -2.1020603  2.064439369
#> 2015-05-21 16:51:00 0.98915436 -2.0384880  2.133522889
#> 2015-05-21 16:52:00 0.99365550 -2.0109472  1.831156407
#> 2015-05-21 16:53:00 1.00671647 -2.0359643  2.055651884
#> 2015-05-21 19:00:00 0.81755175 -1.8058706  1.569979553
#> 2015-05-21 19:01:00 0.72018524 -1.6837107  1.515291888
#> 2015-05-21 19:02:00 0.81359657 -1.7639597  1.203153254
#> 2015-05-21 19:03:00 1.10729228 -1.8247496  1.917763850
#> 2015-05-21 19:04:00 0.89392404 -1.9042127  1.793617043
#> 2015-05-21 19:05:00 0.86365140 -2.0245418  1.802304189
#> 2015-05-21 19:06:00 0.88699213 -1.9988913  1.712875615
#> 2015-05-21 19:07:00 1.00903331 -2.1098070  1.884803853
#> 2015-05-21 19:08:00 0.77730253 -1.8249261  1.752137213
#> 2015-05-21 19:09:00 0.81438394 -1.8809429  1.623568059
#> 2015-05-21 19:10:00 1.04893064 -1.9666296  2.048178769
#> 2015-05-21 19:11:00 0.93419108 -2.1457537  1.834002650
#> 2015-05-21 19:12:00 0.84268744 -1.7223003  1.563053870
#> 2015-05-21 19:13:00 1.02541520 -1.8254659  1.840349752
#> 2015-05-21 19:14:00 0.82028040 -1.9017626  1.578632836
#> 2015-05-21 19:15:00 0.65497824 -1.7024060  1.311186730
#> 2015-05-21 19:16:00 0.75627748 -1.7914797  1.338717705
#> 2015-05-21 19:17:00 1.01733367 -1.9248773  1.889243216
#> 2015-05-21 19:18:00 0.87832649 -1.9613802  1.711858231
#> 2015-05-21 19:19:00 0.97601134 -1.9656200  1.667648906
#> 2015-05-21 19:20:00 0.66442696 -1.9356111  1.748686943
#> 2015-05-21 19:21:00 0.89920092 -1.8830202  1.842455067
#> 2015-05-21 19:22:00 0.83125102 -1.9219641  1.779290780
#> 2015-05-21 19:23:00 0.80425855 -1.7856445  1.611269421
#> 2015-05-21 19:24:00 0.74695404 -1.8505517  1.608180423
#> 2015-05-21 19:25:00 0.83620145 -1.7586299  1.515907655
#> 2015-05-21 19:26:00 1.00467524 -1.8154335  1.646763848
#> 2015-05-21 19:27:00 0.87347933 -1.8558455  1.780016853
#> 2015-05-21 19:28:00 0.89727662 -1.7791407  1.832867518
#> 2015-05-21 19:29:00 0.83051636 -1.8393408  1.759152168
#> 2015-05-21 19:30:00 0.95112835 -1.9273578  1.803751446
#> 2015-05-21 19:31:00 0.78314479 -1.7303922  1.629390531
#> 2015-05-21 19:32:00 0.88431155 -1.9235228  1.895938008
#> 2015-05-21 19:33:00 0.87156606 -1.9187216  1.586967631
#> 2015-05-21 19:34:00 0.72999227 -1.7250921  1.579358680
#> 2015-05-21 19:35:00 0.98111883 -1.8616226  1.813882889
#> 2015-05-21 19:36:00 0.94360476 -1.9769945  1.876466931
#> 2015-05-21 19:37:00 0.81526566 -1.9652314  1.524252707
#> 2015-05-21 19:38:00 0.95550073 -2.0044687  1.748918577
#> 2015-05-21 19:39:00 0.87673569 -1.9689191  1.915945777
#> 2015-05-21 19:40:00 0.77419958 -1.8291428  1.582846590
#> 2015-05-21 19:41:00 0.77742772 -1.7022368  1.559320670
#> 2015-05-21 19:42:00 0.90204445 -1.7669292  1.499789077
#> 2015-05-21 19:43:00 0.93406671 -1.6227218  1.463404472
#> 2015-05-21 19:44:00 1.02530878 -1.9195846  1.593479357
#> 2015-05-21 19:45:00 0.77759235 -1.7376524  1.745280270
#> 2015-05-21 19:46:00 0.95266234 -1.9252252  1.530980327
#> 2015-05-21 19:47:00 0.71791694 -2.1173722  2.043951376
#> 2015-05-21 19:48:00 1.08443948 -1.9674257  2.112389792
#> 2015-05-21 19:49:00 1.15681849 -2.2620414  2.277310496
#> 2015-05-21 19:50:00 1.15140681 -2.1797100  2.124082852
#> 2015-05-21 19:51:00 1.03041436 -2.2041789  2.108572832
#> 2015-05-21 19:57:00 0.88062360 -1.8395623  1.840938194
#> 2015-05-21 19:58:00 0.82240762 -1.9004325  1.935977005
#> 2015-05-21 19:59:00 0.89040012 -1.8974538  1.905377634
#> 2015-05-21 22:00:00 0.99782989 -2.0978218  1.864502536
#> 2015-05-21 22:01:00 1.05846980 -2.0130280  2.106424912
#> 2015-05-21 22:02:00 1.23590157 -1.8572489  1.961562598
#> 2015-05-21 22:03:00 0.80191839 -2.1172909  2.039968195
#> 2015-05-21 22:09:00 0.86278185 -1.8435420  2.052336942
#> 2015-05-21 22:10:00 0.76897618 -1.9443848  1.973722356
#> 2015-05-21 22:11:00 0.82230147 -1.9902648  1.766371007
#> 2015-05-21 22:12:00 0.97904075 -1.9629667  1.605486425
#> 2015-05-21 22:13:00 0.98604230 -1.6250333  1.412283945
#> 2015-05-21 22:14:00 0.80201681 -1.9368454  1.284207516
#> 2015-05-21 22:15:00 0.70713953 -1.8105801  1.220978791
#> 2015-05-21 22:16:00 0.65920276 -1.5756875  1.037260794
#> 2015-05-21 22:17:00 0.83183299 -1.8801113  1.560320123
#> 2015-05-21 22:18:00 0.78356540 -1.6675355  1.314918306
#> 2015-05-21 22:19:00 0.88019973 -1.7208735  1.624671779
#> 2015-05-21 22:20:00 0.63139844 -1.8538997  1.533628765
#> 2015-05-21 22:21:00 0.90757118 -1.9203441  1.544126494
#> 2015-05-21 22:22:00 1.18303319 -1.9758762  1.462226049
#> 2015-05-21 22:23:00 0.86825011 -1.8541686  1.728325142
#> 2015-05-21 22:24:00 0.82934958 -1.8953316  1.512659734
#> 2015-05-21 22:25:00 1.02816114 -1.7443010  1.560958849
#> 2015-05-21 22:26:00 0.66198536 -1.8383046  1.587747632
#> 2015-05-21 22:27:00 0.76340240 -1.9843266  1.720949854
#> 2015-05-21 22:28:00 0.82551922 -1.6330160  1.588911282
#> 2015-05-21 22:29:00 0.89051708 -1.7647682  1.619629609
#> 2015-05-21 22:30:00 0.82823602 -1.7999658  1.501190742
#> 2015-05-21 22:31:00 0.73195635 -1.9860757  1.780164439
#> 2015-05-21 22:32:00 0.83615123 -1.7126022  1.502885519
#> 2015-05-21 22:33:00 0.73009585 -1.9867883  1.667805205
#> 2015-05-21 22:34:00 0.69575779 -1.8018589  1.707930812
#> 2015-05-21 22:35:00 0.72239465 -1.9054922  1.530985507
#> 2015-05-21 22:36:00 0.97082870 -1.9877843  1.854671922
#> 2015-05-21 22:37:00 1.00688561 -2.0431862  1.692638608
#> 2015-05-21 22:38:00 0.74770026 -1.9950197  1.911576104
#>  [ reached getOption("max.print") -- omitted 681 rows ]
#> 
#> $Alarms
#>                             x         y        z         SPE SPE_Flag        T2
#> 2015-05-17 22:17:00 1.0369163 -2.035376 1.978081  0.61640030        0  9.618046
#> 2015-05-17 22:18:00 1.0293792 -2.048854 1.697991  0.11423605        0  8.152410
#> 2015-05-18 01:04:00 1.0514742 -1.922362 1.939579  0.89899612        0  8.919222
#> 2015-05-18 01:05:00 0.9568433 -2.028744 1.835321  0.33537336        0  8.035936
#> 2015-05-18 01:06:00 1.0003412 -2.042280 1.947223  0.53644503        0  9.114321
#> 2015-05-18 01:07:00 1.1333457 -2.060638 2.186753  1.24158302        0 12.045525
#> 2015-05-18 01:08:00 1.1089685 -2.099762 2.191343  1.08547424        0 12.041831
#> 2015-05-18 01:09:00 1.0735047 -1.918398 2.170390  1.81026878        0 10.385200
#> 2015-05-18 01:10:00 1.2187949 -1.923942 2.114029  2.00189924        1 11.718414
#> 2015-05-18 01:11:00 0.9983502 -1.986144 1.973476  0.76997837        0  8.917823
#> 2015-05-18 01:12:00 1.0345756 -1.932322 1.815329  0.53274153        0  8.169432
#> 2015-05-18 01:42:00 0.9393675 -2.133850 1.991679  0.67851415        0  9.246130
#> 2015-05-18 04:04:00 0.9443516 -2.027298 2.297624  2.36994529        1 10.357405
#> 2015-05-18 04:18:00 0.9714644 -2.036998 2.341055  2.45803074        1 10.963969
#> 2015-05-18 04:19:00 1.1146718 -2.311524 2.426754  1.33570918        0 15.086868
#> 2015-05-18 04:20:00 1.1196047 -1.879849 2.164067  2.05817666        1 10.619607
#> 2015-05-18 04:21:00 1.0208438 -2.046592 2.012078  0.69725295        0  9.700285
#> 2015-05-18 04:22:00 0.8598808 -2.068925 2.002521  1.22505249        0  8.154723
#> 2015-05-18 04:23:00 1.0722052 -2.029795 1.961606  0.60641392        0  9.866564
#> 2015-05-18 04:24:00 0.9042526 -2.130310 2.122042  1.39259978        0  9.564861
#> 2015-05-18 04:25:00 0.9809135 -2.014702 1.812662  0.28540804        0  8.078900
#> 2015-05-18 04:26:00 1.0534324 -1.900701 1.947576  1.01302962        0  8.860166
#> 2015-05-18 04:34:00 1.1232367 -2.188548 2.486186  2.07543168        1 14.714728
#> 2015-05-18 04:35:00 0.9700040 -2.082075 2.650451  4.60359765        1 13.122393
#> 2015-05-18 04:36:00 1.1748230 -1.901306 2.419201  3.35510249        1 12.930600
#> 2015-05-18 04:37:00 1.1826958 -2.303752 2.798384  3.30695111        1 18.644331
#> 2015-05-18 04:38:00 1.3632523 -2.236356 2.702442  2.98195019        1 20.026576
#> 2015-05-18 04:39:00 1.2526921 -2.252782 2.442557  1.46162949        0 16.612752
#> 2015-05-18 04:40:00 1.2857637 -2.140245 2.774829  4.00077468        1 18.647707
#> 2015-05-18 04:41:00 1.1773974 -2.149348 2.580450  2.71633746        1 15.788267
#> 2015-05-18 04:42:00 1.1373855 -2.045131 2.229101  1.48032566        0 12.251260
#> 2015-05-18 04:43:00 1.0577397 -2.103381 2.333672  1.79340254        0 12.335045
#> 2015-05-18 04:44:00 0.9454172 -2.084938 2.304732  2.21416044        1 10.759386
#> 2015-05-18 04:57:00 1.1824171 -2.012846 2.281286  1.93957062        0 12.900798
#> 2015-05-18 07:09:00 0.9342394 -2.027475 2.076116  1.22700873        0  9.041538
#> 2015-05-18 07:10:00 1.1924707 -2.159770 2.449382  1.92160529        0 15.159356
#> 2015-05-18 07:11:00 0.9700245 -2.277420 2.339398  1.78199290        0 12.473589
#> 2015-05-18 07:12:00 1.1109930 -2.292831 2.306834  0.89181175        0 14.101757
#> 2015-05-18 07:13:00 1.0256295 -2.127062 2.444770  2.47709053        1 12.799661
#> 2015-05-18 07:14:00 1.2279853 -2.154774 2.590326  2.70649276        1 16.584944
#> 2015-05-18 07:15:00 1.1863105 -1.982085 2.518440  3.39906209        1 14.272095
#> 2015-05-18 07:16:00 1.2699806 -2.220015 2.471776  1.77387460        0 16.810646
#> 2015-05-18 07:17:00 1.0324882 -2.171213 2.357402  1.76778850        0 12.629028
#> 2015-05-18 07:18:00 1.3865545 -2.196022 2.510740  2.49904147        1 18.562330
#> 2015-05-18 07:19:00 1.1116306 -2.153352 2.429186  1.94756337        0 13.939027
#> 2015-05-18 07:20:00 1.1696782 -2.131135 2.577838  2.81723524        1 15.532533
#> 2015-05-18 07:21:00 1.1082996 -2.014823 2.478895  2.97133379        1 13.265499
#> 2015-05-18 07:22:00 1.0799098 -2.210145 2.243335  0.95185324        0 12.743967
#> 2015-05-18 07:23:00 1.0300555 -2.017127 1.959637  0.61838928        0  9.342180
#> 2015-05-18 07:24:00 1.1104104 -2.031731 2.157104  1.23807634        0 11.417497
#> 2015-05-18 07:25:00 1.1025701 -1.971525 1.949330  0.84208291        0  9.778805
#> 2015-05-18 07:26:00 0.9519305 -1.987131 1.926112  0.68296771        0  8.221720
#> 2015-05-18 07:27:00 0.9639932 -2.015047 2.164065  1.56175646        0  9.751197
#> 2015-05-18 07:28:00 1.3044899 -1.983893 2.244051  2.53927377        1 13.971314
#> 2015-05-18 07:29:00 1.1231300 -2.109201 2.301391  1.49214633        0 12.956032
#> 2015-05-18 07:30:00 1.0779677 -2.032260 2.223918  1.51321452        0 11.447726
#> 2015-05-18 07:31:00 1.2811503 -2.084460 2.327436  2.03101219        1 14.942842
#> 2015-05-18 07:32:00 1.4544443 -2.128197 2.484820  3.38278082        1 18.807453
#> 2015-05-18 07:33:00 1.1884268 -2.221610 2.411465  1.45001737        0 15.296970
#> 2015-05-18 07:34:00 1.2479078 -2.084981 2.618306  3.34934550        1 16.527851
#> 2015-05-18 07:35:00 1.0011898 -2.119827 2.249236  1.50702716        0 11.276275
#> 2015-05-18 07:36:00 1.0901395 -2.119051 2.171268  0.94968614        0 11.825194
#> 2015-05-18 07:37:00 0.9611474 -1.897705 2.150629  1.94477881        0  8.979192
#> 2015-05-18 07:38:00 0.8647467 -2.306112 2.345554  2.75979665        1 11.468536
#> 2015-05-18 07:39:00 0.9817866 -2.208938 2.412849  2.26682933        1 12.617944
#> 2015-05-18 07:40:00 1.1685674 -2.148807 2.368604  1.59704290        0 14.228302
#> 2015-05-18 07:41:00 1.1222889 -2.050021 2.558440  3.26338431        1 14.201414
#> 2015-05-18 07:42:00 0.9949973 -1.957503 2.262636  2.21482397        1 10.289830
#> 2015-05-18 07:43:00 1.1635390 -2.141465 2.272088  1.23059982        0 13.483321
#> 2015-05-18 07:44:00 1.1157557 -2.230921 2.277132  0.94505730        0 13.537779
#> 2015-05-18 07:45:00 1.1077419 -2.090894 2.169172  1.03862646        0 11.835795
#> 2015-05-18 07:46:00 1.1194776 -2.127004 2.411472  1.95606002        0 13.738681
#> 2015-05-18 07:47:00 1.1021973 -2.189118 2.467785  2.03394827        1 14.325365
#> 2015-05-18 07:48:00 1.1571680 -2.303454 2.444097  1.31313722        0 15.709436
#> 2015-05-18 07:49:00 1.2165253 -2.036819 2.600964  3.55830359        1 15.621032
#> 2015-05-18 07:50:00 1.1899725 -2.266453 2.230289  0.64620292        0 14.413624
#> 2015-05-18 07:51:00 1.2132351 -2.287008 2.399755  1.12365285        0 16.029904
#> 2015-05-18 07:52:00 1.2534268 -2.201812 2.649830  2.76354707        1 17.736887
#> 2015-05-18 07:53:00 1.0864200 -2.247781 2.669629  3.15292148        1 15.922463
#> 2015-05-18 07:54:00 1.2451242 -2.095264 2.750341  4.16189767        1 17.525015
#> 2015-05-18 07:55:00 1.3104090 -2.189002 2.672820  3.03024649        1 18.626137
#> 2015-05-18 07:56:00 1.1376027 -2.185549 2.597549  2.72136571        1 15.643431
#> 2015-05-18 07:57:00 1.2442026 -2.106099 2.359367  1.88698879        0 14.834481
#> 2015-05-18 07:58:00 1.1448605 -2.267953 2.699997  2.96416114        1 17.092527
#> 2015-05-18 07:59:00 1.3927853 -2.181877 2.777199  3.91186687        1 20.608547
#> 2015-05-18 10:00:00 1.0004014 -2.054611 2.411314  2.69156190        1 11.820577
#> 2015-05-18 10:01:00 0.9966159 -2.064460 2.163509  1.28431237        0 10.389883
#> 2015-05-18 10:02:00 1.2244041 -1.913525 2.279495  2.70524719        1 12.732299
#> 2015-05-18 10:03:00 1.1220581 -2.261470 2.264013  0.79645252        0 13.741859
#> 2015-05-18 10:04:00 1.2295922 -2.044481 2.377413  2.29207464        1 14.328013
#> 2015-05-18 10:05:00 1.2701797 -2.054880 2.381895  2.38163696        1 14.956016
#> 2015-05-18 10:06:00 1.1544446 -1.980399 2.670229  4.50060652        1 14.867427
#> 2015-05-18 10:07:00 1.3199069 -2.225662 2.638086  2.61355385        1 18.796946
#> 2015-05-18 10:08:00 1.2012091 -2.221095 2.511020  1.91475090        0 16.157309
#> 2015-05-18 10:09:00 1.1724296 -2.216272 2.288316  0.99071849        0 14.220816
#> 2015-05-18 10:10:00 0.9916762 -2.014068 2.588176  4.26405831        1 12.537521
#> 2015-05-18 10:11:00 1.1723480 -2.217513 2.526225  2.04939303        1 15.848118
#> 2015-05-18 10:12:00 1.3699031 -2.091097 2.692017  4.07596116        1 18.848885
#> 2015-05-18 10:13:00 1.2136236 -2.263503 2.986018  4.91003590        1 20.224943
#> 2015-05-18 10:14:00 1.3871627 -2.182234 2.888221  4.60080814        1 21.418399
#> 2015-05-18 10:15:00 1.1491585 -2.190283 2.771508  3.89261229        1 17.071969
#> 2015-05-18 10:16:00 1.1865113 -2.129125 2.432248  1.99789924        1 14.744042
#> 2015-05-18 10:17:00 1.2365675 -1.967517 2.474848  3.33514146        1 14.521319
#> 2015-05-18 10:18:00 1.0816492 -2.070948 2.499195  2.85651635        1 13.450468
#> 2015-05-18 10:19:00 1.2353503 -2.225212 2.654231  2.65609527        1 17.696829
#> 2015-05-18 10:20:00 1.2054139 -2.111259 2.577185  2.90223374        1 15.858257
#> 2015-05-18 10:21:00 1.2443190 -2.254939 2.669017  2.56329364        1 18.169913
#> 2015-05-18 10:22:00 1.0359141 -1.968031 2.344496  2.53801149        1 11.279243
#> 2015-05-18 10:23:00 1.1243436 -2.140775 2.339463  1.52761616        0 13.427867
#> 2015-05-18 10:24:00 1.1703618 -2.156162 2.632779  3.02468539        1 16.111360
#> 2015-05-18 10:25:00 1.1956426 -2.133347 2.409989  1.87189282        0 14.742432
#> 2015-05-18 10:26:00 1.4364457 -2.113734 2.554437  3.60742620        1 18.957651
#> 2015-05-18 10:27:00 1.1889435 -2.088898 2.646418  3.50825242        1 15.955140
#> 2015-05-18 10:28:00 1.1558670 -2.136660 2.234971  1.11019563        0 13.118549
#> 2015-05-18 10:29:00 1.1855170 -2.231249 2.871700  4.28257909        1 18.656704
#> 2015-05-18 10:30:00 1.4751653 -2.298289 2.759893  3.27146026        1 22.772335
#> 2015-05-18 10:31:00 1.4550236 -2.159610 2.814291  4.60887260        1 21.681669
#> 2015-05-18 10:32:00 1.3017333 -2.085895 2.705115  3.97793170        1 17.920469
#> 2015-05-18 10:33:00 1.1428837 -2.196493 2.721711  3.50451275        1 16.673506
#> 2015-05-18 10:34:00 1.1974889 -2.141118 2.493973  2.25038985        1 15.395404
#> 2015-05-18 10:35:00 1.0977893 -2.442989 2.658101  2.30580020        1 17.485880
#> 2015-05-18 10:36:00 1.1603422 -2.172222 2.667411  3.19181488        1 16.341080
#> 2015-05-18 10:37:00 1.0641403 -2.191890 2.613707  3.13329775        1 14.833650
#> 2015-05-18 10:38:00 1.1848704 -2.079080 2.565820  3.04402806        1 15.269226
#> 2015-05-18 10:39:00 1.0694241 -2.044881 2.659267  4.23818576        1 14.160865
#>                     T2_Flag Alarm
#> 2015-05-17 22:17:00       1     1
#> 2015-05-17 22:18:00       1     1
#> 2015-05-18 01:04:00       1     1
#> 2015-05-18 01:05:00       1     1
#> 2015-05-18 01:06:00       1     1
#> 2015-05-18 01:07:00       1     1
#> 2015-05-18 01:08:00       1     1
#> 2015-05-18 01:09:00       1     1
#> 2015-05-18 01:10:00       1     1
#> 2015-05-18 01:11:00       1     1
#> 2015-05-18 01:12:00       1     1
#> 2015-05-18 01:42:00       1     1
#> 2015-05-18 04:04:00       1     1
#> 2015-05-18 04:18:00       1     1
#> 2015-05-18 04:19:00       1     1
#> 2015-05-18 04:20:00       1     1
#> 2015-05-18 04:21:00       1     1
#> 2015-05-18 04:22:00       1     1
#> 2015-05-18 04:23:00       1     1
#> 2015-05-18 04:24:00       1     1
#> 2015-05-18 04:25:00       1     1
#> 2015-05-18 04:26:00       1     1
#> 2015-05-18 04:34:00       1     1
#> 2015-05-18 04:35:00       1     1
#> 2015-05-18 04:36:00       1     1
#> 2015-05-18 04:37:00       1     1
#> 2015-05-18 04:38:00       1     3
#> 2015-05-18 04:39:00       1     1
#> 2015-05-18 04:40:00       1     1
#> 2015-05-18 04:41:00       1     1
#> 2015-05-18 04:42:00       1     1
#> 2015-05-18 04:43:00       1     1
#> 2015-05-18 04:44:00       1     1
#> 2015-05-18 04:57:00       1     1
#> 2015-05-18 07:09:00       1     1
#> 2015-05-18 07:10:00       1     1
#> 2015-05-18 07:11:00       1     1
#> 2015-05-18 07:12:00       1     1
#> 2015-05-18 07:13:00       1     1
#> 2015-05-18 07:14:00       1     1
#> 2015-05-18 07:15:00       1     1
#> 2015-05-18 07:16:00       1     1
#> 2015-05-18 07:17:00       1     1
#> 2015-05-18 07:18:00       1     1
#> 2015-05-18 07:19:00       1     1
#> 2015-05-18 07:20:00       1     1
#> 2015-05-18 07:21:00       1     1
#> 2015-05-18 07:22:00       1     1
#> 2015-05-18 07:23:00       1     1
#> 2015-05-18 07:24:00       1     1
#> 2015-05-18 07:25:00       1     1
#> 2015-05-18 07:26:00       1     1
#> 2015-05-18 07:27:00       1     1
#> 2015-05-18 07:28:00       1     1
#> 2015-05-18 07:29:00       1     1
#> 2015-05-18 07:30:00       1     1
#> 2015-05-18 07:31:00       1     1
#> 2015-05-18 07:32:00       1     1
#> 2015-05-18 07:33:00       1     1
#> 2015-05-18 07:34:00       1     1
#> 2015-05-18 07:35:00       1     1
#> 2015-05-18 07:36:00       1     1
#> 2015-05-18 07:37:00       1     1
#> 2015-05-18 07:38:00       1     1
#> 2015-05-18 07:39:00       1     1
#> 2015-05-18 07:40:00       1     1
#> 2015-05-18 07:41:00       1     1
#> 2015-05-18 07:42:00       1     1
#> 2015-05-18 07:43:00       1     1
#> 2015-05-18 07:44:00       1     1
#> 2015-05-18 07:45:00       1     1
#> 2015-05-18 07:46:00       1     1
#> 2015-05-18 07:47:00       1     1
#> 2015-05-18 07:48:00       1     1
#> 2015-05-18 07:49:00       1     1
#> 2015-05-18 07:50:00       1     1
#> 2015-05-18 07:51:00       1     1
#> 2015-05-18 07:52:00       1     1
#> 2015-05-18 07:53:00       1     1
#> 2015-05-18 07:54:00       1     1
#> 2015-05-18 07:55:00       1     1
#> 2015-05-18 07:56:00       1     3
#> 2015-05-18 07:57:00       1     1
#> 2015-05-18 07:58:00       1     1
#> 2015-05-18 07:59:00       1     1
#> 2015-05-18 10:00:00       1     1
#> 2015-05-18 10:01:00       1     1
#> 2015-05-18 10:02:00       1     1
#> 2015-05-18 10:03:00       1     1
#> 2015-05-18 10:04:00       1     1
#> 2015-05-18 10:05:00       1     1
#> 2015-05-18 10:06:00       1     1
#> 2015-05-18 10:07:00       1     1
#> 2015-05-18 10:08:00       1     1
#> 2015-05-18 10:09:00       1     1
#> 2015-05-18 10:10:00       1     1
#> 2015-05-18 10:11:00       1     1
#> 2015-05-18 10:12:00       1     1
#> 2015-05-18 10:13:00       1     1
#> 2015-05-18 10:14:00       1     3
#> 2015-05-18 10:15:00       1     3
#> 2015-05-18 10:16:00       1     3
#> 2015-05-18 10:17:00       1     3
#> 2015-05-18 10:18:00       1     3
#> 2015-05-18 10:19:00       1     3
#> 2015-05-18 10:20:00       1     3
#> 2015-05-18 10:21:00       1     3
#> 2015-05-18 10:22:00       1     3
#> 2015-05-18 10:23:00       1     1
#> 2015-05-18 10:24:00       1     1
#> 2015-05-18 10:25:00       1     1
#> 2015-05-18 10:26:00       1     1
#> 2015-05-18 10:27:00       1     1
#> 2015-05-18 10:28:00       1     1
#> 2015-05-18 10:29:00       1     1
#> 2015-05-18 10:30:00       1     1
#> 2015-05-18 10:31:00       1     1
#> 2015-05-18 10:32:00       1     1
#> 2015-05-18 10:33:00       1     3
#> 2015-05-18 10:34:00       1     3
#> 2015-05-18 10:35:00       1     3
#> 2015-05-18 10:36:00       1     3
#> 2015-05-18 10:37:00       1     3
#> 2015-05-18 10:38:00       1     3
#> 2015-05-18 10:39:00       1     3
#>  [ reached getOption("max.print") -- omitted 1549 rows ]
#> 
#> $TrainingSpecs
#> $SPE_threshold
#>    99.9% 
#> 1.433843 
#> 
#> $T2_threshold
#>   99.9% 
#> 5.30888 
#> 
#> $projectionMatrix
#>            [,1]
#> [1,]  0.5717481
#> [2,] -0.5789897
#> [3,]  0.5812703
#> 
#> $LambdaInv
#>           [,1]
#> [1,] 0.3543623
#> 
#> $muTrain
#>          x          y          z 
#>  0.6045176 -1.3861014  0.9361741 
#> 
#> $RootPrecisTrain
#>          [,1]     [,2]     [,3]
#> [1,] 3.909243 0.000000 0.000000
#> [2,] 0.000000 2.281766 0.000000
#> [3,] 0.000000 0.000000 1.638021
#> 
#> attr(,"class")
#> [1] "threshold" "pca"      
#>