Simulate NOC Observations from a Single-State or Multi-State Process

This function generates data under normal operating conditions from a single-state or multi-state process model.

processNOCdata(startTime = "2015-05-16 10:00:00 CST", period = 7 * 24 * 60,
  stateDuration = 60, increment = "min", multiState = TRUE,
  autocorellation = 0.75, tLower = 0.01, tUpper = 2, errVar = 0.01)

Arguments

startTime
a POSIXct object specifying the day and time for the starting observation.

period
The observation cycle length. Defaults to one week's worth of minute-level observations (10,080 observations).

stateDuration
The number of observations generated during a stay in each state. Defaults to 60.

increment
The time-sequence base increment. See "Deatails" of the seq.POSIXt() function options. Defaults to "min" for minutes.

multiState
Should the observations be generated from a multi-state process? Defaults to TRUE.

autocorellation
The autocorrelation parameter. Must be less than 1 in absolute value, or the process generated will be nonstationary. Defaults to 0.75 in accordance to Kazor et al (2016).

tLower
Lower bound of the latent $t$ variable. Defaults to 0.01.

tUpper
Upper bound of the latent $t$ variable. Defaults to 2.

errVar
Error variance of the normal white noise process on the feature variables.

Value

An data frame with the following information:

dateTime -- A POSIXct column of times starting at the user-defined `startTime` argument, length given by the `period` argument, and spacing given by the `increment` argument. For example, if the starting value is "2016-01-10", period is 10080, and the incrementation is in minutes, then this sequence will be one week's worth of observations recorded every minute from midnight on the tenth of January.

state -- An integer column of all 1's (when the `multiState` argument is FALSE), or a column of the state values (1, 2 or 3).

x -- A double column of generated values for the first feature.

y -- A double column of generated values for the second feature.

z -- A double column of generated values for the third feature.

Details

This function randomly generates a non-stationary (sinusoidal) and autocorrelated latent variable t with lower and upper bounds given by the arguments "tLower" and "tUpper", respectively, with autocorrelation governed by the "autocorrelation" argument. Necessarily, this coefficient must be less than 1 in absolute value, otherwise the latent variable will be unbounded. Next, this function draws a realization of this random variable t and calculates three functions of it, then jitters these functions with a normal white noise variable (with variance set by "errVar"). These three functions are:

x : x(t) = t + error

y : y(t) = t ^ 2 - 3t + error

z : z(t) = -t ^ 3 + 3t ^ 2 + error

This function is called by the mspProcessData() function. See ?mspProcessData for more details.

See also

Called by: mspProcessData.

Examples

processNOCdata()

#>                t          err1          err2          err3            x
#> 1     0.22570129  6.702976e-03  1.061582e-01 -3.103734e-02  0.232404264
#> 2     0.28149380  1.024840e-01  3.223804e-02 -1.316752e-01  0.383977851
#> 3     0.33912850 -1.543232e-03  1.087214e-01 -5.987215e-02  0.337585266
#> 4     0.37536237 -9.002038e-02 -2.361961e-03 -6.757011e-02  0.285341989
#> 5     0.45833879 -1.380380e-02 -5.863574e-02 -6.202566e-02  0.444534991
#> 6     0.37216162  3.337771e-03 -9.917849e-02  1.821219e-02  0.375499391
#> 7     0.33025635 -1.362761e-01  9.672597e-02 -9.434065e-02  0.193980274
#> 8     0.28217841  4.420524e-02  1.092251e-01  2.931087e-02  0.326383654
#> 9     0.20256072 -2.925275e-01 -5.618754e-02  3.396516e-02 -0.089966817
#> 10    0.25854931  2.149170e-01 -3.061290e-02  1.003956e-02  0.473466335
#> 11    0.14842602  4.655183e-02  1.707343e-02 -8.214724e-02  0.194977859
#> 12    0.20092684 -8.003171e-02  3.922016e-02 -4.742575e-02  0.120895129
#> 13    0.13708037 -8.605258e-02  1.049118e-01  1.673372e-03  0.051027791
#> 14    0.23038259 -1.455934e-01 -1.362689e-01  7.643140e-02  0.084789147
#> 15    0.22024289  1.691198e-01  5.239418e-02 -8.281014e-02  0.389362738
#> 16    0.17606746  1.268650e-01  1.428821e-01 -6.126260e-02  0.302932422
#> 17    0.22326827  2.683342e-02 -1.371678e-01  9.747525e-02  0.250101694
#> 18    0.17173892 -1.239826e-01 -6.289146e-02 -1.538434e-01  0.047756281
#> 19    0.20637471 -1.014562e-01 -9.804492e-02 -1.161752e-01  0.104918547
#> 20    0.13807748 -2.111180e-01 -8.244700e-02  8.760703e-02 -0.073040530
#> 21    0.19948211 -1.557714e-01 -1.741793e-02  1.469939e-01  0.043710751
#> 22    0.20286393 -7.183576e-02  1.419511e-02  8.600498e-02  0.131028172
#> 23    0.18525586 -1.479045e-01  3.864334e-02  7.772177e-02  0.037351377
#> 24    0.19322139  1.290169e-01  4.271285e-02 -2.615263e-01  0.322238323
#> 25    0.23328800 -2.279736e-02 -2.766794e-02  1.654279e-02  0.210490634
#> 26    0.37934947  5.479328e-02 -1.229387e-01 -1.792572e-01  0.434142751
#> 27    0.22900671  3.005670e-02  1.309247e-01 -1.066287e-01  0.259063415
#> 28    0.31823126 -1.475994e-01  5.972637e-02  1.108183e-01  0.170631877
#> 29    0.35106227  1.530585e-01  6.572975e-02 -4.860341e-02  0.504120727
#> 30    0.26251698  1.380289e-01 -4.689652e-02  4.422089e-02  0.400545846
#> 31    0.30105787  5.618629e-02 -9.678772e-03 -2.758692e-02  0.357244161
#> 32    0.32636723 -4.749511e-02  1.067795e-01  1.082112e-02  0.278872121
#> 33    0.36825856 -1.474043e-01 -2.981582e-01 -6.549973e-02  0.220854222
#> 34    0.32291759 -1.481244e-01 -9.070439e-02  6.803034e-02  0.174793178
#> 35    0.31487133 -7.036240e-02  5.358693e-02 -8.875742e-02  0.244508930
#> 36    0.29824501  4.679547e-03 -6.019614e-02 -1.265899e-01  0.302924557
#> 37    0.31084334 -1.210225e-01 -1.187612e-01 -1.127031e-01  0.189820872
#> 38    0.26597388  1.136685e-01 -1.621560e-01  2.217745e-02  0.379642383
#> 39    0.30093690  2.935565e-02 -1.540554e-01  4.913948e-03  0.330292551
#> 40    0.29774802 -3.324486e-02  7.195904e-02  1.157212e-01  0.264503157
#> 41    0.30792509  4.865981e-02  5.355443e-02  8.356151e-03  0.356584900
#> 42    0.27664991  1.845353e-01  1.667044e-01 -8.398879e-02  0.461185165
#> 43    0.17328230 -7.052918e-02  3.521225e-02  7.541417e-02  0.102753125
#> 44    0.31691785  2.132812e-01 -8.658573e-03  2.003137e-01  0.530199037
#> 45    0.36803237 -3.314154e-02 -1.174995e-01  6.454182e-02  0.334890833
#> 46    0.26825053 -5.138265e-03 -1.149051e-01  3.778523e-02  0.263112265
#> 47    0.24481552  2.789078e-02 -4.608391e-03 -1.525382e-01  0.272706296
#> 48    0.27039423  1.052499e-01 -8.730177e-03  1.230339e-01  0.375644102
#> 49    0.34898739  1.428090e-01 -5.930845e-02 -1.089064e-01  0.491796401
#> 50    0.30013794  1.197540e-02 -4.488650e-02  1.955720e-02  0.312113334
#> 51    0.34473444 -8.644677e-02 -9.456488e-03 -7.964398e-02  0.258287671
#> 52    0.37126959 -2.932511e-02  6.043181e-02  1.430998e-02  0.341944482
#> 53    0.35413268 -1.813679e-02  1.059953e-01  5.914000e-02  0.335995890
#> 54    0.30833108 -2.051319e-01  1.100771e-01  3.138313e-02  0.103199173
#> 55    0.23982948  7.491147e-02  8.154037e-02  8.019514e-02  0.314740953
#> 56    0.25797077  5.480557e-02  3.724855e-02 -1.714271e-01  0.312776335
#> 57    0.26276185 -2.165362e-01 -2.231583e-01  8.853430e-02  0.046225617
#> 58    0.28924338 -7.313374e-02  3.308791e-02  9.060663e-02  0.216109648
#> 59    0.29119409  1.212474e-01  1.296888e-01 -6.394147e-03  0.412441533
#> 60    0.27378423  5.237076e-02 -2.187951e-02 -4.871421e-02  0.326154989
#> 61    0.24825212 -1.369712e-01 -3.190105e-02  3.536546e-02  0.111280969
#> 62    0.32053768  6.552318e-02  1.227183e-01 -1.061930e-02  0.386060860
#> 63    0.30789923 -7.127866e-02 -5.304142e-02  1.950347e-01  0.236620575
#> 64    0.27822759  1.283213e-01 -1.146955e-01 -4.182664e-02  0.406548847
#> 65    0.28591879 -8.659067e-02  3.827017e-02  8.835309e-02  0.199328116
#> 66    0.31145147  4.204219e-02  4.082372e-02  2.627094e-02  0.353493670
#> 67    0.32722139  2.152943e-03  3.489398e-02 -2.953446e-02  0.329374336
#> 68    0.38868133  3.361757e-02 -3.263914e-02 -5.968841e-02  0.422298900
#> 69    0.40411348  4.889004e-02 -5.089254e-02  1.408715e-01  0.453003524
#> 70    0.34044825 -1.991237e-01  6.115724e-02 -2.670204e-02  0.141324561
#> 71    0.32290480  5.119010e-02 -6.554219e-02 -4.512614e-02  0.374094897
#> 72    0.26291647  8.136015e-02  8.216075e-02 -1.871398e-02  0.344276618
#> 73    0.22559320  8.173226e-02  1.681556e-02  2.502781e-02  0.307325458
#> 74    0.16365475  1.648885e-01 -1.321555e-01 -3.907820e-02  0.328543271
#> 75    0.23250312 -1.293354e-02 -1.911161e-02 -1.769186e-01  0.219569583
#> 76    0.30994628 -2.127116e-01 -2.377097e-02 -9.748947e-02  0.097234669
#> 77    0.33976192 -3.345485e-02  8.361867e-02 -6.163267e-02  0.306307073
#> 78    0.32231233 -1.759856e-02  9.815332e-02  1.712814e-01  0.304713771
#> 79    0.28596449  4.186086e-02 -9.651246e-02  9.261509e-02  0.327825356
#> 80    0.24037788  6.670348e-02  1.369443e-01  9.371564e-02  0.307081355
#> 81    0.24976188 -8.250228e-03  5.712890e-02 -1.022722e-01  0.241511656
#> 82    0.27440686  1.353837e-01 -4.101318e-02 -1.333935e-02  0.409790552
#> 83    0.27954925  5.339079e-02 -2.510316e-02 -1.498502e-01  0.332940046
#> 84    0.21472795 -1.963714e-01 -1.189776e-01  4.499106e-02  0.018356525
#> 85    0.18543221  5.235935e-02  1.465643e-01  3.987165e-02  0.237791561
#> 86    0.21452758  2.235878e-01 -1.079256e-02 -6.548448e-02  0.438115418
#> 87    0.23138353  8.261921e-03 -8.452030e-02 -7.821172e-02  0.239645456
#> 88    0.32476693  7.910202e-02 -3.891779e-02 -6.267742e-02  0.403868945
#> 89    0.33191422 -4.708716e-02  3.605494e-02  3.692636e-02  0.284827060
#> 90    0.26764899 -2.414888e-02  7.010209e-02  4.540773e-02  0.243500107
#> 91    0.31904006 -8.821381e-02 -1.474701e-02  4.483372e-02  0.230826249
#> 92    0.29629521 -1.004360e-01  6.964468e-02 -8.825033e-02  0.195859192
#> 93    0.21321740  6.019970e-02  4.533152e-02  9.466978e-02  0.273417096
#> 94    0.19751944  3.154009e-02 -2.454536e-01 -7.651278e-02  0.229059530
#> 95    0.18820816 -9.963436e-02 -6.820786e-02  2.686191e-02  0.088573794
#> 96    0.29486459  1.301998e-01  2.691855e-01  7.803404e-02  0.425064409
#> 97    0.31508957  7.460007e-02  7.791234e-02 -5.778725e-02  0.389689644
#> 98    0.38958135 -1.607004e-01  9.782096e-02  3.829372e-02  0.228880969
#> 99    0.37949707  6.462786e-02 -4.283167e-02  1.129872e-01  0.444124932
#> 100   0.33211362 -1.356036e-01 -1.967116e-02  1.129085e-01  0.196510020
#> 101   0.32739618  1.045213e-01 -7.887211e-02  9.291536e-03  0.431917445
#> 102   0.29286577  7.282447e-02  4.802730e-03 -3.009130e-02  0.365690233
#> 103   0.29291572  1.432610e-01 -8.787822e-02  1.996734e-03  0.436176765
#> 104   0.28286133 -4.786951e-02 -3.474327e-03 -6.679191e-02  0.234991819
#> 105   0.27484278 -1.526134e-01 -7.749768e-02 -7.460880e-02  0.122229377
#> 106   0.35848429  6.970272e-02  1.473852e-01 -7.567744e-02  0.428187009
#> 107   0.40742627 -6.218961e-02 -3.843869e-02 -3.416347e-02  0.345236661
#> 108   0.29056566  5.386609e-02  3.586027e-02  3.255916e-02  0.344431745
#> 109   0.20515528  3.235110e-02 -1.177690e-01 -5.417021e-02  0.237506376
#> 110   0.29514094 -4.212599e-02  5.278680e-02  2.774048e-01  0.253014945
#> 111   0.34020172 -8.884173e-02  6.159708e-02 -3.404763e-02  0.251359989
#>                y             z            dateTime state
#> 1     -0.5200046  0.1102884036 2015-05-16 10:00:00     1
#> 2     -0.7330046  0.0837358281 2015-05-16 10:01:00     1
#> 3     -0.7936559  0.2461497319 2015-05-16 10:02:00     1
#> 4     -0.9875522  0.3022332263 2015-05-16 10:03:00     1
#> 5     -1.2235777  0.4719124064 2015-05-16 10:04:00     1
#> 6     -1.0771591  0.3821790323 2015-05-16 10:05:00     1
#> 7     -0.7849738  0.1968463052 2015-05-16 10:06:00     1
#> 8     -0.6576854  0.2457164791 2015-05-16 10:07:00     1
#> 9     -0.6228388  0.1487464566 2015-05-16 10:08:00     1
#> 10    -0.7394131  0.1932993536 2015-05-16 10:09:00     1
#> 11    -0.4061744 -0.0193262498 2015-05-16 10:10:00     1
#> 12    -0.5231888  0.0655772893 2015-05-16 10:11:00     1
#> 13    -0.2875382  0.0554705724 2015-05-16 10:12:00     1
#> 14    -0.7743405  0.2234319972 2015-05-16 10:13:00     1
#> 15    -0.5598276  0.0520273461 2015-05-16 10:14:00     1
#> 16    -0.3543205  0.0262785979 2015-05-16 10:15:00     1
#> 17    -0.7571239  0.2358917678 2015-05-16 10:16:00     1
#> 18    -0.5486140 -0.0704259316 2015-05-16 10:17:00     1
#> 19    -0.6745785  0.0028067563 2015-05-16 10:18:00     1
#> 20    -0.4776141  0.1421707023 2015-05-16 10:19:00     1
#> 21    -0.5760711  0.2584352439 2015-05-16 10:20:00     1
#> 22    -0.5532429  0.2011176777 2015-05-16 10:21:00     1
#> 23    -0.4828045  0.1743230382 2015-05-16 10:22:00     1
#> 24    -0.4996168 -0.1567366056 2015-05-16 10:23:00     1
#> 25    -0.6731086  0.1671163582 2015-05-16 10:24:00     1
#> 26    -1.1170811  0.1978701708 2015-05-16 10:25:00     1
#> 27    -0.5036513  0.0386934725 2015-05-16 10:26:00     1
#> 28    -0.7936963  0.3824040832 2015-05-16 10:27:00     1
#> 29    -0.8642123  0.2778641713 2015-05-16 10:28:00     1
#> 30    -0.7655323  0.2328749872 2015-05-16 10:29:00     1
#> 31    -0.8222165  0.2170339734 2015-05-16 10:30:00     1
#> 32    -0.7658066  0.2956046471 2015-05-16 10:31:00     1
#> 33    -1.2673196  0.2914022222 2015-05-16 10:32:00     1
#> 34    -0.9551814  0.3471851705 2015-05-16 10:33:00     1
#> 35    -0.7918831  0.1774568524 2015-05-16 10:34:00     1
#> 36    -0.8659811  0.1137314611 2015-05-16 10:35:00     1
#> 37    -0.9546677  0.1471328122 2015-05-16 10:36:00     1
#> 38    -0.8893355  0.2155882034 2015-05-16 10:37:00     1
#> 39    -0.9663031  0.2493492436 2015-05-16 10:38:00     1
#> 40    -0.7326311  0.3552863303 2015-05-16 10:39:00     1
#> 41    -0.7754030  0.2636129352 2015-05-16 10:40:00     1
#> 42    -0.5867102  0.1244432814 2015-05-16 10:41:00     1
#> 43    -0.4546079  0.1602913328 2015-05-16 10:42:00     1
#> 44    -0.8589752  0.4697941705 2015-05-16 10:43:00     1
#> 45    -1.0861488  0.4210361185 2015-05-16 10:44:00     1
#> 46    -0.8476983  0.2343574027 2015-05-16 10:45:00     1
#> 47    -0.6791203  0.0125928078 2015-05-16 10:46:00     1
#> 48    -0.7467998  0.3226036258 2015-05-16 10:47:00     1
#> 49    -0.9844784  0.2139662553 2015-05-16 10:48:00     1
#> 50    -0.8552175  0.2627682805 2015-05-16 10:49:00     1
#> 51    -0.9248180  0.2359126530 2015-05-16 10:50:00     1
#> 52    -0.9155358  0.3766570885 2015-05-16 10:51:00     1
#> 53    -0.8309927  0.3909580934 2015-05-16 10:52:00     1
#> 54    -0.7198481  0.2872748575 2015-05-16 10:53:00     1
#> 55    -0.5804299  0.2389551179 2015-05-16 10:54:00     1
#> 56    -0.6701148  0.0110520154 2015-05-16 10:55:00     1
#> 57    -0.9424001  0.2775235911 2015-05-16 10:56:00     1
#> 58    -0.7509805  0.3173932340 2015-05-16 10:57:00     1
#> 59    -0.6590995  0.2232963390 2015-05-16 10:58:00     1
#> 60    -0.7682744  0.1556369388 2015-05-16 10:59:00     1
#> 61    -0.7150283  0.2049532552 2015-05-16 11:00:00     2
#> 62    -0.7361503  0.2646804677 2015-05-16 11:01:00     2
#> 63    -0.8819372  0.4502511074 2015-05-16 11:02:00     2
#> 64    -0.8719677  0.1688673752 2015-05-16 11:03:00     2
#> 65    -0.7377366  0.3102280154 2015-05-16 11:04:00     2
#> 66    -0.7965287  0.2870655779 2015-05-16 11:05:00     2
#> 67    -0.8396964  0.2566502114 2015-05-16 11:06:00     2
#> 68    -1.0476100  0.3348118017 2015-05-16 11:07:00     2
#> 69    -1.0999253  0.5647997519 2015-05-16 11:08:00     2
#> 70    -0.8442825  0.2815533414 2015-05-16 11:09:00     2
#> 71    -0.9299891  0.2340079046 2015-05-16 11:10:00     2
#> 72    -0.6374636  0.1704871057 2015-05-16 11:11:00     2
#> 73    -0.6090717  0.1662237295 2015-05-16 11:12:00     2
#> 74    -0.5963369  0.0368872869 2015-05-16 11:13:00     2
#> 75    -0.6625633 -0.0273141066 2015-05-16 11:14:00     2
#> 76    -0.8575431  0.1609351148 2015-05-16 11:15:00     2
#> 77    -0.8202289  0.2454603295 2015-05-16 11:16:00     2
#> 78    -0.7648984  0.4494536307 2015-05-16 11:17:00     2
#> 79    -0.8726302  0.3145572176 2015-05-16 11:18:00     2
#> 80    -0.5264078  0.2531708117 2015-05-16 11:19:00     2
#> 81    -0.6297758  0.0692904346 2015-05-16 11:20:00     2
#> 82    -0.7889346  0.1918954208 2015-05-16 11:21:00     2
#> 83    -0.7856031  0.0627470019 2015-05-16 11:22:00     2
#> 84    -0.7170534  0.1734146424 2015-05-16 11:23:00     2
#> 85    -0.3753472  0.1366508580 2015-05-16 11:24:00     2
#> 86    -0.6083532  0.0627087635 2015-05-16 11:25:00     2
#> 87    -0.7251326  0.0700154114 2015-05-16 11:26:00     2
#> 88    -0.9077450  0.2194889251 2015-05-16 11:27:00     2
#> 89    -0.8495207  0.3308615066 2015-05-16 11:28:00     2
#> 90    -0.6612089  0.2411423695 2015-05-16 11:29:00     2
#> 91    -0.8700806  0.3177194169 2015-05-16 11:30:00     2
#> 92    -0.7314501  0.1491102169 2015-05-16 11:31:00     2
#> 93    -0.5488590  0.2213615392 2015-05-16 11:32:00     2
#> 94    -0.7989980  0.0328229945 2015-05-16 11:33:00     2
#> 95    -0.5974100  0.1264620776 2015-05-16 11:34:00     2
#> 96    -0.5284631  0.3132323736 2015-05-16 11:35:00     2
#> 97    -0.7680749  0.2087745191 2015-05-16 11:36:00     2
#> 98    -0.9191495  0.4344864253 2015-05-16 11:37:00     2
#> 99    -1.0373049  0.4903868927 2015-05-16 11:38:00     2
#> 100   -0.9057126  0.4071749661 2015-05-16 11:39:00     2
#> 101   -0.9538724  0.2957632804 2015-05-16 11:40:00     2
#> 102   -0.7880242  0.2021005764 2015-05-16 11:41:00     2
#> 103   -0.8808258  0.2342635321 2015-05-16 11:42:00     2
#> 104   -0.7720478  0.1506078020 2015-05-16 11:43:00     2
#> 105   -0.8264875  0.1312456281 2015-05-16 11:44:00     2
#> 106   -0.7995567  0.2637863552 2015-05-16 11:45:00     2
#> 107   -1.0947213  0.3961938259 2015-05-16 11:46:00     2
#> 108   -0.7514083  0.2613123745 2015-05-16 11:47:00     2
#> 109   -0.6911462  0.0634611337 2015-05-16 11:48:00     2
#> 110   -0.7455278  0.5130201309 2015-05-16 11:49:00     2
#> 111   -0.8432709  0.2737899993 2015-05-16 11:50:00     2
#>  [ reached getOption("max.print") -- omitted 9969 rows ]