Three-Dimensional Rotation and Scaling Matrix

Render a 3-Dimensional projection matrix given positive or negative degree changes in yaw, pitch, and / or roll and increment or decrement feature scales.

rotateScale3D(rot_angles = c(0, 0, 0), scale_factors = c(1, 1, 1))

Arguments

rot_angles

a list or vector containg the rotation angles in the order following: yaw, pitch, roll. Defaults to <0,0,0>.

scale_factors

a list or vector containing the values by which to multiply each dimension. Defaults to <1,1,1>.

Value

A 3 x 3 projection matrix corresponding to the degree and scale changes entered.

Details

See the help file of function rotate_3D() for a brief explination of how these angles behave in scatterplot3d functionality (from package scatterplot3d).

This function is used only in data generation in the package vignette (version 1) and the dataStateSwitch() function within the mspProcessData() function. This function calls rotate3D().

See also

Examples

rotateScale3D(rot_angles = list(yaw = -10, pitch = 0, roll = 15),
              scale_factors = c(0.2, 1, 5))
#>            [,1]       [,2]       [,3]
#> [1,] 0.19696155 -0.1677313 -0.2247173
#> [2,] 0.03472964  0.9512512  1.2744350
#> [3,] 0.00000000 -0.2588190  4.8296291