Process Fault Filtering

Flag and filter out observations beyond normal operating conditions, then return the observations within normal operating conditions.

faultFilter(trainData, testData, updateFreq, faultsToTriggerAlarm = 5, ...)

Arguments

trainData
An xts data matrix of initial training observations

testData
The data not included in the training data set

updateFreq
The number of observations from the test data matrix that must be returned to update the training data matrix and move it forward.

faultsToTriggerAlarm
Specifies how many sequential faults will cause an alarm to trigger. Defaults to 5.

...
Lazy dots for additional internal arguments

Value

A list of class "fault_ls" with the following:

faultObj -- An xts flagging matrix with the same number of rows as "testData". This flag matrix has the following five columns:

SPE -- The SPE statistic value for each observation in "testData". This statistic is defined as $$ SPE_i = (\textbf{X}_i - \textbf{Y}_i * \textbf{P}^T) * (\textbf{X}_i - \textbf{Y}_i * \textbf{P}^T)^T, $$ where \(\textbf{X}_i\) is the \(i^{th}\) observation vector, \(\textbf{Y}_i\) is the reduced-feature projection of the observation \(\textbf{X}_i\), and \(\textbf{P}\) is the projection matrix such that \(\textbf{X}_i\textbf{P} = \textbf{Y}_i\).

SPE_Flag -- A vector of SPE indicators recording 0 if the test statistic is less than or equal to the critical value passed through from the threshold object.

T2 -- The T2 statistic value for each observation in "testData". This statistic is defined as $$ T^2_i = \textbf{Y}_i * \textbf{D}^{-1} * \textbf{Y}_i^T, $$ where \(\textbf{Y}_i = \textbf{X}_i\textbf{P}\) is the reduced- feature projection of the observation \(\textbf{X}_i\), and \(\textbf{D}\) is the diagonal matrix of eigenvalues.

T2_Flag -- A vector of T2 fault indicators, defined like SPE_Flag.

Alarm -- A column indicating if there have been five flags in a row for either the SPE or T2 monitoring statistics or both. Alarm states are as follows: 0 = no alarm, 1 = Hotelling's T2 alarm, 2 = Squared Prediction Error alarm, and 3 = both alarms.

nonAlarmedTestObs -- An xts matrix of the first updateFreq number of rows of the training data which were not alarmed.

trainSpecs -- The threshold object returned by the internal threshold() function. See the threshold() function's help file for more details.

Details

This function is essentially a wrapper function to call and organize the output from these other internal functions: faultDetect(), threshold(), and pca(). It is applied over a rolling window, with observation width equal to updateFreq, of the larger full data matrix via the processMonitor() function, wherein the testing and training data sets move forward in time across the entire data matrix.

This internal function is called by processMonitor().

See also

Calls: pca, threshold, faultDetect. Called by: processMonitor.

Examples

nrml <- mspProcessData(faults = "NOC")
# Select the data under state 1
data <- nrml[nrml[,1] == 1]

faultFilter(trainData = data[1:672, -1],
            testData = data[673:3360, -1],
            updateFreq = 336)

#> $faultObj
#>                              SPE SPE_Flag           T2 T2_Flag Alarm
#> 2015-05-17 19:12:00 1.136624e+00        0 5.430296e+00       0     0
#> 2015-05-17 19:13:00 1.739725e+00        0 5.014730e+00       0     0
#> 2015-05-17 19:14:00 3.999711e-01        0 3.357003e+00       0     0
#> 2015-05-17 19:15:00 8.232836e-02        0 3.704983e+00       0     0
#> 2015-05-17 19:16:00 6.170623e-01        0 6.289844e+00       0     0
#> 2015-05-17 19:17:00 6.340335e-01        0 3.338330e+00       0     0
#> 2015-05-17 19:18:00 5.788635e-02        0 2.682427e+00       0     0
#> 2015-05-17 19:19:00 2.288376e-01        0 2.126589e+00       0     0
#> 2015-05-17 19:20:00 5.782369e-01        0 9.354648e-01       0     0
#> 2015-05-17 19:21:00 8.984972e-03        0 1.435090e+00       0     0
#> 2015-05-17 19:22:00 9.714080e-01        0 2.643954e+00       0     0
#> 2015-05-17 19:23:00 3.092313e-01        0 4.108731e+00       0     0
#> 2015-05-17 19:24:00 1.089499e-02        0 7.339955e+00       1     0
#> 2015-05-17 19:25:00 2.804739e-02        0 4.220728e+00       0     0
#> 2015-05-17 19:26:00 1.629388e-01        0 4.211270e+00       0     0
#> 2015-05-17 19:27:00 2.027933e-02        0 1.102558e+00       0     0
#> 2015-05-17 19:28:00 6.324720e-01        0 1.144948e+00       0     0
#> 2015-05-17 19:29:00 4.800746e-01        0 7.090531e-01       0     0
#> 2015-05-17 19:30:00 6.875443e-02        0 1.180586e+00       0     0
#> 2015-05-17 19:31:00 3.761577e-01        0 1.770778e+00       0     0
#> 2015-05-17 19:32:00 1.960025e-01        0 8.802436e-01       0     0
#> 2015-05-17 19:33:00 2.779795e-02        0 2.881728e+00       0     0
#> 2015-05-17 19:34:00 4.432185e-01        0 3.191795e+00       0     0
#> 2015-05-17 19:35:00 6.612049e-01        0 3.630081e+00       0     0
#> 2015-05-17 19:36:00 9.479429e-02        0 2.577928e+00       0     0
#> 2015-05-17 19:37:00 1.242151e+00        0 2.750710e+00       0     0
#> 2015-05-17 19:38:00 1.311500e-01        0 2.205578e+00       0     0
#> 2015-05-17 19:39:00 8.512231e-02        0 3.974765e+00       0     0
#> 2015-05-17 19:40:00 6.116365e-01        0 4.871786e+00       0     0
#> 2015-05-17 19:41:00 2.723357e-01        0 6.104688e+00       0     0
#> 2015-05-17 19:42:00 1.451519e-02        0 5.292736e+00       0     0
#> 2015-05-17 19:43:00 1.236179e+00        0 4.287181e+00       0     0
#> 2015-05-17 19:44:00 7.002865e-01        0 8.093228e+00       1     0
#> 2015-05-17 19:45:00 4.445335e-01        0 1.039150e+01       1     0
#> 2015-05-17 19:46:00 7.027584e-01        0 4.809936e+00       0     0
#> 2015-05-17 19:47:00 9.229196e-01        0 4.889790e+00       0     0
#> 2015-05-17 19:48:00 2.169066e-01        0 6.801263e+00       1     0
#> 2015-05-17 19:49:00 7.191706e-01        0 3.879650e+00       0     0
#> 2015-05-17 19:50:00 4.027967e-01        0 4.716358e+00       0     0
#> 2015-05-17 19:51:00 7.854238e-01        0 3.171250e+00       0     0
#> 2015-05-17 19:52:00 1.453294e-01        0 2.569049e+00       0     0
#> 2015-05-17 19:53:00 3.126835e-01        0 2.367290e+00       0     0
#> 2015-05-17 19:54:00 2.085796e-01        0 2.684751e+00       0     0
#> 2015-05-17 19:55:00 9.073206e-01        0 3.159220e+00       0     0
#> 2015-05-17 19:56:00 5.317495e-01        0 4.042438e+00       0     0
#> 2015-05-17 19:57:00 1.272198e-01        0 6.151828e+00       0     0
#> 2015-05-17 19:58:00 1.524363e-01        0 3.330363e+00       0     0
#> 2015-05-17 19:59:00 6.997605e-01        0 3.836996e+00       0     0
#> 2015-05-17 22:00:00 1.263405e+00        0 2.964566e+00       0     0
#> 2015-05-17 22:01:00 2.249434e-01        0 4.359498e+00       0     0
#> 2015-05-17 22:02:00 4.441618e-01        0 3.143505e+00       0     0
#> 2015-05-17 22:03:00 2.466014e-02        0 1.081318e+00       0     0
#> 2015-05-17 22:04:00 2.311593e-01        0 3.263139e+00       0     0
#> 2015-05-17 22:05:00 2.074836e-01        0 5.832577e+00       0     0
#> 2015-05-17 22:06:00 1.164732e+00        0 4.977363e+00       0     0
#> 2015-05-17 22:07:00 5.409492e-01        0 1.738685e+00       0     0
#> 2015-05-17 22:08:00 5.730048e-01        0 2.889925e+00       0     0
#> 2015-05-17 22:09:00 3.379348e-01        0 2.817233e+00       0     0
#> 2015-05-17 22:10:00 2.803785e-01        0 3.539530e+00       0     0
#> 2015-05-17 22:11:00 3.063330e-02        0 2.489966e+00       0     0
#> 2015-05-17 22:12:00 9.430350e-02        0 5.863724e-01       0     0
#> 2015-05-17 22:13:00 6.503451e-01        0 2.992301e+00       0     0
#> 2015-05-17 22:14:00 6.464723e-01        0 4.437344e+00       0     0
#> 2015-05-17 22:15:00 4.901242e-01        0 2.576431e+00       0     0
#> 2015-05-17 22:16:00 5.558165e-01        0 2.087707e+00       0     0
#> 2015-05-17 22:17:00 2.252517e-01        0 2.801654e+00       0     0
#> 2015-05-17 22:18:00 9.458848e-02        0 3.340948e+00       0     0
#> 2015-05-17 22:19:00 4.693359e-02        0 5.862461e+00       0     0
#> 2015-05-17 22:20:00 2.269112e+00        0 2.919750e+00       0     0
#> 2015-05-17 22:21:00 5.523546e-01        0 5.760525e+00       0     0
#> 2015-05-17 22:22:00 9.444973e-02        0 4.913704e+00       0     0
#> 2015-05-17 22:23:00 2.415017e-01        0 3.268056e+00       0     0
#> 2015-05-17 22:24:00 5.501501e-02        0 3.435803e+00       0     0
#> 2015-05-17 22:25:00 1.224973e-02        0 3.757459e+00       0     0
#> 2015-05-17 22:26:00 5.086943e-01        0 6.242489e+00       0     0
#> 2015-05-17 22:27:00 8.727750e-01        0 6.362021e+00       0     0
#> 2015-05-17 22:28:00 1.181985e+00        0 7.130414e+00       1     0
#> 2015-05-17 22:29:00 2.898835e-01        0 7.056053e+00       1     0
#> 2015-05-17 22:30:00 9.781755e-01        0 7.176734e+00       1     0
#> 2015-05-17 22:31:00 1.094178e+00        0 6.781231e+00       1     0
#> 2015-05-17 22:32:00 5.633325e-01        0 6.231041e+00       0     0
#> 2015-05-17 22:33:00 1.440618e+00        0 5.281827e+00       0     0
#> 2015-05-17 22:34:00 1.760920e+00        0 8.935405e+00       1     0
#> 2015-05-17 22:35:00 2.282591e+00        0 1.376493e+01       1     0
#> 2015-05-17 22:36:00 1.288028e+00        0 1.263673e+01       1     0
#> 2015-05-17 22:37:00 5.788482e-01        0 1.166600e+01       1     0
#> 2015-05-17 22:38:00 1.945189e+00        0 1.435003e+01       1     1
#> 2015-05-17 22:39:00 1.544493e+00        0 1.202989e+01       1     1
#> 2015-05-17 22:40:00 1.533192e-01        0 7.718389e+00       1     1
#> 2015-05-17 22:41:00 2.823898e-01        0 1.086693e+01       1     1
#> 2015-05-17 22:42:00 1.286445e+00        0 8.500627e+00       1     1
#> 2015-05-17 22:43:00 1.732227e-01        0 9.568106e+00       1     1
#> 2015-05-17 22:44:00 1.933068e-01        0 4.259123e+00       0     0
#> 2015-05-17 22:45:00 1.629325e+00        0 1.131114e+01       1     0
#> 2015-05-17 22:46:00 2.051987e-01        0 6.454029e+00       0     0
#> 2015-05-17 22:47:00 1.517123e+00        0 1.129397e+01       1     0
#> 2015-05-17 22:48:00 1.024681e+00        0 9.632094e+00       1     0
#> 2015-05-17 22:49:00 7.287300e-01        0 1.642523e+01       1     0
#> 2015-05-17 22:50:00 1.483273e+00        0 1.107277e+01       1     0
#> 2015-05-17 22:51:00 1.744833e-02        0 6.200694e+00       0     0
#> 2015-05-17 22:52:00 1.258314e+00        0 7.044384e+00       1     0
#> 2015-05-17 22:53:00 1.046145e+00        0 4.538185e+00       0     0
#> 2015-05-17 22:54:00 1.838041e+00        0 4.358803e+00       0     0
#> 2015-05-17 22:55:00 1.861082e+00        0 7.123409e+00       1     0
#> 2015-05-17 22:56:00 1.787293e+00        0 1.109194e+01       1     0
#> 2015-05-17 22:57:00 1.744946e+00        0 1.136434e+01       1     0
#> 2015-05-17 22:58:00 1.502068e+00        0 9.905690e+00       1     0
#> 2015-05-17 22:59:00 1.952463e+00        0 1.172988e+01       1     1
#> 2015-05-18 01:00:00 1.849732e+00        0 7.782592e+00       1     1
#> 2015-05-18 01:01:00 1.301640e+00        0 9.425392e+00       1     1
#> 2015-05-18 01:02:00 2.907138e+00        1 9.028105e+00       1     1
#> 2015-05-18 01:03:00 2.259051e+00        0 1.420117e+01       1     1
#> 2015-05-18 01:04:00 1.973610e+00        0 1.513986e+01       1     1
#> 2015-05-18 01:05:00 1.570406e+00        0 1.551188e+01       1     1
#> 2015-05-18 01:06:00 9.812105e-01        0 1.225512e+01       1     1
#> 2015-05-18 01:07:00 7.567127e-01        0 1.109659e+01       1     1
#> 2015-05-18 01:08:00 8.290844e-01        0 1.057058e+01       1     1
#> 2015-05-18 01:09:00 7.144063e-01        0 1.065138e+01       1     1
#> 2015-05-18 01:10:00 8.319846e-01        0 6.823785e+00       1     1
#> 2015-05-18 01:11:00 7.681600e-01        0 7.369350e+00       1     1
#> 2015-05-18 01:12:00 1.618429e-01        0 7.904642e+00       1     1
#> 2015-05-18 01:13:00 2.658899e-01        0 7.661116e+00       1     1
#> 2015-05-18 01:14:00 1.123964e+00        0 7.782882e+00       1     1
#> 2015-05-18 01:15:00 1.082590e+00        0 7.457145e+00       1     1
#> 2015-05-18 01:16:00 1.417591e+00        0 1.080617e+01       1     1
#> 2015-05-18 01:17:00 7.606840e-01        0 1.320800e+01       1     1
#> 2015-05-18 01:18:00 3.064139e-01        0 8.010453e+00       1     1
#> 2015-05-18 01:19:00 1.743450e+00        0 6.207429e+00       0     0
#> 2015-05-18 01:20:00 6.454836e-01        0 6.014374e+00       0     0
#> 2015-05-18 01:21:00 1.003871e+00        0 7.495722e+00       1     0
#> 2015-05-18 01:22:00 5.502216e-02        0 5.196881e+00       0     0
#> 2015-05-18 01:23:00 5.510193e-01        0 6.245578e+00       0     0
#> 2015-05-18 01:24:00 6.108979e-01        0 8.008004e+00       1     0
#> 2015-05-18 01:25:00 9.977354e-01        0 8.344467e+00       1     0
#> 2015-05-18 01:26:00 8.825355e-01        0 7.327729e+00       1     0
#> 2015-05-18 01:27:00 1.127416e+00        0 8.129517e+00       1     0
#> 2015-05-18 01:28:00 1.072350e+00        0 1.160122e+01       1     1
#> 2015-05-18 01:29:00 9.888973e-01        0 9.814039e+00       1     1
#> 2015-05-18 01:30:00 1.414259e+00        0 1.085142e+01       1     1
#> 2015-05-18 01:31:00 1.590638e+00        0 1.404599e+01       1     1
#> 2015-05-18 01:32:00 1.384604e+00        0 1.398304e+01       1     1
#> 2015-05-18 01:33:00 1.358609e+00        0 1.033508e+01       1     1
#> 2015-05-18 01:34:00 1.612998e+00        0 1.034035e+01       1     1
#> 2015-05-18 01:35:00 1.455697e+00        0 1.123140e+01       1     1
#> 2015-05-18 01:36:00 1.286267e+00        0 8.421983e+00       1     1
#> 2015-05-18 01:37:00 9.564644e-01        0 9.901347e+00       1     1
#> 2015-05-18 01:38:00 1.847074e+00        0 1.223227e+01       1     1
#> 2015-05-18 01:39:00 1.017560e+00        0 1.613202e+01       1     1
#> 2015-05-18 01:40:00 9.132458e-01        0 9.211775e+00       1     1
#> 2015-05-18 01:41:00 1.417306e+00        0 8.218920e+00       1     1
#> 2015-05-18 01:42:00 2.418964e+00        1 1.024464e+01       1     1
#> 2015-05-18 01:43:00 7.783194e-01        0 1.122777e+01       1     1
#> 2015-05-18 01:44:00 1.502353e+00        0 8.546395e+00       1     1
#> 2015-05-18 01:45:00 3.028257e+00        1 1.056281e+01       1     1
#> 2015-05-18 01:46:00 4.725019e-01        0 8.005538e+00       1     1
#> 2015-05-18 01:47:00 1.880773e+00        0 1.014751e+01       1     1
#> 2015-05-18 01:48:00 1.213983e-01        0 8.898345e+00       1     1
#> 2015-05-18 01:49:00 3.171854e-01        0 5.421781e+00       0     0
#> 2015-05-18 01:50:00 2.062751e+00        0 1.149658e+01       1     0
#> 2015-05-18 01:51:00 1.729173e+00        0 6.529531e+00       0     0
#> 2015-05-18 01:52:00 1.000612e+00        0 7.513107e+00       1     0
#> 2015-05-18 01:53:00 2.676316e+00        1 1.704936e+01       1     0
#> 2015-05-18 01:54:00 2.172680e+00        0 1.444542e+01       1     0
#> 2015-05-18 01:55:00 5.678218e+00        1 1.194139e+01       1     0
#> 2015-05-18 01:56:00 1.557724e+00        0 1.085607e+01       1     1
#> 2015-05-18 01:57:00 3.066841e+00        1 1.978056e+01       1     1
#> 2015-05-18 01:58:00 5.496193e-01        0 1.521407e+01       1     1
#> 2015-05-18 01:59:00 2.659947e+00        1 1.482183e+01       1     1
#> 2015-05-18 04:00:00 8.199025e-01        0 1.349003e+01       1     1
#> 2015-05-18 04:01:00 8.252747e-01        0 1.306108e+01       1     1
#> 2015-05-18 04:02:00 2.273405e+00        0 1.108535e+01       1     1
#> 2015-05-18 04:03:00 1.408997e+00        0 1.024602e+01       1     1
#> 2015-05-18 04:04:00 2.442296e+00        1 8.987079e+00       1     1
#> 2015-05-18 04:05:00 1.458398e+00        0 8.329715e+00       1     1
#> 2015-05-18 04:06:00 1.315418e+00        0 7.992913e+00       1     1
#> 2015-05-18 04:07:00 3.325209e-01        0 1.200626e+01       1     1
#> 2015-05-18 04:08:00 1.743248e+00        0 1.133614e+01       1     1
#> 2015-05-18 04:09:00 8.084739e-01        0 1.406166e+01       1     1
#> 2015-05-18 04:10:00 2.434315e+00        1 1.212091e+01       1     1
#> 2015-05-18 04:11:00 2.618257e+00        1 1.666896e+01       1     1
#> 2015-05-18 04:12:00 2.177245e+00        0 1.944205e+01       1     1
#> 2015-05-18 04:13:00 1.112182e+00        0 9.877308e+00       1     1
#> 2015-05-18 04:14:00 1.358875e+00        0 7.820747e+00       1     1
#> 2015-05-18 04:15:00 1.204777e+00        0 8.338689e+00       1     1
#> 2015-05-18 04:16:00 5.050117e-01        0 7.503430e+00       1     1
#> 2015-05-18 04:17:00 1.330468e+00        0 7.761053e+00       1     1
#> 2015-05-18 04:18:00 9.924907e-01        0 1.143433e+01       1     1
#> 2015-05-18 04:19:00 2.059709e+00        0 7.093532e+00       1     1
#> 2015-05-18 04:20:00 1.642273e+00        0 9.254446e+00       1     1
#> 2015-05-18 04:21:00 7.880690e-01        0 7.181958e+00       1     1
#> 2015-05-18 04:22:00 3.882802e-01        0 6.137784e+00       0     0
#> 2015-05-18 04:23:00 7.303342e-01        0 7.659739e+00       1     0
#> 2015-05-18 04:24:00 1.436888e+00        0 9.518331e+00       1     0
#> 2015-05-18 04:25:00 1.514635e+00        0 1.009707e+01       1     0
#> 2015-05-18 04:26:00 3.524263e+00        1 8.722808e+00       1     0
#> 2015-05-18 04:27:00 5.152817e-01        0 1.222919e+01       1     1
#> 2015-05-18 04:28:00 2.611663e+00        1 6.739396e+00       1     1
#> 2015-05-18 04:29:00 6.464648e-01        0 8.932151e+00       1     1
#> 2015-05-18 04:30:00 1.693332e+00        0 1.319220e+01       1     1
#> 2015-05-18 04:31:00 6.821672e-01        0 1.015799e+01       1     1
#>  [ reached getOption("max.print") -- omitted 2488 rows ]
#> 
#> $nonAlarmedTestObs
#>                             x         y         z
#> 2015-05-17 19:12:00 0.6873707 -1.791209 1.4974422
#> 2015-05-17 19:13:00 0.6493085 -1.714285 1.5448407
#> 2015-05-17 19:14:00 0.6469978 -1.648453 1.2142648
#> 2015-05-17 19:15:00 0.7304903 -1.640529 1.1723993
#> 2015-05-17 19:16:00 0.9701771 -1.677653 1.3047262
#> 2015-05-17 19:17:00 0.7147756 -1.464012 1.2695188
#> 2015-05-17 19:18:00 0.6999505 -1.520508 1.0568807
#> 2015-05-17 19:19:00 0.7194566 -1.566418 0.8063545
#> 2015-05-17 19:20:00 0.5107392 -1.572126 0.6716052
#> 2015-05-17 19:21:00 0.6204801 -1.450096 0.8276680
#> 2015-05-17 19:22:00 0.6311992 -1.405327 1.2685761
#> 2015-05-17 19:23:00 0.8084357 -1.554528 1.2225188
#> 2015-05-17 19:24:00 0.8838028 -1.971292 1.3688204
#> 2015-05-17 19:25:00 0.7726761 -1.691882 1.1798015
#> 2015-05-17 19:26:00 0.7694961 -1.632047 1.2391411
#> 2015-05-17 19:27:00 0.5701776 -1.408883 0.8064593
#> 2015-05-17 19:28:00 0.7195357 -1.203305 0.7697363
#> 2015-05-17 19:29:00 0.4392277 -1.425017 0.8075906
#> 2015-05-17 19:30:00 0.5598485 -1.452792 0.8179733
#> 2015-05-17 19:31:00 0.5427673 -1.573988 0.9702159
#> 2015-05-17 19:32:00 0.4999565 -1.427125 0.7977655
#> 2015-05-17 19:33:00 0.7377298 -1.563519 1.0114877
#> 2015-05-17 19:34:00 0.6374301 -1.619623 1.2129189
#> 2015-05-17 19:35:00 0.6546451 -1.614819 1.3042455
#> 2015-05-17 19:36:00 0.6602555 -1.543411 1.0702915
#> 2015-05-17 19:37:00 0.5303595 -1.652609 1.2361311
#> 2015-05-17 19:38:00 0.6722496 -1.434395 1.0354962
#> 2015-05-17 19:39:00 0.7207151 -1.730719 1.1709362
#> 2015-05-17 19:40:00 0.7126364 -1.728666 1.3942287
#> 2015-05-17 19:41:00 0.8453341 -1.751143 1.4079720
#> 2015-05-17 19:42:00 0.8402281 -1.782150 1.2216861
#> 2015-05-17 19:43:00 0.6677306 -1.609733 1.4480558
#> 2015-05-17 19:44:00 0.9520128 -1.771116 1.5751432
#> 2015-05-17 19:45:00 0.9989320 -1.966126 1.6820503
#> 2015-05-17 19:46:00 0.6800729 -1.793430 1.3740416
#> 2015-05-17 19:47:00 0.7847383 -1.558163 1.4380848
#> 2015-05-17 19:48:00 0.8158729 -1.920954 1.4310153
#> 2015-05-17 19:49:00 0.6278282 -1.761086 1.2741912
#> 2015-05-17 19:50:00 0.7231521 -1.728181 1.3425646
#> 2015-05-17 19:51:00 0.6158450 -1.583592 1.2772546
#> 2015-05-17 19:52:00 0.7576308 -1.509304 0.9370001
#> 2015-05-17 19:53:00 0.7718770 -1.505685 0.8540817
#> 2015-05-17 19:54:00 0.6717851 -1.499251 1.1246133
#> 2015-05-17 19:55:00 0.6079449 -1.575734 1.2945515
#> 2015-05-17 19:56:00 0.6630176 -1.723914 1.2901758
#> 2015-05-17 19:57:00 0.8273284 -1.827589 1.3754307
#> 2015-05-17 19:58:00 0.7929303 -1.549276 1.0566491
#> 2015-05-17 19:59:00 0.6807106 -1.589728 1.3364232
#> 2015-05-17 22:00:00 0.5340951 -1.721241 1.2269357
#> 2015-05-17 22:01:00 0.7569753 -1.650668 1.2774000
#> 2015-05-17 22:02:00 0.8238358 -1.462076 1.0364662
#> 2015-05-17 22:03:00 0.5652898 -1.406465 0.8070044
#> 2015-05-17 22:04:00 0.7101897 -1.541508 1.1841574
#> 2015-05-17 22:05:00 0.8943077 -1.722228 1.2991336
#> 2015-05-17 22:06:00 0.6873059 -1.701029 1.4856917
#> 2015-05-17 22:07:00 0.5448942 -1.670963 0.8636743
#> 2015-05-17 22:08:00 0.6616046 -1.469957 1.2289947
#> 2015-05-17 22:09:00 0.6555024 -1.517950 1.1731105
#> 2015-05-17 22:10:00 0.8292932 -1.639158 0.9663156
#> 2015-05-17 22:11:00 0.6981635 -1.597103 0.9295740
#> 2015-05-17 22:12:00 0.4910135 -1.289020 0.7746922
#> 2015-05-17 22:13:00 0.7030741 -1.419886 1.2358616
#> 2015-05-17 22:14:00 0.8259609 -1.513541 1.3091765
#> 2015-05-17 22:15:00 0.5995841 -1.552654 1.1625016
#> 2015-05-17 22:16:00 0.5787313 -1.728658 0.8758376
#> 2015-05-17 22:17:00 0.7688462 -1.462707 1.0311888
#> 2015-05-17 22:18:00 0.7717248 -1.559865 1.0849381
#> 2015-05-17 22:19:00 0.8829823 -1.829399 1.2232853
#> 2015-05-17 22:20:00 0.4705706 -1.744411 1.2985515
#> 2015-05-17 22:21:00 0.7745122 -1.751987 1.4567091
#> 2015-05-17 22:22:00 0.8197238 -1.698513 1.2529110
#> 2015-05-17 22:23:00 0.6659560 -1.773685 1.0410373
#> 2015-05-17 22:24:00 0.7047781 -1.710512 1.0801909
#> 2015-05-17 22:25:00 0.7752838 -1.667211 1.0856998
#> 2015-05-17 22:26:00 0.8259219 -1.740676 1.4772401
#> 2015-05-17 22:27:00 0.9045056 -1.624660 1.4782377
#> 2015-05-17 22:28:00 0.8376360 -1.727606 1.6371894
#> 2015-05-17 22:29:00 0.8141338 -1.939114 1.4639588
#> 2015-05-17 22:30:00 0.8835678 -1.703311 1.5917342
#> 2015-05-17 22:31:00 0.8706952 -1.661001 1.5801503
#> 2015-05-17 22:32:00 0.7535283 -1.883530 1.4616069
#> 2015-05-17 22:33:00 0.6434480 -1.847445 1.4867084
#> 2015-05-17 22:34:00 0.8111077 -1.922176 1.8090020
#> 2015-05-17 22:35:00 1.1157111 -1.897836 2.0248005
#> 2015-05-17 22:36:00 0.9440925 -2.126254 1.9455236
#> 2015-05-17 22:37:00 0.9729098 -2.087997 1.7977882
#> 2015-05-17 22:44:00 0.7206327 -1.720881 1.2485494
#> 2015-05-17 22:45:00 0.9129170 -2.005222 1.9248928
#> 2015-05-17 22:46:00 0.7978570 -1.929383 1.3875301
#> 2015-05-17 22:47:00 0.9497094 -1.955318 1.9078716
#> 2015-05-17 22:48:00 0.9753383 -1.835147 1.7277986
#> 2015-05-17 22:49:00 1.1643215 -2.201319 1.9917979
#> 2015-05-17 22:50:00 1.0391270 -1.843219 1.8317153
#> 2015-05-17 22:51:00 0.8424709 -1.877755 1.3125281
#> 2015-05-17 22:52:00 0.7331560 -1.926608 1.6088596
#> 2015-05-17 22:53:00 0.6733121 -1.674440 1.4360464
#> 2015-05-17 22:54:00 0.5763495 -1.797985 1.4404493
#> 2015-05-17 22:55:00 0.7230353 -1.865030 1.6981267
#> 2015-05-17 22:56:00 0.9441150 -1.915558 1.9252159
#> 2015-05-17 22:57:00 0.9833681 -1.892660 1.9207431
#> 2015-05-17 22:58:00 0.8180935 -2.074759 1.8077633
#> 2015-05-18 01:19:00 0.6909522 -1.817346 1.6237754
#> 2015-05-18 01:20:00 0.9261847 -1.629942 1.3690254
#> 2015-05-18 01:21:00 0.8262619 -1.813164 1.6414884
#> 2015-05-18 01:22:00 0.7882384 -1.804212 1.2674182
#> 2015-05-18 01:23:00 0.7753097 -1.825698 1.4824372
#> 2015-05-18 01:24:00 0.8976331 -1.826309 1.5994641
#> 2015-05-18 01:25:00 1.0334414 -1.727368 1.5229081
#> 2015-05-18 01:26:00 0.9252127 -1.696120 1.5562170
#> 2015-05-18 01:27:00 0.8622793 -1.811700 1.6930851
#> 2015-05-18 01:49:00 0.8584588 -1.660850 1.3324390
#> 2015-05-18 01:50:00 0.8789496 -2.042254 1.9735400
#> 2015-05-18 01:51:00 0.6998778 -1.850667 1.6396594
#> 2015-05-18 01:52:00 0.8266002 -1.815211 1.6420870
#> 2015-05-18 01:53:00 1.2060599 -2.004485 2.1814326
#> 2015-05-18 01:54:00 1.0702740 -1.996058 2.0965358
#> 2015-05-18 01:55:00 0.8858580 -1.790844 2.2616734
#> 2015-05-18 04:22:00 0.8102239 -1.776640 1.4491594
#> 2015-05-18 04:23:00 0.9417147 -1.736832 1.5494004
#> 2015-05-18 04:24:00 0.9315355 -1.822415 1.7950528
#> 2015-05-18 04:25:00 0.8463991 -2.021957 1.8395865
#> 2015-05-18 04:26:00 0.7240660 -1.903702 1.9369592
#> 2015-05-21 13:47:00 0.7931590 -1.667597 1.5784573
#> 2015-05-21 13:48:00 0.6546512 -1.711271 1.4189895
#> 2015-05-21 13:49:00 0.8469116 -1.983782 1.6213761
#> 2015-05-21 13:50:00 0.9203912 -2.000016 1.9003996
#> 2015-05-21 13:51:00 1.0559446 -2.094671 2.0779185
#> 2015-05-21 13:52:00 1.0492681 -1.970398 2.2059422
#> 2015-05-21 16:01:00 0.7620971 -1.819745 1.4955356
#> 2015-05-21 16:02:00 0.8701804 -1.789754 1.7158017
#> 2015-05-21 16:03:00 0.9766529 -1.906933 1.6682565
#> 2015-05-21 16:04:00 0.8343526 -1.708399 1.6686660
#> 2015-05-21 16:05:00 1.0121636 -1.784280 1.8100107
#> 2015-05-21 16:06:00 0.7302877 -1.878527 1.3763874
#> 2015-05-21 16:07:00 0.8180671 -1.813010 1.5764492
#> 2015-05-21 16:08:00 0.8008947 -1.732593 1.3632223
#> 2015-05-21 16:09:00 0.7405121 -1.683506 1.6155509
#> 2015-05-21 16:10:00 1.0452083 -1.913418 1.7527058
#> 2015-05-21 16:11:00 0.7560100 -1.847542 1.5317769
#> 2015-05-21 16:12:00 0.9653522 -1.904015 1.6927420
#> 2015-05-21 16:13:00 0.9061379 -2.043789 1.6301195
#> 2015-05-21 16:14:00 0.7246120 -1.786006 1.2804510
#> 2015-05-21 16:15:00 0.6863782 -1.792881 1.4469579
#> 2015-05-21 16:16:00 1.0155781 -1.980259 1.8393863
#> 2015-05-21 16:17:00 0.9849544 -1.878636 1.5399417
#> 2015-05-21 16:18:00 0.7915275 -1.637715 1.5083132
#> 2015-05-21 16:19:00 0.6258214 -1.797675 1.5598374
#> 2015-05-21 16:20:00 0.9446357 -2.147654 1.7096866
#> 2015-05-21 16:21:00 1.2353174 -1.970549 1.9294537
#> 2015-05-21 16:22:00 0.8534868 -1.768379 1.8402985
#> 2015-05-21 16:23:00 0.7768931 -1.629950 1.6090377
#> 2015-05-21 16:24:00 0.8159887 -1.996942 1.6688324
#> 2015-05-21 16:25:00 0.8575177 -1.852042 1.6917351
#> 2015-05-21 16:26:00 0.6848043 -2.018246 1.7414330
#> 2015-05-21 16:27:00 0.9809874 -1.880149 2.0024677
#> 2015-05-21 16:36:00 0.7105867 -1.601170 1.3915346
#> 2015-05-21 16:37:00 1.0626846 -2.027586 1.7742845
#> 2015-05-21 16:38:00 0.9270231 -1.880823 1.8631109
#> 2015-05-21 16:39:00 0.8131657 -1.970087 1.4501418
#> 2015-05-21 16:40:00 0.8809513 -1.808222 1.9884279
#> 2015-05-21 16:56:00 0.6421335 -2.047258 1.5397565
#> 2015-05-21 16:57:00 1.1472879 -1.722843 1.5091777
#> 2015-05-21 16:58:00 0.7675197 -1.724869 1.1667958
#> 2015-05-21 16:59:00 0.7137583 -1.625471 1.0664110
#> 2015-05-21 19:00:00 0.9177612 -1.929606 1.7648350
#> 2015-05-21 19:01:00 1.0278010 -1.887630 1.9979193
#> 2015-05-21 19:02:00 0.9142407 -1.685397 1.6264043
#> 2015-05-21 19:03:00 1.1340949 -1.970618 1.7381994
#> 2015-05-21 19:05:00 0.5917956 -1.769912 1.5578276
#> 2015-05-21 19:06:00 0.9367929 -2.070003 2.0374577
#> 2015-05-21 19:07:00 0.9549566 -1.999014 1.8947881
#> 2015-05-21 19:08:00 0.7157944 -1.582646 1.4824779
#> 2015-05-21 19:09:00 0.7811432 -1.724683 1.0845785
#> 2015-05-21 19:10:00 0.7870990 -1.817082 1.1364211
#> 2015-05-21 19:11:00 0.8326438 -1.873407 1.1577146
#> 2015-05-21 19:12:00 0.7458999 -1.775818 1.6258284
#> 2015-05-21 19:13:00 0.9888952 -1.996032 1.7340069
#> 2015-05-21 19:14:00 0.8692414 -1.839422 1.6829403
#> 2015-05-21 19:15:00 0.7733877 -1.773175 1.7403553
#> 2015-05-21 19:16:00 0.9133915 -1.787656 1.6920535
#> 2015-05-21 19:17:00 0.7743057 -1.731923 1.3624950
#> 2015-05-21 19:18:00 0.7905048 -1.840346 1.5860941
#> 2015-05-21 19:19:00 0.8313173 -1.894278 1.4899929
#> 2015-05-21 19:20:00 0.7701154 -1.864945 1.5769282
#> 2015-05-21 19:21:00 0.8553960 -1.923023 1.8192333
#> 2015-05-21 19:23:00 0.7139855 -1.900810 1.5144069
#> 2015-05-21 19:24:00 0.9367638 -1.872560 1.5231379
#> 2015-05-21 19:25:00 0.8046707 -1.746043 1.4033682
#> 2015-05-21 19:26:00 0.9505447 -1.809141 1.5750385
#> 2015-05-21 19:27:00 0.8199438 -1.783576 1.6341677
#> 2015-05-21 19:28:00 0.8175635 -1.770582 1.8602600
#> 2015-05-21 19:29:00 1.0583409 -1.960339 1.8827081
#> 2015-05-21 19:48:00 0.5962058 -1.853551 1.6105077
#> 2015-05-21 19:49:00 0.9493364 -1.932600 1.6665372
#> 2015-05-21 19:50:00 0.8484000 -1.810008 1.5466237
#> 2015-05-21 19:51:00 0.9842180 -1.838770 1.4172263
#> 2015-05-21 19:52:00 0.6624316 -1.851135 1.4373014
#> 2015-05-21 19:53:00 0.8529630 -1.628970 1.3695434
#> 2015-05-21 19:54:00 0.6982323 -1.688669 1.4552079
#> 2015-05-21 19:55:00 0.7388349 -1.912667 1.6477696
#> 2015-05-21 19:56:00 0.9847101 -2.020430 1.9108013
#> 2015-05-21 19:57:00 0.9478521 -1.864108 2.0936456
#> 2015-05-21 19:58:00 0.9685125 -1.789129 1.7908921
#> 2015-05-21 22:00:00 0.8666549 -1.650480 1.5410331
#> 2015-05-21 22:01:00 0.9519226 -1.903925 1.6779579
#> 2015-05-21 22:02:00 0.9894308 -2.011162 1.8098963
#> 2015-05-21 22:03:00 0.8181103 -1.858613 1.7269595
#> 2015-05-21 22:04:00 0.5193554 -1.610917 1.3416785
#> 2015-05-21 22:05:00 0.9103657 -1.623983 1.2747030
#> 2015-05-21 22:06:00 0.7620207 -1.711633 1.4857399
#> 2015-05-21 22:07:00 0.8516683 -1.831322 1.3450163
#> 2015-05-21 22:08:00 0.6777269 -1.471065 1.2157847
#> 2015-05-21 22:09:00 0.7085724 -1.528782 1.1316494
#> 2015-05-21 22:10:00 0.7507638 -1.412470 0.7178784
#> 2015-05-21 22:11:00 0.7459220 -1.525115 1.0345530
#> 2015-05-21 22:12:00 0.7030484 -1.518761 1.2672098
#> 2015-05-21 22:13:00 0.6874107 -1.562064 1.4470339
#> 2015-05-21 22:14:00 0.8828895 -1.716893 1.2963643
#> 2015-05-21 22:15:00 0.9031975 -1.819649 1.5246860
#> 2015-05-21 22:16:00 0.9057044 -1.775341 1.5582163
#> 2015-05-21 22:17:00 0.9172675 -1.745416 1.4320096
#> 2015-05-21 22:18:00 0.7462380 -1.775517 1.4836234
#> 2015-05-21 22:19:00 0.9228434 -2.002632 1.7394128
#> 2015-05-21 22:20:00 0.7360322 -1.663578 1.4018996
#> 2015-05-21 22:21:00 0.6737482 -1.646197 1.2949016
#> 2015-05-21 22:22:00 0.7111991 -1.767547 1.3482867
#> 2015-05-21 22:23:00 0.8261415 -1.776223 1.3597772
#> 2015-05-21 22:24:00 0.8061016 -1.663450 1.1957541
#> 2015-05-21 22:25:00 0.8410621 -1.664531 1.3168395
#> 2015-05-21 22:26:00 0.7303819 -1.930316 1.4377907
#> 2015-05-21 22:27:00 0.9426577 -1.717630 1.3419922
#> 2015-05-21 22:28:00 0.7130371 -1.773459 1.6317982
#> 2015-05-21 22:29:00 0.8843443 -1.874194 1.4345350
#> 2015-05-21 22:30:00 0.7882081 -1.779010 1.2340904
#> 2015-05-21 22:31:00 0.8766557 -1.780778 1.1768332
#> 2015-05-21 22:32:00 0.9103444 -1.696058 1.5903341
#> 2015-05-21 22:33:00 0.9345075 -1.876305 1.5719897
#> 2015-05-21 22:34:00 0.7704347 -1.745099 1.1868534
#> 2015-05-21 22:35:00 0.8418091 -1.590714 1.3139698
#> 2015-05-21 22:36:00 0.9484050 -1.925920 1.5568817
#> 2015-05-21 22:37:00 0.7542364 -1.757141 1.4194220
#> 2015-05-21 22:38:00 0.8092820 -1.928998 1.5639797
#> 2015-05-21 22:39:00 0.8537279 -1.662414 1.2991604
#> 2015-05-21 22:40:00 0.9403684 -1.702315 1.4921554
#> 2015-05-21 22:41:00 0.8755050 -1.806338 1.5447118
#> 2015-05-21 22:42:00 1.0355345 -1.979985 1.6276858
#> 2015-05-21 22:43:00 0.9455628 -1.926048 1.6980703
#> 2015-05-21 22:44:00 0.8353222 -1.536105 1.3250808
#> 2015-05-21 22:45:00 0.7424372 -1.582824 1.1845513
#> 2015-05-21 22:46:00 0.8232610 -1.812355 1.1631593
#> 2015-05-21 22:47:00 0.7945519 -1.781158 1.3156014
#> 2015-05-21 22:48:00 0.9785864 -1.773109 1.6534846
#> 2015-05-21 22:49:00 0.8281519 -1.781618 1.5075292
#> 2015-05-21 22:50:00 0.9157235 -1.979824 1.6713851
#> 2015-05-21 22:51:00 0.9368668 -1.774589 1.4075803
#> 2015-05-21 22:52:00 0.9257608 -1.829623 1.3974936
#> 2015-05-21 22:53:00 0.8690777 -1.821938 1.5431499
#> 2015-05-21 22:55:00 0.6091579 -1.859397 1.5551136
#> 2015-05-21 22:56:00 0.5265891 -1.567116 1.4248438
#> 2015-05-21 22:57:00 0.8245800 -1.743016 1.2767653
#> 2015-05-21 22:58:00 0.7008884 -1.745872 1.3620789
#> 2015-05-21 22:59:00 0.6382861 -1.548476 1.2529358
#> 2015-05-22 01:00:00 0.7201877 -1.612469 1.0828244
#> 2015-05-22 01:01:00 0.7145069 -1.891430 1.3669002
#> 2015-05-22 01:02:00 0.8686795 -1.919742 1.4513525
#> 2015-05-22 01:03:00 0.8278842 -1.784880 1.3966638
#> 2015-05-22 01:04:00 0.9785745 -1.992668 1.6270461
#> 2015-05-22 01:05:00 0.7595316 -1.526279 1.3805424
#> 2015-05-22 01:06:00 0.8126264 -1.574619 1.5831985
#> 2015-05-22 01:07:00 0.8603157 -1.847632 1.5111472
#> 2015-05-22 01:08:00 0.7726486 -1.820360 1.4221581
#> 2015-05-22 01:09:00 0.8934142 -1.752890 1.2759475
#> 2015-05-22 01:10:00 0.7229584 -1.943268 1.8476340
#> 2015-05-22 01:11:00 0.8882222 -1.922504 1.4987542
#> 2015-05-22 01:12:00 0.7990630 -1.597672 1.1276594
#> 2015-05-22 01:13:00 0.7756187 -1.605431 1.2803107
#> 2015-05-22 01:14:00 0.7292108 -1.577669 1.2759458
#> 2015-05-22 01:15:00 0.8117526 -1.766549 1.4638433
#> 2015-05-22 01:16:00 0.9314391 -1.873287 1.4883528
#> 2015-05-22 01:17:00 0.7031592 -1.738438 1.1267612
#> 2015-05-22 01:18:00 0.6672247 -1.656530 1.1359539
#> 2015-05-22 01:19:00 0.6994637 -1.609752 1.0955334
#> 2015-05-22 01:20:00 0.5833013 -1.718685 1.0681277
#> 2015-05-22 01:21:00 0.7456629 -1.772583 1.0992964
#> 2015-05-22 01:22:00 0.7111094 -1.527162 0.9936375
#> 2015-05-22 01:23:00 0.7741547 -1.826316 1.6593946
#> 2015-05-22 01:24:00 0.8416392 -1.825656 1.4255079
#> 2015-05-22 01:25:00 0.5516299 -1.565088 1.1589625
#> 2015-05-22 01:26:00 0.6275398 -1.708813 1.0067915
#> 2015-05-22 01:27:00 0.6028303 -1.532550 0.9311522
#> 2015-05-22 01:28:00 0.7677781 -1.668462 1.2060758
#> 2015-05-22 01:29:00 0.6299759 -1.466700 1.1547045
#> 2015-05-22 01:30:00 0.6250974 -1.558802 0.9001126
#> 2015-05-22 01:31:00 0.7275819 -1.261448 0.8329401
#> 2015-05-22 01:32:00 0.6205109 -1.757841 0.9779636
#> 2015-05-22 01:33:00 0.9741188 -1.495576 1.0095974
#> 2015-05-22 01:34:00 0.6580188 -1.648727 1.1192421
#> 2015-05-22 01:35:00 0.6597531 -1.414226 1.1179695
#> 2015-05-22 01:36:00 0.7931460 -1.488934 1.2133098
#> 2015-05-22 01:37:00 0.7463977 -1.417137 1.0322873
#> 2015-05-22 01:38:00 0.8593817 -1.886582 1.3910798
#> 2015-05-22 01:39:00 0.8253967 -1.542104 1.3383318
#> 2015-05-22 01:40:00 0.7108423 -1.631308 1.5186751
#> 2015-05-22 01:41:00 0.8764260 -1.819364 1.4950682
#> 2015-05-22 01:42:00 0.7806614 -1.884591 1.4875561
#> 2015-05-22 01:43:00 0.7340523 -1.620853 1.3391217
#> 2015-05-22 01:44:00 0.8411787 -1.774856 1.3009442
#> 2015-05-22 01:45:00 0.8227261 -1.718270 1.3187803
#> 2015-05-22 01:46:00 0.7046404 -1.844643 1.2614632
#> 2015-05-22 01:47:00 0.6763743 -1.639554 1.0150575
#> 2015-05-22 01:48:00 0.7495246 -1.553781 0.9761661
#> 2015-05-22 01:49:00 0.6693290 -1.440429 0.7345208
#> 2015-05-22 01:50:00 0.7129654 -1.623017 1.1177058
#> 2015-05-22 01:51:00 0.9058666 -1.553564 0.9711670
#> 2015-05-22 01:52:00 0.5435457 -1.572306 0.7493552
#> 2015-05-22 01:53:00 0.7129022 -1.707143 1.0012998
#> 2015-05-22 01:54:00 0.6822502 -1.750527 1.0616735
#> 2015-05-22 01:55:00 0.6653260 -1.483931 1.2399029
#> 2015-05-22 01:56:00 0.6093645 -1.627823 1.0194405
#> 2015-05-22 01:57:00 0.7167987 -1.439131 1.1919010
#> 2015-05-22 01:58:00 0.6731189 -1.828314 1.0165699
#> 2015-05-22 01:59:00 0.6931670 -1.725571 1.3070977
#> 2015-05-22 04:00:00 0.6236930 -1.549691 0.9154882
#> 2015-05-22 04:01:00 0.5261839 -1.695984 1.1955437
#> 2015-05-22 04:02:00 0.7507922 -1.507362 1.1951771
#> 2015-05-22 04:03:00 0.5898448 -1.689279 1.1518478
#> 2015-05-22 04:04:00 0.7312944 -1.710223 1.0224877
#> 2015-05-22 04:05:00 0.6734644 -1.606019 1.0259969
#> 2015-05-22 04:06:00 0.6213366 -1.301504 0.7362881
#> 2015-05-22 04:07:00 0.8082299 -1.598741 0.8956272
#> 2015-05-22 04:08:00 0.5353085 -1.644716 0.8392323
#> 2015-05-22 04:09:00 0.3574875 -1.274061 0.8161517
#> 2015-05-22 04:10:00 0.5650730 -1.497325 0.5267923
#>  [ reached getOption("max.print") -- omitted 3 rows ]
#> 
#> $trainSpecs
#> $SPE_threshold
#>    99.9% 
#> 2.403837 
#> 
#> $T2_threshold
#>    99.9% 
#> 6.692038 
#> 
#> $projectionMatrix
#>            [,1]
#> [1,]  0.5655850
#> [2,] -0.5836925
#> [3,]  0.5825948
#> 
#> $LambdaInv
#>           [,1]
#> [1,] 0.3699458
#> 
#> $T2
#>   [1] 2.652271e-01 5.744740e-01 4.076017e-01 3.903449e-01 2.468810e-01
#>   [6] 2.507833e-01 1.415489e-01 2.567107e-03 1.471761e-02 6.481263e-02
#>  [11] 1.794327e+00 1.785492e+00 8.369394e-01 1.851245e+00 1.182059e+00
#>  [16] 2.291648e+00 9.974041e-01 9.995228e-01 2.076399e+00 1.578772e+00
#>  [21] 1.896022e+00 1.280118e+00 1.475373e+00 4.211909e-01 7.992232e-01
#>  [26] 2.866881e+00 3.965979e-01 8.277737e-01 5.600134e-01 8.926884e-02
#>  [31] 9.728237e-01 4.609117e-01 5.277945e-01 2.167118e+00 2.049859e+00
#>  [36] 5.303921e-01 9.754922e-02 1.992302e-01 8.729117e-01 1.359308e+00
#>  [41] 2.035444e+00 1.912650e+00 2.606516e+00 1.424349e+00 1.978680e+00
#>  [46] 8.520372e-01 8.964215e-01 9.921221e-01 1.826647e+00 1.803866e+00
#>  [51] 1.253065e+00 6.972076e-02 9.143298e-01 5.291213e-01 3.046726e-01
#>  [56] 1.100163e-01 2.074104e-01 9.963083e-01 1.179218e+00 2.312956e+00
#>  [61] 2.365986e+00 1.316903e+00 2.404831e-01 1.722991e+00 1.665800e+00
#>  [66] 9.452271e-01 3.267967e+00 1.790808e+00 3.640993e-01 1.888159e+00
#>  [71] 8.188424e-01 2.002137e+00 9.874890e-01 2.257955e-01 3.460186e-03
#>  [76] 3.435218e-02 1.750454e-01 6.013370e-02 2.770898e-02 3.411767e-01
#>  [81] 2.248917e-01 3.313447e-01 3.446709e-01 1.932939e-01 5.806350e-01
#>  [86] 6.468605e-01 5.377332e-01 4.938523e-01 1.031596e+00 2.693309e-02
#>  [91] 4.138066e-01 1.468198e-01 6.270041e-01 1.445888e+00 6.700320e-03
#>  [96] 2.439318e-01 1.331690e+00 2.494442e-01 3.184666e-02 3.935957e-03
#> [101] 1.090499e+00 1.521540e+00 5.083933e-01 1.686813e-01 1.548148e-01
#> [106] 8.394128e-01 6.235483e-01 1.846209e+00 2.417662e+00 1.682470e+00
#> [111] 3.345257e+00 1.335542e+00 2.022156e+00 3.098166e+00 1.313634e+00
#> [116] 1.457013e+00 1.305646e+00 1.489704e+00 9.236460e-01 2.355958e-01
#> [121] 1.053285e+00 6.028136e-01 4.801326e-01 7.065292e-01 3.993054e-01
#> [126] 7.454895e-01 2.590806e-01 3.351850e-01 9.537900e-02 7.951686e-03
#> [131] 2.338682e-01 2.792305e-01 1.465258e+00 9.879705e-02 1.178285e+00
#> [136] 1.335052e+00 7.720610e-01 1.908620e+00 4.424014e-02 1.210570e+00
#> [141] 1.000510e+00 3.135457e-01 1.164414e+00 1.469798e+00 9.709378e-01
#> [146] 1.089907e+00 5.566688e-01 4.064718e-01 1.763117e+00 7.650187e-01
#> [151] 9.564962e-01 1.044018e+00 7.120883e-01 1.824095e+00 1.013749e+00
#> [156] 9.664269e-01 2.353848e+00 1.106088e+00 2.982860e+00 2.945822e+00
#> [161] 4.847083e+00 2.759944e+00 1.429160e+00 3.769828e-01 5.460660e-01
#> [166] 5.011222e-02 4.170392e-01 3.198131e-01 1.157828e-01 1.574764e+00
#> [171] 2.229812e+00 2.947831e+00 1.054685e+00 3.364584e-01 2.264285e-01
#> [176] 1.799160e-01 9.675673e-01 9.326589e-01 1.426402e+00 2.224464e+00
#> [181] 3.435722e-01 6.032850e-01 3.707900e-01 1.351489e-01 6.818831e-01
#> [186] 2.937835e-01 9.456792e-04 5.819277e-02 1.021746e-01 1.493456e-02
#> [191] 8.551398e-01 4.492871e-01 1.102307e+00 1.500344e+00 1.508257e+00
#> [196] 1.782901e+00 1.210723e+00 9.847397e-01 3.257262e+00 7.762979e-01
#> [201] 6.940970e-01 3.285424e-01 4.450099e-01 1.058987e+00 3.284285e-01
#> [206] 3.155493e-01 1.221223e-01 4.862164e-01 3.376602e-01 6.739127e-02
#> [211] 1.194999e-01 1.327796e-01 2.172927e-01 1.168328e+00 7.066302e-01
#> [216] 8.814584e-01 3.663747e-01 3.724620e-01 1.240584e+00 1.605729e+00
#> [221] 1.212227e+00 3.389759e-01 2.125615e-01 2.478779e-01 2.682930e-01
#> [226] 5.704394e-02 7.293163e-01 7.368943e-06 7.605065e-02 6.095674e-01
#> [231] 8.877422e-02 1.923775e-01 6.995359e-02 3.451180e-03 4.352276e-01
#> [236] 4.780576e-02 1.557875e-01 4.395734e-01 7.756573e-01 3.170402e-01
#> [241] 3.853394e-01 6.000725e-05 1.779396e-01 3.860261e-02 1.433339e-02
#> [246] 2.897878e-02 1.869531e-01 9.344170e-01 4.537152e-01 4.325642e-03
#> [251] 5.067270e-01 4.514738e-01 8.377743e-01 1.621683e+00 3.127167e+00
#> [256] 2.127348e+00 3.500428e-01 6.403384e-01 5.785041e-01 4.526964e-01
#> [261] 1.125154e-01 2.919139e-01 1.155262e+00 8.749021e-01 1.788977e+00
#> [266] 1.509335e+00 1.607398e+00 6.466533e-01 2.733900e-02 3.329188e-01
#> [271] 1.182050e+00 1.121460e-01 9.584867e-01 5.551277e-01 1.594306e+00
#> [276] 2.034419e-01 1.978289e-01 2.934523e-01 3.681413e-01 1.061244e-03
#> [281] 1.361788e-01 1.916172e-01 3.571184e-01 9.658306e-01 6.111742e-01
#> [286] 6.537558e-01 1.710345e-01 6.574877e-01 2.306410e-01 5.207664e-01
#> [291] 4.795095e-01 8.316496e-01 5.132188e-02 6.159376e-01 1.093802e-01
#> [296] 3.073827e-01 2.437268e-03 1.919619e-01 7.208803e-02 1.736595e-02
#> [301] 1.431896e-02 7.161242e-03 8.086918e-03 7.491452e-01 1.636132e-02
#> [306] 1.762294e-01 1.036057e-02 2.132865e-02 4.600794e-05 5.479273e-02
#> [311] 2.812155e-01 4.190069e-01 5.967155e-02 1.995779e-01 5.976630e-01
#> [316] 2.368907e-01 3.694802e-01 6.112109e-01 4.316350e-02 4.435871e-01
#> [321] 1.337521e-01 1.111278e-01 6.787446e-03 3.867358e-02 1.633579e-02
#> [326] 2.666666e-01 4.464891e-01 6.318083e-01 3.647054e-02 6.672214e-03
#> [331] 2.240231e-02 2.286708e-01 2.332115e-01 4.773770e-01 8.279945e-01
#> [336] 7.111508e-01 4.804044e-01 3.053909e-01 3.611110e-02 1.711813e-01
#> [341] 2.019446e-03 8.006489e-02 6.067980e-01 2.756245e-01 5.181096e-01
#> [346] 2.581461e-01 2.847878e-01 1.688891e-04 1.313714e-01 6.205030e-01
#> [351] 9.699641e-01 7.567568e-01 2.277365e+00 2.150178e+00 1.568188e+00
#> [356] 1.468900e+00 7.840970e-01 2.573417e-01 8.950889e-01 7.714819e-01
#> [361] 5.244105e-01 1.767042e-01 2.112644e-01 7.928131e-01 1.547075e-01
#> [366] 1.294925e-01 4.964654e-02 3.870041e-01 3.250358e-01 5.450260e-04
#> [371] 2.577701e-01 1.164208e-01 1.253332e-01 4.287478e-05 2.567447e-02
#> [376] 7.548589e-01 7.145804e-02 7.074193e-01 7.925131e-02 1.825599e-01
#> [381] 1.360496e-02 9.429358e-02 4.484437e-02 1.313850e+00 1.714678e+00
#> [386] 7.019738e-01 6.396461e-01 5.946295e-01 2.590361e-01 8.425501e-01
#> [391] 5.882805e-01 2.166632e-01 6.532189e-02 2.168236e-01 2.518982e-01
#> [396] 1.653008e-01 1.242447e-01 9.682272e-01 2.644186e-02 3.055347e-01
#> [401] 1.128728e+00 1.949174e-01 3.473695e-01 2.848979e-01 2.235178e+00
#> [406] 6.950976e-01 7.033157e-01 2.182639e-03 7.111987e-01 3.130455e-02
#> [411] 3.869905e-05 3.361206e-01 3.264225e-01 4.021264e-01 8.898731e-03
#> [416] 2.097130e-02 3.142912e-02 1.487772e-01 4.943599e-02 5.576307e-02
#> [421] 3.936460e-02 1.634086e-01 1.447644e-02 6.516330e-03 1.945954e-03
#> [426] 5.290777e-03 1.167454e-01 4.137545e-02 8.922817e-03 5.553660e-03
#> [431] 2.066683e-02 1.801377e-01 2.664406e-02 6.047108e-01 1.104255e+00
#> [436] 1.355942e-01 3.509435e-02 3.312046e-02 2.709020e-01 5.515907e-01
#> [441] 1.352540e-02 6.385000e-01 8.502143e-01 7.018838e-02 5.615691e-02
#> [446] 8.509495e-01 5.457990e-01 8.004963e-01 1.762887e+00 8.014486e-01
#> [451] 6.930757e-04 4.732761e-01 1.109954e-03 4.466091e-01 9.544868e-01
#> [456] 6.956226e-01 3.974079e-01 8.365079e-01 7.570461e-01 7.289166e-02
#> [461] 2.594062e-01 6.037124e-01 6.929014e-02 5.063761e-02 1.668838e-01
#> [466] 8.452328e-05 2.336607e-01 1.399652e+00 1.386056e-01 8.169700e-02
#> [471] 4.308329e-02 5.133683e-01 1.579971e+00 7.717703e-01 1.546186e+00
#> [476] 1.015414e+00 1.714674e+00 1.285077e-01 4.992759e-01 2.753205e-01
#> [481] 4.821242e-01 3.232491e-02 4.039603e-01 8.363383e-01 3.878745e-01
#> [486] 1.205481e-01 1.510585e+00 2.134815e-01 2.325786e-03 8.471784e-02
#> [491] 1.672921e-02 7.108503e-01 6.467500e-03 1.229658e+00 3.944560e-01
#> [496] 2.605315e-01 5.070116e-01 4.257392e-01 7.959909e-02 2.148083e-01
#> [501] 8.359053e-01 3.243696e-01 6.316793e-02 2.392802e-06 3.776147e-02
#> [506] 1.086536e+00 7.461621e-01 1.642139e+00 9.687744e-01 1.063107e+00
#> [511] 5.304093e-01 1.139347e-01 1.041184e+00 3.145342e-01 2.321155e-02
#> [516] 7.585798e-02 4.429813e-01 1.111343e-01 1.064695e+00 8.271406e-01
#> [521] 1.102877e-01 4.238313e-04 2.630908e-01 2.187558e-01 1.516255e-01
#> [526] 1.302551e-02 4.212074e-01 2.999160e-01 2.993378e-02 4.765854e-02
#> [531] 2.450072e-03 1.289214e-01 7.102475e-01 6.226848e-01 1.935193e+00
#> [536] 1.897556e+00 9.591125e-01 5.165559e-01 9.057158e-01 6.180497e-01
#> [541] 2.746488e+00 2.397058e+00 2.943562e+00 2.274120e+00 5.704166e+00
#> [546] 2.154016e+00 2.755864e+00 1.182249e+00 1.274210e+00 1.638452e+00
#> [551] 8.265061e-01 1.653555e+00 9.211437e-01 1.598180e+00 1.399936e+00
#> [556] 1.525653e+00 4.121681e-01 1.412633e+00 1.323153e+00 3.243654e-01
#> [561] 1.258097e-01 3.023167e-01 5.310134e-05 9.858989e-01 8.385229e-01
#> [566] 5.778677e-01 8.623550e-01 1.515412e+00 2.500599e+00 1.810356e+00
#> [571] 1.962283e+00 5.771835e-01 1.230412e+00 2.188290e+00 6.675510e-01
#> [576] 1.553656e+00 7.616109e-01 7.586506e-01 1.843022e+00 5.223536e-01
#> [581] 6.636322e-01 5.315091e-01 2.061896e-01 6.029496e-01 4.933990e-01
#> [586] 2.819606e+00 2.767983e+00 4.554420e+00 1.975137e+00 3.559413e+00
#> [591] 1.836052e+00 3.735189e+00 1.957655e+00 3.674015e+00 3.719526e+00
#> [596] 1.505154e+00 1.188443e+00 6.977748e-01 1.584943e+00 3.154913e+00
#> [601] 5.654406e+00 2.211169e+00 1.614546e+00 5.338490e-01 1.254388e+00
#> [606] 7.310052e-01 2.451878e+00 1.579216e+00 1.133630e+00 1.210623e+00
#> [611] 1.056016e+00 9.866266e-01 2.487547e+00 1.483695e+00 1.382647e+00
#> [616] 3.558082e+00 3.917504e+00 2.614535e+00 3.292130e+00 2.431960e+00
#> [621] 2.161381e+00 2.057390e+00 4.609886e+00 3.286963e+00 1.288118e+00
#> [626] 2.055980e+00 1.515722e+00 3.318457e+00 3.599509e+00 3.347833e+00
#> [631] 3.023622e+00 2.083731e+00 3.820821e+00 5.005213e+00 4.520157e+00
#> [636] 3.001332e+00 1.447232e+00 1.267830e+00 1.624660e+00 2.791369e+00
#> [641] 2.722144e+00 3.464506e+00 3.471779e+00 3.943317e+00 2.857360e+00
#> [646] 3.678159e+00 2.873106e+00 2.676950e+00 4.408340e+00 3.151772e+00
#> [651] 1.426377e+00 1.171395e+00 3.133459e+00 2.553508e+00 5.028784e+00
#> [656] 3.275259e+00 2.802289e+00 3.607093e+00 2.393521e+00 1.904402e+00
#> [661] 6.012926e+00 5.074446e+00 6.067994e+00 6.534811e+00 5.420967e+00
#> [666] 2.263354e+00 4.918636e+00 5.450641e+00 5.343304e+00 5.109840e+00
#> [671] 5.027397e+00 6.714980e+00
#> 
#> $SPE
#>   [1] 0.0785098882 0.1553836503 0.1101395632 0.3646311603 0.1343461196
#>   [6] 0.0392751945 0.3077523926 0.0540291736 0.0538249739 0.5781686005
#>  [11] 0.3492858379 0.0898066781 0.1809008780 0.1987569805 0.4735468602
#>  [16] 0.4963021691 0.3992408700 0.3813501417 0.1488396074 0.1123715756
#>  [21] 0.3463147277 0.2637040447 0.6524375970 0.5149517483 0.3526842260
#>  [26] 1.1196095944 0.0522611413 0.2747520594 0.1792696476 0.5926706339
#>  [31] 0.1410069730 0.2132730369 0.6532882372 0.2608482825 0.1748058466
#>  [36] 0.8382881041 0.2744990058 0.9215894856 0.0602317764 0.0180229121
#>  [41] 0.1708520476 0.4905693312 0.1424967149 0.0900807471 0.6517856067
#>  [46] 0.2052333224 0.4321056470 0.0005624770 0.1060196304 0.0368901567
#>  [51] 0.0578537818 0.2034313501 0.8526311505 0.2628979455 0.0654201443
#>  [56] 0.2174584924 0.1955781853 0.2754370903 0.0649343184 0.5691605760
#>  [61] 0.2940680900 0.2514844223 1.1848529055 0.4476503922 0.2683403674
#>  [66] 0.0736332942 0.0679463751 0.3019557516 0.1461245894 0.2423093755
#>  [71] 0.0732741267 0.8553878002 0.0540222994 0.7131969385 0.4529226486
#>  [76] 0.0601305379 0.1139212832 0.1062338642 0.0434899702 0.0566379652
#>  [81] 0.0140571981 0.2606105452 0.0590338838 0.6232326649 0.1390217950
#>  [86] 0.1206618271 0.0272735557 0.0070239341 0.0397079287 0.1279298419
#>  [91] 0.1718974149 0.4752339220 1.1337696221 0.0005404816 0.0482857219
#>  [96] 0.4032892624 0.2695850204 0.4646534596 0.0759499606 0.1870479643
#> [101] 0.0940249791 0.2314434524 0.4187073235 0.0361077721 0.0019719681
#> [106] 0.0696107794 0.3922459065 0.6222550078 0.1741129292 0.2884397345
#> [111] 1.0097356030 1.3385939518 0.1315743674 0.1954305196 0.2282268907
#> [116] 0.3700912011 0.0714312643 0.3693302569 2.0956438173 0.1047276487
#> [121] 0.2632251439 0.0028371030 0.1013389912 0.1164528576 0.2831168608
#> [126] 0.4299690371 0.0102929035 0.1064948616 0.0334245673 0.1367489545
#> [131] 0.1965208453 0.0576609939 0.0728142939 0.0690699966 0.0055774798
#> [136] 0.1483010381 0.4962882840 0.1427409145 0.0004579461 0.2673571328
#> [141] 0.2123896255 0.8045320860 0.5311625034 0.0108923701 0.9833384421
#> [146] 0.3761896128 0.2875434857 0.0671578730 1.0919605658 0.2878302196
#> [151] 0.2673231487 0.6606392362 0.0441898271 0.6971686331 0.0333844841
#> [156] 0.0864830245 0.0225515553 0.7551596459 0.1907097732 0.9573786007
#> [161] 0.7842911350 0.0567711957 0.0476610228 0.1005215884 0.2020496299
#> [166] 0.0859904206 0.3609373799 0.1749963210 0.5032398850 0.3191921056
#> [171] 1.2744439889 0.2055929355 0.0303886273 0.1118916345 0.1475416871
#> [176] 0.4726776515 0.5698936668 1.1071546231 0.0961599066 0.7453158779
#> [181] 0.1805929740 0.3943223423 0.5960640688 0.1314287597 0.0329503743
#> [186] 0.0918417054 2.0483354450 0.5605818673 0.2054722239 0.2714416115
#> [191] 0.9491647170 0.2562786010 0.1286742654 0.2172581424 0.0615971241
#> [196] 1.1184748652 0.1192906890 0.3243780000 0.1381748543 0.1390336037
#> [201] 0.1080576739 0.1133375962 0.0213665510 0.7730676764 0.4614786798
#> [206] 0.3130801621 0.7111007872 0.0498364378 0.0018429599 0.1460577210
#> [211] 0.1462279417 0.6882445460 0.1571582491 0.0200936498 0.2873238093
#> [216] 0.6489275553 0.5498321914 0.0900111444 0.1520717060 0.1335330464
#> [221] 0.6866888269 0.0388189053 0.0920593627 0.0829704882 0.1092903684
#> [226] 0.0813515494 0.0862445073 0.1404487682 0.5027685160 0.8127181648
#> [231] 0.2239107270 0.4421550040 0.3805246258 0.0176108322 0.0017408956
#> [236] 0.2409034123 0.1619636814 0.2195615308 0.0967271309 0.2007242254
#> [241] 0.4163787758 0.4850507882 0.2455660551 0.1819293376 0.1201893971
#> [246] 0.8239744781 0.2955911676 0.4812429020 0.2416840715 1.3053105057
#> [251] 0.1703883913 0.0460386640 0.0137240986 0.0903323749 0.2133484686
#> [256] 0.7535637668 0.5937692812 0.1154001178 0.4874374793 0.1754050520
#> [261] 0.4182192319 0.1785430974 0.3751022286 0.0697626148 0.1669686675
#> [266] 1.1087835722 0.2979091135 0.1566706990 0.3043954908 0.0442421352
#> [271] 0.1468643014 0.1538392481 0.1251148308 0.2851722218 0.0404865858
#> [276] 0.2273029267 0.0169488899 0.0895158144 0.5647876806 0.0812023602
#> [281] 0.1091750885 0.1428828414 0.0426832019 0.1377042186 0.0592765093
#> [286] 0.0561327041 0.1498127920 0.1170579453 0.4954149095 0.2068198221
#> [291] 0.0569181587 0.0158394785 0.3503646708 0.1507935098 0.7433699570
#> [296] 0.5419000830 0.0273545781 0.0105578008 0.3146804136 0.2875988600
#> [301] 1.3691692175 0.0565689394 0.1277192759 0.0655699356 1.0441790970
#> [306] 0.0865053518 0.1627310549 0.2608699768 0.2535819725 1.0564216063
#> [311] 0.0921952781 0.3786423643 0.1305965836 0.2036616033 1.0057079093
#> [316] 0.1410054437 0.4925362543 0.0718184865 0.1451966834 0.6330910347
#> [321] 0.2078816188 0.0828092574 0.0536897418 0.2000799756 0.0143435397
#> [326] 0.1301427661 0.5273906722 0.2367568531 0.7917119207 0.1525741376
#> [331] 0.0790715362 0.1124046454 0.4260578302 0.3317583105 0.0132390251
#> [336] 0.0221368603 0.0906484807 0.9614072923 0.0314246099 0.0887342972
#> [341] 0.4658745142 0.2762994836 0.1012582906 0.8292653326 0.6149491532
#> [346] 0.3011299594 0.2258116579 0.1007640825 0.6946826453 0.3748844805
#> [351] 0.0720639057 1.5993501585 0.1120842045 0.0735386091 0.0480594215
#> [356] 0.1380550357 0.0695580653 0.1154276202 0.5707979258 0.1461853275
#> [361] 0.1250774039 0.0340552923 0.0895914851 0.0364447983 0.3348711374
#> [366] 0.1767560451 0.1304370765 0.0039000728 0.1848760155 1.3341227511
#> [371] 0.2111884212 0.0352959767 0.1342902678 0.0672500099 0.0404271934
#> [376] 0.1146590129 0.7200135167 0.2905785408 0.2470330207 0.0601870591
#> [381] 0.2659718233 0.2607678198 0.1514931843 0.3122069797 0.1707995727
#> [386] 0.1247135993 0.0847030687 0.1077010304 0.3595460805 0.1904389172
#> [391] 1.2765460322 0.0665989765 0.9776719363 0.0124387916 0.8932061961
#> [396] 0.0474456709 0.1775249479 1.0215236453 0.4133024545 0.0141139859
#> [401] 0.3121360105 0.0067389816 0.0817709522 0.2734285663 0.2569800419
#> [406] 0.4580471582 0.0091273077 0.2386106136 0.4316041699 0.3124795371
#> [411] 0.4842445552 0.3426913453 0.8317755292 0.2562420179 0.1527749650
#> [416] 0.9669026644 0.0802986332 0.6963491813 0.3285935513 0.0475294008
#> [421] 0.1501239022 0.0414249235 0.0515601957 0.6426797486 0.1687207505
#> [426] 0.1399722179 0.7065943818 0.3065553600 0.0606562112 0.1360159110
#> [431] 0.4050077807 0.0554547741 1.2922164191 0.2107607062 0.1348037404
#> [436] 0.2319007729 0.1815925300 0.0321857285 0.0305588388 0.4465829375
#> [441] 0.2412440444 0.0680018615 0.0321317163 0.0072926058 0.3969372831
#> [446] 0.2721722232 0.0117117217 0.0318751950 0.0922255960 0.0384149319
#> [451] 0.1549575692 0.3942644792 0.2479815740 0.1177627918 0.0871688921
#> [456] 0.0186537687 0.4337978080 0.1626492149 0.4649861323 1.1895854411
#> [461] 0.0398007525 0.2357306604 0.0850893023 0.0492570287 0.0112408497
#> [466] 0.2656279149 0.4029852626 0.3210910997 0.5257050831 1.0979842125
#> [471] 0.2513170385 0.1862907131 0.0452443767 0.0572872187 0.0929984779
#> [476] 0.4159626258 0.3095240348 0.3460760326 0.1100413938 0.1394311816
#> [481] 0.8763417072 1.2665357534 1.0264112739 0.1158929018 0.7527372818
#> [486] 0.0931679210 0.3012368086 0.3157387830 0.4117150131 0.3064254623
#> [491] 0.4752465852 0.2615420372 0.1641314750 0.1119266260 1.2973709292
#> [496] 0.4465261925 0.1471736589 0.1338477707 0.3542383353 0.1288256008
#> [501] 0.0906784280 0.1637217336 0.0166853852 0.1474467989 0.0943142162
#> [506] 0.4753879810 0.4947738144 0.1626979734 0.1726681157 0.1593560654
#> [511] 0.0682672897 1.0749703935 0.1972566632 0.1539888148 0.0723430367
#> [516] 0.1966920344 0.1786962540 0.0548013229 0.2467395489 0.1206924892
#> [521] 0.7097121044 0.2270120146 0.3855043415 0.1660106585 0.1084933492
#> [526] 0.4272856947 0.5583167922 0.1526477738 2.2319720565 0.0659845231
#> [531] 0.1875639734 0.6611667710 0.1222994020 0.1688241542 0.2940644032
#> [536] 0.2691742480 0.5625923433 0.2740012821 0.0368909618 0.2209317932
#> [541] 0.0101735809 0.0204969028 0.0990622559 0.4213524327 0.2645828391
#> [546] 0.0133173865 0.0228008393 0.4726477376 0.6219215138 0.0731504158
#> [551] 0.5743628636 0.2072859929 0.0949055005 0.0414092824 0.0789050569
#> [556] 0.1049059920 0.3509623313 0.2075272208 0.1535083182 0.0537814311
#> [561] 0.6002191919 0.3306246731 0.3738728338 0.0440131660 0.7038292834
#> [566] 0.9470367229 0.5249497314 0.1824915117 0.0613576761 0.4297494876
#> [571] 0.5064514080 0.3039448746 0.1308689977 0.2017678410 0.0014508670
#> [576] 0.3206083980 0.2477366631 0.0052212949 0.3832097000 0.1325883617
#> [581] 0.1324070426 0.1474542339 0.1098781600 0.3227387567 0.1622094893
#> [586] 0.1411711450 0.2778222674 0.3020352447 0.0271187118 0.0130782690
#> [591] 0.0509924663 0.4671870640 0.1629653172 0.3298815034 0.1856600187
#> [596] 0.0095835440 0.0182218005 0.1322148345 0.0211633983 0.1414819620
#> [601] 0.7437040352 0.9977280615 0.0230144535 0.0780226585 0.3893349122
#> [606] 1.1203519069 0.2145154976 0.0522294981 0.0343621526 0.1644996220
#> [611] 0.4646872806 0.3436914730 0.1714508135 0.1018808344 0.2381532263
#> [616] 0.5461005253 0.0358521471 0.0098179560 0.1873636206 0.6003525305
#> [621] 0.0275321568 0.1418412048 0.0975344076 0.1291327152 0.2424992382
#> [626] 0.1905189682 0.6749442855 0.2164874019 0.4376096644 0.1852835595
#> [631] 0.0208497989 0.0501529602 0.5492849851 0.2319199306 1.0371568216
#> [636] 0.0180737023 0.1254194796 0.1717874758 0.2838432020 0.0150516945
#> [641] 0.0545856786 0.0954223239 0.2707191426 0.0052581222 0.0250884123
#> [646] 0.5703295713 0.3916356158 1.0822024757 0.0062919838 0.2757766805
#> [651] 0.1584457400 0.6158000254 0.0269976747 0.0568604922 1.1056638973
#> [656] 0.0681749857 0.6556681625 0.3883169806 0.0103969162 0.2103671415
#> [661] 0.8587447342 0.1691300016 0.1463391956 0.0155023853 2.4147203076
#> [666] 0.0060548767 0.1332901864 0.4312738742 0.3581653944 0.2963465056
#> [671] 0.6081862333 0.3134971616
#> 
#> $muTrain
#>          x          y          z 
#>  0.3958936 -1.0042242  0.4559867 
#> 
#> $RootPrecisTrain
#>          [,1]     [,2]     [,3]
#> [1,] 4.988573 0.000000 0.000000
#> [2,] 0.000000 2.721066 0.000000
#> [3,] 0.000000 0.000000 2.898963
#> 
#> attr(,"class")
#> [1] "threshold" "pca"      
#> 
#> attr(,"class")
#> [1] "fault_ls"