Process Fault Filtering

Flag and filter out observations beyond normal operating conditions, then return the observations within normal operating conditions.

faultFilter(trainData, testData, updateFreq, faultsToTriggerAlarm = 5, ...)

Arguments

trainData

An xts data matrix of initial training observations

testData

The data not included in the training data set

updateFreq

The number of observations from the test data matrix that must be returned to update the training data matrix and move it forward.

faultsToTriggerAlarm

Specifies how many sequential faults will cause an alarm to trigger. Defaults to 5.

...

Lazy dots for additional internal arguments

Value

A list of class "fault_ls" with the following:

faultObj -- An xts flagging matrix with the same number of rows as "testData". This flag matrix has the following five columns:

SPE -- The SPE statistic value for each observation in "testData". This statistic is defined as $$ SPE_i = (\textbf{X}_i - \textbf{Y}_i * \textbf{P}^T) * (\textbf{X}_i - \textbf{Y}_i * \textbf{P}^T)^T, $$ where \(\textbf{X}_i\) is the \(i^{th}\) observation vector, \(\textbf{Y}_i\) is the reduced-feature projection of the observation \(\textbf{X}_i\), and \(\textbf{P}\) is the projection matrix such that \(\textbf{X}_i\textbf{P} = \textbf{Y}_i\).

SPE_Flag -- A vector of SPE indicators recording 0 if the test statistic is less than or equal to the critical value passed through from the threshold object.

T2 -- The T2 statistic value for each observation in "testData". This statistic is defined as $$ T^2_i = \textbf{Y}_i * \textbf{D}^{-1} * \textbf{Y}_i^T, $$ where \(\textbf{Y}_i = \textbf{X}_i\textbf{P}\) is the reduced- feature projection of the observation \(\textbf{X}_i\), and \(\textbf{D}\) is the diagonal matrix of eigenvalues.

T2_Flag -- A vector of T2 fault indicators, defined like SPE_Flag.

Alarm -- A column indicating if there have been five flags in a row for either the SPE or T2 monitoring statistics or both. Alarm states are as follows: 0 = no alarm, 1 = Hotelling's T2 alarm, 2 = Squared Prediction Error alarm, and 3 = both alarms.

nonAlarmedTestObs -- An xts matrix of the first updateFreq number of rows of the training data which were not alarmed.

trainSpecs -- The threshold object returned by the internal threshold() function. See the threshold() function's help file for more details.

Details

This function is essentially a wrapper function to call and organize the output from these other internal functions: faultDetect(), threshold(), and pca(). It is applied over a rolling window, with observation width equal to updateFreq, of the larger full data matrix via the processMonitor() function, wherein the testing and training data sets move forward in time across the entire data matrix.

This internal function is called by processMonitor().

See also

Examples

nrml <- mspProcessData(faults = "NOC")
# Select the data under state 1
data <- nrml[nrml[,1] == 1]

faultFilter(trainData = data[1:672, -1],
            testData = data[673:3360, -1],
            updateFreq = 336)
#> Warning: In density.default(spe, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(t2, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> $faultObj
#>                            SPE SPE_Flag        T2 T2_Flag Alarm
#> 2015-05-17 19:12:00 0.29086880        0 4.3858902       0     0
#> 2015-05-17 19:13:00 0.29774156        0 3.6585249       0     0
#> 2015-05-17 19:14:00 0.19893653        0 3.4264311       0     0
#> 2015-05-17 19:15:00 0.09379590        0 3.6034199       0     0
#> 2015-05-17 19:16:00 0.63025051        0 4.6187686       0     0
#> 2015-05-17 19:17:00 0.51853316        0 6.3960901       0     0
#> 2015-05-17 19:18:00 0.28789758        0 7.2939946       1     0
#> 2015-05-17 19:19:00 0.21015939        0 8.7081985       1     0
#> 2015-05-17 19:20:00 0.28702258        0 7.2339339       1     0
#> 2015-05-17 19:21:00 0.74784522        0 4.0292617       0     0
#>                 ...                                            
#> 2015-05-23 07:50:00 0.03693008        0 0.3047079       0     0
#> 2015-05-23 07:51:00 0.06074398        0 0.4672710       0     0
#> 2015-05-23 07:52:00 1.03738977        0 0.5338110       0     0
#> 2015-05-23 07:53:00 0.01691076        0 0.9691651       0     0
#> 2015-05-23 07:54:00 0.04086803        0 1.1263796       0     0
#> 2015-05-23 07:55:00 0.16346131        0 1.7325452       0     0
#> 2015-05-23 07:56:00 0.05522202        0 1.8232238       0     0
#> 2015-05-23 07:57:00 0.25271214        0 2.0105645       0     0
#> 2015-05-23 07:58:00 0.74048845        0 2.6418004       0     0
#> 2015-05-23 07:59:00 0.28590256        0 1.1445834       0     0
#> 
#> $nonAlarmedTestObs
#>                             x         y         z
#> 2015-05-17 19:12:00 0.8605188 -1.631811 1.2696867
#> 2015-05-17 19:13:00 0.7359875 -1.615228 1.3156059
#> 2015-05-17 19:14:00 0.7120464 -1.643703 1.2661628
#> 2015-05-17 19:15:00 0.7638631 -1.808493 1.0671185
#> 2015-05-17 19:16:00 0.9150534 -1.586458 1.2737103
#> 2015-05-17 19:17:00 0.8886312 -1.758097 1.5571639
#> 2015-05-17 19:18:00 0.8486165 -1.989691 1.5886541
#> 2015-05-17 19:19:00 0.9876192 -1.977999 1.6162074
#> 2015-05-17 19:20:00 0.8732797 -1.927404 1.5923600
#> 2015-05-17 19:21:00 0.6450746 -1.822960 1.3790341
#>                 ...                              
#> 2015-05-22 01:53:00 0.7615198 -1.568808 1.4480655
#> 2015-05-22 01:54:00 0.8239800 -1.591415 1.4070706
#> 2015-05-22 01:55:00 0.7502797 -1.897920 1.2827102
#> 2015-05-22 01:56:00 0.7614805 -1.730645 1.2326326
#> 2015-05-22 01:57:00 0.7164834 -1.610266 1.2095013
#> 2015-05-22 01:58:00 0.5421986 -1.623719 0.9535691
#> 2015-05-22 01:59:00 0.6312398 -1.622723 1.0299424
#> 2015-05-22 04:00:00 0.7070070 -1.501437 0.8858914
#> 2015-05-22 04:01:00 0.5957601 -1.582648 0.9969890
#> 2015-05-22 04:02:00 0.7116615 -1.589539 1.0127178
#> 
#> $trainSpecs
#> $SPE_threshold
#>    99.9% 
#> 1.901201 
#> 
#> $T2_threshold
#>    99.9% 
#> 7.085542 
#> 
#> $projectionMatrix
#>            [,1]
#> [1,]  0.5668819
#> [2,] -0.5828616
#> [3,]  0.5821660
#> 
#> $LambdaInv
#>           [,1]
#> [1,] 0.3671282
#> 
#> $T2
#>   [1] 6.149144e-01 3.309311e-03 4.560181e-01 1.871118e-01 9.156986e-01
#>   [6] 6.071631e-01 3.371082e-01 5.251052e-01 6.947958e-02 1.052281e-02
#>  [11] 8.272992e-01 1.313216e+00 1.253174e+00 1.408350e+00 1.173487e+00
#>  [16] 3.857498e+00 2.944363e+00 2.951126e+00 1.340675e+00 1.716901e+00
#>  [21] 1.946051e+00 2.129347e+00 2.851937e+00 1.885773e+00 1.847876e+00
#>  [26] 1.995289e+00 2.390462e+00 1.302842e+00 9.780833e-01 3.490404e-01
#>  [31] 1.007928e+00 6.156492e-01 1.118174e+00 2.090830e+00 8.709526e-01
#>  [36] 3.703330e-01 7.900172e-01 1.256562e+00 1.115406e+00 1.923260e+00
#>  [41] 2.206121e+00 1.085021e+00 9.452858e-01 8.218028e-01 1.237123e+00
#>  [46] 1.622819e+00 1.204853e+00 6.172626e-01 3.197839e-01 1.113026e+00
#>  [51] 3.760312e-01 6.097023e-01 8.712974e-01 5.181309e-01 4.350591e-01
#>  [56] 8.209693e-01 4.825903e-01 3.831575e-01 3.990502e-01 8.888486e-01
#>  [61] 6.179813e-03 1.937840e-01 2.195463e-01 7.185544e-01 1.231991e+00
#>  [66] 1.139976e+00 8.716956e-01 7.559263e-01 1.450026e+00 3.481964e-01
#>  [71] 6.784416e-03 7.666186e-01 3.180959e-01 1.265716e+00 8.171757e-01
#>  [76] 7.644848e-03 3.359014e-02 7.983046e-02 1.283988e-04 1.193848e-01
#>  [81] 3.638112e-01 3.711513e-02 3.993010e-02 1.004492e+00 2.096444e+00
#>  [86] 7.969902e-01 1.760722e+00 2.838727e+00 1.566556e+00 1.607054e+00
#>  [91] 1.264387e+00 2.108757e+00 5.125554e-01 6.302816e-01 4.148022e-01
#>  [96] 2.589337e-01 4.296419e-01 9.069752e-01 3.994107e-01 2.467858e-01
#> [101] 5.543999e-01 1.283826e-01 7.418423e-01 1.126374e+00 1.102956e+00
#> [106] 1.671562e-01 1.701103e-03 6.382017e-01 6.696914e-04 3.327843e-01
#> [111] 1.658233e+00 1.082998e-01 7.599397e-01 2.429521e+00 3.553521e+00
#> [116] 3.987532e+00 3.198844e+00 1.453696e+00 2.566877e+00 8.456823e-01
#> [121] 7.794829e-01 2.431335e-01 1.654287e-01 2.847040e-01 3.932250e-01
#> [126] 5.060061e-01 3.028733e-01 1.043179e-02 4.711762e-04 2.463626e-01
#> [131] 3.030780e-02 1.283639e-01 1.911569e-01 1.913033e-01 5.945414e-01
#> [136] 5.810484e-01 5.898170e-01 6.488233e-01 1.407014e+00 5.145990e-01
#> [141] 6.037995e-01 1.742214e+00 1.743807e+00 6.010002e-01 5.114983e-02
#> [146] 2.109196e-01 1.114315e+00 1.229503e-01 7.471206e-01 5.686867e-01
#> [151] 2.730158e+00 1.160967e+00 7.814308e-01 6.274517e-01 1.220456e+00
#> [156] 2.515282e-01 8.302148e-01 1.219594e+00 2.014383e+00 1.460899e+00
#> [161] 1.189900e+00 5.851908e-01 2.837840e-01 1.319999e+00 7.019749e-01
#> [166] 4.183928e-01 1.097558e+00 6.730289e-01 6.104882e-01 1.327645e+00
#> [171] 1.064938e+00 9.796157e-01 3.817377e-01 4.766242e-01 2.430527e-01
#> [176] 1.615558e-02 3.709872e-06 4.262940e-01 4.407461e-01 6.105800e-01
#> [181] 1.834318e+00 2.859393e+00 3.466168e+00 2.059565e+00 1.400990e+00
#> [186] 2.019564e+00 1.418070e+00 2.954299e-01 9.536756e-01 5.810167e-01
#> [191] 4.787751e-01 6.496470e-01 1.426933e+00 6.062870e-01 4.885165e-01
#> [196] 8.328192e-02 1.158013e+00 6.501381e-01 6.889627e-01 8.720827e-01
#> [201] 6.918445e-01 7.516499e-01 8.873868e-01 5.713734e-01 3.428620e-01
#> [206] 1.948923e-01 4.510875e-01 9.113948e-01 1.590286e+00 2.129492e+00
#> [211] 1.759303e+00 1.034059e+00 9.575070e-01 9.601459e-01 6.753386e-01
#> [216] 7.764022e-01 9.619040e-01 1.311794e-07 1.750479e-02 1.870794e-01
#> [221] 1.245027e-01 6.850282e-01 4.570091e-01 8.455514e-03 4.108823e-02
#> [226] 1.035886e-01 5.205107e-01 5.630295e-02 2.634861e-02 1.519007e-02
#> [231] 1.076585e-01 2.238668e-01 3.473433e-01 9.685394e-01 1.606830e+00
#> [236] 9.745286e-01 1.031519e+00 8.200069e-01 5.089254e-01 1.189858e+00
#> [241] 6.329323e-01 2.312293e+00 1.689328e+00 3.251460e-01 6.584771e-01
#> [246] 4.910815e-01 2.585812e-01 2.778086e-01 6.752317e-04 1.040174e-02
#> [251] 1.269217e-01 3.014597e-01 2.439069e-02 5.243859e-01 9.993678e-01
#> [256] 1.333527e+00 2.343227e+00 1.092266e+00 8.949474e-02 6.872261e-03
#> [261] 1.911686e-01 1.279201e-02 3.400866e-02 5.516345e-02 5.838382e-03
#> [266] 1.070990e-01 1.471796e-01 3.758104e-01 3.999089e-01 6.150973e-01
#> [271] 1.242941e-01 2.062823e+00 1.199179e+00 1.845257e+00 1.456669e+00
#> [276] 2.489010e+00 7.183500e-01 9.902449e-01 8.779601e-01 4.446346e-01
#> [281] 8.676247e-01 2.304024e-03 1.615615e-05 1.578737e-01 1.245342e-01
#> [286] 1.717298e-01 1.335026e-01 3.345850e-01 2.512927e-02 1.430051e-01
#> [291] 1.766990e-01 1.706065e-01 4.025330e-01 1.544168e-02 9.759061e-05
#> [296] 6.322961e-01 5.571377e-01 4.213132e-01 1.910491e-01 7.747682e-01
#> [301] 3.018114e-01 1.452342e-01 1.039446e-01 4.392625e-01 2.286608e-03
#> [306] 1.054684e-03 1.108055e-02 4.751544e-01 1.701813e+00 3.643297e-01
#> [311] 4.873139e-02 2.450712e-01 1.347982e+00 1.027168e+00 1.523807e+00
#> [316] 5.813033e-01 1.040274e+00 1.203150e+00 1.063718e+00 5.564130e-01
#> [321] 3.471514e-01 9.190554e-01 6.302727e-01 3.571316e-01 5.184083e-01
#> [326] 6.200680e-01 2.138071e-01 7.373152e-02 1.709251e-01 1.882329e-01
#> [331] 7.596995e-02 1.405570e-01 8.617484e-01 3.347002e-01 1.374286e-01
#> [336] 6.887043e-01 2.418911e-02 1.111910e-01 5.266721e-01 1.402341e-01
#> [341] 5.483723e-02 2.318085e-01 1.781520e-01 1.444381e-02 1.124539e-01
#> [346] 4.165136e-01 7.586738e-01 5.489363e-01 8.016844e-01 2.988441e-01
#> [351] 6.967300e-03 2.749529e-03 2.570913e-03 2.071278e-01 8.221024e-02
#> [356] 1.792616e-03 1.862407e-01 6.465129e-01 3.837843e-01 2.517901e-02
#> [361] 3.839917e-01 1.961833e-01 6.898884e-01 2.743379e-01 5.167553e-01
#> [366] 2.283135e-03 1.994869e-04 8.942630e-03 1.557260e-01 1.124908e-02
#> [371] 2.347126e-02 5.284768e-03 4.939143e-02 1.897905e-01 1.353947e-03
#> [376] 1.254406e-01 2.221468e-02 3.869440e-04 4.399604e-01 5.037802e-01
#> [381] 6.790908e-02 8.931579e-02 2.130498e-01 5.162847e-01 1.544265e-01
#> [386] 4.232723e-02 2.585324e-04 2.647096e-03 8.583511e-01 8.627667e-01
#> [391] 1.621347e-01 4.743111e-01 1.914905e-03 2.403000e-01 6.313470e-01
#> [396] 9.690881e-01 4.854840e-01 1.225534e+00 3.990992e-01 1.992026e-01
#> [401] 2.892202e-01 2.340786e-01 5.720469e-02 6.566623e-02 1.145496e-01
#> [406] 8.378750e-02 1.501521e-01 4.290771e-01 4.274419e-01 2.062959e-01
#> [411] 1.787063e-01 7.171875e-04 1.308400e-02 3.132854e-02 5.816747e-03
#> [416] 1.239712e-01 2.765606e-01 6.249362e-02 3.044162e-01 2.764405e-01
#> [421] 5.738418e-01 2.336195e-02 3.393169e-02 2.116321e-02 2.287707e-01
#> [426] 2.079831e-01 4.750000e-02 4.149119e-01 1.851239e-01 3.933830e-02
#> [431] 2.119387e-01 2.107377e-01 5.165369e-02 1.366479e-01 5.248771e-02
#> [436] 1.773894e-01 3.349723e-03 3.801316e-01 7.780958e-02 2.198802e-01
#> [441] 6.406453e-01 1.563181e-02 1.044079e-01 4.766587e-02 3.437111e-07
#> [446] 2.642932e-01 2.198895e-01 1.194240e-01 3.171110e-02 1.787845e-03
#> [451] 1.595444e-01 1.586751e-02 2.000692e-01 4.665799e-01 1.233439e+00
#> [456] 1.323809e+00 5.183579e-01 2.981328e-01 1.292081e-01 2.479562e-01
#> [461] 1.416380e-01 5.681333e-02 1.462120e+00 1.500313e-01 4.067369e-01
#> [466] 3.281255e-01 2.816378e-01 1.321925e-02 6.329649e-01 2.362219e-01
#> [471] 3.542826e-04 1.425279e-01 8.326387e-06 1.091364e-01 1.663654e-01
#> [476] 3.082456e-02 3.234088e-02 2.098194e-01 1.329823e-01 1.497518e-01
#> [481] 7.000100e-01 1.491126e+00 9.032689e-01 5.556094e-01 1.480839e-01
#> [486] 7.989562e-01 6.884883e-01 2.315791e-01 3.576017e-01 8.966338e-01
#> [491] 7.229556e-01 6.082878e-01 3.303268e+00 1.604237e+00 7.193066e-01
#> [496] 1.378173e+00 5.740333e-01 8.436148e-01 1.517283e+00 5.947307e-01
#> [501] 6.960417e-01 1.523323e+00 1.302694e+00 3.598621e-01 9.417150e-01
#> [506] 7.887998e-01 1.254362e+00 1.156720e+00 1.336846e+00 6.571263e-01
#> [511] 4.683575e-01 4.051730e-01 1.420846e-01 6.025706e-01 5.029451e-01
#> [516] 4.920007e-01 1.112292e+00 2.090696e+00 1.094495e+00 1.285968e+00
#> [521] 1.578850e-01 1.125807e-02 4.922186e-01 1.341977e+00 8.193037e-01
#> [526] 7.245309e-01 1.826910e+00 1.453460e+00 4.453295e-01 3.772543e-01
#> [531] 1.365922e-01 6.420780e-01 1.309654e+00 1.339801e+00 7.187654e-01
#> [536] 3.859750e-01 3.968268e-01 4.096378e-02 5.934181e-01 1.936592e-02
#> [541] 4.310530e-01 1.273424e+00 9.459488e-01 6.279474e-01 3.510583e-01
#> [546] 4.639027e-01 1.477019e+00 1.174194e+00 2.528650e+00 3.350774e+00
#> [551] 4.602937e+00 2.988489e+00 2.247486e+00 3.559216e+00 1.771102e+00
#> [556] 9.034346e-01 2.265326e+00 2.722366e+00 8.935110e-01 2.025167e+00
#> [561] 2.570602e+00 4.680858e-01 1.075387e-04 6.952441e-02 7.694460e-01
#> [566] 1.667224e+00 1.084794e+00 9.223480e-01 1.334765e+00 1.116494e+00
#> [571] 1.303143e+00 9.568529e-01 2.293809e+00 1.775337e+00 4.760068e-01
#> [576] 1.467641e+00 1.792813e+00 2.796328e+00 6.908559e-01 7.219573e-01
#> [581] 3.075433e-01 2.103568e+00 5.397859e+00 3.108872e+00 2.100485e+00
#> [586] 3.019239e+00 2.982611e+00 6.243558e-01 1.494521e+00 1.877462e+00
#> [591] 1.545891e+00 2.348969e+00 2.059795e+00 1.238854e+00 1.491738e+00
#> [596] 1.165434e+00 2.322781e+00 2.879342e+00 1.766323e+00 1.983200e+00
#> [601] 2.664654e+00 3.028553e+00 2.104900e+00 1.626503e+00 1.071690e+00
#> [606] 1.532791e+00 2.915867e-01 3.545073e-01 1.591378e+00 2.947028e+00
#> [611] 1.274825e+00 4.357792e-01 7.996028e-01 1.586078e+00 5.726309e-01
#> [616] 1.792904e+00 2.879341e+00 8.191996e-01 3.064709e+00 5.961352e+00
#> [621] 1.831742e+00 2.017583e+00 2.995893e+00 5.672502e+00 2.935956e+00
#> [626] 4.237202e+00 4.604629e+00 5.473507e+00 7.016387e+00 3.458992e+00
#> [631] 3.282345e+00 2.985770e+00 3.818422e+00 3.828148e+00 1.037037e+00
#> [636] 3.732784e+00 9.278979e-01 2.081955e+00 5.239099e-02 2.594552e+00
#> [641] 3.224661e+00 2.963977e+00 2.030904e+00 2.016899e+00 1.162525e+00
#> [646] 1.895692e+00 2.455563e+00 1.665732e+00 1.744011e+00 2.695390e+00
#> [651] 3.984931e+00 4.525632e+00 4.553379e+00 2.559259e+00 5.517966e+00
#> [656] 6.908382e+00 3.654676e+00 2.307891e+00 2.252777e+00 3.053576e+00
#> [661] 4.070367e+00 2.028582e+00 5.934423e+00 5.103880e+00 3.738981e+00
#> [666] 2.997191e+00 3.804849e+00 4.657281e+00 6.208788e+00 6.244071e+00
#> [671] 5.044030e+00 7.079761e+00
#> 
#> $SPE
#>   [1] 0.3186121667 0.2700186886 1.7318184870 0.1788704656 0.1576808492
#>   [6] 0.3214218139 0.3467891365 0.2621219008 0.2694911846 0.2143139668
#>  [11] 0.1629880537 0.2255234714 0.2989102438 0.3145770765 0.0534613616
#>  [16] 0.4243640508 0.2447735584 0.2941006225 0.0985651748 0.1212732385
#>  [21] 0.1241227671 0.2330459558 0.8045697949 0.0279058584 0.2876656080
#>  [26] 0.9230701964 0.0551183340 0.0455591192 0.1480760931 0.0207334574
#>  [31] 0.4600410212 0.8060744353 0.0320597674 0.0918652533 0.1409951610
#>  [36] 0.3951149976 0.0563449126 0.2621473253 0.1878872109 0.5183488771
#>  [41] 0.6112765489 0.0345649231 0.1008217405 0.1447108599 0.1214408100
#>  [46] 0.1763717344 0.6522774673 0.2581184643 0.5173551407 0.9529687390
#>  [51] 0.0826794408 0.2052954686 1.0054296523 0.0579659459 0.1013123331
#>  [56] 0.2191880893 0.6635984593 0.1118633785 0.1902220203 0.4690899524
#>  [61] 0.1050899729 0.0296494334 0.0007100642 0.2676576709 0.0903470707
#>  [66] 0.0634572923 0.0418119314 0.3872804178 0.2281523794 0.0480954558
#>  [71] 0.2754454414 0.1761327679 0.3387783807 0.0070709525 1.1161265619
#>  [76] 0.3246973002 0.1747635573 0.5555318990 0.1518238190 0.1143308934
#>  [81] 0.2998293317 0.0085034642 0.0336504498 0.1354017606 0.4293628537
#>  [86] 0.2654448620 0.1950662834 0.0418122426 0.1293470572 0.5728880212
#>  [91] 0.0256217871 0.5494530821 0.2707635902 0.3165539297 0.1958601789
#>  [96] 0.2938590286 0.0219520656 0.4088021476 0.0182590974 0.7071169670
#> [101] 0.2661013757 0.0050789859 0.0986538305 0.1450669721 0.2444272856
#> [106] 0.1124959081 0.0557846739 0.1827179753 0.2326186562 0.0116523434
#> [111] 0.1678261697 1.1562592242 0.0762049636 0.2403318277 0.4925611920
#> [116] 0.2224758844 0.3012120308 0.2618827784 0.1517486546 0.0047290017
#> [121] 0.0422617098 0.1254039295 0.3327216445 0.4514303743 0.3514378188
#> [126] 0.0544284961 0.1380536027 0.0424002828 0.0560406337 0.4644525665
#> [131] 0.4435212537 0.0328618867 0.1406625816 0.3619128244 0.3595726232
#> [136] 0.1279841825 0.0637744443 0.0094551312 0.3224494024 0.1052967566
#> [141] 1.0124928678 0.1379762764 0.6167192629 0.0018804486 0.0399651839
#> [146] 0.4491520698 0.3557776224 0.0208904206 0.0735480144 0.2402358687
#> [151] 1.4009620317 0.5268219329 0.0037077506 0.2580078049 0.3177918373
#> [156] 0.3063258093 0.4962941011 0.2864111107 0.9094261731 0.3039424423
#> [161] 0.4082165831 0.1429369060 0.0368817205 0.1864369413 0.1331081925
#> [166] 0.0103722780 0.3687920705 0.0125853386 0.2736169867 0.2034100763
#> [171] 0.1019334911 0.0459751104 0.2773064801 0.3544472269 0.2112939216
#> [176] 0.0306980458 0.6852724665 0.1947343941 0.0510556382 0.0819331658
#> [181] 0.0326075272 1.0454799531 0.4469969507 0.2109506827 0.1513220911
#> [186] 0.0714751151 0.0106893347 0.2611015089 0.1195765011 0.0011639488
#> [191] 0.8363059363 0.0481412524 0.0283463126 0.0379470551 0.4386483920
#> [196] 0.1018005059 0.1781810165 0.4684569835 0.0727193628 0.3623103972
#> [201] 0.4418825022 0.5481873729 0.0824202806 0.1093052916 0.3111071076
#> [206] 0.6821131198 0.2631265595 0.0507560088 0.1246752290 0.1277042497
#> [211] 0.2797125012 0.2462393457 0.1079767335 0.1841041352 0.1278078197
#> [216] 0.1190187087 0.0858780023 0.1310968133 0.1377376628 0.7941756170
#> [221] 0.1961573709 0.1258948082 0.4903183695 0.1229927543 0.0707281909
#> [226] 1.3076895286 0.1351064774 0.0579600782 0.0739360116 0.3364291079
#> [231] 0.0068369897 0.1069010644 0.0608813810 0.0382598791 0.7539068612
#> [236] 0.0334471267 0.2355450911 0.9013871458 0.0651654452 0.0278256392
#> [241] 0.2288444017 0.1077847087 0.5272494372 0.0084820326 0.0535199291
#> [246] 0.2080849217 0.4924586307 0.1256068772 0.4343429666 0.0885248122
#> [251] 0.2922248870 0.1898771366 1.9111736601 0.6517398878 0.0349443881
#> [256] 0.0570778241 0.2466996155 0.1110003622 0.3889120226 0.0069732631
#> [261] 0.0122021730 0.6156311408 0.3005205355 0.1735375994 0.1129072606
#> [266] 0.0806255173 0.2917544054 1.5272266100 0.6251455349 0.1791278418
#> [271] 0.5420105460 0.4731163318 0.0703100481 0.5787221193 0.3010303181
#> [276] 0.4847308702 0.4615711743 0.0657964171 0.4483235812 0.1842948892
#> [281] 0.0740697735 0.1778033797 0.2341211036 0.6825216213 0.2165805482
#> [286] 0.0106183115 0.3474932640 0.2526853053 0.0322710421 0.0980563534
#> [291] 0.0586723415 0.0099057490 0.0343762165 0.0136120262 0.1140671898
#> [296] 0.3240529546 0.2652268998 0.1663816005 0.2531006501 0.3534811892
#> [301] 0.3302669734 0.0668929404 0.1866230559 0.0311889705 0.0619728144
#> [306] 0.2314688038 0.0018714257 0.4379691479 0.0052891686 0.1959300055
#> [311] 0.4568374938 0.0267445283 0.6492669200 0.0629468460 0.2164156793
#> [316] 0.0143474046 0.0698012560 0.3961288812 0.1645425621 0.0358890439
#> [321] 0.1971691956 0.2962273701 0.0512524193 0.5608304106 0.3993536325
#> [326] 0.0020743310 0.1239411161 0.0024191526 0.0027587842 0.5111851153
#> [331] 0.0421039908 0.1013600438 0.2401078448 0.0217637494 0.3815663109
#> [336] 0.2479358618 0.0727336536 0.1533839959 0.3186740658 1.2574657026
#> [341] 0.2912246341 0.0921666103 0.1796247813 0.2229918945 0.2780751272
#> [346] 0.0465820985 0.1411618274 0.0290263415 0.2347902919 0.0976229711
#> [351] 0.0283700388 0.0927445013 0.1365029622 0.8038968435 0.1998903662
#> [356] 0.4521535421 0.5533660405 0.1477483039 0.4218540588 0.3066162974
#> [361] 0.0741273426 0.0088645002 0.7366198967 0.2140949583 0.0075743463
#> [366] 0.2791408434 0.0729979351 0.1629805801 0.1383112350 0.0775010555
#> [371] 0.1651323340 0.0594983782 0.0993084210 0.0188610920 0.0396302875
#> [376] 0.0569949900 0.4768325189 0.0200694968 0.1126431094 0.2064054979
#> [381] 0.2736178733 0.4680139502 0.7983389862 0.1140682582 0.3471190062
#> [386] 0.2663883412 0.5205431114 0.0535276258 0.2878105120 0.1398816359
#> [391] 0.0012221884 0.0806674432 0.2099580743 0.1166406028 0.9269195712
#> [396] 0.7123142910 0.1470726008 0.4039899602 0.5011740844 0.0394147089
#> [401] 0.2493737620 0.0897067266 0.3042937516 0.0205010798 0.8194906687
#> [406] 0.6977330420 0.3079354962 0.1293786647 0.3942681279 0.1205791603
#> [411] 0.5262746989 0.2959723671 0.0894768523 0.0794688706 0.3317213102
#> [416] 1.8135615975 0.2713088056 0.2262263151 0.1278848348 0.8412503274
#> [421] 0.0046263556 0.0839475975 0.2622381813 0.1086336801 1.2470058268
#> [426] 1.0107032493 0.0117951513 0.1363447354 0.1833417495 0.2277101753
#> [431] 0.1466574654 0.4705458340 0.3581244069 0.7611073937 0.0766222903
#> [436] 0.2166566939 1.2564509154 0.0229571775 0.1324437922 0.1977692650
#> [441] 0.1747233063 0.2348691513 0.0785247127 0.3570536405 0.0101120402
#> [446] 0.6203791579 0.1988824507 0.1477515108 0.6817894334 0.4657557155
#> [451] 0.3115569191 0.2848434966 0.6388364637 0.3313234487 0.1808212034
#> [456] 0.0677672629 1.0121486319 0.0632863556 0.1861139515 0.0337600421
#> [461] 0.1322450454 0.0021689478 0.1906276813 0.6143470729 0.0306695410
#> [466] 0.0600106420 0.2527173529 0.0515146213 0.1984864867 0.6562905866
#> [471] 0.3878026207 0.0591452509 0.2672453212 0.6570917609 0.2252811121
#> [476] 0.0602234836 0.3188465191 0.1351678986 0.5175928591 0.6967310317
#> [481] 0.0063220740 0.0981006111 0.2380387364 0.1876592421 0.1227085848
#> [486] 0.1109578257 0.7883312299 0.0868179165 0.3348729346 0.0261494003
#> [491] 0.2428413251 0.0652101791 0.6530146083 0.2938901242 0.1374971850
#> [496] 0.2865725688 0.4444552954 0.2096722751 0.4391412721 0.0147681638
#> [501] 0.2424170708 0.1182816226 0.1399613285 0.1914058023 0.1824519621
#> [506] 0.0757214625 0.0174173406 0.0290530085 0.3951414221 0.2089493304
#> [511] 0.2503864516 0.1531628399 0.1702945750 0.3214003472 0.1879753153
#> [516] 0.5577368334 0.0087743121 0.5977254426 0.0565435647 0.6959611514
#> [521] 0.1452837319 0.5217440582 0.7972330882 0.0362238523 0.1633853863
#> [526] 0.0208102787 0.0169593438 0.2708631107 0.1402104647 0.0303995693
#> [531] 0.0357019185 0.4459738787 0.7414584615 0.8565659818 0.1184594017
#> [536] 0.1051383094 0.3149030577 0.1451273724 0.1192567439 0.3079359171
#> [541] 0.0514016693 0.1557013914 0.3384736964 0.4205987073 0.1370401840
#> [546] 0.0428483200 0.0545825798 0.4642509651 0.2057141385 0.2988846665
#> [551] 0.3704184505 0.3497241742 0.0626521195 0.0847969987 0.0765315957
#> [556] 0.3785393716 0.4141491848 0.1048129579 0.3908185177 0.4425860813
#> [561] 0.7608551995 0.3825471044 0.7351732444 0.3476146520 0.1551004509
#> [566] 0.6232957868 0.2752213338 0.3170992590 0.0763030986 0.0488269790
#> [571] 0.0708126559 0.1464858313 0.3531455380 0.0707643418 1.0822225899
#> [576] 0.2116697219 0.5223674673 0.0168484112 0.4821282622 0.2716400334
#> [581] 0.6843164033 0.1449420473 0.3994214197 0.0061489980 0.3123798653
#> [586] 1.1150839269 0.4406574152 0.2275857425 0.0542067123 0.5229192286
#> [591] 0.7170026314 0.0024162582 0.0197555530 0.9509097365 0.0454359959
#> [596] 0.0037688503 0.0583098991 0.1037885513 0.2659223874 0.2422116398
#> [601] 0.7773812116 0.0023664401 0.1317233494 0.0214858797 0.4779859097
#> [606] 0.0827952159 0.0930104726 0.2395688380 0.2772974960 0.1268527302
#> [611] 0.1303051143 0.7773422891 0.1282113469 1.6480625316 0.2389714207
#> [616] 0.0134456029 0.0021610779 0.0459456371 0.1249046641 0.5806382731
#> [621] 0.0166888755 0.1034576015 0.9008986783 0.0237614168 0.2466159040
#> [626] 0.0246407259 0.3488818733 0.5671581327 0.7468473810 0.6167445663
#> [631] 0.1010456177 0.1781917258 0.4910284740 0.2661873602 0.4147736062
#> [636] 0.2994242737 0.1088314484 0.2529295887 0.1556680600 0.3764643414
#> [641] 0.0819979363 0.0697930642 0.3842482295 0.1872323245 0.2317421636
#> [646] 0.0233041068 0.1664500131 0.1058827843 0.0907375347 0.0037352586
#> [651] 0.3399562716 0.6781590916 0.0875852330 0.4517226852 0.6931053090
#> [656] 0.8266439931 0.6306616086 0.9011030642 0.1445885242 0.0045839738
#> [661] 0.1601814391 0.1033140846 0.5265738017 0.3546128070 0.5889913957
#> [666] 0.0648501491 0.0698911967 0.0390100812 0.4312298341 1.2243129509
#> [671] 0.9771619822 0.1086498963
#> 
#> $muTrain
#>          x          y          z 
#>  0.4010323 -1.0116578  0.4718293 
#> 
#> $RootPrecisTrain
#>          [,1]    [,2]    [,3]
#> [1,] 4.913747 0.00000 0.00000
#> [2,] 0.000000 2.55895 0.00000
#> [3,] 0.000000 0.00000 2.69437
#> 
#> attr(,"class")
#> [1] "threshold" "pca"      
#> 
#> attr(,"class")
#> [1] "fault_ls"