Process Fault Filtering

Flag and filter out observations beyond normal operating conditions, then return the observations within normal operating conditions.

faultFilter(trainData, testData, updateFreq, faultsToTriggerAlarm = 5, ...)

Arguments

trainData
An xts data matrix of initial training observations

testData
The data not included in the training data set

updateFreq
The number of observations from the test data matrix that must be returned to update the training data matrix and move it forward.

faultsToTriggerAlarm
Specifies how many sequential faults will cause an alarm to trigger. Defaults to 5.

...
Lazy dots for additional internal arguments

Value

A list of class "fault_ls" with the following:

faultObj -- An xts flagging matrix with the same number of rows as "testData". This flag matrix has the following five columns:

SPE -- The SPE statistic value for each observation in "testData". This statistic is defined as $$ SPE_i = (\textbf{X}_i - \textbf{Y}_i * \textbf{P}^T) * (\textbf{X}_i - \textbf{Y}_i * \textbf{P}^T)^T, $$ where \(\textbf{X}_i\) is the \(i^{th}\) observation vector, \(\textbf{Y}_i\) is the reduced-feature projection of the observation \(\textbf{X}_i\), and \(\textbf{P}\) is the projection matrix such that \(\textbf{X}_i\textbf{P} = \textbf{Y}_i\).

SPE_Flag -- A vector of SPE indicators recording 0 if the test statistic is less than or equal to the critical value passed through from the threshold object.

T2 -- The T2 statistic value for each observation in "testData". This statistic is defined as $$ T^2_i = \textbf{Y}_i * \textbf{D}^{-1} * \textbf{Y}_i^T, $$ where \(\textbf{Y}_i = \textbf{X}_i\textbf{P}\) is the reduced- feature projection of the observation \(\textbf{X}_i\), and \(\textbf{D}\) is the diagonal matrix of eigenvalues.

T2_Flag -- A vector of T2 fault indicators, defined like SPE_Flag.

Alarm -- A column indicating if there have been three flags in a row for either the SPE or T2 monitoring statistics or both. Alarm states are as follows: 0 = no alarm, 1 = Hotelling's T2 alarm, 2 = Squared Prediction Error alarm, and 3 = both alarms.

nonAlarmedTestObs -- An xts matrix of the first updateFreq number of rows of the training data which were not alarmed.

trainSpecs -- The threshold object returned by the internal threshold() function. See the threshold() function's help file for more details.

Details

This function is essentially a wrapper function to call and organize the output from these other internal functions: faultDetect(), threshold(), and pca(). It is applied over a rolling window, with observation width equal to updateFreq, of the larger full data matrix via the processMonitor() function, wherein the testing and training data sets move forward in time across the entire data matrix.

This internal function is called by processMonitor().

See also

Calls: pca, threshold, faultDetect. Called by: processMonitor.

Examples

nrml <- mspProcessData(faults = "NOC")
# Select the data under state 1
data <- nrml[nrml[,1] == 1]

faultFilter(trainData = data[1:672, -1],
            testData = data[673:3360, -1],
            updateFreq = 336)

#> $faultObj
#>                              SPE SPE_Flag           T2 T2_Flag Alarm
#> 2015-05-17 19:12:00 1.136624e+00        0 5.430296e+00       0     0
#> 2015-05-17 19:13:00 1.739725e+00        0 5.014730e+00       0     0
#> 2015-05-17 19:14:00 3.999711e-01        0 3.357003e+00       0     0
#> 2015-05-17 19:15:00 8.232836e-02        0 3.704983e+00       0     0
#> 2015-05-17 19:16:00 6.170623e-01        0 6.289844e+00       0     0
#> 2015-05-17 19:17:00 6.340335e-01        0 3.338330e+00       0     0
#> 2015-05-17 19:18:00 5.788635e-02        0 2.682427e+00       0     0
#> 2015-05-17 19:19:00 2.288376e-01        0 2.126589e+00       0     0
#> 2015-05-17 19:20:00 5.782369e-01        0 9.354648e-01       0     0
#> 2015-05-17 19:21:00 8.984972e-03        0 1.435090e+00       0     0
#> 2015-05-17 19:22:00 9.714080e-01        0 2.643954e+00       0     0
#> 2015-05-17 19:23:00 3.092313e-01        0 4.108731e+00       0     0
#> 2015-05-17 19:24:00 1.089499e-02        0 7.339955e+00       1     0
#> 2015-05-17 19:25:00 2.804739e-02        0 4.220728e+00       0     0
#> 2015-05-17 19:26:00 1.629388e-01        0 4.211270e+00       0     0
#> 2015-05-17 19:27:00 2.027933e-02        0 1.102558e+00       0     0
#> 2015-05-17 19:28:00 6.324720e-01        0 1.144948e+00       0     0
#> 2015-05-17 19:29:00 4.800746e-01        0 7.090531e-01       0     0
#> 2015-05-17 19:30:00 6.875443e-02        0 1.180586e+00       0     0
#> 2015-05-17 19:31:00 3.761577e-01        0 1.770778e+00       0     0
#> 2015-05-17 19:32:00 1.960025e-01        0 8.802436e-01       0     0
#> 2015-05-17 19:33:00 2.779795e-02        0 2.881728e+00       0     0
#> 2015-05-17 19:34:00 4.432185e-01        0 3.191795e+00       0     0
#> 2015-05-17 19:35:00 6.612049e-01        0 3.630081e+00       0     0
#> 2015-05-17 19:36:00 9.479429e-02        0 2.577928e+00       0     0
#> 2015-05-17 19:37:00 1.242151e+00        0 2.750710e+00       0     0
#> 2015-05-17 19:38:00 1.311500e-01        0 2.205578e+00       0     0
#> 2015-05-17 19:39:00 8.512231e-02        0 3.974765e+00       0     0
#> 2015-05-17 19:40:00 6.116365e-01        0 4.871786e+00       0     0
#> 2015-05-17 19:41:00 2.723357e-01        0 6.104688e+00       0     0
#> 2015-05-17 19:42:00 1.451519e-02        0 5.292736e+00       0     0
#> 2015-05-17 19:43:00 1.236179e+00        0 4.287181e+00       0     0
#> 2015-05-17 19:44:00 7.002865e-01        0 8.093228e+00       1     0
#> 2015-05-17 19:45:00 4.445335e-01        0 1.039150e+01       1     0
#> 2015-05-17 19:46:00 7.027584e-01        0 4.809936e+00       0     0
#> 2015-05-17 19:47:00 9.229196e-01        0 4.889790e+00       0     0
#> 2015-05-17 19:48:00 2.169066e-01        0 6.801263e+00       1     0
#> 2015-05-17 19:49:00 7.191706e-01        0 3.879650e+00       0     0
#> 2015-05-17 19:50:00 4.027967e-01        0 4.716358e+00       0     0
#> 2015-05-17 19:51:00 7.854238e-01        0 3.171250e+00       0     0
#> 2015-05-17 19:52:00 1.453294e-01        0 2.569049e+00       0     0
#> 2015-05-17 19:53:00 3.126835e-01        0 2.367290e+00       0     0
#> 2015-05-17 19:54:00 2.085796e-01        0 2.684751e+00       0     0
#> 2015-05-17 19:55:00 9.073206e-01        0 3.159220e+00       0     0
#> 2015-05-17 19:56:00 5.317495e-01        0 4.042438e+00       0     0
#> 2015-05-17 19:57:00 1.272198e-01        0 6.151828e+00       0     0
#> 2015-05-17 19:58:00 1.524363e-01        0 3.330363e+00       0     0
#> 2015-05-17 19:59:00 6.997605e-01        0 3.836996e+00       0     0
#> 2015-05-17 22:00:00 1.263405e+00        0 2.964566e+00       0     0
#> 2015-05-17 22:01:00 2.249434e-01        0 4.359498e+00       0     0
#> 2015-05-17 22:02:00 4.441618e-01        0 3.143505e+00       0     0
#> 2015-05-17 22:03:00 2.466014e-02        0 1.081318e+00       0     0
#> 2015-05-17 22:04:00 2.311593e-01        0 3.263139e+00       0     0
#> 2015-05-17 22:05:00 2.074836e-01        0 5.832577e+00       0     0
#> 2015-05-17 22:06:00 1.164732e+00        0 4.977363e+00       0     0
#> 2015-05-17 22:07:00 5.409492e-01        0 1.738685e+00       0     0
#> 2015-05-17 22:08:00 5.730048e-01        0 2.889925e+00       0     0
#> 2015-05-17 22:09:00 3.379348e-01        0 2.817233e+00       0     0
#> 2015-05-17 22:10:00 2.803785e-01        0 3.539530e+00       0     0
#> 2015-05-17 22:11:00 3.063330e-02        0 2.489966e+00       0     0
#> 2015-05-17 22:12:00 9.430350e-02        0 5.863724e-01       0     0
#> 2015-05-17 22:13:00 6.503451e-01        0 2.992301e+00       0     0
#> 2015-05-17 22:14:00 6.464723e-01        0 4.437344e+00       0     0
#> 2015-05-17 22:15:00 4.901242e-01        0 2.576431e+00       0     0
#> 2015-05-17 22:16:00 5.558165e-01        0 2.087707e+00       0     0
#> 2015-05-17 22:17:00 2.252517e-01        0 2.801654e+00       0     0
#> 2015-05-17 22:18:00 9.458848e-02        0 3.340948e+00       0     0
#> 2015-05-17 22:19:00 4.693359e-02        0 5.862461e+00       0     0
#> 2015-05-17 22:20:00 2.269112e+00        0 2.919750e+00       0     0
#> 2015-05-17 22:21:00 5.523546e-01        0 5.760525e+00       0     0
#> 2015-05-17 22:22:00 9.444973e-02        0 4.913704e+00       0     0
#> 2015-05-17 22:23:00 2.415017e-01        0 3.268056e+00       0     0
#> 2015-05-17 22:24:00 5.501501e-02        0 3.435803e+00       0     0
#> 2015-05-17 22:25:00 1.224973e-02        0 3.757459e+00       0     0
#> 2015-05-17 22:26:00 5.086943e-01        0 6.242489e+00       0     0
#> 2015-05-17 22:27:00 8.727750e-01        0 6.362021e+00       0     0
#> 2015-05-17 22:28:00 1.181985e+00        0 7.130414e+00       1     0
#> 2015-05-17 22:29:00 2.898835e-01        0 7.056053e+00       1     0
#> 2015-05-17 22:30:00 9.781755e-01        0 7.176734e+00       1     0
#> 2015-05-17 22:31:00 1.094178e+00        0 6.781231e+00       1     0
#> 2015-05-17 22:32:00 5.633325e-01        0 6.231041e+00       0     0
#> 2015-05-17 22:33:00 1.440618e+00        0 5.281827e+00       0     0
#> 2015-05-17 22:34:00 1.760920e+00        0 8.935405e+00       1     0
#> 2015-05-17 22:35:00 2.282591e+00        0 1.376493e+01       1     0
#> 2015-05-17 22:36:00 1.288028e+00        0 1.263673e+01       1     0
#> 2015-05-17 22:37:00 5.788482e-01        0 1.166600e+01       1     0
#> 2015-05-17 22:38:00 1.945189e+00        0 1.435003e+01       1     1
#> 2015-05-17 22:39:00 1.544493e+00        0 1.202989e+01       1     1
#> 2015-05-17 22:40:00 1.533192e-01        0 7.718389e+00       1     1
#> 2015-05-17 22:41:00 2.823898e-01        0 1.086693e+01       1     1
#> 2015-05-17 22:42:00 1.286445e+00        0 8.500627e+00       1     1
#> 2015-05-17 22:43:00 1.732227e-01        0 9.568106e+00       1     1
#> 2015-05-17 22:44:00 1.933068e-01        0 4.259123e+00       0     0
#> 2015-05-17 22:45:00 1.629325e+00        0 1.131114e+01       1     0
#> 2015-05-17 22:46:00 2.051987e-01        0 6.454029e+00       0     0
#> 2015-05-17 22:47:00 1.517123e+00        0 1.129397e+01       1     0
#> 2015-05-17 22:48:00 1.024681e+00        0 9.632094e+00       1     0
#> 2015-05-17 22:49:00 7.287300e-01        0 1.642523e+01       1     0
#> 2015-05-17 22:50:00 1.483273e+00        0 1.107277e+01       1     0
#> 2015-05-17 22:51:00 1.744833e-02        0 6.200694e+00       0     0
#> 2015-05-17 22:52:00 1.258314e+00        0 7.044384e+00       1     0
#> 2015-05-17 22:53:00 1.046145e+00        0 4.538185e+00       0     0
#> 2015-05-17 22:54:00 1.838041e+00        0 4.358803e+00       0     0
#> 2015-05-17 22:55:00 1.861082e+00        0 7.123409e+00       1     0
#> 2015-05-17 22:56:00 1.787293e+00        0 1.109194e+01       1     0
#> 2015-05-17 22:57:00 1.744946e+00        0 1.136434e+01       1     0
#> 2015-05-17 22:58:00 1.502068e+00        0 9.905690e+00       1     0
#> 2015-05-17 22:59:00 1.952463e+00        0 1.172988e+01       1     1
#> 2015-05-18 01:00:00 1.849732e+00        0 7.782592e+00       1     1
#> 2015-05-18 01:01:00 1.301640e+00        0 9.425392e+00       1     1
#> 2015-05-18 01:02:00 2.907138e+00        1 9.028105e+00       1     1
#> 2015-05-18 01:03:00 2.259051e+00        0 1.420117e+01       1     1
#> 2015-05-18 01:04:00 1.973610e+00        0 1.513986e+01       1     1
#> 2015-05-18 01:05:00 1.570406e+00        0 1.551188e+01       1     1
#> 2015-05-18 01:06:00 9.812105e-01        0 1.225512e+01       1     1
#> 2015-05-18 01:07:00 7.567127e-01        0 1.109659e+01       1     1
#> 2015-05-18 01:08:00 8.290844e-01        0 1.057058e+01       1     1
#> 2015-05-18 01:09:00 7.144063e-01        0 1.065138e+01       1     1
#> 2015-05-18 01:10:00 8.319846e-01        0 6.823785e+00       1     1
#> 2015-05-18 01:11:00 7.681600e-01        0 7.369350e+00       1     1
#> 2015-05-18 01:12:00 1.618429e-01        0 7.904642e+00       1     1
#> 2015-05-18 01:13:00 2.658899e-01        0 7.661116e+00       1     1
#> 2015-05-18 01:14:00 1.123964e+00        0 7.782882e+00       1     1
#> 2015-05-18 01:15:00 1.082590e+00        0 7.457145e+00       1     1
#> 2015-05-18 01:16:00 1.417591e+00        0 1.080617e+01       1     1
#> 2015-05-18 01:17:00 7.606840e-01        0 1.320800e+01       1     1
#> 2015-05-18 01:18:00 3.064139e-01        0 8.010453e+00       1     1
#> 2015-05-18 01:19:00 1.743450e+00        0 6.207429e+00       0     0
#> 2015-05-18 01:20:00 6.454836e-01        0 6.014374e+00       0     0
#> 2015-05-18 01:21:00 1.003871e+00        0 7.495722e+00       1     0
#> 2015-05-18 01:22:00 5.502216e-02        0 5.196881e+00       0     0
#> 2015-05-18 01:23:00 5.510193e-01        0 6.245578e+00       0     0
#> 2015-05-18 01:24:00 6.108979e-01        0 8.008004e+00       1     0
#> 2015-05-18 01:25:00 9.977354e-01        0 8.344467e+00       1     0
#> 2015-05-18 01:26:00 8.825355e-01        0 7.327729e+00       1     0
#> 2015-05-18 01:27:00 1.127416e+00        0 8.129517e+00       1     0
#> 2015-05-18 01:28:00 1.072350e+00        0 1.160122e+01       1     1
#> 2015-05-18 01:29:00 9.888973e-01        0 9.814039e+00       1     1
#> 2015-05-18 01:30:00 1.414259e+00        0 1.085142e+01       1     1
#> 2015-05-18 01:31:00 1.590638e+00        0 1.404599e+01       1     1
#> 2015-05-18 01:32:00 1.384604e+00        0 1.398304e+01       1     1
#> 2015-05-18 01:33:00 1.358609e+00        0 1.033508e+01       1     1
#> 2015-05-18 01:34:00 1.612998e+00        0 1.034035e+01       1     1
#> 2015-05-18 01:35:00 1.455697e+00        0 1.123140e+01       1     1
#> 2015-05-18 01:36:00 1.286267e+00        0 8.421983e+00       1     1
#> 2015-05-18 01:37:00 9.564644e-01        0 9.901347e+00       1     1
#> 2015-05-18 01:38:00 1.847074e+00        0 1.223227e+01       1     1
#> 2015-05-18 01:39:00 1.017560e+00        0 1.613202e+01       1     1
#> 2015-05-18 01:40:00 9.132458e-01        0 9.211775e+00       1     1
#> 2015-05-18 01:41:00 1.417306e+00        0 8.218920e+00       1     1
#> 2015-05-18 01:42:00 2.418964e+00        1 1.024464e+01       1     1
#> 2015-05-18 01:43:00 7.783194e-01        0 1.122777e+01       1     1
#> 2015-05-18 01:44:00 1.502353e+00        0 8.546395e+00       1     1
#> 2015-05-18 01:45:00 3.028257e+00        1 1.056281e+01       1     1
#> 2015-05-18 01:46:00 4.725019e-01        0 8.005538e+00       1     1
#> 2015-05-18 01:47:00 1.880773e+00        0 1.014751e+01       1     1
#> 2015-05-18 01:48:00 1.213983e-01        0 8.898345e+00       1     1
#> 2015-05-18 01:49:00 3.171854e-01        0 5.421781e+00       0     0
#> 2015-05-18 01:50:00 2.062751e+00        0 1.149658e+01       1     0
#> 2015-05-18 01:51:00 1.729173e+00        0 6.529531e+00       0     0
#> 2015-05-18 01:52:00 1.000612e+00        0 7.513107e+00       1     0
#> 2015-05-18 01:53:00 2.676316e+00        1 1.704936e+01       1     0
#> 2015-05-18 01:54:00 2.172680e+00        0 1.444542e+01       1     0
#> 2015-05-18 01:55:00 5.678218e+00        1 1.194139e+01       1     0
#> 2015-05-18 01:56:00 1.557724e+00        0 1.085607e+01       1     1
#> 2015-05-18 01:57:00 3.066841e+00        1 1.978056e+01       1     1
#> 2015-05-18 01:58:00 5.496193e-01        0 1.521407e+01       1     1
#> 2015-05-18 01:59:00 2.659947e+00        1 1.482183e+01       1     1
#> 2015-05-18 04:00:00 8.199025e-01        0 1.349003e+01       1     1
#> 2015-05-18 04:01:00 8.252747e-01        0 1.306108e+01       1     1
#> 2015-05-18 04:02:00 2.273405e+00        0 1.108535e+01       1     1
#> 2015-05-18 04:03:00 1.408997e+00        0 1.024602e+01       1     1
#> 2015-05-18 04:04:00 2.442296e+00        1 8.987079e+00       1     1
#> 2015-05-18 04:05:00 1.458398e+00        0 8.329715e+00       1     1
#> 2015-05-18 04:06:00 1.315418e+00        0 7.992913e+00       1     1
#> 2015-05-18 04:07:00 3.325209e-01        0 1.200626e+01       1     1
#> 2015-05-18 04:08:00 1.743248e+00        0 1.133614e+01       1     1
#> 2015-05-18 04:09:00 8.084739e-01        0 1.406166e+01       1     1
#> 2015-05-18 04:10:00 2.434315e+00        1 1.212091e+01       1     1
#> 2015-05-18 04:11:00 2.618257e+00        1 1.666896e+01       1     1
#> 2015-05-18 04:12:00 2.177245e+00        0 1.944205e+01       1     1
#> 2015-05-18 04:13:00 1.112182e+00        0 9.877308e+00       1     1
#> 2015-05-18 04:14:00 1.358875e+00        0 7.820747e+00       1     1
#> 2015-05-18 04:15:00 1.204777e+00        0 8.338689e+00       1     1
#> 2015-05-18 04:16:00 5.050117e-01        0 7.503430e+00       1     1
#> 2015-05-18 04:17:00 1.330468e+00        0 7.761053e+00       1     1
#> 2015-05-18 04:18:00 9.924907e-01        0 1.143433e+01       1     1
#> 2015-05-18 04:19:00 2.059709e+00        0 7.093532e+00       1     1
#> 2015-05-18 04:20:00 1.642273e+00        0 9.254446e+00       1     1
#> 2015-05-18 04:21:00 7.880690e-01        0 7.181958e+00       1     1
#> 2015-05-18 04:22:00 3.882802e-01        0 6.137784e+00       0     0
#> 2015-05-18 04:23:00 7.303342e-01        0 7.659739e+00       1     0
#> 2015-05-18 04:24:00 1.436888e+00        0 9.518331e+00       1     0
#> 2015-05-18 04:25:00 1.514635e+00        0 1.009707e+01       1     0
#> 2015-05-18 04:26:00 3.524263e+00        1 8.722808e+00       1     0
#> 2015-05-18 04:27:00 5.152817e-01        0 1.222919e+01       1     1
#> 2015-05-18 04:28:00 2.611663e+00        1 6.739396e+00       1     1
#> 2015-05-18 04:29:00 6.464648e-01        0 8.932151e+00       1     1
#> 2015-05-18 04:30:00 1.693332e+00        0 1.319220e+01       1     1
#> 2015-05-18 04:31:00 6.821672e-01        0 1.015799e+01       1     1
#>  [ reached getOption("max.print") -- omitted 2488 rows ]
#> 
#> $nonAlarmedTestObs
#>                             x         y         z
#> 2015-05-17 19:12:00 0.6873707 -1.791209 1.4974422
#> 2015-05-17 19:13:00 0.6493085 -1.714285 1.5448407
#> 2015-05-17 19:14:00 0.6469978 -1.648453 1.2142648
#> 2015-05-17 19:15:00 0.7304903 -1.640529 1.1723993
#> 2015-05-17 19:16:00 0.9701771 -1.677653 1.3047262
#> 2015-05-17 19:17:00 0.7147756 -1.464012 1.2695188
#> 2015-05-17 19:18:00 0.6999505 -1.520508 1.0568807
#> 2015-05-17 19:19:00 0.7194566 -1.566418 0.8063545
#> 2015-05-17 19:20:00 0.5107392 -1.572126 0.6716052
#> 2015-05-17 19:21:00 0.6204801 -1.450096 0.8276680
#> 2015-05-17 19:22:00 0.6311992 -1.405327 1.2685761
#> 2015-05-17 19:23:00 0.8084357 -1.554528 1.2225188
#> 2015-05-17 19:24:00 0.8838028 -1.971292 1.3688204
#> 2015-05-17 19:25:00 0.7726761 -1.691882 1.1798015
#> 2015-05-17 19:26:00 0.7694961 -1.632047 1.2391411
#> 2015-05-17 19:27:00 0.5701776 -1.408883 0.8064593
#> 2015-05-17 19:28:00 0.7195357 -1.203305 0.7697363
#> 2015-05-17 19:29:00 0.4392277 -1.425017 0.8075906
#> 2015-05-17 19:30:00 0.5598485 -1.452792 0.8179733
#> 2015-05-17 19:31:00 0.5427673 -1.573988 0.9702159
#> 2015-05-17 19:32:00 0.4999565 -1.427125 0.7977655
#> 2015-05-17 19:33:00 0.7377298 -1.563519 1.0114877
#> 2015-05-17 19:34:00 0.6374301 -1.619623 1.2129189
#> 2015-05-17 19:35:00 0.6546451 -1.614819 1.3042455
#> 2015-05-17 19:36:00 0.6602555 -1.543411 1.0702915
#> 2015-05-17 19:37:00 0.5303595 -1.652609 1.2361311
#> 2015-05-17 19:38:00 0.6722496 -1.434395 1.0354962
#> 2015-05-17 19:39:00 0.7207151 -1.730719 1.1709362
#> 2015-05-17 19:40:00 0.7126364 -1.728666 1.3942287
#> 2015-05-17 19:41:00 0.8453341 -1.751143 1.4079720
#> 2015-05-17 19:42:00 0.8402281 -1.782150 1.2216861
#> 2015-05-17 19:43:00 0.6677306 -1.609733 1.4480558
#> 2015-05-17 19:44:00 0.9520128 -1.771116 1.5751432
#> 2015-05-17 19:45:00 0.9989320 -1.966126 1.6820503
#> 2015-05-17 19:46:00 0.6800729 -1.793430 1.3740416
#> 2015-05-17 19:47:00 0.7847383 -1.558163 1.4380848
#> 2015-05-17 19:48:00 0.8158729 -1.920954 1.4310153
#> 2015-05-17 19:49:00 0.6278282 -1.761086 1.2741912
#> 2015-05-17 19:50:00 0.7231521 -1.728181 1.3425646
#> 2015-05-17 19:51:00 0.6158450 -1.583592 1.2772546
#> 2015-05-17 19:52:00 0.7576308 -1.509304 0.9370001
#> 2015-05-17 19:53:00 0.7718770 -1.505685 0.8540817
#> 2015-05-17 19:54:00 0.6717851 -1.499251 1.1246133
#> 2015-05-17 19:55:00 0.6079449 -1.575734 1.2945515
#> 2015-05-17 19:56:00 0.6630176 -1.723914 1.2901758
#> 2015-05-17 19:57:00 0.8273284 -1.827589 1.3754307
#> 2015-05-17 19:58:00 0.7929303 -1.549276 1.0566491
#> 2015-05-17 19:59:00 0.6807106 -1.589728 1.3364232
#> 2015-05-17 22:00:00 0.5340951 -1.721241 1.2269357
#> 2015-05-17 22:01:00 0.7569753 -1.650668 1.2774000
#> 2015-05-17 22:02:00 0.8238358 -1.462076 1.0364662
#> 2015-05-17 22:03:00 0.5652898 -1.406465 0.8070044
#> 2015-05-17 22:04:00 0.7101897 -1.541508 1.1841574
#> 2015-05-17 22:05:00 0.8943077 -1.722228 1.2991336
#> 2015-05-17 22:06:00 0.6873059 -1.701029 1.4856917
#> 2015-05-17 22:07:00 0.5448942 -1.670963 0.8636743
#> 2015-05-17 22:08:00 0.6616046 -1.469957 1.2289947
#> 2015-05-17 22:09:00 0.6555024 -1.517950 1.1731105
#> 2015-05-17 22:10:00 0.8292932 -1.639158 0.9663156
#> 2015-05-17 22:11:00 0.6981635 -1.597103 0.9295740
#> 2015-05-17 22:12:00 0.4910135 -1.289020 0.7746922
#> 2015-05-17 22:13:00 0.7030741 -1.419886 1.2358616
#> 2015-05-17 22:14:00 0.8259609 -1.513541 1.3091765
#> 2015-05-17 22:15:00 0.5995841 -1.552654 1.1625016
#> 2015-05-17 22:16:00 0.5787313 -1.728658 0.8758376
#> 2015-05-17 22:17:00 0.7688462 -1.462707 1.0311888
#> 2015-05-17 22:18:00 0.7717248 -1.559865 1.0849381
#> 2015-05-17 22:19:00 0.8829823 -1.829399 1.2232853
#> 2015-05-17 22:20:00 0.4705706 -1.744411 1.2985515
#> 2015-05-17 22:21:00 0.7745122 -1.751987 1.4567091
#> 2015-05-17 22:22:00 0.8197238 -1.698513 1.2529110
#> 2015-05-17 22:23:00 0.6659560 -1.773685 1.0410373
#> 2015-05-17 22:24:00 0.7047781 -1.710512 1.0801909
#> 2015-05-17 22:25:00 0.7752838 -1.667211 1.0856998
#> 2015-05-17 22:26:00 0.8259219 -1.740676 1.4772401
#> 2015-05-17 22:27:00 0.9045056 -1.624660 1.4782377
#> 2015-05-17 22:28:00 0.8376360 -1.727606 1.6371894
#> 2015-05-17 22:29:00 0.8141338 -1.939114 1.4639588
#> 2015-05-17 22:30:00 0.8835678 -1.703311 1.5917342
#> 2015-05-17 22:31:00 0.8706952 -1.661001 1.5801503
#> 2015-05-17 22:32:00 0.7535283 -1.883530 1.4616069
#> 2015-05-17 22:33:00 0.6434480 -1.847445 1.4867084
#> 2015-05-17 22:34:00 0.8111077 -1.922176 1.8090020
#> 2015-05-17 22:35:00 1.1157111 -1.897836 2.0248005
#> 2015-05-17 22:36:00 0.9440925 -2.126254 1.9455236
#> 2015-05-17 22:37:00 0.9729098 -2.087997 1.7977882
#> 2015-05-17 22:44:00 0.7206327 -1.720881 1.2485494
#> 2015-05-17 22:45:00 0.9129170 -2.005222 1.9248928
#> 2015-05-17 22:46:00 0.7978570 -1.929383 1.3875301
#> 2015-05-17 22:47:00 0.9497094 -1.955318 1.9078716
#> 2015-05-17 22:48:00 0.9753383 -1.835147 1.7277986
#> 2015-05-17 22:49:00 1.1643215 -2.201319 1.9917979
#> 2015-05-17 22:50:00 1.0391270 -1.843219 1.8317153
#> 2015-05-17 22:51:00 0.8424709 -1.877755 1.3125281
#> 2015-05-17 22:52:00 0.7331560 -1.926608 1.6088596
#> 2015-05-17 22:53:00 0.6733121 -1.674440 1.4360464
#> 2015-05-17 22:54:00 0.5763495 -1.797985 1.4404493
#> 2015-05-17 22:55:00 0.7230353 -1.865030 1.6981267
#> 2015-05-17 22:56:00 0.9441150 -1.915558 1.9252159
#> 2015-05-17 22:57:00 0.9833681 -1.892660 1.9207431
#> 2015-05-17 22:58:00 0.8180935 -2.074759 1.8077633
#> 2015-05-18 01:19:00 0.6909522 -1.817346 1.6237754
#> 2015-05-18 01:20:00 0.9261847 -1.629942 1.3690254
#> 2015-05-18 01:21:00 0.8262619 -1.813164 1.6414884
#> 2015-05-18 01:22:00 0.7882384 -1.804212 1.2674182
#> 2015-05-18 01:23:00 0.7753097 -1.825698 1.4824372
#> 2015-05-18 01:24:00 0.8976331 -1.826309 1.5994641
#> 2015-05-18 01:25:00 1.0334414 -1.727368 1.5229081
#> 2015-05-18 01:26:00 0.9252127 -1.696120 1.5562170
#> 2015-05-18 01:27:00 0.8622793 -1.811700 1.6930851
#> 2015-05-18 01:49:00 0.8584588 -1.660850 1.3324390
#> 2015-05-18 01:50:00 0.8789496 -2.042254 1.9735400
#> 2015-05-18 01:51:00 0.6998778 -1.850667 1.6396594
#> 2015-05-18 01:52:00 0.8266002 -1.815211 1.6420870
#> 2015-05-18 01:53:00 1.2060599 -2.004485 2.1814326
#> 2015-05-18 01:54:00 1.0702740 -1.996058 2.0965358
#> 2015-05-18 01:55:00 0.8858580 -1.790844 2.2616734
#> 2015-05-18 04:22:00 0.8102239 -1.776640 1.4491594
#> 2015-05-18 04:23:00 0.9417147 -1.736832 1.5494004
#> 2015-05-18 04:24:00 0.9315355 -1.822415 1.7950528
#> 2015-05-18 04:25:00 0.8463991 -2.021957 1.8395865
#> 2015-05-18 04:26:00 0.7240660 -1.903702 1.9369592
#> 2015-05-21 13:47:00 0.7931590 -1.667597 1.5784573
#> 2015-05-21 13:48:00 0.6546512 -1.711271 1.4189895
#> 2015-05-21 13:49:00 0.8469116 -1.983782 1.6213761
#> 2015-05-21 13:50:00 0.9203912 -2.000016 1.9003996
#> 2015-05-21 13:51:00 1.0559446 -2.094671 2.0779185
#> 2015-05-21 13:52:00 1.0492681 -1.970398 2.2059422
#> 2015-05-21 16:01:00 0.7620971 -1.819745 1.4955356
#> 2015-05-21 16:02:00 0.8701804 -1.789754 1.7158017
#> 2015-05-21 16:03:00 0.9766529 -1.906933 1.6682565
#> 2015-05-21 16:04:00 0.8343526 -1.708399 1.6686660
#> 2015-05-21 16:05:00 1.0121636 -1.784280 1.8100107
#> 2015-05-21 16:06:00 0.7302877 -1.878527 1.3763874
#> 2015-05-21 16:07:00 0.8180671 -1.813010 1.5764492
#> 2015-05-21 16:08:00 0.8008947 -1.732593 1.3632223
#> 2015-05-21 16:09:00 0.7405121 -1.683506 1.6155509
#> 2015-05-21 16:10:00 1.0452083 -1.913418 1.7527058
#> 2015-05-21 16:11:00 0.7560100 -1.847542 1.5317769
#> 2015-05-21 16:12:00 0.9653522 -1.904015 1.6927420
#> 2015-05-21 16:13:00 0.9061379 -2.043789 1.6301195
#> 2015-05-21 16:14:00 0.7246120 -1.786006 1.2804510
#> 2015-05-21 16:15:00 0.6863782 -1.792881 1.4469579
#> 2015-05-21 16:16:00 1.0155781 -1.980259 1.8393863
#> 2015-05-21 16:17:00 0.9849544 -1.878636 1.5399417
#> 2015-05-21 16:18:00 0.7915275 -1.637715 1.5083132
#> 2015-05-21 16:19:00 0.6258214 -1.797675 1.5598374
#> 2015-05-21 16:20:00 0.9446357 -2.147654 1.7096866
#> 2015-05-21 16:21:00 1.2353174 -1.970549 1.9294537
#> 2015-05-21 16:22:00 0.8534868 -1.768379 1.8402985
#> 2015-05-21 16:23:00 0.7768931 -1.629950 1.6090377
#> 2015-05-21 16:24:00 0.8159887 -1.996942 1.6688324
#> 2015-05-21 16:25:00 0.8575177 -1.852042 1.6917351
#> 2015-05-21 16:26:00 0.6848043 -2.018246 1.7414330
#> 2015-05-21 16:27:00 0.9809874 -1.880149 2.0024677
#> 2015-05-21 16:36:00 0.7105867 -1.601170 1.3915346
#> 2015-05-21 16:37:00 1.0626846 -2.027586 1.7742845
#> 2015-05-21 16:38:00 0.9270231 -1.880823 1.8631109
#> 2015-05-21 16:39:00 0.8131657 -1.970087 1.4501418
#> 2015-05-21 16:40:00 0.8809513 -1.808222 1.9884279
#> 2015-05-21 16:56:00 0.6421335 -2.047258 1.5397565
#> 2015-05-21 16:57:00 1.1472879 -1.722843 1.5091777
#> 2015-05-21 16:58:00 0.7675197 -1.724869 1.1667958
#> 2015-05-21 16:59:00 0.7137583 -1.625471 1.0664110
#> 2015-05-21 19:00:00 0.9177612 -1.929606 1.7648350
#> 2015-05-21 19:01:00 1.0278010 -1.887630 1.9979193
#> 2015-05-21 19:02:00 0.9142407 -1.685397 1.6264043
#> 2015-05-21 19:03:00 1.1340949 -1.970618 1.7381994
#> 2015-05-21 19:05:00 0.5917956 -1.769912 1.5578276
#> 2015-05-21 19:06:00 0.9367929 -2.070003 2.0374577
#> 2015-05-21 19:07:00 0.9549566 -1.999014 1.8947881
#> 2015-05-21 19:08:00 0.7157944 -1.582646 1.4824779
#> 2015-05-21 19:09:00 0.7811432 -1.724683 1.0845785
#> 2015-05-21 19:10:00 0.7870990 -1.817082 1.1364211
#> 2015-05-21 19:11:00 0.8326438 -1.873407 1.1577146
#> 2015-05-21 19:12:00 0.7458999 -1.775818 1.6258284
#> 2015-05-21 19:13:00 0.9888952 -1.996032 1.7340069
#> 2015-05-21 19:14:00 0.8692414 -1.839422 1.6829403
#> 2015-05-21 19:15:00 0.7733877 -1.773175 1.7403553
#> 2015-05-21 19:16:00 0.9133915 -1.787656 1.6920535
#> 2015-05-21 19:17:00 0.7743057 -1.731923 1.3624950
#> 2015-05-21 19:18:00 0.7905048 -1.840346 1.5860941
#> 2015-05-21 19:19:00 0.8313173 -1.894278 1.4899929
#> 2015-05-21 19:20:00 0.7701154 -1.864945 1.5769282
#> 2015-05-21 19:21:00 0.8553960 -1.923023 1.8192333
#> 2015-05-21 19:23:00 0.7139855 -1.900810 1.5144069
#> 2015-05-21 19:24:00 0.9367638 -1.872560 1.5231379
#> 2015-05-21 19:25:00 0.8046707 -1.746043 1.4033682
#> 2015-05-21 19:26:00 0.9505447 -1.809141 1.5750385
#> 2015-05-21 19:27:00 0.8199438 -1.783576 1.6341677
#> 2015-05-21 19:28:00 0.8175635 -1.770582 1.8602600
#> 2015-05-21 19:29:00 1.0583409 -1.960339 1.8827081
#> 2015-05-21 19:48:00 0.5962058 -1.853551 1.6105077
#> 2015-05-21 19:49:00 0.9493364 -1.932600 1.6665372
#> 2015-05-21 19:50:00 0.8484000 -1.810008 1.5466237
#> 2015-05-21 19:51:00 0.9842180 -1.838770 1.4172263
#> 2015-05-21 19:52:00 0.6624316 -1.851135 1.4373014
#> 2015-05-21 19:53:00 0.8529630 -1.628970 1.3695434
#> 2015-05-21 19:54:00 0.6982323 -1.688669 1.4552079
#> 2015-05-21 19:55:00 0.7388349 -1.912667 1.6477696
#> 2015-05-21 19:56:00 0.9847101 -2.020430 1.9108013
#> 2015-05-21 19:57:00 0.9478521 -1.864108 2.0936456
#> 2015-05-21 19:58:00 0.9685125 -1.789129 1.7908921
#> 2015-05-21 22:00:00 0.8666549 -1.650480 1.5410331
#> 2015-05-21 22:01:00 0.9519226 -1.903925 1.6779579
#> 2015-05-21 22:02:00 0.9894308 -2.011162 1.8098963
#> 2015-05-21 22:03:00 0.8181103 -1.858613 1.7269595
#> 2015-05-21 22:04:00 0.5193554 -1.610917 1.3416785
#> 2015-05-21 22:05:00 0.9103657 -1.623983 1.2747030
#> 2015-05-21 22:06:00 0.7620207 -1.711633 1.4857399
#> 2015-05-21 22:07:00 0.8516683 -1.831322 1.3450163
#> 2015-05-21 22:08:00 0.6777269 -1.471065 1.2157847
#> 2015-05-21 22:09:00 0.7085724 -1.528782 1.1316494
#> 2015-05-21 22:10:00 0.7507638 -1.412470 0.7178784
#> 2015-05-21 22:11:00 0.7459220 -1.525115 1.0345530
#> 2015-05-21 22:12:00 0.7030484 -1.518761 1.2672098
#> 2015-05-21 22:13:00 0.6874107 -1.562064 1.4470339
#> 2015-05-21 22:14:00 0.8828895 -1.716893 1.2963643
#> 2015-05-21 22:15:00 0.9031975 -1.819649 1.5246860
#> 2015-05-21 22:16:00 0.9057044 -1.775341 1.5582163
#> 2015-05-21 22:17:00 0.9172675 -1.745416 1.4320096
#> 2015-05-21 22:18:00 0.7462380 -1.775517 1.4836234
#> 2015-05-21 22:19:00 0.9228434 -2.002632 1.7394128
#> 2015-05-21 22:20:00 0.7360322 -1.663578 1.4018996
#> 2015-05-21 22:21:00 0.6737482 -1.646197 1.2949016
#> 2015-05-21 22:22:00 0.7111991 -1.767547 1.3482867
#> 2015-05-21 22:23:00 0.8261415 -1.776223 1.3597772
#> 2015-05-21 22:24:00 0.8061016 -1.663450 1.1957541
#> 2015-05-21 22:25:00 0.8410621 -1.664531 1.3168395
#> 2015-05-21 22:26:00 0.7303819 -1.930316 1.4377907
#> 2015-05-21 22:27:00 0.9426577 -1.717630 1.3419922
#> 2015-05-21 22:28:00 0.7130371 -1.773459 1.6317982
#> 2015-05-21 22:29:00 0.8843443 -1.874194 1.4345350
#> 2015-05-21 22:30:00 0.7882081 -1.779010 1.2340904
#> 2015-05-21 22:31:00 0.8766557 -1.780778 1.1768332
#> 2015-05-21 22:32:00 0.9103444 -1.696058 1.5903341
#> 2015-05-21 22:33:00 0.9345075 -1.876305 1.5719897
#> 2015-05-21 22:34:00 0.7704347 -1.745099 1.1868534
#> 2015-05-21 22:35:00 0.8418091 -1.590714 1.3139698
#> 2015-05-21 22:36:00 0.9484050 -1.925920 1.5568817
#> 2015-05-21 22:37:00 0.7542364 -1.757141 1.4194220
#> 2015-05-21 22:38:00 0.8092820 -1.928998 1.5639797
#> 2015-05-21 22:39:00 0.8537279 -1.662414 1.2991604
#> 2015-05-21 22:40:00 0.9403684 -1.702315 1.4921554
#> 2015-05-21 22:41:00 0.8755050 -1.806338 1.5447118
#> 2015-05-21 22:42:00 1.0355345 -1.979985 1.6276858
#> 2015-05-21 22:43:00 0.9455628 -1.926048 1.6980703
#> 2015-05-21 22:44:00 0.8353222 -1.536105 1.3250808
#> 2015-05-21 22:45:00 0.7424372 -1.582824 1.1845513
#> 2015-05-21 22:46:00 0.8232610 -1.812355 1.1631593
#> 2015-05-21 22:47:00 0.7945519 -1.781158 1.3156014
#> 2015-05-21 22:48:00 0.9785864 -1.773109 1.6534846
#> 2015-05-21 22:49:00 0.8281519 -1.781618 1.5075292
#> 2015-05-21 22:50:00 0.9157235 -1.979824 1.6713851
#> 2015-05-21 22:51:00 0.9368668 -1.774589 1.4075803
#> 2015-05-21 22:52:00 0.9257608 -1.829623 1.3974936
#> 2015-05-21 22:53:00 0.8690777 -1.821938 1.5431499
#> 2015-05-21 22:55:00 0.6091579 -1.859397 1.5551136
#> 2015-05-21 22:56:00 0.5265891 -1.567116 1.4248438
#> 2015-05-21 22:57:00 0.8245800 -1.743016 1.2767653
#> 2015-05-21 22:58:00 0.7008884 -1.745872 1.3620789
#> 2015-05-21 22:59:00 0.6382861 -1.548476 1.2529358
#> 2015-05-22 01:00:00 0.7201877 -1.612469 1.0828244
#> 2015-05-22 01:01:00 0.7145069 -1.891430 1.3669002
#> 2015-05-22 01:02:00 0.8686795 -1.919742 1.4513525
#> 2015-05-22 01:03:00 0.8278842 -1.784880 1.3966638
#> 2015-05-22 01:04:00 0.9785745 -1.992668 1.6270461
#> 2015-05-22 01:05:00 0.7595316 -1.526279 1.3805424
#> 2015-05-22 01:06:00 0.8126264 -1.574619 1.5831985
#> 2015-05-22 01:07:00 0.8603157 -1.847632 1.5111472
#> 2015-05-22 01:08:00 0.7726486 -1.820360 1.4221581
#> 2015-05-22 01:09:00 0.8934142 -1.752890 1.2759475
#> 2015-05-22 01:10:00 0.7229584 -1.943268 1.8476340
#> 2015-05-22 01:11:00 0.8882222 -1.922504 1.4987542
#> 2015-05-22 01:12:00 0.7990630 -1.597672 1.1276594
#> 2015-05-22 01:13:00 0.7756187 -1.605431 1.2803107
#> 2015-05-22 01:14:00 0.7292108 -1.577669 1.2759458
#> 2015-05-22 01:15:00 0.8117526 -1.766549 1.4638433
#> 2015-05-22 01:16:00 0.9314391 -1.873287 1.4883528
#> 2015-05-22 01:17:00 0.7031592 -1.738438 1.1267612
#> 2015-05-22 01:18:00 0.6672247 -1.656530 1.1359539
#> 2015-05-22 01:19:00 0.6994637 -1.609752 1.0955334
#> 2015-05-22 01:20:00 0.5833013 -1.718685 1.0681277
#> 2015-05-22 01:21:00 0.7456629 -1.772583 1.0992964
#> 2015-05-22 01:22:00 0.7111094 -1.527162 0.9936375
#> 2015-05-22 01:23:00 0.7741547 -1.826316 1.6593946
#> 2015-05-22 01:24:00 0.8416392 -1.825656 1.4255079
#> 2015-05-22 01:25:00 0.5516299 -1.565088 1.1589625
#> 2015-05-22 01:26:00 0.6275398 -1.708813 1.0067915
#> 2015-05-22 01:27:00 0.6028303 -1.532550 0.9311522
#> 2015-05-22 01:28:00 0.7677781 -1.668462 1.2060758
#> 2015-05-22 01:29:00 0.6299759 -1.466700 1.1547045
#> 2015-05-22 01:30:00 0.6250974 -1.558802 0.9001126
#> 2015-05-22 01:31:00 0.7275819 -1.261448 0.8329401
#> 2015-05-22 01:32:00 0.6205109 -1.757841 0.9779636
#> 2015-05-22 01:33:00 0.9741188 -1.495576 1.0095974
#> 2015-05-22 01:34:00 0.6580188 -1.648727 1.1192421
#> 2015-05-22 01:35:00 0.6597531 -1.414226 1.1179695
#> 2015-05-22 01:36:00 0.7931460 -1.488934 1.2133098
#> 2015-05-22 01:37:00 0.7463977 -1.417137 1.0322873
#> 2015-05-22 01:38:00 0.8593817 -1.886582 1.3910798
#> 2015-05-22 01:39:00 0.8253967 -1.542104 1.3383318
#> 2015-05-22 01:40:00 0.7108423 -1.631308 1.5186751
#> 2015-05-22 01:41:00 0.8764260 -1.819364 1.4950682
#> 2015-05-22 01:42:00 0.7806614 -1.884591 1.4875561
#> 2015-05-22 01:43:00 0.7340523 -1.620853 1.3391217
#> 2015-05-22 01:44:00 0.8411787 -1.774856 1.3009442
#> 2015-05-22 01:45:00 0.8227261 -1.718270 1.3187803
#> 2015-05-22 01:46:00 0.7046404 -1.844643 1.2614632
#> 2015-05-22 01:47:00 0.6763743 -1.639554 1.0150575
#> 2015-05-22 01:48:00 0.7495246 -1.553781 0.9761661
#> 2015-05-22 01:49:00 0.6693290 -1.440429 0.7345208
#> 2015-05-22 01:50:00 0.7129654 -1.623017 1.1177058
#> 2015-05-22 01:51:00 0.9058666 -1.553564 0.9711670
#> 2015-05-22 01:52:00 0.5435457 -1.572306 0.7493552
#> 2015-05-22 01:53:00 0.7129022 -1.707143 1.0012998
#> 2015-05-22 01:54:00 0.6822502 -1.750527 1.0616735
#> 2015-05-22 01:55:00 0.6653260 -1.483931 1.2399029
#> 2015-05-22 01:56:00 0.6093645 -1.627823 1.0194405
#> 2015-05-22 01:57:00 0.7167987 -1.439131 1.1919010
#> 2015-05-22 01:58:00 0.6731189 -1.828314 1.0165699
#> 2015-05-22 01:59:00 0.6931670 -1.725571 1.3070977
#> 2015-05-22 04:00:00 0.6236930 -1.549691 0.9154882
#> 2015-05-22 04:01:00 0.5261839 -1.695984 1.1955437
#> 2015-05-22 04:02:00 0.7507922 -1.507362 1.1951771
#> 2015-05-22 04:03:00 0.5898448 -1.689279 1.1518478
#> 2015-05-22 04:04:00 0.7312944 -1.710223 1.0224877
#> 2015-05-22 04:05:00 0.6734644 -1.606019 1.0259969
#> 2015-05-22 04:06:00 0.6213366 -1.301504 0.7362881
#> 2015-05-22 04:07:00 0.8082299 -1.598741 0.8956272
#> 2015-05-22 04:08:00 0.5353085 -1.644716 0.8392323
#> 2015-05-22 04:09:00 0.3574875 -1.274061 0.8161517
#> 2015-05-22 04:10:00 0.5650730 -1.497325 0.5267923
#>  [ reached getOption("max.print") -- omitted 3 rows ]
#> 
#> $trainSpecs
#> $SPE_threshold
#>    99.9% 
#> 2.403837 
#> 
#> $T2_threshold
#>    99.9% 
#> 6.692038 
#> 
#> $projectionMatrix
#>            [,1]
#> [1,]  0.5655850
#> [2,] -0.5836925
#> [3,]  0.5825948
#> 
#> $LambdaInv
#>           [,1]
#> [1,] 0.3699458
#> 
#> $muTrain
#>          x          y          z 
#>  0.3958936 -1.0042242  0.4559867 
#> 
#> $RootPrecisTrain
#>          [,1]     [,2]     [,3]
#> [1,] 4.988573 0.000000 0.000000
#> [2,] 0.000000 2.721066 0.000000
#> [3,] 0.000000 0.000000 2.898963
#> 
#> attr(,"class")
#> [1] "threshold" "pca"      
#> 
#> attr(,"class")
#> [1] "fault_ls"