Process Alarms

Trigger an alarm, if necessary, for incoming multivariate process observations.

mspWarning(mspMonitor_object, faultsToTriggerAlarm = 5)

Arguments

mspMonitor_object
An xts matrix returned by the mspMonitor() function

faultsToTriggerAlarm
Specifies how many sequential faults will cause an alarm to trigger. Defaults to 5.

Value

An xts matrix of the same dimensions as mspMonitor_object, with a recorded negative or positive and type-specific alarm status. Alarm codes are: 0 = no alarm, 1 = Hotelling's T2 alarm, 2 = Squared Prediction Error alarm, and 3 = both alarms.

Details

This function and the mspMonitor() function are designed to be ran via a scheduled task through Windows "Task Scheduler" or Macintosh OX "launchd" application suites. The file flow is as follows: at each time interval, run the mspMonitor() function on the matrix of daily observations to add a flag status to the most recent incoming observation in the matrix, and return this new xts matrix. Then, pass this updated daily observation matrix to the mspWarning() function, which will check if the process has recorded three or more sequential monitoring statistic flags in a row. Of note, because these functions are expected to be repeatedly called in real time, this function will only check for an alarm within the last row of the xts matrix. To check multiple rows for an alarm state, please use the mspTrain() function, which was designed to check multiple past observations.

This function requires an xts matrix returned by the mspMonitor() function.

See also

Pipe flow: mspTrain into mspMonitor into mspWarning.

Examples

nrml <- mspProcessData(faults = "NOC")
n <- nrow(nrml)

# Calculate the training summary, but save five observations for monitoring.
trainResults_ls <- mspTrain(data = nrml[1:(n - 5), -1],
                            labelVector = nrml[1:(n - 5), 1],
                            trainObs = 4320)

# While training, we included 1 lag (the default), so we will also lag the
# observations we will test.
testObs <- nrml[(n - 6):n, -1]
testObs <- xts:::lag.xts(testObs, 0:1)
testObs <- testObs[-1,]
testObs <- cbind(nrml[(n - 5):n, 1], testObs)

# Run the monitoring function.
dataAndFlags <- mspMonitor(observations = testObs[, -1],
                           labelVector = testObs[, 1],
                           trainingSummary = trainResults_ls$TrainingSpecs)

# Alarm check the last row of the matrix returned by the mspMonitor function
mspWarning(dataAndFlags)

#>                             x          y            z       x.1        y.1
#> 2015-05-23 09:54:00 0.1199682 0.00247014  0.041183509 0.1321447 0.03136360
#> 2015-05-23 09:55:00 0.1138535 0.01537853  0.015611471 0.1199682 0.00247014
#> 2015-05-23 09:56:00 0.1702870 0.02349061  0.079247525 0.1138535 0.01537853
#> 2015-05-23 09:57:00 0.1982922 0.04833615  0.081870953 0.1702870 0.02349061
#> 2015-05-23 09:58:00 0.2006981 0.04154680 -0.005721085 0.1982922 0.04833615
#> 2015-05-23 09:59:00 0.2079140 0.02794911 -0.057844524 0.2006981 0.04154680
#>                              z.1      SPE SPE_Flag       T2 T2_Flag Alarm
#> 2015-05-23 09:54:00 -0.020886126 9.263416        1 18.09932       0    NA
#> 2015-05-23 09:55:00  0.041183509 9.878314        1 18.21206       0    NA
#> 2015-05-23 09:56:00  0.015611471 6.173057        1 15.37955       0    NA
#> 2015-05-23 09:57:00  0.079247525 4.984703        0 13.25881       0    NA
#> 2015-05-23 09:58:00  0.081870953 4.458017        0 13.57401       0    NA
#> 2015-05-23 09:59:00 -0.005721085 3.654876        0 14.00404       0     0