PCA for Data Scatter Matrix

Calculate the principal component analysis for a data matrix, and also find the squared prediction error (SPE) and Hotelling's T2 test statistic values for each observation in this data matrix.

pca(data, var.amnt = 0.9, ...)

Arguments

data
A centred-and-scaled data matrix or xts matrix

var.amnt
The energy proportion to preserve in the projection, which dictates the number of principal components to keep. Defaults to 0.90.

...
Lazy dots for additional internal arguments

Value

A list of class "pca" with the following:

projectionMatrix -- the q eigenvectors corresponding to the q largest eigenvalues as a p x q projection matrix

LambdaInv -- the diagonal matrix of inverse eigenvalues

SPE -- the vector of SPE test statistic values for each of the n observations contained in "data"

T2 -- the vector of Hotelling's T2 test statistic for each of the same n observations

Details

This function takes in a training data matrix, without the label column, and the energy preservation proportion, which defaults to 95 percent per Kazor et al (2016). This proportion is the sum of the q largest eigenvalues divided by the sum of all p eigenvalues, where q is the number of columns of the p x q projection matrix P. This function then returns the projection matrix P, a diagonal matrix of the reciprocal eigenvalues (LambdaInv), a vector of the SPE test statistic values corresponding to the rows of the data matrix, and a T2 test statistic vector similar to the SPE vector.

This internal function is called by faultFilter().

See also

Called by: faultFilter.

Examples

nrml <- mspProcessData(faults = "NOC")
scaledData <- scale(nrml[,-1])
pca(scaledData)

#> $projectionMatrix
#>            [,1]       [,2]
#> [1,]  0.6622875  0.4342830
#> [2,]  0.1440786 -0.8734944
#> [3,] -0.7352663  0.2200133
#> 
#> $LambdaInv
#>           [,1]      [,2]
#> [1,] 0.6240677 0.0000000
#> [2,] 0.0000000 0.8227832
#> 
#> $SPE
#>    [1] 1.870147e-01 2.541410e-01 2.101616e-01 1.762330e-01 2.369429e-01
#>    [6] 1.841018e-01 2.992353e-01 3.830962e-01 3.123520e-01 2.626422e-01
#>   [11] 1.277476e-01 1.478048e-01 2.434756e-01 1.221912e-01 1.058476e-01
#>   [16] 1.388826e-01 2.403637e-01 1.883175e-01 1.071056e-01 1.386541e-01
#>   [21] 3.175448e-01 1.186901e-01 1.614009e-01 1.750542e-01 3.280677e-01
#>   [26] 2.301228e-01 3.650358e-01 9.093149e-02 8.567475e-02 3.525789e-01
#>   [31] 5.958476e-02 1.830012e-01 1.381206e-01 1.536931e-01 1.898184e-01
#>   [36] 2.197033e-01 2.076220e-01 2.097887e-01 1.631708e-01 1.686510e-01
#>   [41] 2.071342e-01 5.028729e-01 2.160136e-01 1.722073e-01 2.514365e-01
#>   [46] 1.944264e-01 2.089707e-01 2.018510e-01 1.979503e-01 1.539143e-01
#>   [51] 2.465474e-01 1.625719e-01 3.035538e-01 1.788196e-01 1.658159e-01
#>   [56] 3.175309e-01 4.072363e-01 2.968117e-01 2.524894e-01 2.931289e-01
#>   [61] 2.794747e-01 3.807085e-01 2.599048e-01 3.620732e-01 2.732920e-01
#>   [66] 3.239620e-01 2.383515e-01 2.302591e-01 1.368289e-01 2.073852e-01
#>   [71] 3.066856e-01 3.062116e-01 2.237676e-01 3.668345e-01 2.265491e-01
#>   [76] 4.177578e-01 4.682874e-01 5.022207e-01 3.724097e-01 3.725118e-01
#>   [81] 3.581564e-01 1.787136e-01 2.996505e-01 1.328517e-01 1.579055e-01
#>   [86] 2.885402e-01 2.547951e-01 2.439148e-01 2.545744e-01 3.223350e-01
#>   [91] 1.319059e-01 1.969134e-01 3.060953e-01 1.657595e-01 2.009490e-01
#>   [96] 3.152515e-01 2.237270e-01 2.346606e-01 1.876814e-01 3.002698e-01
#>  [101] 2.138297e-01 1.768283e-01 1.719776e-01 4.543856e-01 2.655358e-01
#>  [106] 8.738729e-02 3.502261e-01 4.071641e-01 3.606030e-01 2.061436e-01
#>  [111] 2.400192e-01 2.117777e-01 2.344797e-01 3.322022e-01 2.828182e-01
#>  [116] 1.994018e-01 2.694450e-01 2.088362e-01 3.975527e-01 2.465959e-01
#>  [121] 9.852624e-03 6.426360e-03 6.438215e-03 1.704232e-02 2.013899e-02
#>  [126] 1.175156e-02 1.124356e-02 3.592003e-03 9.741199e-03 1.332930e-02
#>  [131] 2.172382e-02 2.808922e-03 1.149466e-02 9.565131e-05 9.720563e-03
#>  [136] 1.526583e-02 7.172875e-03 9.796183e-03 4.696962e-03 4.672382e-03
#>  [141] 4.384638e-03 1.263956e-03 1.217133e-05 6.345938e-03 3.757337e-03
#>  [146] 2.210692e-03 1.238756e-02 4.694190e-03 2.097544e-02 2.801255e-02
#>  [151] 1.933410e-02 1.153407e-02 7.121356e-03 2.174340e-03 1.231643e-02
#>  [156] 1.391193e-02 4.246171e-03 1.074381e-02 1.231208e-02 5.684411e-03
#>  [161] 2.555159e-02 1.324984e-02 6.224955e-03 3.657207e-02 1.458213e-02
#>  [166] 2.510936e-02 3.221374e-02 4.506960e-02 2.552265e-02 8.856788e-03
#>  [171] 9.146657e-03 8.518280e-03 1.321929e-02 1.016666e-02 2.032362e-03
#>  [176] 7.698642e-03 1.989731e-03 2.256057e-03 2.100794e-04 1.508663e-02
#>  [181] 1.527642e-01 1.796704e-01 1.281795e-01 3.615707e-01 2.928639e-01
#>  [186] 1.542435e-01 2.462619e-01 2.225295e-01 6.499531e-02 8.282918e-02
#>  [191] 1.511088e-01 4.760574e-02 9.547513e-02 1.257537e-01 2.932164e-01
#>  [196] 1.647449e-01 1.452835e-01 1.400159e-01 1.952689e-01 3.515618e-01
#>  [201] 1.253846e-01 9.739097e-02 1.273517e-01 1.689178e-01 1.618529e-01
#>  [206] 3.599461e-01 1.063879e-01 2.173176e-01 1.998545e-01 2.460855e-01
#>  [211] 1.430067e-01 2.045297e-01 1.697087e-01 3.677994e-01 1.378438e-01
#>  [216] 1.568667e-01 2.179248e-01 1.794137e-01 1.755617e-01 1.498813e-01
#>  [221] 1.715744e-01 1.315221e-01 2.206399e-01 2.429315e-01 1.206237e-01
#>  [226] 1.800034e-01 1.795657e-01 8.718382e-02 2.184104e-01 1.782235e-01
#>  [231] 2.186728e-01 2.611588e-01 3.952215e-01 1.686323e-01 1.747863e-01
#>  [236] 2.011211e-01 1.083513e-01 1.484204e-01 1.837102e-01 1.082053e-01
#>  [241] 1.357707e-01 2.911317e-01 2.526588e-01 4.484140e-01 3.227252e-01
#>  [246] 4.040221e-01 5.376958e-01 2.350703e-01 3.231578e-01 2.844059e-01
#>  [251] 2.500121e-01 2.799886e-01 4.174605e-01 3.596017e-01 3.966721e-01
#>  [256] 1.606319e-01 3.268234e-01 3.829123e-01 2.409372e-01 2.979713e-01
#>  [261] 1.307745e-01 1.560118e-01 1.254726e-01 3.951157e-01 3.179472e-01
#>  [266] 3.476874e-01 2.253851e-01 3.016624e-01 3.724874e-01 3.202625e-01
#>  [271] 2.365164e-01 2.566233e-01 3.258365e-01 3.191789e-01 2.549537e-01
#>  [276] 3.469645e-01 2.262952e-01 4.285522e-01 4.595391e-01 1.288633e-01
#>  [281] 3.699067e-01 2.910941e-01 2.158065e-01 1.588175e-01 1.980530e-01
#>  [286] 2.670581e-01 1.882153e-01 2.742590e-01 1.819113e-01 5.512737e-01
#>  [291] 2.794599e-01 1.349492e-01 1.377067e-01 2.276095e-01 2.453315e-01
#>  [296] 2.505226e-01 3.138389e-01 4.726422e-01 4.322256e-01 3.332152e-01
#>  [301] 1.076663e-02 2.313398e-02 9.442787e-03 1.108927e-02 4.335149e-03
#>  [306] 1.232109e-02 2.025548e-02 1.467075e-02 1.677928e-02 1.730808e-02
#>  [311] 1.362397e-02 1.035097e-02 1.879836e-02 1.217106e-02 1.711409e-02
#>  [316] 2.039021e-02 2.118076e-02 3.806939e-02 6.760291e-03 1.728727e-02
#>  [321] 5.992039e-03 2.690075e-02 5.915753e-03 1.499461e-02 1.426326e-02
#>  [326] 8.268294e-03 2.032227e-02 1.093890e-02 3.086004e-02 3.885145e-02
#>  [331] 2.743736e-02 1.724243e-02 1.082412e-03 5.501552e-03 1.952585e-02
#>  [336] 2.118856e-02 1.080902e-02 1.071165e-02 1.240326e-02 3.230704e-02
#>  [341] 2.073685e-02 2.145652e-02 2.005048e-02 2.677360e-02 1.632468e-02
#>  [346] 8.107387e-03 1.274637e-02 1.268061e-02 5.303975e-03 1.541282e-02
#>  [351] 1.773036e-02 1.522204e-02 3.135172e-02 1.720702e-02 1.814126e-02
#>  [356] 4.682309e-02 4.237910e-02 3.286605e-02 4.241066e-02 2.062076e-02
#>  [361] 7.074076e-02 1.084811e-01 3.785369e-01 9.119682e-02 1.407900e-01
#>  [366] 1.507240e-01 9.756271e-02 1.479071e-01 1.361900e-01 2.348655e-01
#>  [371] 2.085000e-01 1.954749e-01 3.576445e-01 1.474406e-01 1.353841e-01
#>  [376] 1.994995e-01 4.037401e-01 2.162832e-01 2.071688e-01 1.840146e-01
#>  [381] 1.415286e-01 2.299595e-01 1.583310e-01 1.320258e-01 1.511328e-01
#>  [386] 2.805426e-01 1.800038e-01 1.566429e-01 1.823453e-01 1.990218e-01
#>  [391] 1.908167e-01 1.881207e-01 1.110237e-01 5.228874e-02 1.776696e-01
#>  [396] 1.466939e-01 1.717067e-01 1.659635e-01 5.370282e-02 7.571759e-02
#>  [401] 1.245800e-01 2.482530e-01 1.877198e-01 3.650430e-01 2.697678e-01
#>  [406] 2.032869e-01 1.883266e-01 2.609527e-01 1.578610e-01 1.940269e-01
#>  [411] 1.196958e-01 1.201875e-01 4.385454e-02 1.718711e-01 1.842648e-01
#>  [416] 1.964228e-01 8.196635e-02 3.271875e-02 1.301305e-01 1.148655e-01
#>  [421] 2.635986e-01 1.159071e-01 3.064925e-01 1.897133e-01 2.712625e-01
#>  [426] 2.787525e-01 1.706146e-01 3.748412e-01 1.524093e-01 2.150685e-01
#>  [431] 1.458663e-01 1.257862e-01 1.436705e-01 3.172500e-01 2.615925e-01
#>  [436] 1.933655e-01 3.102997e-01 3.292484e-01 4.680489e-01 3.338570e-01
#>  [441] 3.329182e-01 1.618352e-01 2.716692e-01 4.524966e-01 3.836289e-01
#>  [446] 2.576119e-01 2.791898e-01 2.240230e-01 2.130899e-01 1.940578e-01
#>  [451] 2.227694e-01 2.599917e-01 4.354498e-01 2.352342e-01 1.543012e-01
#>  [456] 2.301006e-01 3.811206e-01 1.998297e-01 2.967037e-01 2.265043e-01
#>  [461] 2.352511e-01 5.159589e-01 3.428909e-01 4.430617e-01 2.525732e-01
#>  [466] 3.018244e-01 3.417094e-01 3.794067e-01 3.916541e-01 3.270666e-01
#>  [471] 4.667789e-01 4.942337e-01 4.298100e-01 1.363313e-01 3.069020e-01
#>  [476] 3.118731e-01 1.780823e-01 4.848834e-01 4.600173e-01 3.774410e-01
#>  [481] 8.838789e-03 8.899739e-03 6.214477e-03 2.271140e-02 1.544693e-02
#>  [486] 4.504118e-03 8.497399e-03 6.293776e-03 7.719707e-03 7.510692e-03
#>  [491] 2.679999e-03 1.006498e-02 3.091471e-02 4.243185e-03 1.237838e-02
#>  [496] 8.448197e-03 5.305471e-04 7.557374e-04 4.915520e-03 2.522485e-03
#>  [501] 4.285486e-05 6.581256e-03 1.389523e-04 7.464054e-03 1.221452e-02
#>  [506] 1.327687e-02 1.873465e-03 1.730985e-03 8.662616e-03 9.263645e-03
#>  [511] 2.484887e-02 3.950138e-02 3.199231e-02 4.156508e-02 1.257572e-02
#>  [516] 2.566612e-02 2.325757e-02 2.735374e-02 1.159135e-03 1.701681e-02
#>  [521] 1.940255e-02 1.677947e-02 7.103062e-03 4.981808e-03 4.438444e-03
#>  [526] 8.530153e-03 1.183796e-02 1.532337e-02 1.450310e-03 6.816833e-03
#>  [531] 1.235006e-04 1.045261e-02 2.486892e-03 6.910504e-06 1.156723e-03
#>  [536] 7.388511e-03 2.661337e-04 4.130106e-05 5.002524e-03 1.120063e-02
#>  [541] 3.621282e-01 2.090760e-01 1.748816e-01 1.893446e-01 1.990353e-01
#>  [546] 1.415081e-01 3.098309e-01 1.677928e-01 1.750871e-01 1.956764e-01
#>  [551] 1.567421e-01 1.827619e-01 1.241634e-01 2.622430e-01 6.260600e-02
#>  [556] 1.412346e-01 1.402611e-01 2.264150e-01 1.398240e-01 1.669478e-01
#>  [561] 1.156932e-01 2.168312e-01 2.601523e-01 9.773506e-02 2.351206e-01
#>  [566] 1.338932e-01 2.764789e-01 2.550487e-01 1.013385e-01 3.119199e-01
#>  [571] 1.794229e-01 1.455480e-01 6.092898e-02 2.317489e-01 1.097371e-01
#>  [576] 2.150685e-01 3.277554e-01 3.237160e-01 2.413025e-01 2.901797e-01
#>  [581] 1.782532e-01 1.832570e-01 2.102746e-01 1.233826e-01 1.148281e-01
#>  [586] 2.210847e-01 7.179908e-02 1.406566e-01 3.351416e-01 8.998252e-02
#>  [591] 6.268239e-02 1.968291e-01 7.481833e-02 3.083013e-01 2.288757e-01
#>  [596] 1.479082e-01 1.813192e-01 4.609301e-02 9.919461e-02 1.774067e-01
#>  [601] 3.387589e-01 2.468683e-01 4.114289e-01 2.781292e-01 1.489376e-01
#>  [606] 8.702580e-02 1.081644e-01 2.013238e-01 6.756500e-02 2.842566e-01
#>  [611] 2.622676e-01 1.882489e-01 2.409677e-01 2.120039e-01 4.466991e-01
#>  [616] 3.641369e-01 3.699070e-01 2.730169e-01 3.249675e-01 5.481818e-01
#>  [621] 2.833071e-01 3.624671e-01 3.041100e-01 4.376258e-01 4.084951e-01
#>  [626] 3.083541e-01 3.670819e-01 4.691448e-01 3.723515e-01 2.796077e-01
#>  [631] 3.556375e-01 2.209545e-01 5.748334e-01 1.435610e-01 1.756888e-01
#>  [636] 3.743908e-01 3.721640e-01 3.458028e-01 3.887821e-01 2.144019e-01
#>  [641] 2.519475e-01 1.582711e-01 2.799047e-01 2.506500e-01 3.068248e-01
#>  [646] 3.782141e-01 3.141310e-01 2.865121e-01 4.600202e-01 6.672733e-01
#>  [651] 2.866534e-01 3.303207e-01 2.777685e-01 3.648156e-01 3.598131e-01
#>  [656] 3.609448e-01 5.222195e-01 1.683305e-01 4.606910e-01 5.091703e-01
#>  [661] 4.044196e-03 5.105914e-03 2.738831e-03 3.592073e-03 2.156830e-02
#>  [666] 1.439249e-02 5.562237e-03 1.352636e-02 6.090207e-03 2.781322e-02
#>  [671] 1.092719e-02 4.645903e-03 8.150210e-03 6.388341e-03 4.147722e-03
#>  [676] 4.262908e-07 3.444740e-03 2.020829e-03 1.714354e-03 9.104967e-06
#>  [681] 2.578732e-03 1.554641e-02 2.242702e-02 1.673903e-02 1.102903e-02
#>  [686] 1.945295e-02 2.564569e-02 3.594354e-03 1.182232e-02 8.133889e-03
#>  [691] 1.329373e-02 1.087747e-02 1.260615e-02 3.711165e-02 1.296458e-02
#>  [696] 2.639887e-03 9.466055e-04 2.560142e-05 6.707998e-03 5.263636e-03
#>  [701] 3.087259e-03 6.112737e-04 1.707051e-02 9.946261e-05 5.258435e-06
#>  [706] 2.755168e-03 5.625662e-04 7.567496e-04 8.569157e-03 5.057698e-03
#>  [711] 1.448165e-02 5.518915e-03 1.070799e-02 1.469237e-02 2.143882e-02
#>  [716] 4.983566e-02 9.688942e-03 1.692903e-02 9.144765e-03 8.376878e-03
#>  [721] 2.106414e-01 9.818254e-02 2.200384e-01 3.234718e-01 2.420233e-01
#>  [726] 2.742412e-01 9.554705e-02 5.694055e-02 3.055146e-01 3.165264e-01
#>  [731] 1.482303e-01 1.974947e-01 2.701942e-01 3.053237e-01 2.274398e-01
#>  [736] 2.862148e-01 1.826634e-01 7.789335e-02 1.594622e-01 2.121908e-01
#>  [741] 2.460801e-01 1.665183e-01 1.997199e-01 2.058624e-01 4.431351e-02
#>  [746] 2.529604e-01 2.627935e-01 3.372576e-01 3.443868e-01 9.467388e-02
#>  [751] 1.845044e-01 2.754646e-01 1.689536e-01 1.225917e-01 1.302789e-01
#>  [756] 1.394269e-01 1.971915e-01 2.096184e-01 1.315026e-01 2.079597e-01
#>  [761] 2.339922e-01 2.521648e-01 1.278884e-01 2.608577e-01 1.910316e-01
#>  [766] 1.704691e-01 2.029583e-01 3.500760e-01 1.811330e-01 1.807225e-01
#>  [771] 1.518052e-01 3.711979e-01 3.245516e-01 2.387073e-01 2.596254e-01
#>  [776] 2.181077e-01 2.370751e-01 2.771552e-01 1.949550e-01 3.109334e-01
#>  [781] 3.985044e-01 2.145478e-01 2.305889e-01 2.502652e-01 3.750572e-01
#>  [786] 5.464281e-01 3.415115e-01 3.182526e-01 3.181526e-01 3.000088e-01
#>  [791] 4.251654e-01 3.481739e-01 4.191182e-01 3.274811e-01 2.208351e-01
#>  [796] 3.984820e-01 5.715424e-01 4.174267e-01 2.874350e-01 2.836466e-01
#>  [801] 2.009276e-01 3.054462e-01 4.319738e-01 2.773650e-01 2.509543e-01
#>  [806] 4.243026e-01 2.560260e-01 5.479034e-01 6.532479e-01 4.262676e-01
#>  [811] 4.320930e-01 5.026575e-01 2.957232e-01 4.506280e-01 2.516386e-01
#>  [816] 2.575138e-01 2.340578e-01 2.748808e-01 3.960233e-01 3.016943e-01
#>  [821] 3.705463e-01 2.465097e-01 4.325890e-01 2.170478e-01 4.191949e-01
#>  [826] 4.421264e-01 2.524742e-01 2.265885e-01 4.802367e-01 2.763467e-01
#>  [831] 2.763575e-01 2.092719e-01 3.316736e-01 3.797132e-01 4.315861e-01
#>  [836] 2.071387e-01 2.427046e-01 1.586423e-01 4.285389e-01 4.082396e-01
#>  [841] 7.727264e-05 1.595059e-04 4.847883e-03 1.182970e-02 2.685647e-03
#>  [846] 1.557066e-03 6.197760e-03 2.209443e-02 7.478858e-03 8.819411e-03
#>  [851] 9.469973e-03 5.833498e-03 1.393927e-04 6.677142e-04 4.295221e-03
#>  [856] 9.688991e-04 2.417734e-03 2.008735e-02 4.643218e-04 1.984040e-03
#>  [861] 4.846900e-03 1.704012e-02 8.961242e-03 1.975527e-02 2.407884e-03
#>  [866] 9.751966e-04 3.960815e-03 1.158242e-02 2.930721e-05 1.915087e-03
#>  [871] 7.783429e-05 9.372653e-03 2.666975e-02 1.512602e-02 5.229589e-03
#>  [876] 1.108595e-02 1.876252e-02 1.015134e-02 1.393507e-02 3.079157e-02
#>  [881] 8.769920e-03 5.458442e-03 6.570700e-04 5.661611e-04 9.157688e-03
#>  [886] 1.090537e-02 1.720533e-02 3.522606e-03 6.934217e-03 1.267897e-02
#>  [891] 1.170367e-02 1.237928e-02 2.623520e-02 6.730032e-03 8.670769e-03
#>  [896] 1.648577e-02 3.938145e-03 2.651178e-03 1.752095e-02 4.441121e-03
#>  [901] 2.834411e-01 3.274426e-01 1.544751e-01 5.012079e-01 3.235770e-01
#>  [906] 1.968248e-01 2.218097e-01 2.160817e-01 1.770706e-01 2.437885e-01
#>  [911] 9.956878e-02 2.010571e-01 2.943247e-01 1.830671e-01 2.733225e-01
#>  [916] 3.281098e-01 2.217723e-01 2.068379e-01 1.222904e-01 2.126885e-01
#>  [921] 2.992362e-01 9.629396e-02 3.425864e-01 2.792654e-01 3.813038e-01
#>  [926] 1.651043e-01 2.811376e-01 1.376807e-01 2.786591e-01 2.290628e-01
#>  [931] 1.638958e-01 2.823929e-01 2.080761e-01 1.548050e-01 1.028861e-01
#>  [936] 1.978762e-01 1.778393e-01 2.000174e-01 2.307380e-01 2.185158e-01
#>  [941] 3.247182e-01 1.423718e-01 2.423505e-01 2.216844e-01 1.968348e-01
#>  [946] 1.940000e-01 1.962259e-01 2.119101e-01 1.746682e-01 2.644137e-01
#>  [951] 1.617294e-01 2.862309e-01 1.235319e-01 2.376430e-01 1.506016e-01
#>  [956] 4.587062e-01 3.218310e-01 1.813794e-01 1.625008e-01 3.338044e-01
#>  [961] 4.554750e-01 5.208277e-01 2.996391e-01 4.165974e-01 3.353731e-01
#>  [966] 5.191754e-01 3.080907e-01 4.129866e-01 4.432392e-01 4.770081e-01
#>  [971] 3.239313e-01 3.907665e-01 4.724794e-01 1.851198e-01 2.254631e-01
#>  [976] 4.299289e-01 2.989240e-01 4.433840e-01 4.173785e-01 3.382524e-01
#>  [981] 4.328243e-01 3.353951e-01 4.431536e-01 4.051375e-01 2.742342e-01
#>  [986] 4.822825e-01 2.810836e-01 3.853926e-01 2.824711e-01 2.854119e-01
#>  [991] 3.635868e-01 2.526739e-01 3.436998e-01 3.756954e-01 6.409038e-01
#>  [996] 2.546648e-01 4.169629e-01 3.879044e-01 4.009342e-01 3.372373e-01
#>  [ reached getOption("max.print") -- omitted 9080 entries ]
#> 
#> $T2
#>    [1] 0.3923440 0.3262846 0.2801002 0.2606294 0.3043221 0.3809601 0.2908950
#>    [8] 0.4185174 0.3309040 0.4511591 0.3336295 0.2437454 0.3674753 0.3753276
#>   [15] 0.5276608 0.4216238 0.5181409 0.5912107 0.7341085 0.3882809 0.3712014
#>   [22] 0.2866925 0.2270468 0.2673197 0.4160737 0.2925927 0.2997134 0.4020641
#>   [29] 0.4534558 0.3554380 0.4238411 0.3250669 0.2767344 0.3402560 0.5865947
#>   [36] 0.4053636 0.2778908 0.3953259 0.2696738 0.2783899 0.3347041 0.3875430
#>   [43] 0.2794312 0.2752325 0.3509383 0.2752633 0.2802154 0.2238150 0.2517655
#>   [50] 0.4000453 0.2679667 0.3759309 0.4417719 0.3881606 0.2268690 0.3037601
#>   [57] 0.4705146 0.3508513 0.2761554 0.3394343 0.4244760 0.4773080 0.5916395
#>   [64] 0.8076414 0.7755469 0.9000189 1.0523148 0.8439708 0.9670112 0.7022529
#>   [71] 0.8510062 0.5274617 0.4346619 0.6069470 0.8146737 0.6007022 0.5958514
#>   [78] 0.7107734 0.7606128 0.4705137 0.7942745 0.6797649 0.7403043 0.8130488
#>   [85] 0.6850111 0.8876129 0.9923180 1.0595844 0.8651866 0.7889405 0.7222983
#>   [92] 0.8654169 1.2501481 0.7956359 0.7483530 0.9250691 1.0519082 0.9805974
#>   [99] 0.7682423 0.6989169 0.6912463 0.7518078 1.0763763 1.1790796 0.7758385
#>  [106] 0.7923735 0.7616032 0.6764666 0.7276081 0.7600951 0.6619177 0.6781508
#>  [113] 0.9127415 0.6229735 1.1219067 0.8220949 0.6594780 0.8155761 0.9042042
#>  [120] 0.5308027 1.0911964 1.1508660 1.0761055 1.1606629 1.1635459 1.1077922
#>  [127] 1.1202174 1.1156922 1.1541711 1.1192359 1.1448109 1.1429746 1.1162647
#>  [134] 1.1025807 1.0721696 1.0966071 1.1282732 1.1047098 1.0699006 1.1546235
#>  [141] 1.1178498 1.1018162 1.0832206 1.0939434 1.1008875 1.0918794 1.2010173
#>  [148] 1.0994323 1.1279318 1.1703610 1.2131681 1.1803067 1.1444721 1.0782360
#>  [155] 1.1250484 1.1806504 1.1103315 1.1593246 1.1035268 1.1022674 1.1256462
#>  [162] 1.1042724 1.1524508 1.2520086 1.1293844 1.1786905 1.1468090 1.2383482
#>  [169] 1.1678269 1.1355620 1.1000307 1.1281975 1.1369703 1.1196793 1.1348039
#>  [176] 1.0781034 1.1348799 1.0683958 1.1156389 1.1855667 0.4491690 0.3388966
#>  [183] 0.7883086 0.5451737 0.4691087 0.3762479 0.5237537 0.6000887 0.7070324
#>  [190] 0.7499785 0.5624471 0.9362314 0.4591986 0.5110282 0.3108705 0.3124060
#>  [197] 0.2036571 0.2656759 0.4171084 0.4335567 0.3932990 0.4184954 0.5796186
#>  [204] 0.4344289 0.3711702 0.3823878 0.2360357 0.4227188 0.2855917 0.3145241
#>  [211] 0.2950376 0.2885038 0.3486198 0.3618175 0.3431389 0.1993373 0.2379035
#>  [218] 0.3412483 0.5944048 0.5238318 0.3066687 0.7897896 0.6111744 0.3535038
#>  [225] 0.2483685 0.2103495 0.2457286 0.2284461 0.2267383 0.2855582 0.2243245
#>  [232] 0.4282999 0.4704934 0.3217692 0.3044004 0.3865223 0.5863445 1.0647977
#>  [239] 0.4293729 0.6162881 0.9378672 1.0062434 0.8403531 0.8678058 0.5152263
#>  [246] 0.5756913 0.6157386 0.5948950 0.6465605 0.6124838 0.7963941 0.7040083
#>  [253] 0.6778300 0.6315159 0.4577640 0.5357468 0.5402081 0.7966032 0.6073865
#>  [260] 1.0767400 0.9702285 0.7677764 0.5482988 0.7289588 0.6531604 0.5553976
#>  [267] 0.5082222 0.6254486 0.7210226 0.5652199 0.5295463 0.5938628 0.6344453
#>  [274] 0.7480879 0.5644840 0.7135527 0.9238209 0.9275170 0.8939472 0.9629654
#>  [281] 0.7244679 0.9935242 1.1136042 1.0876023 0.8144681 0.7486489 1.1637562
#>  [288] 0.8510733 1.0215088 0.9707573 0.9788757 0.5447364 0.8954481 0.5490403
#>  [295] 0.6388409 0.7474359 0.7699604 0.6325648 0.7058695 0.8555846 1.1793435
#>  [302] 1.1429497 1.1763830 1.1216569 1.1484551 1.1090271 1.1297756 1.1460813
#>  [309] 1.1576650 1.1239752 1.1299652 1.1502880 1.1165390 1.1578616 1.1845946
#>  [316] 1.2171546 1.2200928 1.1612087 1.1326165 1.1214497 1.1583674 1.1437245
#>  [323] 1.1068867 1.1124129 1.1750515 1.1312936 1.1178597 1.1524200 1.2013512
#>  [330] 1.1884591 1.1649072 1.1415194 1.0538316 1.1052294 1.0805641 1.1539174
#>  [337] 1.0627556 1.1280903 1.1533653 1.1951106 1.1875375 1.2000716 1.1020868
#>  [344] 1.1771354 1.1889659 1.1021027 1.1331918 1.1651658 1.1350781 1.1309772
#>  [351] 1.0845266 1.1023492 1.1414040 1.1526978 1.1889237 1.1591829 1.2938765
#>  [358] 1.2111277 1.1943536 1.2011513 0.4058627 0.5037069 0.3839735 0.3282547
#>  [365] 0.4174897 0.4411072 0.2864876 0.6206830 0.4357289 0.4391918 0.4332491
#>  [372] 0.5417095 0.2915434 0.2859149 0.2820333 0.3826264 0.3322630 0.3464695
#>  [379] 0.2461247 0.2480156 0.2051920 0.2936524 0.2794454 0.4151990 0.2465519
#>  [386] 0.3287668 0.4413308 0.2818067 0.2404263 0.5102390 0.6333542 0.2993773
#>  [393] 0.2584932 1.1072712 0.7818927 0.3544052 0.7373463 1.0871266 0.9766416
#>  [400] 0.5683480 0.8101304 0.8480046 0.3477468 0.3097819 0.2991701 0.2834461
#>  [407] 0.2791253 0.5348431 0.3390945 0.2488675 0.3601949 0.6336066 0.8359480
#>  [414] 0.5626030 0.7011545 0.6356769 0.7366977 0.8778690 0.7945198 0.7329372
#>  [421] 1.1805004 1.0166164 1.0104152 1.0369053 0.9328838 1.3462035 1.0807671
#>  [428] 1.1782871 0.8196531 0.9756223 1.1148873 1.1263756 1.1833311 1.1407552
#>  [435] 0.7932518 0.5860322 0.5048843 0.6970897 0.5249355 0.6985476 0.6533571
#>  [442] 0.6336114 0.3858521 0.6412289 0.8208120 0.4918186 0.5211260 0.6516606
#>  [449] 0.8556560 0.8463528 0.7808793 0.9184184 0.8549126 0.8471023 0.9563738
#>  [456] 0.7663659 0.7004733 0.7274011 0.5571577 0.4967476 0.6181437 0.8028299
#>  [463] 0.4259706 0.8334096 1.0206000 0.5444203 0.8977297 0.7609301 0.8230410
#>  [470] 0.6584878 0.6135933 0.5865503 0.4631597 0.5339587 0.4804462 0.5228534
#>  [477] 0.6781262 0.6350575 0.4966129 0.4991094 1.1481632 1.1119738 1.1177842
#>  [484] 1.0907163 1.1321047 1.0575923 1.1330501 1.1739445 1.1298551 1.1572194
#>  [491] 1.1402340 1.1267858 1.1848689 1.0788794 1.1494040 1.0764210 1.1233580
#>  [498] 1.0671270 1.0998363 1.0784619 1.0239998 1.0212151 1.0813273 1.1571751
#>  [505] 1.0722479 1.0874083 1.0906330 1.0771178 1.0919012 1.1351628 1.1672045
#>  [512] 1.2379364 1.2193321 1.1899706 1.1555480 1.2095521 1.1506649 1.2200731
#>  [519] 1.0740177 1.2208530 1.1282068 1.2086086 1.1150295 1.1032060 1.1078190
#>  [526] 1.1268043 1.1667938 1.1482511 1.1191585 1.0574690 1.0646186 1.1538501
#>  [533] 1.0776211 1.1217275 1.0244754 1.1307485 1.1218724 1.1068236 1.1414293
#>  [540] 1.1737244 0.3363172 0.4055799 0.2883573 0.2629755 0.2144104 0.3851099
#>  [547] 0.4109821 0.2264445 0.2845396 0.2288187 0.2630676 0.4350332 0.3638058
#>  [554] 0.3340184 0.1972040 0.2801626 0.2392976 0.3423591 0.5070640 0.7033818
#>  [561] 0.4935206 0.3695312 0.3085796 0.2410375 0.4706088 0.8859100 0.4869823
#>  [568] 0.3654742 0.5698147 0.4721538 0.4943206 0.2621274 0.4888422 0.3281161
#>  [575] 0.5574966 0.3900352 0.2275601 0.3977329 0.2760151 0.7591158 0.3001957
#>  [582] 0.6065930 0.4239231 0.6591536 0.5766989 0.4875831 0.3710799 0.2073259
#>  [589] 0.2733245 0.2947307 0.5735523 0.3761317 0.3966888 0.4807896 0.3359475
#>  [596] 0.3474831 0.5416009 0.7472837 0.4606796 0.3524152 0.7912397 0.6849301
#>  [603] 0.8373071 0.7668110 1.0541028 1.0060859 1.1572243 0.8888626 0.9903048
#>  [610] 1.2276471 0.6235825 0.7365312 0.6800244 0.5405890 0.4266264 0.4936514
#>  [617] 0.4890153 0.5475344 0.7110129 0.7795671 0.5181705 0.6677757 0.3534032
#>  [624] 0.3036965 0.4078750 0.5963949 0.8506029 0.8995105 0.8228692 0.6564934
#>  [631] 0.7450748 0.7902526 0.6911524 0.7933038 0.6964279 0.6449280 0.6202688
#>  [638] 0.9086275 0.9087264 0.8283307 0.5929866 0.9747043 0.6526697 0.7329093
#>  [645] 0.5445057 0.6150186 0.5845843 0.3623388 0.4633492 0.4303025 0.5739001
#>  [652] 0.5485837 0.5384166 0.4330102 0.3848091 0.4752459 0.3525636 0.3234839
#>  [659] 0.5367296 0.4749470 1.0617492 1.1305482 1.1178296 1.0662055 1.1159028
#>  [666] 1.1341288 1.0832489 1.0724965 1.1333920 1.1350539 1.1975927 1.0570316
#>  [673] 1.1472793 1.0942907 1.0733845 1.1020840 1.0297530 1.0349156 1.0758375
#>  [680] 1.0984579 1.1088634 1.1161589 1.1466821 1.1188778 1.1439532 1.1854746
#>  [687] 1.1634658 1.1292350 1.1091853 1.1491484 1.0965282 1.1548254 1.1333432
#>  [694] 1.1712806 1.1215278 1.1466498 1.0267634 1.0445994 1.1390727 1.1437404
#>  [701] 1.1049235 1.0932485 1.1013371 1.0988632 1.0727204 1.1581518 1.0815065
#>  [708] 1.0793179 1.1141791 1.1676261 1.1084499 1.0531512 1.0765531 1.1097292
#>  [715] 1.1531781 1.1881356 1.1067750 1.1884350 1.1152309 1.0962788 0.2901074
#>  [722] 0.2093891 0.2704045 0.2999453 0.3160192 0.3043727 0.2983132 0.2467894
#>  [729] 0.3693259 0.2979769 0.2254481 0.3323726 0.4029017 0.3022168 0.2675184
#>  [736] 0.2884389 0.3201853 0.1325553 0.2626956 0.2996064 0.2226760 0.3140111
#>  [743] 0.3044280 0.2337613 0.2030424 0.4041352 0.3610416 0.4128312 0.2971281
#>  [750] 0.2783445 0.2351974 0.2647350 0.4875611 0.5569640 0.2654678 0.6330080
#>  [757] 0.3525319 0.7199208 0.2749718 0.4834903 0.2651629 0.3864624 0.2510255
#>  [764] 0.3104933 0.1886159 0.2623256 0.3860738 0.3979184 0.3622122 0.4194126
#>  [771] 0.2077380 0.5177733 0.4268675 0.3659035 0.3600201 0.3487952 0.2812739
#>  [778] 0.2839733 0.3430673 0.4246043 0.7647354 0.7425627 0.5635601 0.4414883
#>  [785] 0.6134703 0.8081562 0.4618027 0.5653172 0.6033355 0.3512968 0.4028363
#>  [792] 0.4665707 0.4619358 0.4371133 0.3788572 0.6424817 0.6158077 0.5104371
#>  [799] 0.7413106 0.6221697 0.5026634 0.5660220 0.7551586 0.6493325 0.7774258
#>  [806] 0.4997924 0.5034688 0.5279845 0.5565341 0.4959105 0.4267924 0.4800800
#>  [813] 0.7415478 0.5335774 0.4081975 0.4260278 0.4053552 0.6409132 0.5666486
#>  [820] 0.3079295 0.4092305 0.3309868 0.4384991 0.4412707 0.6670385 0.7519498
#>  [827] 0.4097928 0.6569512 0.5711101 0.6323206 0.5301111 0.5211038 0.5962182
#>  [834] 0.5541381 0.9076410 0.8058778 0.8306927 1.0779427 0.6599122 0.4945599
#>  [841] 1.0950402 1.1276192 1.0949144 1.1716053 1.1081884 1.0508820 1.1196611
#>  [848] 1.2217040 1.0885044 1.1224770 1.1756024 1.0960763 1.0396366 1.1254098
#>  [855] 1.1455623 1.0711344 1.1250356 1.1515330 1.0714908 1.1523672 1.0963572
#>  [862] 1.1507359 1.2008713 1.1605331 1.0912429 1.0857167 1.0749842 1.0591309
#>  [869] 1.0799421 1.1231780 1.0774325 1.1380122 1.1919902 1.1605892 1.1437959
#>  [876] 1.1691590 1.1628979 1.1441980 1.1616972 1.1642058 1.0979719 1.1256607
#>  [883] 1.0666281 1.0939903 1.1719282 1.1003786 1.1161810 1.1223432 1.1177638
#>  [890] 1.1123678 1.1600839 1.1935958 1.1470246 1.1434467 1.1408953 1.1230143
#>  [897] 1.1048173 1.0884716 1.1590719 1.1290576 0.2335962 0.4454549 0.3817030
#>  [904] 0.4625712 0.3383355 0.3273365 0.3194966 0.3130161 0.2645484 0.2319886
#>  [911] 0.2149164 0.4005369 0.3058530 0.2588315 0.3514523 0.3831784 0.3017118
#>  [918] 0.2555443 0.2694154 0.2918067 0.3386770 0.2495666 0.4065702 0.3106212
#>  [925] 0.5729391 0.3855022 0.4833597 0.2716256 0.3511413 0.3511010 0.2358284
#>  [932] 0.2778183 0.4765985 0.3168481 0.3299844 0.2469269 0.3657923 0.3014448
#>  [939] 0.3003136 0.3227275 0.4054753 0.2552110 0.3371863 0.2319652 0.2969319
#>  [946] 0.2408583 0.2816527 0.2713713 0.2581927 0.3115259 0.4162270 0.6334520
#>  [953] 0.4998268 0.6376886 0.2927938 0.6072288 0.5578452 0.2495265 0.2090927
#>  [960] 0.3903365 0.4761728 0.3806643 0.6295848 0.5992207 0.4320182 0.5390088
#>  [967] 0.3821356 0.3406015 0.3731040 0.3886694 0.5903157 0.6492255 0.5908043
#>  [974] 0.7182455 0.6429272 0.6556458 0.6791702 1.0227838 0.5406930 0.7059690
#>  [981] 0.6766085 0.7127746 0.6169989 0.6172891 0.4160394 0.4879716 0.5228864
#>  [988] 0.6690032 0.4314899 0.5618086 0.4532387 0.4018414 0.3879460 0.4298722
#>  [995] 0.8422254 0.5021693 0.4241005 0.5185456 0.4474673 0.3753938
#>  [ reached getOption("max.print") -- omitted 9080 entries ]
#> 
#> attr(,"class")
#> [1] "pca"