Adaptive Process Training

Apply Adaptive-Dynamic PCA to state-specific data matrices.

processMonitor(
  data,
  trainObs,
  updateFreq = ceiling(0.5 * trainObs),
  faultsToTriggerAlarm = 5,
  ...
)

Arguments

data

An xts data matrix

trainObs

The number of training observations to be used

updateFreq

The number of non-flagged observations to collect before the function updates. Defaults to half as many observations as the number of training observations.

faultsToTriggerAlarm

The number of sequential faults needed to trigger an alarm. Defaults to 5.

...

Lazy dots for additional internal arguments

Value

A list with the following components:

FaultChecks -- a class-specific xts flagging matrix with the same number of rows as "data". This flag matrix has the following five columns:

SPE -- the SPE statistic value for each observation in "data"

SPE_Flag -- a vector of SPE indicators recording 0 if the test statistic is less than or equal to the critical value passed through from the threshold object

T2 -- the T2 statistic value for each observation in "data"

T2_Flag -- a vector of T2 fault indicators, defined like SPE_Flag

Alarm -- a column indicating if there have been five flags in a row for either the SPE or T2 monitoring statistics or both. Alarm states are as follows: 0 = no alarm, 1 = Hotelling's T2 alarm, 2 = Squared Prediction Error alarm, and 3 = both alarms.

Non_Alarmed_Obs -- a class-specific xts data matrix of all the non-alarmed observations (observations with alarm state equal to 0)

Alarms -- a class-specific xts data matrix of the features and specific alarms of Alarmed observations, where the alarm codes are listed above

trainSpecs -- a threshold object returned by the internal threshold() function. See the threshold() function's help file for more details.

Details

This function is the class-specific implementation of the Adaptive- Dynamic PCA described in the details of the mspTrain() function. See the mspTrain() function's help file for further details.

This internal function is called by mspTrain(). This function calls the faultFilter() function.

See also

Examples

nrml <- mspProcessData(faults = "NOC")
data <- nrml[nrml[,1] == 1]

processMonitor(data = data[,-1], trainObs = 672)
#> Warning: In density.default(spe, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(t2, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(spe, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(t2, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(spe, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(t2, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> $FaultChecks
#>                             SPE SPE_Flag        T2 T2_Flag Alarm
#> 2015-05-17 19:12:00 0.017099512        0 12.377689       1     0
#> 2015-05-17 19:13:00 0.178948306        0  4.182877       0     0
#> 2015-05-17 19:14:00 0.144254436        0  4.191625       0     0
#> 2015-05-17 19:15:00 0.040794627        0  3.591733       0     0
#> 2015-05-17 19:16:00 0.187026482        0  2.636932       0     0
#> 2015-05-17 19:17:00 0.006268561        0  1.916088       0     0
#> 2015-05-17 19:18:00 0.039742207        0  3.991707       0     0
#> 2015-05-17 19:19:00 0.187852481        0  5.715088       0     0
#> 2015-05-17 19:20:00 0.060112960        0  6.866835       0     0
#> 2015-05-17 19:21:00 0.696066014        0  7.405213       0     0
#>                 ...                                             
#> 2015-05-23 07:50:00 0.349494671        0  2.273096       0     0
#> 2015-05-23 07:51:00 0.115128121        0  2.063812       0     0
#> 2015-05-23 07:52:00 0.100532672        0  1.482424       0     0
#> 2015-05-23 07:53:00 0.073972147        0  2.767771       0     0
#> 2015-05-23 07:54:00 0.324144542        0  2.772795       0     0
#> 2015-05-23 07:55:00 0.484832762        0  3.461854       0     0
#> 2015-05-23 07:56:00 0.065964019        0  4.655891       0     0
#> 2015-05-23 07:57:00 1.075060821        0  2.959837       0     0
#> 2015-05-23 07:58:00 0.319744985        0  3.612641       0     0
#> 2015-05-23 07:59:00 0.809103687        0  4.325981       0     0
#> 
#> $Non_Alarmed_Obs
#>                             x          y          z
#> 2015-05-17 19:12:00 0.5256947 -1.7879543 1.37258581
#> 2015-05-17 19:13:00 0.7528171 -1.6582566 1.30012301
#> 2015-05-17 19:14:00 0.7587905 -1.6742932 1.29129494
#> 2015-05-17 19:15:00 0.6301363 -1.5950822 1.15233382
#> 2015-05-17 19:16:00 0.7125945 -1.5064361 1.14970031
#> 2015-05-17 19:17:00 0.6854033 -1.5021981 0.95487285
#> 2015-05-17 19:18:00 0.5708402 -1.5477285 1.12113834
#> 2015-05-17 19:19:00 0.8199780 -1.7864237 1.42414440
#> 2015-05-17 19:20:00 0.9272340 -1.6747013 1.14709677
#> 2015-05-17 19:21:00 0.7794067 -1.7224379 1.59611851
#>                 ...                                
#> 2015-05-23 07:50:00 0.2528888 -1.1350589 0.33600138
#> 2015-05-23 07:51:00 0.4156016 -0.9650732 0.31834752
#> 2015-05-23 07:52:00 0.4496043 -1.0303973 0.44941646
#> 2015-05-23 07:53:00 0.2805888 -0.9594964 0.31745127
#> 2015-05-23 07:54:00 0.2989254 -0.8530296 0.43304507
#> 2015-05-23 07:55:00 0.2590342 -0.7633141 0.38508626
#> 2015-05-23 07:56:00 0.1994790 -0.7778623 0.08331553
#> 2015-05-23 07:57:00 0.4757331 -0.7435511 0.11551501
#> 2015-05-23 07:58:00 0.3590397 -0.7705276 0.08944005
#> 2015-05-23 07:59:00 0.3554212 -0.6151355 0.07974210
#> 
#> $Alarms
#>                             x         y        z        SPE SPE_Flag       T2
#> 2015-05-17 22:59:00 1.0785893 -2.006377 1.933348 1.15336315        0 12.19903
#> 2015-05-18 01:00:00 1.2184138 -1.899103 1.675517 0.87189048        0 15.93482
#> 2015-05-18 01:12:00 0.8334583 -2.053873 1.823581 0.37496372        0 13.38613
#> 2015-05-18 01:24:00 1.0620434 -2.015476 2.416602 3.58505413        1 18.92469
#> 2015-05-18 01:25:00 0.8927476 -2.069960 1.864947 0.48317937        0 12.90862
#> 2015-05-18 01:26:00 0.8808866 -1.915003 1.875607 1.02830519        0 11.59861
#> 2015-05-18 01:37:00 1.0297944 -1.902196 1.893896 1.38639810        1 10.76836
#> 2015-05-18 01:38:00 1.1460500 -1.826738 2.067035 2.93989041        1 12.87132
#> 2015-05-18 01:39:00 0.9617742 -2.106569 2.219072 1.67343484        1 18.53735
#> 2015-05-18 01:40:00 0.8339301 -1.901641 1.933422 1.22973902        1 13.61365
#>                 ...                                                          
#> 2015-05-21 16:32:00 1.1372151 -1.824746 2.102799 3.15120400        1 12.90588
#> 2015-05-21 16:33:00 0.7670237 -2.138714 1.740186 0.06230597        0 14.91259
#> 2015-05-21 16:43:00 0.9323193 -2.004335 2.020997 1.27280692        1 14.05318
#> 2015-05-21 16:53:00 1.0716646 -1.975622 2.133658 2.19734224        1 13.91230
#> 2015-05-21 16:54:00 1.1021654 -1.993571 2.090444 1.93850749        1 13.65621
#> 2015-05-21 16:55:00 1.0997405 -2.284974 2.467440 2.16908160        1 22.94133
#> 2015-05-21 16:56:00 1.0416322 -2.121402 2.268852 1.98323237        1 17.88508
#> 2015-05-21 16:57:00 0.9341825 -2.086657 2.313031 2.18690199        1 21.33499
#> 2015-05-21 16:58:00 0.9617286 -2.182209 2.060458 0.74675835        0 16.33869
#> 2015-05-21 16:59:00 1.0323102 -2.254298 2.239553 1.17823482        0 19.25092
#>                     T2_Flag Alarm
#> 2015-05-17 22:59:00       1     1
#> 2015-05-18 01:00:00       1     1
#> 2015-05-18 01:12:00       1     1
#> 2015-05-18 01:24:00       1     1
#> 2015-05-18 01:25:00       1     1
#> 2015-05-18 01:26:00       1     1
#> 2015-05-18 01:37:00       1     1
#> 2015-05-18 01:38:00       1     1
#> 2015-05-18 01:39:00       1     1
#> 2015-05-18 01:40:00       1     1
#>                 ...              
#> 2015-05-21 16:32:00       1     1
#> 2015-05-21 16:33:00       1     1
#> 2015-05-21 16:43:00       1     1
#> 2015-05-21 16:53:00       1     1
#> 2015-05-21 16:54:00       1     1
#> 2015-05-21 16:55:00       1     1
#> 2015-05-21 16:56:00       1     1
#> 2015-05-21 16:57:00       1     3
#> 2015-05-21 16:58:00       1     1
#> 2015-05-21 16:59:00       1     1
#> 
#> $TrainingSpecs
#> $SPE_threshold
#>    99.9% 
#> 2.905109 
#> 
#> $T2_threshold
#>    99.9% 
#> 7.743168 
#> 
#> $projectionMatrix
#>            [,1]
#> [1,]  0.5677389
#> [2,] -0.5786468
#> [3,]  0.5855258
#> 
#> $LambdaInv
#>           [,1]
#> [1,] 0.3631605
#> 
#> $T2
#>    [1] 1.476127e+00 2.437339e+00 1.834104e+00 3.704796e+00 1.814638e+00
#>    [6] 1.717015e+00 3.804953e+00 2.095315e+00 3.018441e+00 1.901498e+00
#>   [11] 1.431233e+00 3.994513e-01 2.952042e+00 2.879588e+00 4.205777e+00
#>   [16] 1.739842e+00 1.842364e+00 1.404226e+00 2.559148e+00 1.928700e+00
#>   [21] 1.500385e+00 1.034767e+00 1.593346e+00 2.200737e+00 8.639411e-01
#>   [26] 1.131922e+00 8.875439e-01 1.640837e+00 9.076006e-01 2.115942e+00
#>   [31] 6.623396e-01 7.995639e-01 4.265147e-01 7.368226e-01 1.549935e+00
#>   [36] 8.232875e-01 9.634665e-01 5.034774e-01 1.097286e+00 2.320390e+00
#>   [41] 1.406047e+00 1.562894e+00 1.599075e+00 1.840960e+00 1.059159e+00
#>   [46] 5.551157e-01 6.670686e-01 2.704154e-01 1.085160e+00 1.097650e+00
#>   [51] 4.855942e-01 1.312622e-01 1.528194e-01 8.226220e-01 7.120145e-01
#>   [56] 2.064244e-01 1.473789e+00 1.699317e+00 9.339424e-01 9.366308e-01
#>   [61] 1.976323e+00 8.425945e-01 1.987588e-03 8.559300e-01 1.832355e-01
#>   [66] 8.272133e-01 6.081390e-01 1.416798e+00 1.162590e+00 2.607065e+00
#>   [71] 1.455422e+00 1.983443e+00 1.557910e+00 2.402330e+00 2.668392e+00
#>   [76] 1.378260e+00 3.166804e+00 1.675690e+00 1.783500e+00 2.202270e+00
#>   [81] 2.203220e+00 8.241336e-01 1.372843e+00 6.365069e-01 2.232977e+00
#>   [86] 1.861094e+00 7.719470e-01 2.205162e-03 5.864289e-02 4.116368e-03
#>   [91] 1.161213e-01 2.709120e-01 3.070968e-01 5.954703e-01 6.481650e-01
#>   [96] 1.816845e+00 1.895204e+00 2.156304e+00 1.501007e+00 8.619766e-01
#>  [101] 1.509808e-01 6.053313e-01 4.122814e-01 9.013275e-01 6.161066e-01
#>  [106] 2.996306e-01 3.288214e-01 3.987746e-01 2.001073e+00 4.784485e-01
#>  [111] 6.626789e-01 8.802357e-01 1.172110e+00 1.362773e-01 6.437102e-02
#>  [116] 5.666161e-01 5.980517e-01 1.481719e+00 5.373150e-01 8.007826e-01
#>  [121] 1.596311e+00 8.701777e-01 1.496032e+00 1.677040e+00 1.637825e+00
#>  [126] 9.010541e-01 1.081726e-01 6.826454e-01 4.538025e-02 3.878856e-01
#>  [131] 2.900169e-01 1.689716e+00 2.022502e+00 5.870216e-01 6.374097e-01
#>  [136] 2.082005e+00 3.599958e-01 5.208637e-01 7.165310e-01 6.524769e-01
#>  [141] 6.476933e-01 1.785833e-01 1.370813e-01 2.762095e-01 9.972747e-02
#>  [146] 1.036071e+00 2.827693e-03 1.872683e-02 7.086048e-02 1.296197e+00
#>  [151] 4.739444e-01 1.264499e-01 1.749086e-01 1.185791e+00 6.512840e-01
#>  [156] 3.181047e-01 1.978089e-01 5.136757e-01 1.882808e-01 1.610586e-02
#>  [161] 3.631201e-01 1.002970e+00 3.211182e-01 6.097202e-01 1.031723e+00
#>  [166] 1.713020e-01 1.390467e-02 3.016640e-01 8.806586e-03 1.719047e+00
#>  [171] 1.037233e+00 8.164265e-01 1.048861e+00 1.288746e+00 1.316592e+00
#>  [176] 1.357232e+00 1.174167e-01 1.271987e-01 1.746518e-02 2.965812e-01
#>  [181] 5.546146e-02 2.482491e-01 9.631309e-02 5.839875e-02 7.611003e-01
#>  [186] 9.442533e-02 1.257583e+00 3.460445e-01 1.074898e+00 1.579525e+00
#>  [191] 4.616013e-01 5.440211e-01 8.226432e-01 2.737546e-01 3.176104e-01
#>  [196] 1.917261e-02 8.440693e-02 5.625643e-02 1.028584e-01 4.386759e-01
#>  [201] 2.708730e-02 6.922434e-01 9.215374e-01 3.031453e+00 2.843610e-01
#>  [206] 7.510328e-02 7.380791e-02 2.645438e-03 4.960184e-01 8.538476e-02
#>  [211] 7.811864e-02 4.646031e-02 9.695853e-02 7.271631e-07 6.777684e-01
#>  [216] 5.855728e-01 9.900366e-02 9.406250e-05 1.054916e-01 4.874218e-02
#>  [221] 2.067434e-01 4.994546e-01 6.814316e-01 8.647652e-02 1.290800e-01
#>  [226] 1.344157e-01 2.311302e-02 5.718770e-02 1.788420e-02 2.787583e-01
#>  [231] 2.573024e-01 1.451522e-02 2.941865e-01 4.114870e-01 1.799482e-01
#>  [236] 6.098278e-01 1.846351e-01 3.600431e-01 6.152495e-01 1.950102e-01
#>  [241] 7.226554e-03 1.023861e-02 2.968560e-04 1.318597e-01 2.565681e-01
#>  [246] 2.957141e-04 1.309124e-01 3.879704e-02 3.029046e-02 1.341182e-02
#>  [251] 2.750819e-02 6.475213e-01 3.663542e-02 2.425993e-01 1.822377e-02
#>  [256] 6.691408e-02 2.963259e-01 3.542427e-02 2.529544e-03 7.427007e-01
#>  [261] 1.847515e-01 2.857209e-01 2.412880e-01 1.195830e-01 9.716399e-01
#>  [266] 2.183334e-01 1.486988e+00 1.954315e+00 6.974896e-03 9.594611e-02
#>  [271] 3.271611e-01 1.787703e-01 3.563458e-01 8.350482e-01 4.409038e-02
#>  [276] 5.169759e-02 2.110230e-01 4.881317e-01 2.829247e-01 1.098871e-01
#>  [281] 3.344274e-01 2.257339e-03 1.965427e-01 5.423982e-03 3.177653e-01
#>  [286] 2.786173e-01 5.439154e-01 7.818065e-01 6.462314e-02 6.550462e-01
#>  [291] 1.519607e-01 4.764036e-01 8.398391e-01 3.969152e-01 1.386803e-01
#>  [296] 2.971006e-01 4.400745e-05 4.682125e-04 2.698193e-02 4.088869e-04
#>  [301] 9.146965e-03 2.952505e-03 4.084802e-01 5.144655e-02 7.230404e-02
#>  [306] 1.060832e-01 3.471983e-02 1.542795e-01 3.280317e-01 2.036371e-01
#>  [311] 2.606821e-01 1.373537e-01 1.484176e-01 2.228860e-01 7.264392e-03
#>  [316] 1.806602e-01 5.314798e-01 5.140570e-02 1.743461e-02 1.734150e-02
#>  [321] 5.186620e-03 1.020255e+00 4.068841e-01 2.292272e-01 1.090419e+00
#>  [326] 1.254722e+00 2.525057e-01 1.715217e-02 6.329281e-02 1.234439e-01
#>  [331] 1.837742e-01 3.786810e-02 1.044065e-01 5.469942e-02 4.657846e-02
#>  [336] 4.253468e-02 1.420989e-01 3.104833e-01 3.338874e-01 3.375560e-02
#>  [341] 4.909674e-02 1.124002e-03 1.940956e-04 7.968281e-01 5.630151e-01
#>  [346] 7.756264e-01 4.783368e-01 7.959201e-02 1.651992e-01 2.708418e-04
#>  [351] 1.914171e-01 1.097125e+00 1.555214e-01 2.095169e-01 1.400253e-01
#>  [356] 2.576150e-01 3.313245e-01 2.380594e-01 5.170315e-01 1.209474e+00
#>  [361] 5.196221e-01 1.392779e+00 2.499139e+00 1.148404e+00 2.162838e-01
#>  [366] 5.676498e-03 3.227043e-02 4.233766e-01 2.788613e-01 1.529493e-02
#>  [371] 1.995427e-02 5.858212e-01 1.887729e-01 6.089209e-02 4.039925e-02
#>  [376] 3.577714e-01 8.350524e-02 5.179714e-02 1.724825e-01 3.953482e-01
#>  [381] 2.482001e-02 5.247587e-02 1.044977e-01 1.612857e-01 6.160805e-01
#>  [386] 6.307492e-01 9.892027e-01 1.201726e+00 1.418207e+00 9.023722e-01
#>  [391] 1.470640e+00 1.106733e+00 6.277637e-01 1.077606e+00 2.559188e+00
#>  [396] 1.160895e-01 7.400504e-01 7.357253e-01 1.366086e-01 2.902645e-01
#>  [401] 2.847623e-01 1.184916e-01 2.183482e-01 2.176784e+00 1.715905e+00
#>  [406] 6.789378e-01 1.333663e+00 7.629183e-01 6.176957e-01 8.101403e-01
#>  [411] 1.145206e+00 5.831766e-01 2.497801e+00 2.335712e+00 3.014270e+00
#>  [416] 1.033176e+00 3.347808e-01 3.011829e-04 1.472349e-01 1.120659e+00
#>  [421] 4.076388e-01 5.525031e-01 1.631926e+00 2.009851e-01 1.622967e+00
#>  [426] 3.797712e-01 7.403274e-01 1.402022e+00 2.163002e+00 8.837420e-01
#>  [431] 1.110429e+00 3.467452e-01 7.089872e-01 2.167690e+00 1.203699e+00
#>  [436] 1.480453e+00 7.622990e-01 1.063953e-01 1.996506e+00 2.717050e+00
#>  [441] 6.084949e+00 1.912010e+00 3.829289e+00 1.037197e+00 1.107222e+00
#>  [446] 4.774028e-01 1.674894e+00 1.143868e+00 1.184476e+00 2.516535e+00
#>  [451] 2.348988e+00 3.276772e+00 2.154858e+00 2.609630e+00 1.304967e+00
#>  [456] 1.935772e+00 1.899779e+00 2.796150e+00 1.411863e+00 1.321387e+00
#>  [461] 1.635467e+00 1.854838e+00 3.468596e+00 2.833199e+00 4.174147e+00
#>  [466] 2.043023e+00 3.587955e+00 3.029297e+00 3.838248e+00 1.426981e+00
#>  [471] 1.181780e+00 1.936034e+00 1.820067e+00 2.019709e+00 2.751481e+00
#>  [476] 1.219321e+00 1.259474e+00 2.008983e+00 1.684721e+00 1.416996e+00
#>  [481] 2.726972e+00 3.036942e+00 1.883928e+00 1.146007e+00 9.222122e-01
#>  [486] 1.441938e+00 1.663459e+00 1.296801e+00 3.161556e+00 2.677480e+00
#>  [491] 1.355878e+00 1.436977e+00 1.010284e+00 3.146316e+00 4.297834e+00
#>  [496] 7.756199e-01 2.709752e+00 1.317339e+00 2.524167e+00 4.004986e+00
#>  [501] 1.768470e+00 3.381179e+00 2.477076e+00 4.811330e+00 3.357939e+00
#>  [506] 2.482466e+00 2.221486e+00 2.693101e+00 7.846593e+00 3.619120e+00
#>  [511] 3.732351e+00 1.646820e+00 2.001010e+00 4.450645e+00 3.146851e+00
#>  [516] 6.104030e+00 5.232305e+00 2.106316e+00 2.300966e+00 4.494347e+00
#>  [521] 3.776667e+00 4.587946e+00 1.812418e+00 2.693589e+00 2.265967e+00
#>  [526] 2.561782e+00 3.247259e+00 2.184488e+00 3.065945e+00 2.342076e+00
#>  [531] 2.944886e+00 3.491850e+00 2.385416e+00 2.686395e+00 2.803562e+00
#>  [536] 2.738686e+00 2.158247e+00 3.877688e+00 2.798689e+00 2.109950e+00
#>  [541] 4.023004e+00 2.058736e+00 1.853144e+00 2.261939e+00 1.476480e+00
#>  [546] 7.099580e-01 1.570591e+00 4.915123e+00 3.515058e+00 2.845965e+00
#>  [551] 2.414353e+00 4.393101e+00 3.011627e+00 1.854215e+00 1.659773e+00
#>  [556] 1.783780e+00 1.419808e+00 1.637372e+00 1.542279e+00 1.482103e+00
#>  [561] 4.558121e-01 2.390055e+00 1.223009e+00 1.199978e+00 5.309631e-01
#>  [566] 2.209093e-01 1.917324e+00 1.679644e+00 3.099462e-01 7.796072e-01
#>  [571] 1.547573e+00 1.769736e+00 1.308198e+00 2.911399e+00 1.361920e+00
#>  [576] 2.193799e+00 3.092974e+00 1.796763e+00 6.741277e-01 2.041915e+00
#>  [581] 1.373546e+00 1.577406e+00 5.095728e-01 1.965947e+00 2.015102e+00
#>  [586] 2.712318e+00 2.985168e+00 1.839151e+00 2.151419e+00 7.372275e-01
#>  [591] 2.304641e-01 1.331537e-01 5.458765e-01 4.280373e-01 7.763642e-01
#>  [596] 3.503463e-01 1.503721e+00 6.475561e-01 4.507984e-01 1.153305e+00
#>  [601] 1.072856e+00 2.262160e-01 8.710644e-01 1.073432e-01 1.019616e+00
#>  [606] 2.549097e-01 2.811744e-01 2.486147e-01 7.504861e-06 2.312925e-01
#>  [611] 4.243142e+00 1.602978e+00 2.857709e+00 2.883971e+00 2.055300e+00
#>  [616] 1.120749e+00 7.138493e-01 7.345240e-01 3.291253e-01 6.934424e-01
#>  [621] 1.816454e+00 5.000362e-01 1.687426e+00 1.740874e+00 1.555875e+00
#>  [626] 3.383123e+00 3.110519e+00 1.846868e+00 1.512880e+00 8.728197e-01
#>  [631] 1.418301e-01 3.222964e-01 5.415105e-01 3.305159e-01 3.094488e-01
#>  [636] 4.928703e-02 3.947152e-01 5.443310e-01 1.957439e+00 8.916332e-01
#>  [641] 5.535988e-01 3.773123e-01 1.222849e+00 7.711779e-01 7.259587e-01
#>  [646] 4.468411e-01 1.344176e-03 3.870649e-05 6.311470e-01 5.213332e-01
#>  [651] 4.558369e-01 1.365015e-02 2.780793e-01 5.139425e-01 4.789405e-01
#>  [656] 6.806283e-01 3.433119e-01 6.968061e-01 1.298085e-02 1.775081e-01
#>  [661] 2.924304e-01 2.584196e-01 3.067185e-01 4.778827e-05 8.659780e-02
#>  [666] 1.666278e-02 2.353687e-01 1.876983e-05 4.069473e-01 6.622988e-01
#>  [671] 2.761687e-01 4.032136e-02 3.504809e-02 3.597904e-03 1.025081e-01
#>  [676] 6.523490e-03 1.938609e-05 1.908867e-04 1.932096e-01 1.228572e-01
#>  [681] 3.884210e-01 1.179382e-01 1.093588e+00 1.560692e-01 1.792563e-01
#>  [686] 6.980841e-01 2.273307e-01 3.961182e-01 2.849760e-01 1.926746e-02
#>  [691] 3.080474e-01 1.633457e-01 7.455803e-01 6.904736e-02 1.411128e-01
#>  [696] 3.311659e-02 6.589914e-03 7.064194e-01 3.388130e-01 1.771544e-01
#>  [701] 1.051318e+00 4.336302e-01 1.460683e-01 1.558605e-01 2.849251e-01
#>  [706] 5.079868e-01 1.467624e-02 3.619368e-01 1.468254e-02 1.496083e-01
#>  [711] 1.643396e-01 1.208175e-03 1.047666e-01 1.509631e-02 9.242205e-03
#>  [716] 1.089574e-01 8.353918e-02 1.786259e-01 8.541580e-02 3.750364e-01
#>  [721] 1.785144e-01 4.999879e-03 1.497561e-01 1.246628e-02 1.515084e-03
#>  [726] 2.506416e-02 3.516844e-03 2.230214e-03 9.073457e-05 2.491341e-03
#>  [731] 1.551337e-01 3.661307e-02 1.421810e-01 1.379151e-03 2.735516e-02
#>  [736] 4.225547e-02 1.048584e-01 2.142977e-06 8.226299e-02 7.236358e-06
#>  [741] 3.580084e-01 1.986815e-02 1.175976e-01 3.883166e-01 9.777697e-04
#>  [746] 8.159121e-02 5.696933e-04 2.428772e-02 1.011539e-02 1.944039e-01
#>  [751] 3.300580e-01 7.561234e-01 5.280573e-01 5.822311e-01 3.527258e-01
#>  [756] 5.645395e-02 2.201128e-03 1.272197e-01 4.732807e-03 2.578444e-03
#>  [761] 2.108970e-03 3.168419e-02 3.145390e-02 2.956032e-02 4.833397e-03
#>  [766] 4.880583e-02 5.670597e-02 2.597123e-01 5.215165e-01 4.451856e-01
#>  [771] 4.339617e-01 1.376265e-01 1.484900e-02 2.635385e-01 4.522761e-01
#>  [776] 4.269491e-01 2.900734e-01 5.498999e-01 1.883161e-01 2.240426e-02
#>  [781] 5.394121e-03 1.798851e-02 9.518292e-06 7.810543e-03 1.103478e-01
#>  [786] 6.236698e-01 5.273538e-02 4.078209e-01 5.684297e-02 5.844300e-01
#>  [791] 7.819394e-01 2.163015e-01 6.617992e-01 8.533300e-02 1.096183e+00
#>  [796] 2.807220e-01 8.855062e-01 3.768592e-01 3.658009e-01 1.817346e-01
#>  [801] 1.940790e-01 8.758620e-01 1.909941e-01 9.465095e-04 5.578173e-02
#>  [806] 1.019932e-02 9.211452e-05 4.214575e-05 1.637792e-01 8.707868e-02
#>  [811] 4.094660e-03 6.599523e-02 1.668609e-01 1.407158e-01 1.954706e-01
#>  [816] 1.592754e+00 8.659925e-03 6.662880e-01 6.285718e-02 4.026673e-03
#>  [821] 3.571072e-01 1.120294e-03 1.354005e-02 1.842594e-01 8.986321e-03
#>  [826] 1.875018e-01 2.541020e-02 8.938896e-01 6.769344e-01 7.801297e-02
#>  [831] 6.566581e-03 1.321238e-01 3.560949e-01 5.796781e-04 2.274408e-01
#>  [836] 4.691038e-01 7.895746e-01 6.574029e-01 9.529832e-01 7.219465e-01
#>  [841] 8.072840e-01 1.962372e-03 2.233318e-01 8.072462e-01 6.548162e-01
#>  [846] 6.827951e-01 5.853865e-01 4.224586e-01 1.246652e+00 6.361496e-01
#>  [851] 5.683846e-01 7.497705e-01 1.708791e-01 8.349255e-06 7.067436e-01
#>  [856] 1.294611e+00 1.078341e+00 5.586346e-01 3.393877e-01 9.232257e-02
#>  [861] 5.206015e-01 1.588383e+00 1.391251e+00 2.882045e+00 1.501534e+00
#>  [866] 3.932157e-01 7.221290e-01 1.517350e+00 8.996047e-01 1.845964e-01
#>  [871] 1.075083e+00 5.919967e-01 1.240607e+00 6.675763e-01 1.063370e-01
#>  [876] 1.255875e-01 8.916576e-05 2.584383e+00 7.102406e-01 4.369355e-01
#>  [881] 3.057915e-01 2.184258e-01 1.550333e-01 1.164668e+00 1.383043e+00
#>  [886] 1.647664e+00 2.738737e+00 2.202738e+00 3.572652e+00 1.707481e+00
#>  [891] 1.130616e+00 9.261105e-01 4.352680e-01 1.409569e+00 2.333154e+00
#>  [896] 1.801073e+00 3.043350e-01 4.549433e-01 6.678954e-03 6.397736e-01
#>  [901] 7.477250e-01 1.670299e+00 9.961167e-01 1.222175e+00 1.352814e+00
#>  [906] 1.711033e+00 7.125287e-01 1.321784e+00 2.908444e+00 1.137092e-01
#>  [911] 2.549942e+00 1.819019e+00 1.009325e+00 1.005334e+00 5.091671e-01
#>  [916] 1.026020e+00 1.957322e+00 7.903264e-01 6.912001e-01 1.322107e+00
#>  [921] 1.436508e+00 2.316617e+00 1.676504e+00 1.544807e+00 2.533417e+00
#>  [926] 2.600309e+00 3.630111e-01 3.457166e-01 9.228997e-01 2.999964e+00
#>  [931] 2.848594e+00 3.528439e+00 2.708462e+00 2.867178e+00 2.484494e+00
#>  [936] 1.460893e+00 1.640969e+00 1.004780e+00 1.966748e+00 2.277215e+00
#>  [941] 1.873008e+00 2.373153e+00 1.935935e+00 7.177495e-01 2.284930e+00
#>  [946] 6.072723e-01 3.135125e-01 1.585828e+00 5.339234e-01 6.449023e-01
#>  [951] 9.571660e-01 3.238238e-01 2.889046e-02 4.297557e-01 6.389331e-01
#>  [956] 8.551382e-02 5.752237e-01 8.698419e-01 1.406500e+00 8.889461e-01
#>  [961] 4.971983e-01 4.378692e-01 2.911935e-01 2.347467e-01 1.718736e-01
#>  [966] 7.787279e-01 1.291073e+00 2.018758e+00 1.154352e+00 1.086108e+00
#>  [971] 4.908750e+00 3.228673e+00 2.088471e+00 2.667058e+00 1.241048e+00
#>  [976] 8.388335e-01 4.593701e+00 2.595643e+00 2.712426e+00 3.707303e+00
#>  [981] 2.141434e+00 2.853014e+00 3.274882e+00 2.619622e+00 3.173551e+00
#>  [986] 2.896382e+00 2.897876e+00 2.714371e+00 4.118221e+00 4.329172e+00
#>  [991] 2.905485e+00 2.738724e+00 2.189575e+00 2.712050e+00 3.335230e+00
#>  [996] 9.327493e-01 2.106635e+00 1.787349e+00 2.570095e+00 1.598732e+00
#> [1001] 2.231572e+00 2.876739e+00 5.209973e+00 1.732149e+00 3.640825e-01
#> [1006] 4.397635e-01 4.842466e-01 1.039494e+00
#> 
#> $SPE
#>    [1] 0.1332349123 0.0106165677 0.0854717702 0.1175768272 0.1232772217
#>    [6] 0.4657078495 0.1665710597 1.5269734158 0.2434548232 0.7188012513
#>   [11] 0.4866839728 0.0772585623 0.0074034976 0.1830419090 0.8378884530
#>   [16] 0.3961631610 0.2918505455 0.9380317916 0.1833167086 0.2010265868
#>   [21] 0.0670709063 0.4137597194 0.0529479337 0.1019546697 0.6562315555
#>   [26] 0.1405475787 0.2611700242 0.0356875597 0.1412441785 0.0718654761
#>   [31] 0.2726622978 0.1059776526 0.5840802286 0.0937502449 0.3515839324
#>   [36] 0.1462575002 0.0565564637 0.0087161649 0.1291659898 0.0699389563
#>   [41] 0.0032918897 0.0295581736 0.3155911947 0.0020519875 0.2172023179
#>   [46] 0.4895610358 0.0228819656 0.2136351769 0.1704069185 0.1115711080
#>   [51] 0.2213388191 0.3133389212 0.2448827098 0.1848503846 0.1252680482
#>   [56] 0.0377240379 0.4025494603 0.0529682797 0.0056415997 0.1475667083
#>   [61] 0.0478656891 0.4532669166 0.2853162858 0.1908642939 0.1480450674
#>   [66] 0.1670563960 0.7927629500 0.2038795001 0.0097026688 0.5767501520
#>   [71] 0.1168087303 0.3713803932 0.0012030733 0.6124599373 0.5990919937
#>   [76] 0.0227011412 0.0420065038 0.1095733955 0.0779623801 0.0612168929
#>   [81] 0.0118804414 0.2304865315 0.0868403233 0.3542310214 0.1279349939
#>   [86] 0.2024230742 0.0888794600 0.0152487182 0.0602129310 0.0137434313
#>   [91] 0.0266451636 0.0849521602 0.0142048468 1.4459840341 0.2695495328
#>   [96] 0.1610795761 0.2349115134 0.1136167327 0.4355614608 0.1526129394
#>  [101] 0.0567097044 0.0770794306 0.0626770615 0.0346055665 0.2004899386
#>  [106] 0.4530626408 0.0003175181 0.3730297054 0.0824535107 0.0391824965
#>  [111] 0.0317578303 0.0057503621 0.1460246694 0.6633990545 0.3572689267
#>  [116] 1.4646447533 0.0654563331 0.8836762937 0.1835418159 0.1027634027
#>  [121] 0.5298097167 0.0377154350 0.3186111505 0.1226981133 0.3870178220
#>  [126] 0.0031488709 0.1608835667 0.0068936679 0.2716499479 0.3560111286
#>  [131] 0.9530269559 0.7076248269 0.0631950267 0.6602257059 0.3162165791
#>  [136] 0.5854263999 0.0958676948 0.3176346674 0.0609592219 0.1316174359
#>  [141] 0.2545412067 0.1178199832 0.0661384979 0.1301117797 0.9497636845
#>  [146] 0.1668753785 0.2737065213 0.1154152278 0.0621766383 0.3188826390
#>  [151] 0.3695475832 0.4934136578 0.1656932446 0.2454854730 0.0378002140
#>  [156] 0.2201249945 0.0557081688 0.1635981321 0.7285592677 0.4665423818
#>  [161] 0.0532295819 1.0341697900 0.4368642689 0.0121656112 0.7936245591
#>  [166] 0.3014026807 0.0961236459 0.1704856809 0.2381384474 0.2353154333
#>  [171] 0.1952540118 0.1101493768 0.0042803633 0.3965905790 0.0172460907
#>  [176] 0.0982169011 0.1949328432 0.0801564751 0.0058307630 0.1766716634
#>  [181] 0.1687656930 0.0070422987 0.0450632897 0.7621597429 0.1129984648
#>  [186] 0.2368102566 0.1099340971 0.6337255085 0.0108423014 0.0107907579
#>  [191] 0.5647922512 0.0384781121 0.0237766273 0.6190064377 0.0598352369
#>  [196] 0.1796011953 0.2415731761 0.7410060491 0.1331857534 0.8753795678
#>  [201] 0.0411160367 0.2412486764 0.2048593455 0.5675060877 0.0525375443
#>  [206] 0.4341474084 0.0517084963 0.1806407321 0.0225015604 0.3603203516
#>  [211] 0.7803305680 0.0903096835 0.0628282566 0.2629729850 0.1437956394
#>  [216] 0.1401573057 0.0791813182 0.0583763929 0.2986046292 0.5032991907
#>  [221] 0.1419542897 0.0164650265 0.1471210611 0.1974987293 0.3680604949
#>  [226] 0.3272568459 0.0469040947 0.6685591273 0.1364750434 0.0923351356
#>  [231] 0.0110681696 0.2105763589 0.0724686224 0.2079204037 0.0152007458
#>  [236] 0.0545691302 0.6474608847 0.0652598599 0.3795518625 0.1473434143
#>  [241] 0.1635958772 0.4094228581 0.0891633043 0.2623181489 0.5431065436
#>  [246] 0.2633289515 0.0642335049 0.7999675342 0.3895085616 0.2684068524
#>  [251] 0.1012741955 0.5733574701 0.0340540104 0.0372621028 0.3020306845
#>  [256] 0.1402525563 0.0618861383 0.2781696342 0.3045561289 0.1137419346
#>  [261] 0.4577933180 0.0022000254 0.0981883728 0.1657371222 0.0345548874
#>  [266] 1.0060814970 0.1340388311 0.1784246346 0.4603184770 0.0180764299
#>  [271] 0.9307728135 0.3174781871 0.0162043554 0.0640038984 0.0404571144
#>  [276] 0.3554991259 0.3763316841 0.0796485627 0.8791448771 1.3455272017
#>  [281] 0.0272613892 0.0095893880 0.0714116670 0.0938118932 0.0393941431
#>  [286] 0.1159333657 0.1308063072 0.0205154214 0.0094232094 0.5256370408
#>  [291] 0.0447374264 0.2024909569 0.1778825057 0.0688606850 0.0029067235
#>  [296] 0.0865562807 0.1418824059 0.0665174180 0.0959669125 0.3569998938
#>  [301] 0.0716946543 0.4640492178 0.2122360456 0.0401709752 0.1386191926
#>  [306] 0.2239684637 0.2266459929 0.1496069996 0.4853636054 0.0319555409
#>  [311] 0.5280856281 0.0626148888 0.5401957230 0.2321764218 1.3000872091
#>  [316] 0.1634980883 0.1719926647 0.0105217188 0.6611286556 0.1714035210
#>  [321] 0.6104735039 0.2816974634 0.2722793555 0.1074312889 0.0059856265
#>  [326] 0.1899039647 0.0126318620 0.0044504033 0.2483977100 0.0876215538
#>  [331] 0.4460787077 0.0741068925 0.4609130329 0.1193577360 0.4783631058
#>  [336] 0.0313626776 1.3056857951 0.0016281966 0.0003405627 0.0954938273
#>  [341] 0.0392009397 0.0481805301 0.2083439416 0.0052212468 0.6373959705
#>  [346] 0.1557497659 0.0816264465 0.4178568329 0.5945256220 0.4207433137
#>  [351] 0.3155075469 0.3213404872 0.2288422330 0.4407694529 0.1104720308
#>  [356] 0.1376156799 0.1419191136 0.0957932717 0.2265212485 0.2949179632
#>  [361] 0.1652303711 0.0922695423 0.0479317470 0.4826820145 0.0725809942
#>  [366] 0.0561815895 0.3279734922 0.2747903608 0.1530260264 0.2953454543
#>  [371] 0.2831826441 0.0518292186 0.1003424366 0.0376167632 0.3076241510
#>  [376] 0.0297293536 0.1089340281 0.6485593422 0.1170453207 0.2178741562
#>  [381] 0.5552230280 0.0254023230 0.0640255566 0.2590665978 0.0407815673
#>  [386] 0.4393146204 0.1306932161 0.4713187349 0.3000176699 0.3881166733
#>  [391] 0.1585475664 0.0494810080 0.2508502249 0.7008145391 0.1735568509
#>  [396] 2.1918296116 0.0185857766 0.1623801744 0.1517985143 0.1126229969
#>  [401] 0.3762079619 0.1575061093 0.0462427771 0.0232459683 0.0047773356
#>  [406] 0.2017547919 0.0954131825 0.4728109445 0.5488076430 0.0676933655
#>  [411] 0.0290344503 0.0088793094 0.0122756474 0.1960703560 0.4172827417
#>  [416] 0.0710481093 0.1209502498 0.2793490541 0.2857059166 0.1429566111
#>  [421] 0.1338617350 0.0830068191 0.1151876742 0.0420955442 0.0663403965
#>  [426] 0.1166045049 0.3662136222 0.1610596143 0.0486256896 0.1754151541
#>  [431] 0.0674738334 0.0266765116 0.4421241313 0.1889156678 0.1436676817
#>  [436] 0.1959153302 0.1853526997 0.1776032108 0.1856967261 0.1106649085
#>  [441] 0.2855651531 0.7874520731 0.0394575183 0.1710669922 0.1600186616
#>  [446] 0.7514770190 0.3058147427 0.1459400246 0.1048142826 0.1517717616
#>  [451] 0.1935816621 0.0032624912 0.4800125414 0.0694438462 0.1579086325
#>  [456] 0.0713950540 0.1257310494 0.0051583273 0.0654837163 0.2414955824
#>  [461] 0.0404151601 0.4387759925 0.0660593140 0.2021828269 0.5038813351
#>  [466] 0.7753131015 0.0074623181 0.0860736459 0.1935848529 0.2974046044
#>  [471] 0.5566980502 0.7192684074 0.0999291650 0.1928085167 0.1463675804
#>  [476] 0.2394952989 0.7193698467 0.1981198090 0.3212692861 0.0629349538
#>  [481] 0.0570327647 0.2264892597 0.0383003938 0.5338349646 0.0186151476
#>  [486] 0.3266052292 0.1146260543 0.1118020914 0.1044396068 0.1116559153
#>  [491] 0.1776474297 0.0787499228 0.0009764707 0.0220863350 0.1810803682
#>  [496] 0.0112248618 0.3871754030 0.8042191778 0.3489114007 0.2212250070
#>  [501] 0.0264097842 0.0796541647 0.2103545647 0.2307565643 0.5731696874
#>  [506] 0.0562142487 0.1405969501 0.8221617443 0.8906549659 0.8566309459
#>  [511] 0.0224841884 0.0524909635 0.2156682001 0.1911436290 0.1864193081
#>  [516] 0.0059132085 0.0766423445 0.0265644604 0.1153844879 0.1414472644
#>  [521] 0.1691921006 0.6921633923 0.1430116165 0.1058059215 0.0875639548
#>  [526] 0.0453828346 0.2079453281 0.2020713597 0.1701615013 0.5726989912
#>  [531] 0.1682327613 0.0190704197 0.0891864280 0.2238250137 0.0403755297
#>  [536] 0.0841144332 0.2720138455 0.1673961029 0.7174460642 0.3042237764
#>  [541] 0.0147167175 0.9980820658 0.3758274895 0.0788938636 0.0104966870
#>  [546] 0.1736516748 0.2495210079 0.1836916937 0.3730297476 0.2602516979
#>  [551] 0.0371717011 0.0627058187 0.1641259796 0.2872242050 0.2947850020
#>  [556] 0.0043411083 0.1752514024 0.0560815997 0.1188644201 0.0678925149
#>  [561] 0.5698672225 0.3287827808 0.1689576012 0.2518106634 0.1597862588
#>  [566] 0.2830654584 0.0634499822 0.2493706386 1.3153823638 0.6074940388
#>  [571] 0.0041509350 0.6225500973 0.1428639078 0.2817538952 0.4482355890
#>  [576] 0.0779837432 0.1166268824 0.7350386327 1.1094864933 0.2173928947
#>  [581] 0.5644959027 0.3261031593 0.2841989264 0.0813902427 0.1507935482
#>  [586] 0.2939031026 0.2658256143 0.3857353805 0.2500279725 0.0387217824
#>  [591] 0.2645134186 0.4152529765 0.1713296747 0.0641201426 0.1990012519
#>  [596] 0.1024399556 0.2718124187 1.8658310950 0.0066873613 0.0463064732
#>  [601] 0.0821526298 0.4129376989 0.4301484151 0.2569348291 0.1588913638
#>  [606] 0.1849850662 0.2207233998 0.3247389955 0.0455028807 0.0912834239
#>  [611] 0.0669210746 2.9496405465 0.3424311949 0.2375433347 0.0316535254
#>  [616] 0.2931333280 0.0342612933 0.2490345811 0.1008500084 0.0173565871
#>  [621] 0.2676189023 0.1489771279 0.0696006990 0.5320213236 0.0591944837
#>  [626] 0.2496063085 0.0394127056 0.1292911848 0.5850955136 0.0042756728
#>  [631] 0.5457043693 0.0413446458 0.1829852394 0.0468681512 0.1696322288
#>  [636] 0.0789064307 0.0951484385 0.0931802628 0.0899466575 0.0610768966
#>  [641] 0.0468322869 0.0616711735 0.2023905484 0.0970951947 0.5106835922
#>  [646] 0.1566753087 0.0406423365 0.3133604126 0.2193690773 0.2667605292
#>  [651] 0.1383363577 0.0467951700 0.0710628341 0.4505040527 0.2663640248
#>  [656] 0.0727117728 0.0373790311 0.1124109043 0.6842591919 0.0101194794
#>  [661] 0.6308141324 0.0043956856 0.0547335412 0.0414860644 0.3973376643
#>  [666] 0.0072617059 0.2362367544 0.1914281492 0.1734050826 0.0273263650
#>  [671] 0.0095077495 0.1138442080 0.0279742705 0.0621691417 0.3484184051
#>  [676] 0.3666525032 0.1727238815 0.0945327310 0.2945968186 0.0354209441
#>  [681] 0.4227172296 0.1430691252 0.1464605100 0.1516185608 0.3896861123
#>  [686] 0.0608785054 0.2960785831 0.0875105510 0.0570744957 0.4010996775
#>  [691] 0.4496445173 0.2400755206 0.1061301904 0.0038327375 0.0006452050
#>  [696] 0.0857461432 0.3014383893 0.0350097576 0.0877182173 0.0819553536
#>  [701] 0.0150267624 0.2371352099 0.1777009435 0.1349559048 0.4086185963
#>  [706] 0.1626787742 0.3055110756 0.2019185945 0.2726369821 0.3848832918
#>  [711] 0.6186646929 0.3480284000 0.1932138661 0.0067683207 0.1696055857
#>  [716] 0.2009629439 0.1902429624 0.1171296190 0.1946420340 0.8053971069
#>  [721] 0.1328483528 0.1057919457 0.3020122599 0.6771774321 0.1038118157
#>  [726] 0.1851873022 0.0923608028 0.4022201403 0.1865624041 0.7658841396
#>  [731] 0.1702547444 0.1196671245 0.0592733498 0.5010003619 0.5870853025
#>  [736] 0.2328123862 0.0970692060 0.0885847289 0.0205767875 0.0319306722
#>  [741] 0.0516993904 0.2659575695 0.1885348413 0.0123653698 0.1148564630
#>  [746] 0.3478553926 0.3610550911 0.2521007110 0.0088109633 0.1928584458
#>  [751] 0.4081589939 0.3694881678 0.2863106852 0.0339987567 0.0097231068
#>  [756] 0.0775456566 0.0480395732 0.1472168197 0.4083212217 0.4993099187
#>  [761] 0.2147701415 0.2468645994 0.2056494459 0.1357820685 0.3260165370
#>  [766] 0.0183745864 0.0194922910 0.0456249590 0.4337545516 0.0835597914
#>  [771] 0.5672617306 0.2473759561 0.4007577664 0.2231063098 0.0987175637
#>  [776] 0.0140487874 0.2052376325 0.0525940460 0.3406422697 0.0479183707
#>  [781] 0.5523474345 0.0109005783 0.6779869417 0.0363942366 0.1417455513
#>  [786] 0.0434597105 0.1271411543 0.1229892653 0.0243139008 0.1856474716
#>  [791] 0.4144618942 0.3650884626 0.3223720576 0.1827657785 0.1635200833
#>  [796] 0.0725693933 0.0274812320 0.2467720701 0.5148229486 0.2935502517
#>  [801] 0.0864759298 0.0705856829 0.3383358705 0.1165247571 0.0584829181
#>  [806] 0.0314411010 0.0259657099 0.2483335807 0.3016340766 0.0545458196
#>  [811] 0.4641496996 0.0705467757 0.2698994309 0.2568547911 0.3643269240
#>  [816] 0.4699641006 0.2822061538 0.0297411829 0.1335531626 0.1792632242
#>  [821] 0.0274266812 0.2626723518 0.9110440834 0.5203784510 0.1787471335
#>  [826] 0.0931067683 0.6572261852 0.1373040235 0.0177548765 0.0282815257
#>  [831] 1.0649984654 0.2478276891 0.0637850466 0.2735296276 0.2914333548
#>  [836] 0.2622070860 0.0309198408 0.0075235976 0.0691770491 0.3118122431
#>  [841] 0.3255563406 0.1283346459 0.6259337008 1.5737102835 0.0997596939
#>  [846] 0.4423116629 0.5868113109 0.0112057978 0.0075160892 0.0630355184
#>  [851] 0.0026223880 0.2494437363 0.1652924461 0.0620362280 0.2602720526
#>  [856] 0.3072985190 0.7612308417 0.0165674507 0.2109827018 0.0642305655
#>  [861] 0.8172013908 0.1854934752 0.0890939807 0.6222559544 0.3349102112
#>  [866] 0.2544214013 0.4704436321 0.0566360862 0.6143838502 0.4889851988
#>  [871] 0.0151301868 0.3974322853 0.0705850221 0.0352201270 0.0982745636
#>  [876] 0.2448073836 0.5165435282 0.2182392663 0.0112605925 0.0123631434
#>  [881] 0.0425479312 0.9385670590 0.0559221902 0.0181404322 0.8857849530
#>  [886] 0.8890026976 0.1869455137 0.7000920915 0.3790618354 0.2032087898
#>  [891] 0.6487262324 0.0583389179 0.0012823038 0.1088112443 0.5311240947
#>  [896] 0.0237810314 0.0898155949 0.0415574962 0.4623319043 0.3823561660
#>  [901] 0.0742825210 0.0165158939 0.2334167977 0.0712861172 0.2222262240
#>  [906] 0.6420114703 0.5084149693 0.1835541864 0.5853202624 0.1017416629
#>  [911] 0.1918298291 0.0442418009 0.0384006271 0.1380903411 0.4732633015
#>  [916] 0.1738905486 0.2579343152 0.3352760131 0.3205070309 0.0809224501
#>  [921] 0.1073786135 0.0599191013 0.0102690709 0.0766485471 0.2351243072
#>  [926] 0.1032316764 1.5189227296 0.3194923541 0.2343341720 0.6428414472
#>  [931] 0.3733017839 0.0485466062 0.0503609766 0.0515256232 0.5099567021
#>  [936] 0.0105378333 0.0957674979 0.0443520222 0.1460184601 0.0264537632
#>  [941] 0.1386998841 0.1191837015 0.4163718881 0.4533670437 0.0984957457
#>  [946] 0.3119735222 0.2210686270 0.3851594337 0.2689672778 0.5019744808
#>  [951] 0.2191669182 0.0187226855 0.0483456705 0.0874348613 0.2583637662
#>  [956] 0.4130980077 0.0125017661 0.0223617702 0.2837178901 0.3768503906
#>  [961] 0.2071111292 0.2199217301 0.3969185487 0.0349070551 0.4632489678
#>  [966] 0.1338947634 0.2225726168 0.0337963491 0.7047273368 1.0509787766
#>  [971] 0.3921360278 0.5534709666 0.3873344737 0.3551650213 0.0918290881
#>  [976] 0.4130002130 0.1046071164 0.3456394730 0.6553654956 0.2107909834
#>  [981] 0.0942409748 1.2968967971 0.1567430267 0.1546102213 1.0825612142
#>  [986] 0.1770059171 0.0406180487 0.3425333638 0.1640250075 0.1938323054
#>  [991] 0.2304854468 0.1115546206 0.1253318742 0.2325686207 0.0232297241
#>  [996] 0.5855367918 0.2773144563 0.0754764300 0.2007226531 0.2636155732
#> [1001] 0.4832045397 0.0726477746 0.5378109783 0.9448706759 0.1339557267
#> [1006] 0.1106288170 0.2106522059 0.8550758401
#> 
#> $muTrain
#>          x          y          z 
#>  0.6467465 -1.4887594  1.0164689 
#> 
#> $RootPrecisTrain
#>          [,1]     [,2]     [,3]
#> [1,] 4.772481 0.000000 0.000000
#> [2,] 0.000000 3.070339 0.000000
#> [3,] 0.000000 0.000000 2.023623
#> 
#> attr(,"class")
#> [1] "threshold" "pca"      
#>