Adaptive Process Training

Apply Adaptive-Dynamic PCA to state-specific data matrices.

processMonitor(
  data,
  trainObs,
  updateFreq = ceiling(0.5 * trainObs),
  faultsToTriggerAlarm = 5,
  ...
)

Arguments

data

An xts data matrix

trainObs

The number of training observations to be used

updateFreq

The number of non-flagged observations to collect before the function updates. Defaults to half as many observations as the number of training observations.

faultsToTriggerAlarm

The number of sequential faults needed to trigger an alarm. Defaults to 5.

...

Lazy dots for additional internal arguments

Value

A list with the following components:

FaultChecks --

a class-specific xts flagging matrix with the same number of rows as "data". This flag matrix has the following five columns:

SPE --

the SPE statistic value for each observation in "data"

SPE_Flag --

a vector of SPE indicators recording 0 if the test statistic is less than or equal to the critical value passed through from the threshold object

T2 --

the T2 statistic value for each observation in "data"

T2_Flag --

a vector of T2 fault indicators, defined like SPE_Flag

Alarm --

a column indicating if there have been five flags in a row for either the SPE or T2 monitoring statistics or both. Alarm states are as follows: 0 = no alarm, 1 = Hotelling's T2 alarm, 2 = Squared Prediction Error alarm, and 3 = both alarms.

Non_Alarmed_Obs --

a class-specific xts data matrix of all the non-alarmed observations (observations with alarm state equal to 0)

Alarms --

a class-specific xts data matrix of the features and specific alarms of Alarmed observations, where the alarm codes are listed above

trainSpecs --

a threshold object returned by the internal threshold() function. See the threshold() function's help file for more details.

Details

This function is the class-specific implementation of the Adaptive- Dynamic PCA described in the details of the mspTrain() function. See the mspTrain() function's help file for further details.

This internal function is called by mspTrain(). This function calls the faultFilter() function.

See also

Examples

nrml <- mspProcessData(faults = "NOC")
data <- nrml[nrml[,1] == 1]

processMonitor(data = data[,-1], trainObs = 672)
#> Warning: In density.default(spe, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(t2, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(spe, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(t2, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(spe, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(t2, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(spe, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> Warning: In density.default(t2, bw = "SJ", kernel = "gaussian", from = 0, 
#>     old.coords = TRUE) :
#>  extra argument ‘old.coords’ will be disregarded
#> $FaultChecks
#>                              SPE SPE_Flag       T2 T2_Flag Alarm
#> 2015-05-17 19:12:00 0.2146559077        0 4.200818       0     0
#> 2015-05-17 19:13:00 0.6380915270        0 2.850634       0     0
#> 2015-05-17 19:14:00 0.8898350891        0 3.664350       0     0
#> 2015-05-17 19:15:00 0.4268370464        0 5.352311       0     0
#> 2015-05-17 19:16:00 0.1118096173        0 4.037671       0     0
#> 2015-05-17 19:17:00 0.0139106292        0 3.501590       0     0
#> 2015-05-17 19:18:00 0.2919562343        0 8.121417       0     0
#> 2015-05-17 19:19:00 0.0002492626        0 3.332557       0     0
#> 2015-05-17 19:20:00 0.0228709143        0 5.057245       0     0
#> 2015-05-17 19:21:00 0.0207263285        0 7.290043       0     0
#>                 ...                                             
#> 2015-05-23 07:50:00 0.0494955780        0 3.147854       0     0
#> 2015-05-23 07:51:00 0.0199056071        0 2.027382       0     0
#> 2015-05-23 07:52:00 0.3459172608        0 3.923452       0     0
#> 2015-05-23 07:53:00 0.5728132012        0 4.108266       0     0
#> 2015-05-23 07:54:00 0.8190963493        0 4.895518       0     0
#> 2015-05-23 07:55:00 0.5210449599        0 2.331959       0     0
#> 2015-05-23 07:56:00 0.1795723427        0 1.867592       0     0
#> 2015-05-23 07:57:00 0.7261271005        0 2.979553       0     0
#> 2015-05-23 07:58:00 0.1838012431        0 3.041245       0     0
#> 2015-05-23 07:59:00 0.3293486957        0 1.201111       0     0
#> 
#> $Non_Alarmed_Obs
#>                              x          y            z
#> 2015-05-17 19:12:00 0.79508067 -1.6801452  1.368118539
#> 2015-05-17 19:13:00 0.68120031 -1.4501698  1.318548226
#> 2015-05-17 19:14:00 0.62032976 -1.4195376  1.367870857
#> 2015-05-17 19:15:00 0.75799048 -1.7032673  1.494940657
#> 2015-05-17 19:16:00 0.72550094 -1.6882391  1.312557489
#> 2015-05-17 19:17:00 0.68720111 -1.6878400  1.199866491
#> 2015-05-17 19:18:00 0.62974295 -1.7385183  1.479408001
#> 2015-05-17 19:19:00 0.74259793 -1.7203661  1.175296325
#> 2015-05-17 19:20:00 0.88315772 -1.7700153  1.128251838
#> 2015-05-17 19:21:00 0.63447990 -1.8212262  1.353382338
#>                 ...                                   
#> 2015-05-23 07:50:00 0.14282172 -0.6348665  0.026185952
#> 2015-05-23 07:51:00 0.23570827 -0.8015552  0.185338228
#> 2015-05-23 07:52:00 0.09675956 -0.4202202  0.072878269
#> 2015-05-23 07:53:00 0.18578031 -0.2888310 -0.027990068
#> 2015-05-23 07:54:00 0.09520905 -0.1804724  0.009430122
#> 2015-05-23 07:55:00 0.29364195 -0.5109547  0.276808465
#> 2015-05-23 07:56:00 0.19564060 -0.9534731  0.165475733
#> 2015-05-23 07:57:00 0.02534980 -0.6968567  0.290874541
#> 2015-05-23 07:58:00 0.09608343 -0.7017346  0.092497846
#> 2015-05-23 07:59:00 0.46016184 -0.8718582  0.139005248
#> 
#> $Alarms
#>                             x         y        z       SPE SPE_Flag       T2
#> 2015-05-18 01:33:00 1.0155996 -2.122771 2.057074 0.8778236        0 13.21164
#> 2015-05-18 01:34:00 0.9536295 -2.083550 1.963331 0.6857127        0 12.13851
#> 2015-05-18 04:02:00 1.0873694 -2.089024 2.196024 1.6199480        0 14.23551
#> 2015-05-18 04:13:00 1.1247107 -2.039869 2.197057 1.8878983        0 14.00576
#> 2015-05-18 04:14:00 1.0548835 -2.006629 2.049914 1.3270768        0 12.00706
#> 2015-05-18 04:15:00 1.0663219 -1.974234 2.120785 1.8209568        0 12.40443
#> 2015-05-18 04:16:00 1.2944240 -1.824835 2.416445 4.9408863        1 17.79086
#> 2015-05-18 04:17:00 1.1241443 -2.105161 2.418755 2.7761932        1 16.90591
#> 2015-05-18 04:18:00 1.0799186 -2.208456 2.313845 1.5975689        0 17.06334
#> 2015-05-18 04:19:00 1.0258793 -2.297492 2.345297 1.3112212        0 19.71970
#>                 ...                                                         
#> 2015-05-21 16:23:00 0.8222212 -1.950874 2.056454 1.5451572        0 14.40829
#> 2015-05-21 16:24:00 1.0115607 -1.892902 2.152206 2.4292174        1 11.87251
#> 2015-05-21 16:25:00 1.1608309 -1.749670 2.043894 2.7181209        1 12.72421
#> 2015-05-21 16:26:00 0.9550974 -2.154985 2.040951 0.7005870        0 14.05528
#> 2015-05-21 16:35:00 1.0041145 -2.035746 2.080907 1.3180894        0 12.59801
#> 2015-05-21 16:36:00 1.0913996 -2.163229 1.965569 0.4841286        0 12.81883
#> 2015-05-21 16:37:00 0.8676514 -2.114098 2.244093 1.6590134        0 19.25098
#> 2015-05-21 16:38:00 1.1241026 -2.016478 2.124006 1.6416028        0 13.27821
#> 2015-05-21 16:55:00 1.1061490 -1.967717 2.153565 2.0411854        0 12.93308
#> 2015-05-21 16:56:00 1.0062318 -2.073561 2.120063 1.3209556        0 13.47380
#>                     T2_Flag Alarm
#> 2015-05-18 01:33:00       1     1
#> 2015-05-18 01:34:00       1     1
#> 2015-05-18 04:02:00       1     1
#> 2015-05-18 04:13:00       1     1
#> 2015-05-18 04:14:00       1     1
#> 2015-05-18 04:15:00       1     1
#> 2015-05-18 04:16:00       1     1
#> 2015-05-18 04:17:00       1     1
#> 2015-05-18 04:18:00       1     1
#> 2015-05-18 04:19:00       1     1
#>                 ...              
#> 2015-05-21 16:23:00       1     1
#> 2015-05-21 16:24:00       1     1
#> 2015-05-21 16:25:00       1     1
#> 2015-05-21 16:26:00       1     1
#> 2015-05-21 16:35:00       1     1
#> 2015-05-21 16:36:00       1     1
#> 2015-05-21 16:37:00       1     1
#> 2015-05-21 16:38:00       1     1
#> 2015-05-21 16:55:00       1     1
#> 2015-05-21 16:56:00       1     1
#> 
#> $TrainingSpecs
#> $SPE_threshold
#>    99.9% 
#> 1.237889 
#> 
#> $T2_threshold
#>    99.9% 
#> 5.942678 
#> 
#> $projectionMatrix
#>            [,1]
#> [1,]  0.5715648
#> [2,] -0.5786680
#> [3,]  0.5817706
#> 
#> $LambdaInv
#>           [,1]
#> [1,] 0.3555542
#> 
#> $T2
#>    [1] 3.927158e-01 8.427202e-01 1.209525e+00 1.537829e+00 1.147469e+00
#>    [6] 8.499752e-01 3.905890e-01 1.854229e+00 1.636122e+00 3.404065e+00
#>   [11] 8.363408e-01 6.588028e-01 3.252623e-01 2.308283e-01 1.598686e-01
#>   [16] 2.763177e-03 1.309476e-01 1.433080e-01 7.409778e-01 1.403492e+00
#>   [21] 9.160467e-01 1.302712e+00 5.424521e-01 3.011382e-01 3.469390e-01
#>   [26] 7.791746e-01 4.083867e-01 1.278819e+00 1.933828e+00 7.739038e-01
#>   [31] 6.729003e-01 5.437048e-01 6.388350e-01 2.997681e-01 3.067769e-01
#>   [36] 7.878922e-01 2.452660e-02 1.116801e+00 7.828062e-01 4.883271e-01
#>   [41] 1.767260e+00 2.312457e+00 2.163391e+00 1.045806e+00 1.125930e+00
#>   [46] 3.208543e-01 2.386781e-01 8.822420e-01 4.555602e-01 3.178711e-01
#>   [51] 2.302705e-01 8.641561e-01 1.933796e-01 1.088406e-01 9.412242e-01
#>   [56] 5.886794e-01 2.314207e-01 4.442516e-01 5.166719e-01 2.285572e-01
#>   [61] 9.335159e-01 1.016846e+00 1.308666e+00 9.427434e-01 9.435883e-01
#>   [66] 1.582723e+00 1.431486e+00 1.742910e+00 1.059482e+00 3.000397e-01
#>   [71] 1.654253e+00 1.763168e+00 4.547511e-01 3.476169e-01 1.033069e-01
#>   [76] 5.251862e-01 5.090505e-01 4.695240e-01 9.451721e-01 7.145540e-01
#>   [81] 1.370238e+00 8.817237e-01 2.113595e-01 3.701834e-02 7.119435e-01
#>   [86] 8.738652e-02 1.818886e-01 2.199040e-01 2.177214e-01 1.209276e-01
#>   [91] 1.874632e-01 1.175911e-01 1.436282e-01 4.051185e-01 8.269923e-01
#>   [96] 7.713298e-01 1.884885e-01 3.998720e-02 4.411305e-01 4.090771e-01
#>  [101] 1.297868e-02 2.702974e-01 3.242688e-01 4.028168e-01 3.238251e-01
#>  [106] 2.794026e-01 8.380268e-02 5.969090e-04 3.200007e-02 1.622019e-01
#>  [111] 7.855345e-02 1.022037e-01 6.118542e-01 1.703529e-01 2.721229e-03
#>  [116] 1.083438e-01 4.403263e-01 1.198590e-02 2.580112e-01 1.055458e-01
#>  [121] 1.119280e-01 3.081429e-01 2.079106e-01 9.112236e-01 2.212544e+00
#>  [126] 1.186926e+00 1.003654e+00 6.880047e-01 8.613585e-01 4.107331e-01
#>  [131] 1.499194e-01 1.352583e-01 1.549147e-01 2.204182e-01 3.447585e-01
#>  [136] 5.937321e-01 1.161985e+00 1.612375e+00 2.889773e-01 6.952605e-01
#>  [141] 3.116960e-01 6.734872e-01 3.947430e-01 3.734732e-02 8.994471e-02
#>  [146] 8.043589e-03 2.259771e-01 8.192880e-02 1.764440e-01 2.043873e-03
#>  [151] 3.463993e-06 9.928169e-04 1.607810e-01 3.059305e-02 2.172060e-02
#>  [156] 1.834581e-01 1.260583e-01 2.072182e-01 6.107710e-01 5.835974e-01
#>  [161] 3.515155e-01 7.596558e-01 5.305774e-01 8.218580e-01 3.041469e-01
#>  [166] 7.836135e-01 8.073004e-01 4.176075e-01 3.225223e-01 5.256565e-01
#>  [171] 6.438086e-02 8.269540e-03 2.368333e-02 6.052228e-02 3.094833e-01
#>  [176] 7.092890e-01 5.816888e-05 4.195904e-02 5.882289e-02 2.316658e-03
#>  [181] 4.311330e-01 1.688944e-01 4.067827e-01 7.340767e-01 6.209586e-01
#>  [186] 9.145094e-01 1.293454e-02 2.586493e-01 4.489451e-01 8.983159e-02
#>  [191] 2.624585e-01 5.072920e-01 8.272152e-02 2.851142e-01 5.285353e-02
#>  [196] 1.840420e-01 3.084999e-01 4.870648e-01 2.406619e-03 1.842419e-01
#>  [201] 9.655726e-02 2.758682e-02 3.442751e-01 1.176042e-01 4.686274e-01
#>  [206] 2.609808e-01 1.321366e-01 1.743432e-01 1.108678e-02 2.021816e-03
#>  [211] 3.419421e-02 1.552707e-01 1.408847e-01 8.821721e-01 8.370771e-02
#>  [216] 1.287065e-01 1.757243e-01 2.793002e-02 7.129050e-03 1.030132e-02
#>  [221] 6.555777e-03 1.352809e-01 8.175381e-03 6.158729e-01 4.493187e-01
#>  [226] 3.155533e-02 6.414693e-02 4.342268e-03 7.418588e-05 3.875078e-02
#>  [231] 6.356209e-02 8.133655e-02 6.147851e-03 5.756827e-04 8.253816e-02
#>  [236] 4.598931e-02 1.335639e-02 2.350782e-01 7.207764e-01 3.716868e-01
#>  [241] 2.444069e-01 5.145766e-02 1.111307e-01 1.577825e-01 1.387600e-02
#>  [246] 5.843660e-02 5.829086e-02 6.399284e-02 1.558703e-03 3.875076e-04
#>  [251] 8.826762e-09 1.123445e+00 3.055677e-01 2.536458e-01 3.599162e-01
#>  [256] 1.130111e-01 1.894867e-01 3.515465e-02 5.794975e-02 7.565433e-02
#>  [261] 2.690979e-01 3.406837e-02 6.997269e-02 1.159141e-01 3.866492e-01
#>  [266] 1.302211e-01 1.313241e-06 9.687193e-02 9.659331e-02 4.859879e-02
#>  [271] 3.646706e-03 1.556171e-01 1.133396e+00 7.078944e-01 2.115250e+00
#>  [276] 7.984501e-01 1.036322e+00 6.554488e-01 5.640964e-01 4.524801e-02
#>  [281] 1.079375e+00 2.877863e-01 4.944963e-01 3.308456e-01 1.334110e-01
#>  [286] 1.606767e-01 1.792279e-05 2.278854e-02 9.718802e-02 4.031548e-02
#>  [291] 7.648121e-02 2.890302e-04 1.381868e-01 1.945918e-01 3.908653e-01
#>  [296] 6.933810e-01 6.437096e-01 8.063677e-01 4.630906e-01 1.489553e-01
#>  [301] 4.648701e-01 1.402415e+00 1.075684e+00 1.964749e+00 6.757243e-01
#>  [306] 7.694461e-01 3.524031e-01 5.939670e-01 6.357191e-02 1.107153e-01
#>  [311] 4.432816e-01 1.073220e-01 2.527905e-02 2.509208e-02 7.908817e-02
#>  [316] 3.662078e-03 2.798436e-01 1.430436e-03 1.355819e-01 2.088959e-01
#>  [321] 1.334195e-01 3.894165e-02 1.424825e-02 1.280703e-03 1.622624e-01
#>  [326] 1.557658e-01 1.003922e+00 1.768690e-02 3.900500e-01 5.136029e-01
#>  [331] 1.851317e+00 1.267856e+00 1.200490e+00 8.149870e-01 5.327717e-01
#>  [336] 1.016795e-01 5.130262e-01 1.232358e-01 7.522451e-02 5.768825e-01
#>  [341] 3.574080e-01 2.312090e-01 3.700149e-01 3.207043e-01 5.862288e-01
#>  [346] 3.691088e-01 7.382094e-01 1.942029e+00 5.044623e-01 1.386993e-01
#>  [351] 5.434490e-01 7.011469e-01 7.402721e-01 6.764806e-01 4.507795e-01
#>  [356] 7.941969e-01 3.856617e-01 3.014455e-01 5.038782e-01 6.232606e-01
#>  [361] 4.079059e-01 6.574925e-01 6.349509e-01 1.214960e+00 1.475616e+00
#>  [366] 7.486496e-01 5.455982e-01 4.979653e-01 9.871337e-01 2.828848e-01
#>  [371] 4.593258e-01 1.047571e+00 1.970277e+00 2.151469e+00 6.446140e-01
#>  [376] 5.606725e-01 1.605111e+00 2.241673e+00 9.287481e-01 1.057407e+00
#>  [381] 6.693402e-01 8.951218e-01 8.511506e-01 5.271797e-01 1.477123e+00
#>  [386] 7.892575e-01 1.678889e+00 1.008955e+00 3.310457e-01 8.379606e-01
#>  [391] 1.016180e+00 1.325927e+00 1.365192e+00 8.691582e-01 2.184178e+00
#>  [396] 4.573525e-01 7.687947e-01 1.081231e+00 1.861364e-01 6.885874e-01
#>  [401] 2.639641e-01 2.100196e-01 9.523340e-01 6.974344e-01 2.413741e+00
#>  [406] 2.992082e+00 2.571670e+00 1.762276e+00 1.518933e+00 9.897247e-01
#>  [411] 6.924212e-01 1.215146e+00 6.840744e-01 1.337217e+00 7.348871e-01
#>  [416] 2.287132e+00 1.967834e+00 9.426859e-01 3.081603e-02 1.269259e+00
#>  [421] 1.119211e+00 1.692879e+00 9.568199e-01 4.935647e-01 1.079195e+00
#>  [426] 1.955513e+00 1.510620e+00 1.904033e+00 1.183477e+00 1.332951e+00
#>  [431] 1.510343e+00 1.475945e+00 2.545875e+00 1.680495e+00 4.538989e-01
#>  [436] 6.184551e-01 1.391537e+00 1.203396e+00 4.988997e-01 1.810849e+00
#>  [441] 1.348664e+00 1.597544e+00 1.218457e+00 1.033210e+00 1.167215e+00
#>  [446] 2.219977e+00 1.348008e+00 1.331102e+00 1.304796e+00 2.499948e+00
#>  [451] 7.986506e-01 2.079104e+00 1.505765e+00 9.590768e-01 3.718102e-01
#>  [456] 5.695310e-01 1.229932e+00 3.561545e-01 9.216023e-01 5.776873e-01
#>  [461] 8.324450e-01 1.691526e+00 1.866443e+00 1.404930e+00 2.092837e+00
#>  [466] 1.913668e+00 7.412406e-01 1.868622e+00 1.525556e+00 2.365740e+00
#>  [471] 1.415779e+00 5.198998e-01 8.403019e-01 1.882714e+00 3.184725e+00
#>  [476] 2.599321e+00 3.211596e+00 1.968754e+00 1.208512e+00 2.912056e+00
#>  [481] 2.754905e+00 2.080546e+00 1.978078e+00 1.353159e+00 1.055333e+00
#>  [486] 1.374256e+00 2.316671e+00 3.263275e+00 2.039848e+00 1.851743e+00
#>  [491] 4.842231e-01 1.190356e+00 1.489723e+00 1.357260e+00 2.800345e+00
#>  [496] 2.242676e+00 2.131976e+00 2.796161e+00 2.280116e+00 1.785283e+00
#>  [501] 1.403728e+00 3.831969e+00 2.861291e+00 2.946566e+00 2.035088e+00
#>  [506] 1.751931e+00 2.020120e+00 1.637906e+00 1.688765e+00 1.817831e+00
#>  [511] 3.255075e+00 3.239190e+00 3.606052e+00 3.087617e+00 1.295565e+00
#>  [516] 3.278577e+00 4.356690e+00 2.804829e+00 2.784482e+00 2.008551e+00
#>  [521] 2.006847e+00 2.203359e+00 2.462415e+00 1.913025e+00 1.663687e+00
#>  [526] 1.443763e+00 3.643833e+00 3.641561e+00 4.391506e+00 4.445671e+00
#>  [531] 2.964458e+00 3.509805e+00 3.919311e+00 5.911707e+00 1.987419e+00
#>  [536] 2.587696e+00 2.352271e+00 4.285917e+00 2.071309e+00 5.045121e+00
#>  [541] 6.116703e+00 5.750248e+00 2.866106e+00 2.129200e+00 1.487505e+00
#>  [546] 7.886272e-01 1.877363e+00 3.604502e+00 2.288439e+00 2.683884e+00
#>  [551] 2.264527e+00 1.367714e+00 1.078572e+00 2.312934e+00 2.272563e+00
#>  [556] 2.652394e+00 2.766509e+00 2.896663e+00 3.434815e+00 3.409725e+00
#>  [561] 2.268068e+00 2.376333e+00 2.432046e+00 2.884024e+00 3.040080e+00
#>  [566] 3.128059e+00 1.549051e+00 2.093639e+00 2.166701e+00 2.238026e+00
#>  [571] 2.167399e+00 2.878352e+00 3.955367e+00 2.570456e+00 1.865747e+00
#>  [576] 2.338411e+00 1.836826e+00 2.487662e+00 1.536796e+00 1.473795e+00
#>  [581] 2.036823e+00 1.528227e+00 2.100376e+00 2.522295e+00 2.587882e+00
#>  [586] 2.056673e+00 2.408315e+00 3.022173e+00 3.830589e+00 3.352225e+00
#>  [591] 2.244027e+00 4.023728e+00 2.031521e+00 9.741797e-01 1.166085e+00
#>  [596] 7.655295e-01 9.752582e-01 1.023012e+00 1.335920e+00 2.107548e+00
#>  [601] 1.710730e+00 1.357196e+00 1.888186e+00 3.139522e+00 8.603990e-01
#>  [606] 1.364687e+00 1.563676e+00 1.486800e+00 2.071448e+00 1.610854e+00
#>  [611] 2.441521e+00 1.108150e+00 9.003750e-01 9.181425e-01 6.212946e-01
#>  [616] 1.904717e-01 8.235482e-02 5.551688e-02 1.962015e-01 2.338412e-01
#>  [621] 7.808971e-01 3.973837e-01 7.776421e-01 1.982895e+00 2.371603e+00
#>  [626] 1.368681e+00 1.020345e+00 1.178783e+00 1.011396e+00 1.543693e+00
#>  [631] 1.641193e+00 2.432008e+00 1.166463e+00 2.564287e+00 1.098666e+00
#>  [636] 1.939568e+00 2.203907e+00 2.528085e+00 2.717251e+00 1.330173e+00
#>  [641] 2.639315e+00 1.609869e+00 3.064395e+00 2.514102e+00 1.957813e+00
#>  [646] 1.667215e+00 2.444185e+00 2.387729e+00 1.343000e+00 1.561060e+00
#>  [651] 5.691882e-01 9.094443e-01 4.722218e-01 1.643077e+00 2.466192e+00
#>  [656] 2.788445e+00 2.307705e+00 9.394107e-01 9.129855e-01 1.024894e+00
#>  [661] 1.686229e+00 1.711642e+00 1.511257e+00 1.442679e+00 2.148300e+00
#>  [666] 1.926631e+00 1.709200e+00 8.385580e-01 1.036247e+00 1.026407e+00
#>  [671] 3.724850e-01 8.640388e-01 1.303926e+00 1.871259e+00 1.868109e+00
#>  [676] 1.817225e+00 1.963761e+00 1.230442e+00 5.024156e-01 5.477020e-01
#>  [681] 3.958962e-01 5.972655e-01 5.544126e-01 8.259148e-01 9.296945e-01
#>  [686] 5.496036e-01 5.290272e-01 1.037418e+00 1.240569e+00 6.165357e-01
#>  [691] 4.544225e-01 1.318806e+00 1.757794e+00 2.713447e+00 2.255933e+00
#>  [696] 1.089572e+00 1.231233e+00 6.296427e-01 1.212059e+00 4.125572e-01
#>  [701] 6.578002e-01 3.526030e-01 4.440392e-01 9.939434e-01 6.186135e-01
#>  [706] 5.474164e-01 4.878278e-01 1.102624e+00 3.098687e-01 9.285129e-01
#>  [711] 4.695236e-01 1.540640e-01 1.628915e+00 2.204648e+00 1.234810e+00
#>  [716] 8.618330e-01 7.317554e-01 8.546644e-01 8.309248e-01 1.208240e+00
#>  [721] 1.538771e-01 2.855242e-01 2.635922e-01 8.599773e-01 5.166844e-01
#>  [726] 5.571377e-01 5.019940e-01 2.602114e-01 1.292126e-01 2.941541e-01
#>  [731] 1.461355e-01 1.265510e-03 1.603553e-01 3.839365e-02 1.833094e-02
#>  [736] 4.971977e-02 9.117810e-02 5.325120e-01 4.115184e-01 1.991964e-01
#>  [741] 3.132880e-01 1.606818e-01 1.400310e-01 2.277021e-02 1.242865e-01
#>  [746] 9.441577e-04 2.397077e-01 3.388158e-01 1.361506e+00 9.752120e-01
#>  [751] 4.736262e-01 1.039552e+00 1.277271e+00 1.077045e+00 1.204272e+00
#>  [756] 6.314434e-01 6.269111e-01 1.560422e-01 9.293973e-02 3.588627e-01
#>  [761] 5.875647e-01 9.947456e-01 1.225922e+00 5.640600e-01 1.125552e+00
#>  [766] 1.819505e+00 4.305765e-01 2.480828e-01 1.465725e+00 4.694057e-01
#>  [771] 6.094199e-01 3.348223e-01 4.033156e-01 8.430255e-07 2.512170e-01
#>  [776] 3.288273e-01 1.045811e-03 4.536677e-02 2.838899e-01 1.062488e-02
#>  [781] 3.750884e-02 2.055442e-01 1.963809e-01 2.522486e-05 1.066890e-02
#>  [786] 8.689139e-02 2.462489e-01 4.243831e-02 4.673598e-02 7.292624e-02
#>  [791] 2.112008e-03 1.431456e-02 1.193524e-01 1.962406e-01 3.990168e-02
#>  [796] 2.012744e-01 4.794973e-01 2.135709e-02 8.555050e-02 3.158116e-03
#>  [801] 9.710062e-02 8.533993e-02 2.198567e-01 1.200610e-01 2.806819e-01
#>  [806] 2.534443e-01 5.791516e-02 3.327522e-01 8.141046e-02 1.349149e-01
#>  [811] 1.080359e-01 6.906999e-02 1.345178e-02 3.150316e-01 3.100023e-01
#>  [816] 1.231558e-01 2.715490e-02 3.229361e-01 1.003075e-02 6.152997e-03
#>  [821] 2.398409e-02 3.379029e-01 4.902947e-01 2.598492e-01 1.150014e-01
#>  [826] 4.533621e-01 3.498731e-01 2.358427e-07 2.505682e-03 1.024248e-01
#>  [831] 1.726839e-01 8.460368e-02 4.749333e-02 2.281947e-02 4.168316e-02
#>  [836] 6.494397e-02 1.014344e-01 7.879749e-02 2.451457e-01 4.111766e-01
#>  [841] 7.445104e-02 1.024797e-01 2.685438e-01 1.282550e-02 1.294763e-02
#>  [846] 8.439346e-02 1.400843e-02 1.607363e-06 8.298663e-02 5.101300e-03
#>  [851] 1.343388e-02 3.205501e-02 3.320988e-06 2.789724e-04 9.995667e-02
#>  [856] 2.861746e-01 7.538474e-01 4.103739e-01 5.088523e-01 1.004559e-01
#>  [861] 2.249694e-01 1.810536e-02 1.208677e-01 3.636959e-01 4.588683e-02
#>  [866] 7.341471e-04 3.228620e-08 1.894488e-01 3.982117e-02 1.363964e-02
#>  [871] 2.885345e-01 3.752204e-01 3.935490e-05 1.843222e-01 1.015001e-01
#>  [876] 1.263479e-01 5.028715e-01 1.259731e-01 5.879832e-03 5.180410e-03
#>  [881] 4.709964e-02 2.345858e-01 4.836820e-01 2.426275e-02 2.767865e-01
#>  [886] 3.232477e-02 4.785018e-03 5.218991e-01 3.360717e-01 1.473901e-02
#>  [891] 6.768034e-04 1.355509e-03 3.707893e-02 1.989680e-01 4.894931e-02
#>  [896] 7.685916e-02 1.376082e-01 3.133791e-01 1.933588e-02 4.202742e-04
#>  [901] 2.373774e-01 7.145683e-04 3.117858e-02 1.909879e-02 5.908070e-02
#>  [906] 4.162149e-01 2.839091e-01 4.645435e-01 6.478551e-01 3.621844e-01
#>  [911] 4.258756e-04 2.745655e-01 1.931926e-01 2.581586e-04 3.957811e-02
#>  [916] 4.751024e-02 4.235777e-02 1.709248e-01 5.211172e-02 1.734126e-02
#>  [921] 3.154270e-01 2.815287e-01 3.993538e-01 7.891334e-01 6.188371e-01
#>  [926] 8.987493e-02 5.986370e-01 1.680073e-02 1.702310e-01 1.349100e-01
#>  [931] 2.507384e-01 6.985984e-01 3.622484e-01 4.383122e-01 2.383334e-02
#>  [936] 1.022332e-02 4.268628e-03 3.214474e-01 1.763158e-01 1.453674e-01
#>  [941] 1.986219e-01 8.304561e-01 8.240093e-02 1.484568e-01 2.593339e-01
#>  [946] 2.018069e-02 3.371068e-01 5.052934e-01 6.506094e-01 2.679389e-01
#>  [951] 1.377145e+00 5.330044e-01 1.057280e-01 8.549803e-03 2.682476e-01
#>  [956] 5.953093e-01 3.149092e-01 1.538878e+00 8.618735e-02 4.238354e-03
#>  [961] 3.890562e-02 4.331468e-04 1.167805e-01 7.186629e-02 9.941478e-02
#>  [966] 4.609208e-01 4.521759e-01 5.952990e-01 6.831721e-01 3.595469e-01
#>  [971] 6.712191e-01 1.609713e-01 3.554175e-01 2.531042e-02 3.143781e-03
#>  [976] 8.007839e-02 2.716405e-01 5.749320e-01 4.634284e-01 9.349926e-01
#>  [981] 2.226872e-01 2.997665e-01 1.947906e+00 2.350108e-01 1.024575e-01
#>  [986] 5.271679e-01 3.922720e-01 1.584801e-01 3.826495e-01 1.991185e-01
#>  [991] 3.714645e-01 1.121629e-01 4.988275e-01 7.454917e-01 3.915091e-01
#>  [996] 3.896048e-01 5.896770e-01 2.003286e-01 3.965271e-01 1.933952e+00
#> [1001] 1.535046e+00 1.826704e+00 1.568014e+00 1.214781e+00 5.253163e-01
#> [1006] 9.884207e-01 1.160215e+00 5.730913e-01 8.316456e-01 4.637722e-01
#> [1011] 7.583690e-01 1.057163e-02 3.201676e-01 1.058646e-01 6.004526e-03
#> [1016] 7.656338e-01 7.676555e-01 7.265496e-03 5.073202e-02 3.429151e-01
#> [1021] 5.718206e-01 1.456273e-01 1.621938e+00 1.470770e+00 1.636278e+00
#> [1026] 1.118187e+00 6.470519e-01 3.049563e-01 2.130886e-01 6.604658e-01
#> [1031] 5.217534e-01 3.087634e-01 1.390346e+00 5.269597e-01 1.059071e+00
#> [1036] 9.886717e-01 1.172523e-01 3.108206e-01 5.616089e-01 7.039252e-01
#> [1041] 1.193299e+00 1.443948e+00 2.032811e+00 1.240839e+00 2.371448e+00
#> [1046] 2.754454e+00 1.981953e+00 7.699315e-01 7.616148e-01 2.028299e+00
#> [1051] 2.921992e-01 9.092185e-01 7.723957e-01 3.005928e-01 1.539071e-01
#> [1056] 1.549222e-01 6.125947e-01 1.518927e+00 3.236900e+00 2.247844e+00
#> [1061] 2.039950e+00 9.391415e-01 2.566472e+00 1.998728e+00 1.476368e+00
#> [1066] 9.583362e-01 4.491523e-01 8.017679e-02 3.636372e-01 2.461546e-01
#> [1071] 6.974685e-01 7.836349e-01 1.287476e+00 1.199384e+00 2.461444e+00
#> [1076] 2.781156e+00 2.464653e+00 2.077040e+00 1.876611e+00 2.008750e+00
#> [1081] 1.140416e+00 1.124764e+00 5.121500e-01 2.781664e-01 8.572939e-02
#> [1086] 8.726324e-02 5.265604e-01 4.207790e-01 5.394312e-01 3.331716e-01
#> [1091] 3.907160e-01 7.054267e-01 5.377672e-01 4.919270e-01 6.294852e-01
#> [1096] 9.785033e-01 2.861914e-01 5.420617e-01 3.948780e-01 5.028152e-01
#> [1101] 9.783926e-01 1.330946e+00 1.348340e+00 1.079352e+00 1.353671e+00
#> [1106] 1.349192e+00 1.233866e+00 8.757219e-01 2.990027e-01 1.918991e-01
#> [1111] 3.014498e-02 9.830629e-03 3.015408e-03 2.543195e-01 3.155476e-03
#> [1116] 6.271310e-01 6.372227e-01 8.003859e-01 9.357799e-01 2.108889e-01
#> [1121] 3.728354e-01 1.929305e-02 1.716523e-01 1.260186e+00 6.070425e-01
#> [1126] 9.779823e-01 1.808912e+00 1.172299e+00 1.205697e+00 4.863289e-01
#> [1131] 1.796428e+00 1.374794e+00 1.369725e+00 2.101319e+00 1.667259e+00
#> [1136] 2.234406e+00 1.477698e+00 9.719484e-01 8.811550e-01 1.931544e+00
#> [1141] 1.734885e+00 1.320143e+00 8.557491e-01 1.184803e+00 8.831500e-01
#> [1146] 1.491852e+00 8.129553e-01 1.770326e+00 9.811630e-01 1.684619e+00
#> [1151] 8.454490e-01 1.177744e+00 1.167542e+00 1.596945e+00 2.347827e+00
#> [1156] 1.476457e+00 3.528821e+00 2.024948e+00 1.312495e+00 7.956439e-01
#> [1161] 1.303137e+00 1.545646e+00 5.666545e-01 2.341192e+00 2.386691e+00
#> [1166] 2.136552e+00 2.353997e+00 2.120297e+00 2.820999e+00 3.323212e+00
#> [1171] 2.449872e+00 2.482524e+00 1.850257e+00 1.249666e+00 1.684353e+00
#> [1176] 1.609224e+00 1.424447e+00 1.861663e+00 7.691614e-01 9.411984e-01
#> [1181] 2.031417e+00 9.835658e-01 1.064877e+00 1.647894e+00 2.186224e+00
#> [1186] 6.685423e-01 4.184766e-01 2.040058e-01 4.570117e-01 7.249563e-01
#> [1191] 3.794160e-01 2.034617e-01 4.101111e-02 7.024318e-01 1.452468e+00
#> [1196] 8.100414e-01 5.437032e-01 2.018143e+00 1.089699e+00 2.294840e+00
#> [1201] 2.974649e+00 3.228221e+00 2.265184e+00 3.092109e+00 2.700618e+00
#> [1206] 3.586681e+00 2.161539e+00 1.625919e+00 1.315008e+00 1.228710e+00
#> [1211] 1.319493e+00 1.724883e+00 7.048532e-01 1.966364e+00 3.372589e+00
#> [1216] 2.381214e+00 2.356819e+00 5.738221e-01 1.926339e+00 1.665831e+00
#> [1221] 2.143248e+00 2.309206e+00 1.140632e+00 3.622845e+00 4.654883e+00
#> [1226] 1.999323e+00 2.061614e+00 1.843378e+00 2.005193e+00 1.626087e+00
#> [1231] 1.953882e+00 1.792200e+00 1.522680e+00 2.727836e+00 2.172534e+00
#> [1236] 1.479055e+00 1.712481e+00 9.903492e-01 1.052407e+00 1.431516e+00
#> [1241] 2.431474e-01 2.661685e-01 5.934603e-01 9.143241e-01 1.795456e+00
#> [1246] 3.349383e+00 3.201674e+00 4.351571e+00 1.946510e+00 1.272512e+00
#> [1251] 8.242738e-01 2.108298e+00 1.833099e+00 4.084098e+00 2.100890e+00
#> [1256] 3.549264e+00 2.065307e+00 2.883767e+00 1.549949e+00 3.049636e+00
#> [1261] 2.748557e+00 3.477963e+00 3.523137e+00 3.451195e+00 2.353249e+00
#> [1266] 4.297028e+00 2.630043e+00 2.320127e+00 2.578368e+00 2.103694e+00
#> [1271] 9.520066e-01 1.274239e+00 2.245368e+00 1.242118e+00 2.359215e+00
#> [1276] 2.962306e+00 2.659753e+00 1.975563e+00 1.791026e+00 1.300499e+00
#> [1281] 1.861923e+00 1.239872e+00 2.800645e+00 1.280815e+00 1.803686e+00
#> [1286] 6.631684e-01 3.729712e-01 7.108576e-01 9.775373e-01 7.563295e-01
#> [1291] 1.064370e+00 1.755050e-01 7.242775e-01 1.610297e+00 8.608546e-01
#> [1296] 1.789372e+00 1.944214e+00 1.492290e+00 2.294998e+00 1.552743e+00
#> [1301] 3.003931e-01 1.368991e+00 1.051865e+00 2.210329e+00 1.685688e+00
#> [1306] 2.074259e+00 1.271215e+00 1.611129e+00 1.557953e+00 2.371317e+00
#> [1311] 1.174831e+00 1.510683e+00 8.410252e-01 1.995814e+00 1.295088e+00
#> [1316] 1.722665e+00 2.441953e+00 2.715093e+00 2.118638e+00 2.349474e+00
#> [1321] 1.582981e+00 2.350692e+00 2.936899e+00 2.551314e+00 1.741813e+00
#> [1326] 1.401379e+00 2.688071e+00 1.116691e+00 3.370134e+00 3.746688e+00
#> [1331] 2.373704e+00 1.170489e+00 1.835565e+00 1.843209e+00 1.651302e+00
#> [1336] 2.359087e+00 2.099231e+00 1.584640e+00 1.966521e+00 2.353624e+00
#> [1341] 2.352204e+00 2.263781e+00 1.500512e+00 8.041446e-01
#> 
#> $SPE
#>    [1] 0.0550841842 0.6025320609 0.2537610476 0.3794953182 0.0477819282
#>    [6] 0.6587765606 0.4233749013 0.2184164197 0.1076921185 0.1171102902
#>   [11] 0.0145433034 0.0044442452 0.2738381596 0.0034021755 0.0184890591
#>   [16] 0.1418207239 0.1175297914 0.4330384933 0.2441897871 0.1105171812
#>   [21] 0.1791804650 0.0073093558 0.2224761390 0.0125670524 0.3565928408
#>   [26] 0.0641866938 0.1544635869 0.1581956344 0.0036556952 0.1180354975
#>   [31] 0.0430475443 0.0315511115 0.3596940165 0.3907494028 0.3418520947
#>   [36] 0.1414464365 0.2788359377 0.0307131959 0.0825644649 0.0882474566
#>   [41] 0.0254966231 0.0476725310 0.3080037128 0.1591073099 0.2224222693
#>   [46] 0.1679751831 0.0207005441 0.0125796106 0.5138945201 0.0332075731
#>   [51] 0.1560742681 0.3393450959 0.6395178236 0.3836434500 0.0716429638
#>   [56] 0.0933423142 0.1890973499 0.0677248831 0.0132010008 0.1207626679
#>   [61] 0.1741233724 0.2133052524 0.0758336088 0.0271156876 0.1145785198
#>   [66] 0.1172597054 0.0137826760 0.3081673151 0.0129093324 0.0609062600
#>   [71] 0.0600493564 0.0442054451 0.0480658948 0.0270199484 0.1864068405
#>   [76] 0.1835165111 0.0321807280 0.0056066354 0.2248050644 0.3898766229
#>   [81] 0.0297899568 0.0664967947 0.0908328513 0.5532731815 0.4202256551
#>   [86] 0.0819867425 0.7079704278 0.0642734436 0.5722995305 0.0010525314
#>   [91] 0.1260799941 0.4706476403 0.0039732897 0.0111098549 0.3068524952
#>   [96] 0.0071038330 0.0313192614 0.3669259445 0.3506936131 0.0165895664
#>  [101] 0.7520944567 0.0536325056 0.0927826802 0.0151849735 0.1310149436
#>  [106] 0.3718836683 0.1949171545 0.2376975853 0.0186913959 0.0575037311
#>  [111] 0.0474815018 0.1771693263 0.7063032723 0.1901757509 0.0934424785
#>  [116] 0.1971126306 0.0358580611 0.0274439024 0.0647939158 0.3097406511
#>  [121] 0.0993209581 0.3850450291 0.2505692904 0.0263553563 0.0679506181
#>  [126] 0.0396446941 0.5689317761 0.0362800696 0.0924968185 0.0583120314
#>  [131] 0.0990569316 0.1157340487 0.0844471562 0.1658105537 0.0668010213
#>  [136] 0.3114078171 0.2388592341 0.0239224422 0.6167138682 0.0314948895
#>  [141] 0.2971889591 0.0693633830 0.1078559044 0.1123385045 0.1086169219
#>  [146] 0.0425644049 0.0085229695 0.4372027730 0.1160127836 0.2310407649
#>  [151] 0.0096255982 0.2913421171 0.2105286464 0.1784015008 0.0189901593
#>  [156] 0.2180782348 0.2848996215 0.2767239333 0.0501620660 0.0280898177
#>  [161] 0.3604138619 0.6830722032 0.0030486074 0.0201612453 0.4941726641
#>  [166] 0.2517400905 0.2125000082 0.0637438996 0.2888998295 0.3200086174
#>  [171] 0.3569593224 0.3749466233 0.0043451212 0.0939874008 0.1607539642
#>  [176] 0.2632424395 0.1803260195 0.0647482519 0.4195556296 0.1852535407
#>  [181] 0.0575602716 0.5368708436 0.1512522812 0.0096439450 0.2049396866
#>  [186] 0.0397621849 0.2264726142 0.0091856390 0.5200236805 0.1296057528
#>  [191] 0.1585411747 0.0530029458 0.0423226928 0.2778550274 0.0422165924
#>  [196] 0.1738592072 0.2953559904 0.0884701612 0.7991965946 0.0772412147
#>  [201] 0.0894288551 0.0717469312 0.1551314914 0.0751841887 0.0639804115
#>  [206] 0.0358914153 0.0281448011 0.0479223476 0.0603709767 0.1711997529
#>  [211] 0.0072459754 0.5739263875 0.2644018941 0.2795177989 1.0454868082
#>  [216] 0.4603603230 0.8015191625 0.2147770184 0.0341284698 0.5662154438
#>  [221] 0.6114512256 0.1817195992 0.0523761609 0.7398522653 0.0323988348
#>  [226] 0.0089441010 0.1801639958 0.0944115598 0.0367948839 0.1307625362
#>  [231] 0.0395571040 0.0735674205 0.0088252448 0.2697846711 0.7780582660
#>  [236] 0.1458530535 0.1374910613 0.0207647174 0.9299368009 0.0954085124
#>  [241] 0.0548198313 0.1028629683 0.0552770503 0.0805458487 0.0202332502
#>  [246] 0.1272805255 0.0302240239 0.3046483617 0.0536227708 0.0123835701
#>  [251] 0.2364826992 0.5265535858 0.0584463668 0.0034876888 0.3354174263
#>  [256] 0.3264998927 0.1563823588 0.0966845536 0.0310661701 0.0031675616
#>  [261] 0.1748607592 0.2697963565 0.3413979648 0.0316799426 0.1777167093
#>  [266] 0.1303789214 0.9159318112 0.2226867743 0.1613279128 0.0295381104
#>  [271] 0.2020129281 0.1560751991 0.1057489002 0.0406103625 0.1895553665
#>  [276] 0.0614658200 0.0848572711 0.0760061542 0.7440352933 0.4263920769
#>  [281] 0.6938652644 0.0557265936 0.0500812268 0.0208560352 0.1495828095
#>  [286] 0.3546918842 0.0014881827 0.1082023235 0.2233036543 0.4766624995
#>  [291] 0.0405791874 0.1059299441 0.0280561517 0.1163141443 0.2730091883
#>  [296] 0.0657139911 0.0153767855 0.0270717366 0.0737166663 0.2182618368
#>  [301] 0.1242387842 0.0441351084 0.0336838870 0.1310494074 0.1265756328
#>  [306] 0.0321190150 0.1307643974 0.0322056343 0.0200396151 0.0946083419
#>  [311] 0.0167516555 0.7279468626 0.0308266823 0.0035971909 0.0858875301
#>  [316] 0.2012760600 0.1081280175 0.2768714203 0.0186007528 0.0870725674
#>  [321] 0.1938438305 0.2358947314 0.2476862883 0.1356797640 0.2945384874
#>  [326] 0.0604139695 0.4106863199 0.7313202665 0.0760724992 0.0824170018
#>  [331] 0.0722102474 0.0652226781 0.0089547521 0.1754915603 0.2210319859
#>  [336] 0.2554646495 0.0025827534 0.1280909256 0.2855684973 0.0543618176
#>  [341] 0.0252745074 0.0720426312 0.3865913959 0.0717767025 0.3250662900
#>  [346] 0.4029649450 0.2249945250 0.2582519919 0.3729495497 0.0054590781
#>  [351] 0.0207091356 0.2525344445 0.0763342611 0.0205075607 0.5647459286
#>  [356] 0.2402909516 0.3516377563 0.1213787906 0.0581024562 0.1017278501
#>  [361] 0.0842642067 0.0619144291 0.5137047247 0.0509498615 0.0886097124
#>  [366] 0.5064880906 0.0100802010 0.1839472484 0.0036611061 0.0180645058
#>  [371] 0.2530957110 0.0214942419 0.0018734317 0.0003711754 0.2211190120
#>  [376] 0.0359369912 0.1336773155 0.1815287522 0.0534567511 0.0213394226
#>  [381] 0.0743283605 0.0961021738 0.2806142814 0.0627818141 0.0422993545
#>  [386] 0.0195706171 0.3270188391 0.5702919958 0.1384192900 0.5067462388
#>  [391] 0.0974473306 0.2981334154 0.1836633666 0.0277730855 0.3544121702
#>  [396] 0.0178358904 0.0450715881 0.0547946017 0.0997915801 0.0673663068
#>  [401] 0.0621026030 0.0093203542 0.1126031194 0.0298506063 0.0855807287
#>  [406] 0.2022731499 0.0127158272 0.3601011122 0.3690243596 0.0745764457
#>  [411] 0.5329910139 0.0165802750 0.0107337901 0.0783679418 0.0500244600
#>  [416] 0.0860885552 0.3673113453 0.1553707982 0.6940183870 0.1982089976
#>  [421] 0.1037349386 0.0714819130 0.0929095983 0.1774621867 0.0271997371
#>  [426] 0.1243529264 0.1126249317 0.0424998790 0.1182642070 0.0352644502
#>  [431] 0.0202792726 0.0551227593 0.3836057461 0.1203363123 0.2903500097
#>  [436] 0.1655937598 0.0516254183 0.1311045486 0.0697925342 0.2900875013
#>  [441] 0.0741688468 0.0693887257 0.2546074414 0.2565525986 0.1516216615
#>  [446] 0.2112472618 0.0646162857 0.3457761460 0.0280118848 0.2232502298
#>  [451] 0.3476649594 0.0918280889 0.0419868977 0.2722904496 0.0450483307
#>  [456] 0.1052200806 0.1050071787 0.5689726323 0.0244522813 0.1248578082
#>  [461] 0.2811765352 0.0455720864 0.0139220535 0.1801763704 0.0587427560
#>  [466] 0.1219543118 0.1123403014 0.0147573410 0.1247411009 0.1565410684
#>  [471] 0.1252388999 0.1253728397 0.0223355426 0.3100378955 0.1736859115
#>  [476] 0.0438996953 0.1284635922 0.0728089726 0.3163127102 0.0612783228
#>  [481] 0.2988489341 0.3929058127 0.1755688472 0.0682648204 0.2206973234
#>  [486] 0.1346940531 0.1567616262 0.0873590530 0.1260772898 0.2196519353
#>  [491] 0.1273158600 0.1594269767 0.4029255959 0.6076353018 0.0294357027
#>  [496] 0.0962954039 0.1713300576 0.2247258525 0.9478576310 0.0142214163
#>  [501] 0.4158846576 0.4743867437 0.2148928145 0.1585167908 0.1477699155
#>  [506] 0.0151034560 0.1254246865 0.0737390711 0.0525466821 0.2077766561
#>  [511] 0.1513933992 0.0053211587 0.0865864237 0.0253967249 0.2338516835
#>  [516] 0.3381513690 0.7716411394 0.4546512167 0.1217886780 0.0557072513
#>  [521] 0.2544039678 0.1179709352 0.0054527149 0.1411933888 0.0591682832
#>  [526] 0.2601571716 0.4075801624 0.0952244075 0.1391148045 0.6143263301
#>  [531] 0.0506710356 0.0811419593 0.1786189094 0.3639543240 0.3775754816
#>  [536] 0.0622870592 0.1316932065 0.7172575116 0.0431391223 1.2317472771
#>  [541] 1.1087501496 0.6080095594 0.1611906943 0.1535163340 0.1224546144
#>  [546] 0.2981972523 0.2524819360 0.1088555775 0.1982351720 0.2924699533
#>  [551] 0.0440980160 0.0484726784 0.0538872461 0.2734242953 0.5145649494
#>  [556] 0.0986794906 0.0371635626 0.2384155136 0.2936392501 0.2767440947
#>  [561] 0.0415979289 0.0557997428 0.5305625944 0.1224679118 0.3114778973
#>  [566] 0.2381082797 0.2074393005 0.0362297701 0.0241673216 0.0312153929
#>  [571] 0.4234122744 0.3000315825 0.0844732022 0.5365185911 0.1392138031
#>  [576] 0.0474706566 0.2325960880 0.2410411951 0.1691057570 0.1885935847
#>  [581] 0.1168282773 0.6379594188 0.3801140223 0.2928702063 0.0151220714
#>  [586] 0.1383441427 0.2234928288 0.1896868807 0.2658377540 0.2592771924
#>  [591] 0.3595472582 0.0802006402 0.1871092777 0.2799006982 0.5744919928
#>  [596] 0.1894771093 0.0290312037 0.3319343820 0.2381506658 0.0979263901
#>  [601] 0.0123373281 0.0017932182 0.0849100273 0.0097247762 0.6728756103
#>  [606] 0.0101721231 0.0366425783 0.1999140904 0.1996184564 0.2792611214
#>  [611] 0.7328341381 0.0895091079 0.4048275825 0.1568891781 0.0103871968
#>  [616] 0.0923017517 0.1375800314 0.1759295504 0.1464547141 0.1428858566
#>  [621] 0.2484289882 0.0747126558 0.0658822724 0.0414676596 0.0590069768
#>  [626] 0.2846814663 0.0133963373 0.3274175989 0.1644129534 0.1682196708
#>  [631] 0.1841803421 0.6000831077 0.1803361695 0.0304627819 0.4587284778
#>  [636] 0.0220241495 0.1076811473 0.1298625350 0.6383241502 0.0614399705
#>  [641] 0.2181595039 0.0027366083 0.3300567656 0.2389746064 0.2949976453
#>  [646] 0.3008941616 0.2915493494 0.0211112117 0.1834507549 0.0360861579
#>  [651] 0.1719151070 0.0091850605 0.7572794734 0.1277955188 0.2546887999
#>  [656] 0.4081998125 0.0446130161 0.0055108857 0.0157936787 0.1288579698
#>  [661] 0.2397701464 0.0668600456 0.1981094345 0.0145837124 0.0032040896
#>  [666] 0.0488270496 0.0711799212 0.4072475830 0.0219814110 0.0249870832
#>  [671] 0.3305559985 0.0409291486 0.1327249888 0.0903928102 0.1742038219
#>  [676] 0.0448884031 0.0558058014 0.0846047290 0.2345684975 0.0028238428
#>  [681] 0.5793689038 0.0333949544 0.0793317636 0.1110593775 0.0807034703
#>  [686] 0.0269586642 0.1906831456 0.3833229388 0.0023963479 0.1306646267
#>  [691] 0.1771923298 0.0366514100 0.4214470674 0.0832438150 0.0359987930
#>  [696] 0.1255665111 0.1426736950 0.2413695080 0.1926090794 0.0482162751
#>  [701] 0.2044216429 0.7506388311 0.0860017681 0.4458216746 0.0352342584
#>  [706] 0.0474883993 0.1422083944 0.0120421199 0.1390921988 0.1479033868
#>  [711] 0.1177687553 0.4471420665 0.0299503604 0.1637971843 0.2275044733
#>  [716] 0.0890176378 0.0349952273 0.0714659339 0.2331543480 0.2085977295
#>  [721] 0.1047035454 0.1840865311 0.0493599383 0.0108982264 0.0800588719
#>  [726] 0.1226086915 0.0235879090 0.1735245511 0.0157071581 0.0197179866
#>  [731] 0.4536193281 0.2319771962 0.0824376926 0.1429427998 0.4081461149
#>  [736] 0.5185706660 0.1259722331 0.4764102033 0.2020189853 0.2015196270
#>  [741] 0.0108519008 0.2568506879 0.1178153132 0.7617692545 0.0689874119
#>  [746] 0.0029730884 0.2593545508 0.1157326698 0.1118373015 0.6501148795
#>  [751] 0.0978442397 0.1031595153 0.1445570852 0.0385688723 0.2767707781
#>  [756] 0.3080959860 0.0140890310 0.0841833874 0.2300490249 0.1887515642
#>  [761] 0.0257198590 0.4857686871 0.0137514059 0.2882043548 0.2975056556
#>  [766] 0.0034348109 0.0704202657 0.1022949410 0.1569139574 0.1702044936
#>  [771] 0.0752210445 0.0279514716 0.1512232802 0.3785169152 0.1868682810
#>  [776] 0.4170572770 0.1872709723 0.1550308428 0.0520173389 0.0120201246
#>  [781] 0.1210022036 0.2577550401 0.0560482449 0.4024608473 0.2890824786
#>  [786] 0.2546885763 0.0063231391 0.0983237468 0.0779988013 0.0752211405
#>  [791] 0.0027464850 0.0173354046 0.0401378960 0.2312386934 0.0116648906
#>  [796] 0.0383199218 0.0119657585 0.0430817133 0.0365142678 0.1316480045
#>  [801] 0.2111138025 0.2022418943 0.1811307839 0.1345972064 0.0495849149
#>  [806] 0.1091565164 0.0370763046 0.0431278005 0.0188936253 0.2881440824
#>  [811] 0.0201070876 0.1490255187 0.2073667610 0.0532965129 0.0220508941
#>  [816] 0.1309916619 0.2134297313 0.0817084264 0.1506994417 0.0519896274
#>  [821] 0.0533354711 0.0430858923 0.3039103931 0.0142178468 0.2665899178
#>  [826] 0.2445245006 0.4752913356 0.2012266313 0.0912790593 0.0499624595
#>  [831] 0.1376477665 0.3464298880 0.5501868747 0.1484520047 0.2444610499
#>  [836] 0.0827708453 0.0868424227 0.1249898902 0.2515163060 0.1070700515
#>  [841] 0.1046938390 0.2817682667 0.0753292589 0.1668874620 0.2025643875
#>  [846] 0.0794737031 0.1235775886 0.3668453276 0.7132767333 0.0224841984
#>  [851] 0.2566628058 0.0228408960 0.0498644209 0.1837656864 0.0393323540
#>  [856] 0.1630246937 0.0099532223 0.1154356303 0.0964011657 0.4400302322
#>  [861] 0.0619954288 0.2006626034 0.0683293599 0.0189328982 0.4983138645
#>  [866] 0.3107685948 0.1329197198 0.0292830240 0.6815006464 0.0015039795
#>  [871] 0.3801001319 0.0279688490 0.1347041561 0.1524984629 0.0538604962
#>  [876] 0.0925764096 0.1245312024 0.0535804400 0.2405027125 0.1087703644
#>  [881] 0.1601034711 0.1513367290 0.8174359379 0.3207213451 0.0915449540
#>  [886] 0.2202670458 0.1645147897 0.2037904040 0.1550040380 0.9312464402
#>  [891] 0.0122420068 0.1007113498 0.0998930358 0.0531507881 0.0720803602
#>  [896] 0.4234369076 0.0710022288 0.2307889585 0.0829576242 0.1005089819
#>  [901] 0.5794752354 0.0040936151 0.0178858158 0.0222461456 0.3333133900
#>  [906] 0.3058525808 0.0005607745 0.0196626019 0.0109727476 0.0547283111
#>  [911] 0.0043043344 0.1882145413 0.0658433112 0.5211639154 0.1949157817
#>  [916] 0.0085591194 0.4821547390 0.3027037697 0.1451023234 0.0253475143
#>  [921] 0.0044059262 0.2292694934 0.4409501019 0.0557023577 0.0368233978
#>  [926] 0.1830712438 0.3992713813 0.1071833533 0.2248848562 0.0464788774
#>  [931] 0.1336323767 0.0700716020 0.0046776271 0.1024516365 0.2606886137
#>  [936] 0.1208862024 0.1381802483 0.0025001209 0.2315332344 0.0102904075
#>  [941] 0.1116098636 0.3407658922 0.0072356289 0.2357699777 0.3500691805
#>  [946] 0.2562694282 0.3738281320 0.1920753887 0.1065968864 0.0061143170
#>  [951] 0.0599193262 0.2257907934 0.2904139271 0.8080850386 0.4984722518
#>  [956] 0.0337509834 0.1109701881 0.0475179532 0.2208633699 0.1518724190
#>  [961] 0.0095664435 1.0283951975 0.2730530922 0.1149560059 0.1797505720
#>  [966] 0.0875726356 0.0044742384 0.0440811584 0.0602757945 0.2774774515
#>  [971] 0.1126610986 0.2114970391 0.5061524777 0.0740268136 0.0655270521
#>  [976] 0.1020803556 0.2465169696 0.2525065016 0.1131183183 0.1588478046
#>  [981] 0.3261482534 0.0949562503 0.4929443566 0.0261893885 0.4302589371
#>  [986] 0.0051956305 0.0378542919 0.2175224646 0.4254489577 0.3284310868
#>  [991] 0.2658241115 0.0034771729 0.0414797475 0.0652936747 0.1037818310
#>  [996] 0.1701733518 0.0079777132 0.4510755098 0.1570509926 0.1628175147
#> [1001] 0.2295708369 0.0060371041 0.1253676132 0.2071662843 0.0308700001
#> [1006] 0.0878962004 0.1477592621 0.0207536470 0.0472728977 0.1794027345
#> [1011] 0.0142316576 0.1569701127 0.2921343405 0.1644769823 0.0867039685
#> [1016] 0.0358183099 0.0443735655 0.0857286322 0.0155689146 0.0969250861
#> [1021] 0.0056614834 0.2707293156 0.1542683583 0.1734767055 0.6053330212
#> [1026] 0.5235073448 0.2437566727 0.1114410107 0.2421181774 0.1157304838
#> [1031] 0.0544438469 0.2858987408 0.3283116071 0.0032886721 0.0773792332
#> [1036] 0.0980516072 0.2383283316 0.2639349678 0.1224595160 0.0246593357
#> [1041] 0.0459528602 0.0116627374 0.5287174463 0.0228777797 0.2686901975
#> [1046] 0.3257552799 0.1998275447 0.0026615444 0.0559842766 0.1849333939
#> [1051] 0.4679037451 0.0240864669 0.6311233816 0.0365688501 0.0055059997
#> [1056] 0.1445077063 0.1244927187 0.0534094418 0.0669917197 0.0437529034
#> [1061] 0.0299102042 0.1090772017 0.4107422803 0.0003386438 0.0644565462
#> [1066] 0.0519799206 0.0324633961 0.2131452794 0.2758986074 0.0255623815
#> [1071] 0.0725005196 0.1623500905 0.0949808965 0.2449604622 0.3756507094
#> [1076] 0.5247116832 0.1550337968 0.2098913478 0.4136821352 0.0499827612
#> [1081] 0.0495890870 0.1506312088 0.0359899593 0.0567539983 0.0370064744
#> [1086] 0.4830005519 0.0980004534 0.0742610151 0.2030436567 0.1726686678
#> [1091] 0.0178988399 0.0903722118 0.3906863544 0.1072249237 0.0844245280
#> [1096] 0.1630851110 0.1803623880 0.0279071735 0.1653914819 0.0197214522
#> [1101] 0.1035266268 0.1771462481 0.1182510320 0.2634033495 0.0737907800
#> [1106] 0.2035226173 0.1375272817 0.0184159298 0.1415004226 0.4409615742
#> [1111] 0.2519960926 0.1478241156 0.4831540588 0.1252399321 0.1697607689
#> [1116] 0.0586238028 0.2199342441 0.0236020825 0.0590471685 0.1852460321
#> [1121] 0.0289140918 0.9696972438 0.0622858378 0.1735192985 0.0464586373
#> [1126] 0.3083672241 0.0844996668 0.2370403100 0.0834295085 0.0692841250
#> [1131] 0.1538662874 0.0331983705 0.2484268581 0.2736010956 0.1261318500
#> [1136] 0.0166280241 0.4085448658 0.1183082464 0.0884100805 0.6531810236
#> [1141] 0.0063166211 0.0745789346 1.1809846303 0.3403069771 0.1342617210
#> [1146] 0.2847668086 0.8781773912 0.0854267212 0.0437367378 0.1855971604
#> [1151] 0.0835442832 0.2189284445 0.1626582845 0.1847544610 0.0627373447
#> [1156] 0.1047464693 0.3146482741 0.0116523075 0.0884089168 0.0532269511
#> [1161] 0.0434505880 0.5222515557 0.1298357147 0.2634707063 0.0405182505
#> [1166] 0.0665493346 1.0592486097 0.3162241772 0.3234220662 0.1690253272
#> [1171] 0.0574231131 0.3080699044 0.0261284493 0.0440435317 0.1745173679
#> [1176] 0.0891822238 0.0966225066 0.4882896508 0.1775681313 0.2163752670
#> [1181] 0.0299737789 0.0355186867 0.3102721598 0.0761044495 0.0917775055
#> [1186] 0.1834746799 0.0415163700 0.1710809734 0.0004051926 0.1705205861
#> [1191] 0.8234889486 0.0045467336 0.0838961451 0.1404621398 0.1017006630
#> [1196] 0.1881877912 0.0970361833 0.0493038678 0.2433188636 0.2302714328
#> [1201] 0.4620602619 0.3491551481 0.3996933428 0.4490846968 0.3433755187
#> [1206] 0.3580773955 0.1443896916 0.0284478681 0.1596591311 0.0107585306
#> [1211] 0.1480432544 0.2003513600 0.0406923648 0.0403110350 0.4289011947
#> [1216] 0.3102859186 0.2185719899 0.2439882401 0.1017534676 0.1444743536
#> [1221] 0.0899230943 0.2403324500 0.3184566781 0.2406662324 0.9984236626
#> [1226] 0.1989312750 0.3619190480 0.1493888151 0.1561758713 0.0865675006
#> [1231] 0.3512719650 0.1230503562 0.1626191684 0.3658831162 0.3155008316
#> [1236] 0.2468113631 0.0820190665 0.0384964567 0.0712744987 0.1927712520
#> [1241] 0.2316595627 0.0292056283 0.0436191244 0.2363172895 0.0196378553
#> [1246] 0.0229080950 0.2613617857 1.3957273233 0.2518628540 0.0287322009
#> [1251] 0.0514534120 0.3394076275 0.3321532380 0.5045645498 0.3747569485
#> [1256] 0.2336148488 0.3791903011 0.2445841261 0.0754612839 0.5458188266
#> [1261] 0.0096128911 0.5603765597 0.5660611575 0.1741852158 0.3067913580
#> [1266] 0.3874166914 0.2026035093 0.4426318065 0.4893851925 0.3558390773
#> [1271] 0.2402499937 0.0352749112 0.4356083313 0.6220223585 0.2332261746
#> [1276] 0.3401381235 0.3262398429 0.1411463539 0.0401058723 0.0716545676
#> [1281] 0.0538502825 0.5992078196 0.0236015896 0.7603103792 0.0260710813
#> [1286] 0.1479696243 0.7710305997 0.4338046062 0.3443950219 0.0018064990
#> [1291] 0.0008191928 0.0973254928 0.0096223389 0.0821424483 0.2313221354
#> [1296] 0.2684765654 0.3764640842 0.1461426859 0.0216155398 0.0129630995
#> [1301] 0.1280112276 0.0393404357 0.1495004737 0.1192296383 0.2824989025
#> [1306] 0.1335985776 0.1178427858 0.0707565079 0.0035896251 0.1341615895
#> [1311] 0.2285809942 0.1249438043 0.0098537499 0.1238144845 0.1243110118
#> [1316] 0.2116638808 0.4279052414 0.3236541645 0.2716199921 0.4352440987
#> [1321] 0.0457335290 1.0626210544 0.1224681823 0.0087479889 0.1193823438
#> [1326] 1.1104131442 0.0217517070 0.4795846639 0.2398350833 0.3828664305
#> [1331] 0.1292233697 0.0149544712 0.0222206670 1.0071851087 0.0039409472
#> [1336] 0.0343235688 0.0627151918 0.0637879649 0.1238842903 0.7251229702
#> [1341] 0.0111960151 0.1851468051 0.1579605120 0.4532981256
#> 
#> $muTrain
#>          x          y          z 
#>  0.6103488 -1.3978814  0.9503775 
#> 
#> $RootPrecisTrain
#>          [,1]     [,2]     [,3]
#> [1,] 3.901159 0.000000 0.000000
#> [2,] 0.000000 2.329768 0.000000
#> [3,] 0.000000 0.000000 1.681929
#> 
#> attr(,"class")
#> [1] "threshold" "pca"      
#>