Simulate NOC Observations from a Single-State or Multi-State Process

This function generates data under normal operating conditions from a single-state or multi-state process model.

processNOCdata(startTime = "2015-05-16 10:00:00 CST", period = 7 * 24 * 60,
  stateDuration = 60, increment = "min", multiState = TRUE,
  autocorellation = 0.75, tLower = 0.01, tUpper = 2, errVar = 0.01)

Arguments

startTime
a POSIXct object specifying the day and time for the starting observation.

period
The observation cycle length. Defaults to one week's worth of minute-level observations (10,080 observations).

stateDuration
The number of observations generated during a stay in each state. Defaults to 60.

increment
The time-sequence base increment. See "Deatails" of the seq.POSIXt() function options. Defaults to "min" for minutes.

multiState
Should the observations be generated from a multi-state process? Defaults to TRUE.

autocorellation
The autocorrelation parameter. Must be less than 1 in absolute value, or the process generated will be nonstationary. Defaults to 0.75 in accordance to Kazor et al (2016).

tLower
Lower bound of the latent $t$ variable. Defaults to 0.01.

tUpper
Upper bound of the latent $t$ variable. Defaults to 2.

errVar
Error variance of the normal white noise process on the feature variables.

Value

An data frame with the following information:

dateTime -- A POSIXct column of times starting at the user-defined `startTime` argument, length given by the `period` argument, and spacing given by the `increment` argument. For example, if the starting value is "2016-01-10", period is 10080, and the incrementation is in minutes, then this sequence will be one week's worth of observations recorded every minute from midnight on the tenth of January.

state -- An integer column of all 1's (when the `multiState` argument is FALSE), or a column of the state values (1, 2 or 3).

x -- A double column of generated values for the first feature.

y -- A double column of generated values for the second feature.

z -- A double column of generated values for the third feature.

Details

This function randomly generates a non-stationary (sinusoidal) and autocorrelated latent variable t with lower and upper bounds given by the arguments "tLower" and "tUpper", respectively, with autocorrelation governed by the "autocorrelation" argument. Necessarily, this coefficient must be less than 1 in absolute value, otherwise the latent variable will be unbounded. Next, this function draws a realization of this random variable t and calculates three functions of it, then jitters these functions with a normal white noise variable (with variance set by "errVar"). These three functions are:

x : x(t) = t + error

y : y(t) = t ^ 2 - 3t + error

z : z(t) = -t ^ 3 + 3t ^ 2 + error

This function is called by the mspProcessData() function. See ?mspProcessData for more details.

See also

Called by: mspProcessData.

Examples

processNOCdata()

#>               t          err1          err2          err3            x
#> 1     0.2621546  1.556737e-01  7.409917e-02  1.614683e-01  0.417828240
#> 2     0.2810781  6.543933e-02  1.329804e-02 -5.478606e-02  0.346517450
#> 3     0.3097706  1.328268e-01 -1.114990e-01  1.407444e-01  0.442597429
#> 4     0.4166743  1.753963e-01 -9.362201e-02 -9.569020e-02  0.592070598
#> 5     0.3738819  9.580106e-02 -1.545517e-01 -1.207556e-01  0.469683003
#> 6     0.3120113  2.481486e-02  3.933051e-02 -1.010512e-02  0.336826199
#> 7     0.2846599  1.338470e-02  8.634444e-02 -3.976804e-02  0.298044585
#> 8     0.3070906 -1.027115e-01  3.644002e-02  3.970822e-03  0.204379158
#> 9     0.3075152 -5.720748e-03 -7.936919e-02 -8.872414e-02  0.301794456
#> 10    0.1877876  4.637578e-03 -1.858858e-02  2.150638e-01  0.192425187
#> 11    0.1915215 -1.633681e-01 -7.629084e-03 -5.588749e-02  0.028153430
#> 12    0.2267525  7.785308e-02  8.399060e-04 -3.799926e-02  0.304605563
#> 13    0.1958829  2.218336e-02 -2.573029e-02 -2.876974e-02  0.218066277
#> 14    0.1966178 -7.814078e-02 -4.712114e-02  1.219358e-01  0.118477055
#> 15    0.2817221 -8.934600e-02  4.841127e-03 -8.762674e-02  0.192376115
#> 16    0.2932674  2.823691e-02 -1.769191e-02 -1.882930e-01  0.321504352
#> 17    0.2750029 -6.036474e-03  1.058584e-01  9.870400e-02  0.268966400
#> 18    0.2893122 -7.506703e-02 -1.086936e-01  1.712609e-01  0.214245191
#> 19    0.3541019 -1.206695e-01  1.777013e-01  6.336923e-02  0.233432345
#> 20    0.3551669  6.414373e-03  1.200188e-01 -1.094893e-02  0.361581235
#> 21    0.3415772 -9.203733e-03 -2.337975e-02 -2.553604e-01  0.332373449
#> 22    0.2618967 -1.331609e-01 -2.542631e-02 -8.335188e-02  0.128735797
#> 23    0.1999524 -1.038177e-01  1.899202e-01 -9.049042e-03  0.096134767
#> 24    0.1085961 -5.830696e-02  1.557290e-01  2.119990e-02  0.050289102
#> 25    0.1816967 -1.939800e-01 -1.569853e-01 -4.355222e-02 -0.012283290
#> 26    0.3006585  4.431008e-02  2.249992e-03 -1.168099e-01  0.344968597
#> 27    0.3622971  3.889652e-02 -3.928399e-02 -7.069205e-02  0.401193669
#> 28    0.3623883 -2.012162e-02 -1.322148e-01  3.448471e-02  0.342266723
#> 29    0.4579149  2.639431e-02  4.639887e-02  6.378251e-02  0.484309186
#> 30    0.4772984  4.515955e-02 -1.197222e-02 -2.226163e-01  0.522457917
#> 31    0.3705314 -5.340362e-02  2.269685e-02 -3.447137e-02  0.317127796
#> 32    0.3746222  9.857638e-02  8.467148e-02  2.996266e-02  0.473198617
#> 33    0.3576002 -3.266562e-03 -1.013966e-01 -9.012724e-02  0.354333683
#> 34    0.2680562  1.133526e-01  1.138693e-01  2.099133e-01  0.381408776
#> 35    0.2266005  1.022098e-01 -2.121010e-01  3.878354e-02  0.328810266
#> 36    0.2021983  1.484039e-01  7.875796e-02  3.691047e-03  0.350602252
#> 37    0.2512437  1.427647e-02  1.519998e-01 -9.649253e-02  0.265520210
#> 38    0.2380779  1.091497e-01 -1.568346e-01 -1.051424e-01  0.347227603
#> 39    0.2744756 -1.179650e-01  8.330503e-02 -8.569613e-02  0.156510581
#> 40    0.2537532 -1.059007e-01 -1.284627e-01  1.238731e-01  0.147852424
#> 41    0.2121664 -8.962362e-02  1.174012e-01 -8.282722e-03  0.122542748
#> 42    0.3054050  9.161304e-02 -6.726598e-02  1.748495e-02  0.397018074
#> 43    0.3149472  5.385451e-03 -3.242405e-02  8.267206e-02  0.320332633
#> 44    0.3883666 -1.146932e-01 -3.781493e-02  2.341427e-01  0.273673376
#> 45    0.3644614 -1.152512e-02 -6.031573e-02  1.457776e-01  0.352936250
#> 46    0.3009223 -6.784486e-02  7.381053e-02  6.796722e-02  0.233077401
#> 47    0.2994560  2.264581e-02 -1.199737e-01  3.130613e-02  0.322101801
#> 48    0.2752426  1.437058e-01  6.066060e-02  2.408405e-01  0.418948432
#> 49    0.2990186  4.305425e-02 -6.189996e-02  7.730895e-02  0.342072813
#> 50    0.3507520 -6.445995e-03 -4.234366e-02  9.964246e-02  0.344306034
#> 51    0.3513061 -1.370873e-01 -2.864727e-01 -4.483531e-03  0.214218772
#> 52    0.4796198  5.158869e-02  1.042629e-01 -1.391563e-02  0.531208527
#> 53    0.4655924 -1.259715e-01  1.296018e-01  2.098706e-02  0.339620916
#> 54    0.4013555 -6.956395e-02  6.695171e-02  1.688812e-03  0.331791539
#> 55    0.3631125  6.494041e-02  2.696281e-02 -2.438363e-02  0.428052901
#> 56    0.4297777 -3.667251e-02 -3.122502e-02  2.161460e-01  0.393105199
#> 57    0.4560553 -8.767217e-02  5.775447e-02 -2.130909e-01  0.368383123
#> 58    0.4671386  9.792012e-02  4.168095e-02  3.989206e-02  0.565058759
#> 59    0.4550410  7.885297e-03 -8.036559e-02 -5.818261e-02  0.462926318
#> 60    0.4665226 -3.776705e-02 -1.200963e-01  5.106864e-02  0.428755540
#> 61    0.4918716 -3.110015e-02 -7.932047e-02  9.407356e-02  0.460771487
#> 62    0.4278860 -3.182142e-02  9.275401e-02 -6.496881e-02  0.396064540
#> 63    0.4471493 -3.061104e-02 -1.527969e-02  1.747498e-01  0.416538266
#> 64    0.3795784  7.080051e-02 -6.707191e-02  3.922149e-02  0.450378939
#> 65    0.4151549 -2.237236e-02 -5.724836e-03  2.611181e-02  0.392782587
#> 66    0.4431297 -2.231043e-02  1.087918e-01 -2.625447e-02  0.420819316
#> 67    0.4565435  3.833727e-02  2.374851e-02 -1.520176e-01  0.494880757
#> 68    0.4212928  9.543294e-02  1.832474e-01 -5.872776e-02  0.516725734
#> 69    0.4030221 -2.879093e-02 -9.116336e-02 -3.278853e-02  0.374231223
#> 70    0.3818498  9.089056e-02 -4.515828e-02  8.248636e-02  0.472740355
#> 71    0.3831896 -9.566812e-02 -5.240532e-02  1.799927e-01  0.287521435
#> 72    0.3976613  6.101292e-02  4.161323e-02 -4.375404e-03  0.458674204
#> 73    0.4431069 -3.193555e-03 -1.473606e-01 -1.078018e-01  0.439913378
#> 74    0.3639068  1.499134e-01  1.866496e-01 -7.238531e-02  0.513820199
#> 75    0.3074750  1.799935e-01 -2.504233e-01 -5.340217e-02  0.487468590
#> 76    0.3546683  7.124027e-04 -7.077853e-02  6.322314e-02  0.355380705
#> 77    0.3086725 -5.963630e-02  1.281694e-03  1.738216e-01  0.249036185
#> 78    0.2821509  1.489273e-02 -9.421424e-02 -4.888821e-04  0.297043644
#> 79    0.2350291  5.043132e-02  1.335973e-02 -1.181462e-01  0.285460379
#> 80    0.2672797 -5.048798e-02 -9.278828e-03 -6.306325e-02  0.216791763
#> 81    0.2405684  4.613904e-02 -4.075136e-02  2.349172e-02  0.286707428
#> 82    0.2462920  9.388798e-03  6.862436e-02 -4.206175e-02  0.255680764
#> 83    0.3103663 -8.698149e-02  2.515804e-02  9.000078e-03  0.223384761
#> 84    0.2498813  1.410005e-02 -4.531219e-02  6.958189e-02  0.263981313
#> 85    0.1883713  7.707290e-02  2.244254e-01  9.325859e-02  0.265444151
#> 86    0.2005669 -2.276790e-01 -1.016068e-01  1.572431e-02 -0.027112069
#> 87    0.1965048 -7.384667e-02  9.176416e-02  1.122059e-01  0.122658132
#> 88    0.2812905  5.670113e-02  8.685694e-03  1.067319e-02  0.337991622
#> 89    0.1942408 -1.374121e-01 -2.268241e-02  1.710657e-01  0.056828767
#> 90    0.2376063 -9.360932e-02 -1.075297e-01 -1.115587e-01  0.143996959
#> 91    0.2762798  7.705067e-02 -1.365146e-01 -1.163018e-01  0.353330517
#> 92    0.2951507 -1.114882e-01  1.320360e-01 -6.795981e-02  0.183662497
#> 93    0.2395659  1.961972e-02 -1.043724e-01 -3.611923e-02  0.259185574
#> 94    0.2809899  1.596728e-01 -2.872406e-02  8.085387e-02  0.440662728
#> 95    0.3649341 -6.728084e-03 -1.376138e-02 -2.980881e-02  0.358206030
#> 96    0.3428558 -1.883813e-02 -5.843415e-02  2.268729e-02  0.324017651
#> 97    0.3412777  3.008093e-02 -1.043592e-01  5.419843e-02  0.371358619
#> 98    0.2732907  3.425193e-04 -2.923389e-02  6.232663e-02  0.273633191
#> 99    0.3271061  1.044998e-01  3.179673e-02  6.866714e-02  0.431605886
#> 100   0.3023399 -1.227076e-01  6.853225e-02 -6.824909e-02  0.179632263
#> 101   0.2443015  3.287718e-02  3.111397e-02 -1.420367e-02  0.277178707
#> 102   0.2216178  2.660203e-02  9.464944e-02  2.093739e-04  0.248219817
#> 103   0.1978545  8.903743e-02  1.241551e-01 -7.043098e-03  0.286891925
#> 104   0.2544010 -3.050439e-01  1.755210e-01 -7.795093e-03 -0.050642876
#> 105   0.2704275 -1.931826e-02  2.500101e-02 -1.686480e-01  0.251109277
#> 106   0.3240033 -1.061917e-01  7.383435e-02  9.261840e-02  0.217811614
#> 107   0.3150604  1.367031e-01 -1.417248e-01 -1.453981e-01  0.451763528
#> 108   0.3577322 -5.481721e-02  9.729882e-02 -2.521349e-02  0.302915001
#> 109   0.3567904 -9.304952e-02  2.763014e-02  1.528538e-01  0.263740918
#> 110   0.2907980 -4.398673e-03 -6.196059e-02  1.352894e-01  0.286399361
#> 111   0.2505814  8.920846e-03  1.279035e-02  1.277894e-01  0.259502280
#>                 y             z            dateTime state
#> 1     -0.64363952  0.3496267903 2015-05-16 10:00:00     1
#> 2     -0.75093141  0.1600221157 2015-05-16 10:01:00     1
#> 3     -0.94485295  0.3988929266 2015-05-16 10:02:00     1
#> 4     -1.17002735  0.3528201921 2015-05-16 10:03:00     1
#> 5     -1.13640981  0.2463433944 2015-05-16 10:04:00     1
#> 6     -0.79935243  0.2515734717 2015-05-16 10:05:00     1
#> 7     -0.68660396  0.1802593624 2015-05-16 10:06:00     1
#> 8     -0.79052720  0.2579246987 2015-05-16 10:07:00     1
#> 9     -0.90734921  0.1658923043 2015-05-16 10:08:00     1
#> 10    -0.54668722  0.3142341374 2015-05-16 10:09:00     1
#> 11    -0.54551317  0.0471288879 2015-05-16 10:10:00     1
#> 12    -0.62800084  0.1045919366 2015-05-16 10:11:00     1
#> 13    -0.57500892  0.0788245591 2015-05-16 10:12:00     1
#> 14    -0.59831607  0.2303105370 2015-05-16 10:13:00     1
#> 15    -0.76095788  0.1281157766 2015-05-16 10:14:00     1
#> 16    -0.81148843  0.0445016664 2015-05-16 10:15:00     1
#> 17    -0.64352365  0.3047862175 2015-05-16 10:16:00     1
#> 18    -0.89292872  0.3981497469 2015-05-16 10:17:00     1
#> 19    -0.75921618  0.3951334529 2015-05-16 10:18:00     1
#> 20    -0.81933824  0.3226795751 2015-05-16 10:19:00     1
#> 21    -0.93143633  0.0548109573 2015-05-16 10:20:00     1
#> 22    -0.74252662  0.1044543543 2015-05-16 10:21:00     1
#> 23    -0.36995607  0.1028995836 2015-05-16 10:22:00     1
#> 24    -0.15826606  0.0552985305 2015-05-16 10:23:00     1
#> 25    -0.66906157  0.0494903331 2015-05-16 10:24:00     1
#> 26    -0.80933001  0.1271985069 2015-05-16 10:25:00     1
#> 27    -0.99491621  0.2755307828 2015-05-16 10:26:00     1
#> 28    -1.08805451  0.3808698916 2015-05-16 10:27:00     1
#> 29    -1.11765973  0.5968222642 2015-05-16 10:28:00     1
#> 30    -1.21605360  0.3520898281 2015-05-16 10:29:00     1
#> 31    -0.95160387  0.3265376591 2015-05-16 10:30:00     1
#> 32    -0.89885340  0.3984129493 2015-05-16 10:31:00     1
#> 33    -1.04631943  0.2477773796 2015-05-16 10:32:00     1
#> 34    -0.61844517  0.4062147072 2015-05-16 10:33:00     1
#> 35    -0.84055468  0.1811914730 2015-05-16 10:34:00     1
#> 36    -0.48695290  0.1180768486 2015-05-16 10:35:00     1
#> 37    -0.53860803  0.0770183517 2015-05-16 10:36:00     1
#> 38    -0.81438730  0.0514063866 2015-05-16 10:37:00     1
#> 39    -0.66478487  0.1196362840 2015-05-16 10:38:00     1
#> 40    -0.82533153  0.3007057700 2015-05-16 10:39:00     1
#> 41    -0.47408331  0.1172104035 2015-05-16 10:40:00     1
#> 42    -0.89020885  0.2688158500 2015-05-16 10:41:00     1
#> 43    -0.87807387  0.3490070832 2015-05-16 10:42:00     1
#> 44    -1.05208609  0.6280517038 2015-05-16 10:43:00     1
#> 45    -1.02086774  0.4958616987 2015-05-16 10:44:00     1
#> 46    -0.73840205  0.3123800628 2015-05-16 10:45:00     1
#> 47    -0.92866779  0.2734744247 2015-05-16 10:46:00     1
#> 48    -0.68930870  0.4472640029 2015-05-16 10:47:00     1
#> 49    -0.86954355  0.3188093767 2015-05-16 10:48:00     1
#> 50    -0.97157276  0.4255714504 2015-05-16 10:49:00     1
#> 51    -1.21697500  0.3224076266 2015-05-16 10:50:00     1
#> 52    -1.10456140  0.5658604954 2015-05-16 10:51:00     1
#> 53    -1.05039913  0.5703865678 2015-05-16 10:52:00     1
#> 54    -0.97602852  0.4202946504 2015-05-16 10:53:00     1
#> 55    -0.93052398  0.3232917869 2015-05-16 10:54:00     1
#> 56    -1.13584926  0.6908888509 2015-05-16 10:55:00     1
#> 57    -1.10242498  0.3160150552 2015-05-16 10:56:00     1
#> 58    -1.14151646  0.5926092847 2015-05-16 10:57:00     1
#> 59    -1.23842632  0.4687825240 2015-05-16 10:58:00     1
#> 60    -1.30202079  0.6024630908 2015-05-16 10:59:00     1
#> 61    -1.31299767  0.7008843800 2015-05-16 11:00:00     2
#> 62    -1.00781746  0.4059502676 2015-05-16 11:01:00     2
#> 63    -1.15678511  0.6851731314 2015-05-16 11:02:00     2
#> 64    -1.06172743  0.4167712738 2015-05-16 11:03:00     2
#> 65    -1.07883605  0.4716192416 2015-05-16 11:04:00     2
#> 66    -1.02423343  0.4758227306 2015-05-16 11:05:00     2
#> 67    -1.13744999  0.3781199680 2015-05-16 11:06:00     2
#> 68    -0.90314338  0.3989608443 2015-05-16 11:07:00     2
#> 69    -1.13780296  0.3890304044 2015-05-16 11:08:00     2
#> 70    -1.04489838  0.4642369071 2015-05-16 11:09:00     2
#> 71    -1.05513974  0.5642300556 2015-05-16 11:10:00     2
#> 72    -0.99323613  0.4071441220 2015-05-16 11:11:00     2
#> 73    -1.28033769  0.3942281476 2015-05-16 11:12:00     2
#> 74    -0.77264262  0.2767076378 2015-05-16 11:13:00     2
#> 75    -1.07830750  0.2011515667 2015-05-16 11:14:00     2
#> 76    -1.00899383  0.3959783647 2015-05-16 11:15:00     2
#> 77    -0.82945706  0.4302478078 2015-05-16 11:16:00     2
#> 78    -0.86105785  0.2158767452 2015-05-16 11:17:00     2
#> 79    -0.63648879  0.0345870794 2015-05-16 11:18:00     2
#> 80    -0.73967959  0.1321580696 2015-05-16 11:19:00     2
#> 81    -0.70458338  0.1831887162 2015-05-16 11:20:00     2
#> 82    -0.60959181  0.1249774396 2015-05-16 11:21:00     2
#> 83    -0.80961351  0.2680849944 2015-05-16 11:22:00     2
#> 84    -0.73251535  0.2413010885 2015-05-16 11:23:00     2
#> 85    -0.30520464  0.1930256618 2015-05-16 11:24:00     2
#> 86    -0.66308048  0.1283373677 2015-05-16 11:25:00     2
#> 87    -0.45913611  0.2204604679 2015-05-16 11:26:00     2
#> 88    -0.75606145  0.2257892896 2015-05-16 11:27:00     2
#> 89    -0.56767544  0.2769255670 2015-05-16 11:28:00     2
#> 90    -0.76389179  0.0443970773 2015-05-16 11:29:00     2
#> 91    -0.88902358  0.0916012199 2015-05-16 11:30:00     2
#> 92    -0.66630228  0.1676703153 2015-05-16 11:31:00     2
#> 93    -0.76567820  0.1223070498 2015-05-16 11:32:00     2
#> 94    -0.79273838  0.2955341622 2015-05-16 11:33:00     2
#> 95    -0.97538681  0.3211211197 2015-05-16 11:34:00     2
#> 96    -0.96945141  0.3350348190 2015-05-16 11:35:00     2
#> 97    -1.01172184  0.3638610374 2015-05-16 11:36:00     2
#> 98    -0.77441811  0.2659785259 2015-05-16 11:37:00     2
#> 99    -0.84252316  0.3546625002 2015-05-16 11:38:00     2
#> 100   -0.74707805  0.1783424565 2015-05-16 11:39:00     2
#> 101   -0.64210737  0.1502653275 2015-05-16 11:40:00     2
#> 102   -0.52108947  0.1366680667 2015-05-16 11:41:00     2
#> 103   -0.43026198  0.1026508162 2015-05-16 11:42:00     2
#> 104   -0.52296219  0.1698997241 2015-05-16 11:43:00     2
#> 105   -0.71315057  0.0309685424 2015-05-16 11:44:00     2
#> 106   -0.79319751  0.3735396265 2015-05-16 11:45:00     2
#> 107   -0.98764309  0.1211173167 2015-05-16 11:46:00     2
#> 108   -0.84792549  0.3129237003 2015-05-16 11:47:00     2
#> 109   -0.91544176  0.4893328202 2015-05-16 11:48:00     2
#> 110   -0.84979120  0.3643889727 2015-05-16 11:49:00     2
#> 111   -0.67616290  0.3004282847 2015-05-16 11:50:00     2
#>  [ reached getOption("max.print") -- omitted 9969 rows ]